Deprecated functions in package bark.

The functions listed below are deprecated and will be defunct in the near future. When possible, alternative functions with similar functionality are also mentioned. Help pages for deprecated functions are available at help("<function>-deprecated").

Usage

bark_mat(
  x.train,
  y.train,
  x.test = matrix(0, 0, 0),
  type = "se",
  classification = FALSE,
  keepevery = 100,
  nburn = 100,
  nkeep = 100,
  printevery = 1000,
  keeptrain = FALSE,
  fixed = list(),
  tune = list(lstep = 0.5, frequL = 0.2, dpow = 1, upow = 0, varphistep = 0.5, phistep =
    1),
  theta = list()
)

sim.Friedman1(n, sd = 1)

sim.Friedman2(n, sd = 125)

sim.Friedman3(n, sd = 0.1)

sim.Circle(n, dim = 5)

Value

List of deprecated functions

bark_mat

Old version with matrix inputs used for testing;

sim.Friedman1

For sim.Friedman1, use sim_Friedman1.

sim.Friedman2

For sim.Friedman2, use sim_Friedman2.

sim.Friedman3

For sim.Friedman3, use sim_Friedman3.

sim.Circle

For sim.Circle, use sim_circle.

See also