Simulate Data from Hyper-Sphere for Classification Problems

The classification problem Circle is described in the BARK paper (2008). Inputs are dim independent variables uniformly distributed on the interval \([-1,1]\), only the first 2 out of these dim are actually signals. Outputs are created according to the formula $$y = 1(x1^2+x2^2 \le 2/\pi)$$

sim_circle(n, dim = 5)

Arguments

n

number of data points to generate

dim

number of dimension of the problem, no less than 2

Value

Returns a list with components

x

input values (independent variables)

y

0/1 output values (dependent variable)

References

Ouyang, Zhi (2008) Bayesian Additive Regression Kernels. Duke University. PhD dissertation, Chapter 3.

See also

Examples

sim_circle(n=100, dim=5)
#> $x
#>               [,1]        [,2]        [,3]        [,4]        [,5]
#>   [1,]  0.13449574 -0.60615928  0.37155924 -0.87327800 -0.48614191
#>   [2,] -0.95306934  0.70658304  0.51550150 -0.20830783 -0.31757636
#>   [3,] -0.98728620 -0.05229903  0.50745406 -0.76196980 -0.42164769
#>   [4,] -0.19004945  0.82966070  0.47963126  0.82003225 -0.28521181
#>   [5,]  0.97751771 -0.16397050 -0.17081952 -0.12663746 -0.05023214
#>   [6,] -0.42269367  0.20505745 -0.85287142  0.02879546 -0.95002964
#>   [7,] -0.60873510  0.30335402 -0.66481540  0.87542733 -0.43593684
#>   [8,] -0.39583235 -0.35694601 -0.18679642  0.24180889 -0.50492992
#>   [9,] -0.16497192  0.10330223 -0.24657328 -0.19996072 -0.25664701
#>  [10,] -0.08674832 -0.98911768  0.63294499  0.08749914  0.30373583
#>  [11,] -0.48781593  0.07679516 -0.04421238  0.52803511  0.87406507
#>  [12,] -0.84386283 -0.83165262  0.74821757 -0.89790251  0.89537199
#>  [13,] -0.96530823  0.71820614  0.10237754 -0.47128319  0.81131881
#>  [14,] -0.09792719 -0.14326316  0.94543017 -0.12250704  0.62386140
#>  [15,] -0.87898100 -0.55903786 -0.66024775  0.39185451  0.09958254
#>  [16,]  0.62062267  0.62744971  0.65734487  0.65851696  0.42293057
#>  [17,]  0.55679804 -0.54496745 -0.87012270 -0.70559899  0.18595405
#>  [18,]  0.45168910 -0.89989068  0.49207661  0.68410872 -0.47471459
#>  [19,] -0.24554486  0.08977244 -0.66725256  0.37613016 -0.86233392
#>  [20,]  0.46770650  0.40404607 -0.32884608 -0.02479862  0.04258888
#>  [21,] -0.73381630  0.01043322  0.38170010 -0.50817739 -0.32314457
#>  [22,]  0.44516169  0.86762685  0.25279421 -0.37719309  0.97357243
#>  [23,] -0.76553994  0.79470518 -0.84370687 -0.19649146 -0.34352766
#>  [24,] -0.78087253  0.90660216  0.96137900 -0.75138348  0.13651433
#>  [25,]  0.37682743 -0.61997399 -0.19581034  0.77914493 -0.19992586
#>  [26,] -0.35045620 -0.10237289  0.93115390  0.63867558  0.01288258
#>  [27,] -0.20728939  0.93535864  0.54361394 -0.68360455  0.77285112
#>  [28,]  0.65442925  0.16551028 -0.54893986  0.93083208 -0.22964150
#>  [29,] -0.69254703  0.68095721  0.97777680 -0.71201202 -0.41622394
#>  [30,]  0.78838407  0.20164604 -0.54097257  0.93384257  0.53814502
#>  [31,] -0.80259939 -0.74650679 -0.04142424 -0.88731535 -0.04675277
#>  [32,] -0.07280195 -0.18874048  0.90433312 -0.08746909  0.30707286
#>  [33,] -0.96579241 -0.21061626 -0.47211612 -0.15992484 -0.87784397
#>  [34,] -0.31593782 -0.75002411  0.56780757  0.99088325  0.68834654
#>  [35,]  0.03440562  0.90067864 -0.07883936 -0.61400764  0.84611055
#>  [36,] -0.81657626 -0.01930710  0.15122924  0.33345954 -0.11942578
#>  [37,] -0.51825346  0.01075978 -0.80990244 -0.49430427  0.10305192
#>  [38,]  0.59925441  0.77684205 -0.71895034  0.72010573  0.58840408
#>  [39,]  0.23997061  0.06055519 -0.76940397 -0.08051992  0.43620225
#>  [40,]  0.70797115 -0.14602761 -0.40705800 -0.05910896 -0.26810148
#>  [41,] -0.35587474  0.78112262 -0.52877047  0.42501842 -0.62220615
#>  [42,]  0.87953895  0.82438508  0.33740664 -0.07708994 -0.49022610
#>  [43,]  0.84888757  0.61015704 -0.33186924 -0.70788242 -0.85566469
#>  [44,]  0.32219866 -0.50044970 -0.83309980  0.11955466  0.77237034
#>  [45,] -0.93850731  0.78084549 -0.08493988 -0.60617578  0.57913575
#>  [46,]  0.43282910 -0.25185649  0.86299215  0.42118642 -0.59234484
#>  [47,] -0.21837432 -0.93412184  0.08724105  0.85776940 -0.95644263
#>  [48,]  0.56452938 -0.37952439 -0.13508395  0.32984916  0.26782952
#>  [49,] -0.86307861  0.46926190  0.57092825 -0.14067029 -0.80376634
#>  [50,] -0.09656088  0.40006939 -0.83118482  0.34747193  0.12035867
#>  [51,]  0.45264546 -0.53750117  0.15137975 -0.26742998 -0.87757187
#>  [52,] -0.76066555  0.64340411 -0.28221980 -0.92702097 -0.78218901
#>  [53,] -0.80122976  0.93619973 -0.88143630  0.28860121 -0.47028993
#>  [54,]  0.99436681 -0.61397348  0.73215047 -0.43541109 -0.16850487
#>  [55,] -0.73534062 -0.66742952 -0.71927858 -0.40217167 -0.70696981
#>  [56,] -0.81366119 -0.26251697 -0.39427331  0.18566825  0.35574437
#>  [57,]  0.75130544 -0.83427299 -0.70374402  0.94304368  0.73391908
#>  [58,]  0.03797811 -0.67228079 -0.01719219 -0.46526233  0.43793111
#>  [59,]  0.59972542 -0.83171576 -0.81028849 -0.24102878 -0.63802359
#>  [60,] -0.15185017  0.43252410 -0.36990223  0.26908547 -0.71765254
#>  [61,] -0.56255852  0.73236396  0.77852442 -0.03327891 -0.25561823
#>  [62,] -0.96760855 -0.83659572 -0.32261223  0.60304765 -0.72786811
#>  [63,]  0.15069585 -0.39158103  0.67640094  0.23598145  0.45234781
#>  [64,]  0.62078383 -0.17705998  0.45095980 -0.05986177 -0.84480786
#>  [65,] -0.41775669  0.07130236 -0.09934399  0.38059048  0.48086277
#>  [66,] -0.38320312  0.39728137  0.41893723  0.14248556 -0.48594559
#>  [67,]  0.73759041  0.78013622 -0.71991662 -0.54011759 -0.21477951
#>  [68,]  0.48355445 -0.32926661  0.81398893 -0.85355164  0.87520024
#>  [69,] -0.34255679  0.52200862  0.35394309 -0.13698922  0.58119843
#>  [70,] -0.55546004  0.63042195 -0.92823027 -0.98882445 -0.50469197
#>  [71,]  0.50774094 -0.52026305  0.64532441  0.37660347 -0.58319988
#>  [72,] -0.41346637  0.91770033 -0.41591382  0.52682265  0.74167729
#>  [73,] -0.60965706  0.50234708 -0.26101315 -0.61477424 -0.42026186
#>  [74,]  0.08570191  0.89900274  0.29953699  0.56067385  0.04177577
#>  [75,]  0.06460475  0.99994418 -0.06181672 -0.90563350 -0.66351970
#>  [76,]  0.52360379 -0.97688922 -0.90620777  0.27054649 -0.66143129
#>  [77,] -0.52418467 -0.54313402 -0.94784539 -0.82056748 -0.35660608
#>  [78,] -0.44049074  0.60484552 -0.34118994 -0.30266137  0.15964823
#>  [79,]  0.89437949 -0.21123994 -0.30078064  0.99837214  0.46581593
#>  [80,]  0.38025989  0.36495824 -0.51472716  0.21277696  0.03253280
#>  [81,] -0.41491885 -0.98569677 -0.38801592  0.88467857 -0.69883635
#>  [82,]  0.39411976 -0.13791578  0.44291579 -0.85023894 -0.77727788
#>  [83,] -0.83031844  0.42178429 -0.18941427  0.39502557  0.63552980
#>  [84,] -0.86266733  0.64053088 -0.50248874  0.17680303  0.08665122
#>  [85,]  0.36231579 -0.93238877  0.76859387 -0.70558581  0.56345296
#>  [86,] -0.52124205 -0.92475619 -0.09768858 -0.89765400 -0.07531246
#>  [87,]  0.71983333 -0.59289434 -0.70680892 -0.68133078  0.77710205
#>  [88,] -0.96659952 -0.80615036  0.03979500 -0.89842598  0.82895342
#>  [89,] -0.76386254  0.61463987  0.76594457 -0.31953588  0.15356705
#>  [90,]  0.36133069  0.45942756 -0.50126844  0.83201796 -0.26181096
#>  [91,] -0.13782286 -0.89980207 -0.34031176  0.45888960  0.08113641
#>  [92,] -0.35079966  0.49612569 -0.24871805  0.23965372 -0.12658431
#>  [93,]  0.31158636 -0.56113203  0.36860272  0.54438486 -0.50846785
#>  [94,] -0.60755571  0.46144979  0.15118757  0.22613589 -0.29835754
#>  [95,]  0.71436638 -0.57993569 -0.57820085 -0.07849737  0.66248309
#>  [96,] -0.52182516  0.35610056  0.03174354  0.19936650 -0.76368459
#>  [97,] -0.04478277  0.64935708  0.21114718  0.22570444 -0.21894830
#>  [98,] -0.65479391 -0.74739538 -0.91351498  0.77380762 -0.59197456
#>  [99,] -0.48321849  0.60626596  0.50765693  0.31802073  0.92150217
#> [100,] -0.29538889  0.93140791 -0.40336811 -0.39898409 -0.15828147
#> 
#> $y
#>   [1] 1 0 0 0 0 1 1 1 1 0 1 0 0 1 0 0 1 0 1 1 1 0 0 0 1 1 0 1 0 0 0 1 0 0 0 0 1
#>  [38] 0 1 1 0 0 0 1 0 1 0 1 0 1 1 0 0 0 0 0 0 1 0 1 0 0 1 1 1 1 0 1 1 0 1 0 1 0
#>  [75] 0 0 1 1 0 1 0 1 0 0 0 0 0 0 0 1 0 1 1 1 0 1 1 0 1 0
#>