Simulate Data from Hyper-Sphere for Classification Problems

The classification problem Circle is described in the BARK paper (2008). Inputs are dim independent variables uniformly distributed on the interval \([-1,1]\), only the first 2 out of these dim are actually signals. Outputs are created according to the formula $$y = 1(x1^2+x2^2 \le 2/\pi)$$

Usage

sim_circle(n, dim = 5)

Arguments

n

number of data points to generate

dim

number of dimension of the problem, no less than 2

Value

Returns a list with components

x

input values (independent variables)

y

0/1 output values (dependent variable)

References

Ouyang, Zhi (2008) Bayesian Additive Regression Kernels. Duke University. PhD dissertation, Chapter 3.

See also

Examples

sim_circle(n=100, dim=5)
#> $x
#>               [,1]        [,2]          [,3]         [,4]         [,5]
#>   [1,]  0.72882729 -0.73435633  2.839336e-01  0.066863726 -0.208299876
#>   [2,]  0.50447454 -0.34178115  5.249405e-01  0.123139915 -0.576751928
#>   [3,] -0.47880345  0.67258264  4.153200e-01 -0.764984316 -0.324054339
#>   [4,]  0.01550109  0.18453234 -9.646156e-03  0.368622507  0.592172379
#>   [5,] -0.02730042 -0.69977044  4.409119e-01  0.250986505  0.960200502
#>   [6,]  0.96607214  0.98159619  8.822043e-01  0.908437020 -0.199412971
#>   [7,] -0.55047848 -0.89564987 -7.696091e-01 -0.559544583  0.383547236
#>   [8,]  0.07159495  0.17239880 -3.023250e-01 -0.751923140  0.891629337
#>   [9,]  0.39046619  0.26317552  9.018012e-01 -0.387845191  0.777177482
#>  [10,] -0.71049126 -0.61054028  1.776672e-01 -0.184586711  0.544026918
#>  [11,] -0.50995181  0.40484485  6.295353e-01 -0.139689184 -0.255600180
#>  [12,] -0.06127934 -0.66134382 -5.249991e-01  0.738445142  0.705950535
#>  [13,]  0.93568621 -0.39476210  5.540913e-01  0.001645978  0.352551008
#>  [14,]  0.17713175 -0.35611686  3.294965e-01 -0.821703730 -0.964925797
#>  [15,]  0.22243929 -0.39489879 -6.243576e-01  0.926391564  0.861596081
#>  [16,]  0.63625548  0.21795936  9.310486e-01  0.028322078 -0.462646553
#>  [17,] -0.10256771 -0.96278434 -9.797976e-01  0.934003985 -0.268646296
#>  [18,] -0.85000303 -0.44496554  5.863648e-01  0.956679148 -0.893379331
#>  [19,]  0.43758208 -0.80056260 -6.858048e-01  0.062530084 -0.913005544
#>  [20,] -0.18048966 -0.80520399 -8.394324e-01  0.386701739 -0.098056225
#>  [21,]  0.24196203 -0.91850691  3.073197e-01  0.393861835  0.130619658
#>  [22,] -0.29632074  0.67579956  1.440295e-01  0.964516072  0.637667800
#>  [23,] -0.31606959  0.57476035 -2.378086e-05  0.096513005  0.141214151
#>  [24,]  0.36871621 -0.39975550  9.949874e-01  0.596841673 -0.916319630
#>  [25,]  0.73181378  0.93171349 -3.972830e-01 -0.409433989 -0.280746938
#>  [26,]  0.95071694 -0.04393731  4.557776e-01 -0.810450914  0.904224833
#>  [27,] -0.44389120 -0.49594792  8.931262e-01 -0.431464628  0.902201521
#>  [28,] -0.93200422 -0.56880556  2.831234e-01  0.903932384 -0.215213221
#>  [29,] -0.29351949 -0.59915949  6.295148e-01  0.849627504  0.884504640
#>  [30,]  0.36633471  0.35032725 -8.335907e-01  0.153415750  0.028163899
#>  [31,]  0.27880609 -0.05939182 -4.152629e-01 -0.863312247  0.070898898
#>  [32,] -0.50203067  0.33566163 -8.409378e-01  0.347291574 -0.991857568
#>  [33,] -0.24880180  0.91001264 -8.662692e-01 -0.425606040 -0.949711926
#>  [34,]  0.65238890 -0.42070938  6.371226e-01  0.438777053  0.625274035
#>  [35,]  0.04028755  0.69205102  2.332989e-01 -0.919356235 -0.316145618
#>  [36,] -0.13905652  0.57612393  2.229240e-01  0.594460655  0.586433974
#>  [37,] -0.14928974 -0.61654384 -3.018581e-02  0.174818233  0.745410573
#>  [38,]  0.93072443  0.15215533 -3.272231e-02 -0.981005999 -0.236021863
#>  [39,] -0.15846636 -0.85214263  6.822076e-01  0.719502739  0.190759801
#>  [40,] -0.79860311 -0.31794909  1.968093e-01  0.940943658 -0.353852015
#>  [41,]  0.81642348  0.36333453 -9.956619e-01 -0.721793973 -0.682522242
#>  [42,] -0.78009609  0.76074466 -9.092332e-01  0.733631191  0.011288888
#>  [43,] -0.27905081 -0.88762314 -4.270435e-01 -0.043843701  0.314071756
#>  [44,] -0.34421108 -0.49650610 -6.178571e-01  0.216819749 -0.241867909
#>  [45,]  0.73829884  0.46709081 -4.003077e-01  0.627903904 -0.783873698
#>  [46,] -0.36567832 -0.62763194 -3.496493e-01 -0.846395136 -0.235803202
#>  [47,]  0.15513038  0.04845993 -7.658201e-01  0.228605295  0.006037361
#>  [48,]  0.47489661  0.14946042 -6.571979e-01  0.990963270 -0.113606492
#>  [49,] -0.39142998 -0.10466478 -8.631037e-01  0.445277378 -0.501437552
#>  [50,] -0.46235943 -0.89650904 -4.505614e-01 -0.148044161  0.061788187
#>  [51,] -0.85242365 -0.05363214  2.850740e-01  0.859438702 -0.693679833
#>  [52,]  0.32685736  0.49237504  5.513228e-01  0.780324427 -0.241804126
#>  [53,] -0.93440982  0.37841976 -1.537651e-01 -0.034234900 -0.602332090
#>  [54,] -0.29899449 -0.07163558  7.688416e-01 -0.037263992 -0.554634374
#>  [55,] -0.11006701 -0.81079195  7.267611e-01  0.002548166 -0.924716692
#>  [56,] -0.39276850 -0.87576983  3.099424e-01 -0.626658470  0.527948773
#>  [57,]  0.54869476  0.05335946  4.818783e-01  0.252041047  0.289746100
#>  [58,] -0.82360544 -0.06438502  2.508984e-01 -0.719143239 -0.838933937
#>  [59,] -0.37410496 -0.32014274 -3.197541e-01 -0.989623528 -0.622424564
#>  [60,] -0.73732456 -0.19038836  9.284595e-01 -0.149325401 -0.305670675
#>  [61,] -0.96913082 -0.99264597  8.044319e-01 -0.443169730  0.343218867
#>  [62,]  0.50588261 -0.91288645 -4.520304e-01  0.756466430  0.125733223
#>  [63,]  0.59192786  0.01860881  6.546125e-01  0.857552283 -0.731103319
#>  [64,] -0.46666698  0.40865166  5.843093e-01 -0.248844845  0.738230116
#>  [65,] -0.50638068  0.05454483  1.169501e-01  0.037885896 -0.023519443
#>  [66,]  0.11097437 -0.21346459 -4.998924e-01 -0.130513816 -0.963728655
#>  [67,] -0.47849018  0.93804525 -5.531340e-01 -0.104539147 -0.948704850
#>  [68,] -0.45523344 -0.77641509  1.231285e-01  0.784867211  0.388774986
#>  [69,] -0.18000366  0.67599462  7.841078e-01 -0.086537403  0.930312099
#>  [70,] -0.38138511  0.80157616 -5.532996e-01 -0.493666063  0.553436319
#>  [71,]  0.78518573  0.94813269  7.380045e-01  0.535064471 -0.742512045
#>  [72,]  0.46985737  0.16454883  8.920283e-01  0.668873898  0.863288323
#>  [73,]  0.31416661 -0.48312691  2.657139e-01  0.941457554 -0.014902081
#>  [74,]  0.60824508 -0.13117692  9.113200e-01  0.775578902  0.558815590
#>  [75,]  0.52183690  0.41590243 -9.057477e-01  0.334755006  0.799985220
#>  [76,]  0.45622251  0.10895993 -3.989710e-03 -0.478785952  0.144294950
#>  [77,]  0.94885526  0.99965128 -7.281233e-01 -0.844442500 -0.056520732
#>  [78,]  0.20784178  0.74961773  1.918770e-01 -0.074969782 -0.916469856
#>  [79,] -0.70415544  0.88183912 -7.108392e-01  0.734219612  0.726515179
#>  [80,] -0.64459327 -0.92739027  7.420317e-01  0.820586410  0.447874776
#>  [81,] -0.13390130  0.95101718 -9.429973e-01  0.842541240  0.656465518
#>  [82,]  0.54353062 -0.68104059 -7.585640e-01  0.385766249  0.538192007
#>  [83,] -0.48879454  0.19357894  8.066122e-01  0.091468863 -0.544086365
#>  [84,]  0.57805941  0.52576822 -8.765540e-01 -0.190779721 -0.228870488
#>  [85,] -0.64658274 -0.88540862  2.116924e-01 -0.938866528 -0.544807444
#>  [86,] -0.45041874 -0.26143097  4.163471e-01 -0.329207197 -0.412341017
#>  [87,] -0.13688144 -0.23493649 -5.041584e-01  0.571108501  0.377129660
#>  [88,] -0.43576688  0.22250923  8.911676e-01  0.228340949  0.635273329
#>  [89,] -0.09986591  0.70542550  4.661263e-01 -0.704314684 -0.665407425
#>  [90,]  0.42432376 -0.09970413 -8.814577e-01 -0.396616505 -0.996344386
#>  [91,] -0.98009684  0.31568726 -1.651867e-01  0.454744353  0.776805078
#>  [92,]  0.32584345  0.31551620 -2.752596e-01 -0.464046908 -0.546260017
#>  [93,]  0.04650258 -0.91987548  3.361970e-02 -0.885348357  0.268576485
#>  [94,]  0.32175673 -0.07926612 -9.597352e-01  0.043717874 -0.787838443
#>  [95,] -0.24516227  0.41955744 -4.774569e-02  0.823997901 -0.928996280
#>  [96,]  0.05790402 -0.08713444  9.665027e-01  0.171830143  0.342883625
#>  [97,] -0.61086730  0.01035248 -6.445135e-02 -0.355429575 -0.519194877
#>  [98,]  0.81482895 -0.31929255 -6.289636e-01 -0.931065536  0.603163836
#>  [99,] -0.39151049 -0.71368780  6.309952e-02 -0.383102001 -0.176605708
#> [100,] -0.54473417 -0.77910321  8.438365e-01  0.401289260 -0.493852561
#> 
#> $y
#>   [1] 0 1 0 1 1 0 0 1 1 0 1 1 0 1 1 1 0 0 0 0 0 1 1 1 0 0 1 0 1 1 1 1 0 1 1 1 1
#>  [38] 0 0 0 0 0 0 1 0 1 1 1 1 0 0 1 0 1 0 0 1 0 1 1 0 0 1 1 1 1 0 0 1 0 0 1 1 1
#>  [75] 1 1 0 1 0 0 0 0 1 1 0 1 1 1 1 1 0 1 0 1 1 1 1 0 0 0
#>