Simulate Data from Hyper-Sphere for Classification Problems

The classification problem Circle is described in the BARK paper (2008). Inputs are dim independent variables uniformly distributed on the interval \([-1,1]\), only the first 2 out of these dim are actually signals. Outputs are created according to the formula $$y = 1(x1^2+x2^2 \le 2/\pi)$$

Usage

sim_circle(n, dim = 5)

Arguments

n

number of data points to generate

dim

number of dimension of the problem, no less than 2

Value

Returns a list with components

x

input values (independent variables)

y

0/1 output values (dependent variable)

References

Ouyang, Zhi (2008) Bayesian Additive Regression Kernels. Duke University. PhD dissertation, Chapter 3.

See also

Examples

sim_circle(n=100, dim=5)
#> $x
#>                [,1]         [,2]        [,3]          [,4]        [,5]
#>   [1,] -0.753849828  0.044161787 -0.41551345 -0.1845852165 -0.98679782
#>   [2,] -0.812749968  0.635493814 -0.41809110 -0.7657177718 -0.55100831
#>   [3,]  0.130080921 -0.605140667  0.97242263 -0.8725806032 -0.34125543
#>   [4,]  0.975956960  0.553862461  0.35956775  0.0976800220  0.13904216
#>   [5,]  0.578395102  0.911328715  0.35560667 -0.3545116619  0.46485778
#>   [6,]  0.223002935 -0.841963086 -0.43834897 -0.7853094498  0.64897128
#>   [7,]  0.364123168 -0.108251631 -0.47488025  0.8403872875 -0.80994858
#>   [8,]  0.178478129  0.261005933 -0.44049132 -0.9984658533  0.46025692
#>   [9,]  0.659961787 -0.263040283 -0.57531402 -0.2250302355  0.97399785
#>  [10,] -0.450548220 -0.997591791 -0.34326271  0.6663312544 -0.50366675
#>  [11,] -0.727019821 -0.758922413  0.87903187 -0.0005424134  0.55305083
#>  [12,]  0.299747780  0.721584757  0.42060824 -0.9376257467  0.65400793
#>  [13,]  0.641611570 -0.794080864  0.36681692  0.6189332251 -0.09292772
#>  [14,]  0.065828737 -0.668824573  0.87132951  0.9982858631 -0.70762081
#>  [15,] -0.300594278 -0.313030266  0.37396710  0.4469967550 -0.41659934
#>  [16,]  0.524626375  0.612795282  0.75762785 -0.3684728779 -0.34038992
#>  [17,]  0.690678331 -0.937050031  0.64924942 -0.2985218680 -0.38740166
#>  [18,]  0.214158744  0.649760266  0.79828222  0.8460070039 -0.73255670
#>  [19,] -0.692494097  0.043684611  0.62264892 -0.4897416360 -0.50180753
#>  [20,]  0.673466758  0.234452296  0.99064333  0.3544681584 -0.23998319
#>  [21,] -0.832450249 -0.558100571  0.84097645  0.7462321362  0.54717977
#>  [22,] -0.669839748 -0.107632620 -0.24882345 -0.2112774365  0.75036350
#>  [23,]  0.370036685 -0.189862621 -0.79359040 -0.4513919801  0.75636372
#>  [24,] -0.602796072  0.799425109 -0.80294706  0.7844434096 -0.31789168
#>  [25,]  0.304673224  0.415587641  0.42637897 -0.4892521631 -0.60640103
#>  [26,]  0.270674882 -0.777146906  0.43460634 -0.9474691781 -0.65106669
#>  [27,]  0.900937285  0.590391591 -0.94134923  0.1476082411  0.25676226
#>  [28,]  0.034450625 -0.778367602  0.03960410 -0.2656162633  0.93349374
#>  [29,] -0.956981108  0.624412060  0.62225174  0.3374757534  0.12547756
#>  [30,]  0.641896528 -0.206326628 -0.69399671 -0.7430123058  0.83818756
#>  [31,]  0.030079482  0.512572926  0.09835880  0.5923477872  0.65872511
#>  [32,] -0.680595271 -0.893745028  0.72430379  0.8986759507  0.20684216
#>  [33,]  0.511070239  0.472732733 -0.67562647 -0.2083882112  0.80211721
#>  [34,] -0.279962411 -0.342248061 -0.96067905  0.1790960324 -0.47101845
#>  [35,]  0.307427207  0.920881989  0.42125083  0.3432982359 -0.83983198
#>  [36,]  0.562002519 -0.663276411  0.36764296  0.1975851981  0.19850330
#>  [37,]  0.930718377 -0.754348148 -0.75260939  0.5496494500  0.83095182
#>  [38,] -0.617358414 -0.051938948  0.73073108 -0.6008825572  0.72628579
#>  [39,]  0.239179984  0.391443789  0.57617645 -0.1610297370  0.79612967
#>  [40,]  0.437392324  0.252979920 -0.36212167  0.8512243740  0.19269649
#>  [41,]  0.001854944 -0.796658118  0.93830464 -0.9094590414 -0.95368864
#>  [42,]  0.023547829 -0.033177441  0.67652331 -0.0741950930  0.02091774
#>  [43,]  0.017969531  0.344759250 -0.11385384 -0.3097533765 -0.94637275
#>  [44,]  0.254127432  0.385273543  0.12474933  0.3574440009 -0.38598117
#>  [45,]  0.380069661  0.954788729 -0.35221758  0.4483178146 -0.90691132
#>  [46,] -0.101762411  0.634150407 -0.84757987  0.0560975610 -0.14867095
#>  [47,] -0.195617525  0.099738319 -0.64241246  0.8257946731 -0.21737337
#>  [48,] -0.926899453  0.986936103  0.62576725 -0.7741410974 -0.16874499
#>  [49,]  0.546756348  0.028321262 -0.41604517 -0.6828317349 -0.18064028
#>  [50,]  0.614999239  0.673827983 -0.54189440  0.0531370668 -0.94215222
#>  [51,] -0.010388093  0.876283143 -0.94833555  0.8941079457  0.51428644
#>  [52,] -0.389683878  0.002653699  0.09703327  0.2749769646  0.53928022
#>  [53,] -0.889491824 -0.987384435 -0.23166470 -0.1024027499  0.76817692
#>  [54,] -0.156292771 -0.552888910  0.42691165 -0.9224653747 -0.29150891
#>  [55,]  0.642329070 -0.954236856  0.89593077 -0.8489854867  0.83007735
#>  [56,]  0.193246397 -0.104384595  0.02534101 -0.7692922871 -0.91051408
#>  [57,] -0.773181595 -0.019717423 -0.44453637 -0.0301440507  0.63542106
#>  [58,] -0.068991744 -0.893029093  0.05050094 -0.5731244590  0.05597063
#>  [59,] -0.552508276  0.210858945 -0.32191617  0.8781461450  0.74586351
#>  [60,]  0.715546057  0.001135941 -0.75039735 -0.5720730810  0.47726677
#>  [61,] -0.510912456  0.392537388  0.03680298 -0.8635691646 -0.39890802
#>  [62,] -0.454356248  0.465259724  0.98298756 -0.1917028707 -0.16107341
#>  [63,]  0.475454241 -0.316243767  0.37247596 -0.2306310600 -0.78031972
#>  [64,]  0.345033772  0.388807089  0.59659341 -0.1982656494  0.05333398
#>  [65,] -0.962489592 -0.584333789  0.94321550  0.3616900789 -0.83929770
#>  [66,] -0.984015703  0.156867405  0.32281071  0.4739046600 -0.51451818
#>  [67,] -0.667472476 -0.454941618 -0.94062761  0.5807208614 -0.80545896
#>  [68,]  0.126059495 -0.786355980 -0.99508853  0.8281530980 -0.49534371
#>  [69,]  0.652318069  0.073784622  0.45144468 -0.9084746605 -0.26581299
#>  [70,]  0.336163447 -0.798528728  0.50908428 -0.1392132444  0.70215638
#>  [71,] -0.557837247  0.550285802 -0.90719735  0.0285660289  0.14129778
#>  [72,] -0.526919291  0.118522769 -0.29919530  0.8035058263 -0.66192816
#>  [73,]  0.093707208  0.768430865 -0.58525269  0.5082988529 -0.65293588
#>  [74,]  0.561984339 -0.788921047  0.07824213  0.3183718529 -0.42304372
#>  [75,] -0.616283384 -0.499071763 -0.77458968 -0.3571467716 -0.15780511
#>  [76,]  0.198646067  0.996793714  0.24104704  0.3403811660 -0.53827011
#>  [77,]  0.854526938  0.490147237 -0.58996248 -0.1281028111 -0.61564884
#>  [78,]  0.799039922  0.415705142 -0.66138977  0.4866837962 -0.38115075
#>  [79,] -0.243774807 -0.920742780 -0.06195412 -0.9526630570 -0.69857154
#>  [80,]  0.931525302 -0.519525132  0.95899116  0.7749334928  0.47519298
#>  [81,]  0.668650641 -0.332115802 -0.40303734  0.7645922317 -0.92488484
#>  [82,] -0.141143296 -0.823521127 -0.70138597  0.2220463110 -0.97885041
#>  [83,] -0.822910749  0.849378454 -0.59401866  0.4978711368  0.05806962
#>  [84,] -0.665299285  0.845453994 -0.18628130  0.7811106308  0.18616323
#>  [85,]  0.831031469  0.333415958 -0.63395262  0.9585460331 -0.20346249
#>  [86,]  0.330811546  0.024566244  0.95963485 -0.1992584672  0.13897403
#>  [87,] -0.490222660  0.492365658 -0.13600129  0.3273299658 -0.89296524
#>  [88,]  0.056171148  0.616835551  0.73017613  0.1931607826  0.56446291
#>  [89,]  0.704292145 -0.139320159 -0.89720289  0.6746348860  0.18161147
#>  [90,]  0.845488121  0.037467274 -0.39994170 -0.2655783389 -0.02849389
#>  [91,]  0.475222867 -0.705550144  0.28661739  0.9887077319  0.70124391
#>  [92,] -0.987170972  0.433464617  0.67237856  0.9914621999 -0.91282848
#>  [93,]  0.009941828  0.430237516  0.81539209  0.4656105428  0.39157798
#>  [94,]  0.021039076  0.356130154 -0.94653623  0.7383182873 -0.37790493
#>  [95,]  0.942632402 -0.132621223  0.32005528  0.2520748652  0.99331406
#>  [96,] -0.097006679 -0.472900466 -0.27042493 -0.9811566146 -0.58090590
#>  [97,]  0.549306822  0.623658898 -0.22187576  0.9589220011  0.11761326
#>  [98,] -0.265587725 -0.895940777  0.79783570 -0.9873871864  0.12729874
#>  [99,]  0.379204390 -0.761267420 -0.48626345 -0.7890875516  0.55550878
#> [100,]  0.433129949 -0.040868045 -0.12059903 -0.1889044801 -0.30597730
#> 
#> $y
#>   [1] 1 0 1 0 0 0 1 1 1 0 0 1 0 1 1 0 0 1 1 1 0 1 1 0 1 0 0 1 0 1 1 0 1 1 0 0 0
#>  [38] 1 1 1 1 1 1 1 0 1 1 0 1 0 0 1 0 1 0 1 1 0 1 1 1 1 1 1 0 0 0 1 1 0 1 1 1 0
#>  [75] 1 0 0 0 0 0 1 0 0 0 0 1 1 1 1 0 0 0 1 1 0 1 0 0 0 1
#>