Simulate Data from Hyper-Sphere for Classification Problems

The classification problem Circle is described in the BARK paper (2008). Inputs are dim independent variables uniformly distributed on the interval \([-1,1]\), only the first 2 out of these dim are actually signals. Outputs are created according to the formula $$y = 1(x1^2+x2^2 \le 2/\pi)$$

sim_circle(n, dim = 5)

Arguments

n

number of data points to generate

dim

number of dimension of the problem, no less than 2

Value

Returns a list with components

x

input values (independent variables)

y

0/1 output values (dependent variable)

References

Ouyang, Zhi (2008) Bayesian Additive Regression Kernels. Duke University. PhD dissertation, Chapter 3.
http://stat.duke.edu/people/theses/OuyangZ.html

See also

Examples

sim_circle(n=100, dim=5)
#> $x
#>               [,1]        [,2]         [,3]        [,4]        [,5]
#>   [1,] -0.49054898  0.64516353 -0.740478680  0.12466039  0.89136894
#>   [2,]  0.86115052 -0.49807986  0.669108315 -0.96754605 -0.44207294
#>   [3,]  0.83437986  0.93931827 -0.312509508 -0.12315668  0.19245253
#>   [4,] -0.18148055  0.27705690 -0.292600456 -0.37340572  0.31911250
#>   [5,]  0.70777028  0.67973280 -0.991186197  0.91231176  0.70738537
#>   [6,]  0.63775171  0.96640497  0.559953541  0.82315536 -0.03763307
#>   [7,]  0.62182072 -0.33051246 -0.688006284  0.15483697 -0.13109227
#>   [8,] -0.03757413  0.35418146 -0.523498928  0.65845561  0.82019897
#>   [9,] -0.24641166  0.90672819 -0.511566260  0.84055934 -0.53503605
#>  [10,] -0.23548459  0.79121144 -0.031740350  0.73182245 -0.11925819
#>  [11,] -0.30771084 -0.28243246 -0.148273742 -0.03637882  0.47020689
#>  [12,]  0.42129193 -0.88275350 -0.575869049 -0.81732217  0.02232247
#>  [13,]  0.93325345 -0.73521553  0.601062640 -0.94959236  0.79630486
#>  [14,] -0.26646719  0.69371219  0.340681326  0.03493922  0.19514706
#>  [15,]  0.02015422 -0.86320850 -0.546171010  0.02216132 -0.30618760
#>  [16,]  0.14691753  0.65781939  0.980757964 -0.42710494 -0.80312652
#>  [17,] -0.40968380 -0.23603347 -0.006278979 -0.99782564  0.07590558
#>  [18,]  0.25413147 -0.72512905  0.458759894 -0.71689540  0.04506134
#>  [19,] -0.13132966 -0.00167060  0.314547290 -0.93121804 -0.47067106
#>  [20,] -0.56899290 -0.33194868  0.137691441 -0.88538906 -0.11538445
#>  [21,]  0.30173566  0.90262539 -0.115637568 -0.65494491  0.26093534
#>  [22,] -0.04821433 -0.09648364  0.226104281 -0.73118631  0.51862911
#>  [23,] -0.29028834 -0.34175911  0.895649328  0.68842393 -0.33228865
#>  [24,]  0.08533624 -0.70671689  0.566307058 -0.91263018 -0.36650533
#>  [25,] -0.21492855  0.99287133 -0.538398626  0.34865355 -0.30213817
#>  [26,]  0.33144451  0.29063179  0.394957645  0.94236005  0.45823295
#>  [27,] -0.80403414 -0.10511257 -0.720226958  0.30160608 -0.14787282
#>  [28,]  0.28996333 -0.86509858 -0.040204106 -0.76649818  0.15910313
#>  [29,] -0.28639193  0.49816799  0.976307451 -0.40339198  0.27851988
#>  [30,] -0.46607529  0.17212160 -0.874723316 -0.12614505  0.82110784
#>  [31,]  0.91142966  0.44073449  0.632839018  0.56566868 -0.61800014
#>  [32,]  0.17518788  0.74234666  0.657191578 -0.93325374  0.36768109
#>  [33,]  0.41660360  0.28833258 -0.258516442  0.03895645 -0.74878089
#>  [34,]  0.77633105 -0.10403356 -0.886607139 -0.66732061 -0.16604128
#>  [35,]  0.08018579  0.44216421  0.225521589 -0.26481068 -0.52060221
#>  [36,] -0.42227843  0.87351595  0.859156157 -0.97345323  0.02678401
#>  [37,] -0.26693135 -0.53674722 -0.135959622 -0.14535966 -0.97333345
#>  [38,] -0.42631888  0.37198635  0.508631432 -0.04186825 -0.17820434
#>  [39,] -0.04325522  0.78849549  0.735876723  0.58690136 -0.93069704
#>  [40,] -0.19469423 -0.47097778 -0.926327489  0.97149654  0.87843982
#>  [41,]  0.56866711 -0.34914947 -0.237043603  0.92413483 -0.13110193
#>  [42,]  0.64248704  0.76966980  0.223755263  0.01849597 -0.88431862
#>  [43,] -0.92817259  0.76801678 -0.972328037  0.76847360 -0.28766404
#>  [44,] -0.94466074  0.93982228  0.672693251 -0.47448906 -0.96281857
#>  [45,]  0.14168855 -0.85450338  0.809237175 -0.64856962 -0.23296170
#>  [46,]  0.03331676  0.32784414  0.753530870 -0.44176309 -0.22864400
#>  [47,] -0.28344255 -0.09554383 -0.814595901 -0.12970252  0.82767227
#>  [48,] -0.10962814  0.42537958 -0.261466803 -0.65593140  0.22295090
#>  [49,]  0.19192766  0.28205670  0.716508450 -0.83549671  0.54803217
#>  [50,] -0.42298430  0.51240401 -0.109979282 -0.85048318 -0.01588618
#>  [51,]  0.98311219  0.30453174  0.871869226 -0.50775788 -0.44811638
#>  [52,] -0.99815284 -0.11150146 -0.079881154  0.32963269 -0.83780264
#>  [53,] -0.49234402 -0.84103004  0.343450233  0.91868297  0.94335206
#>  [54,] -0.87322211  0.19389587  0.359883705 -0.41291406 -0.27572200
#>  [55,] -0.41502543 -0.44485774  0.144688208  0.29808302  0.43194196
#>  [56,] -0.51395801 -0.38805608 -0.549385659  0.56934888  0.33015856
#>  [57,]  0.13274818 -0.83246364  0.208851207 -0.27025163  0.59734355
#>  [58,]  0.72341772 -0.94296473  0.156418301 -0.85295889  0.92611939
#>  [59,]  0.32646600 -0.25468333 -0.981995346  0.88302803 -0.32145316
#>  [60,] -0.56086257 -0.39136597  0.689063127 -0.81448451  0.70425643
#>  [61,]  0.24193744 -0.86820874  0.792296852  0.03002663 -0.89686663
#>  [62,]  0.15851417  0.63326667 -0.090878061 -0.44078397 -0.45233427
#>  [63,]  0.86991412  0.79193869 -0.767847340 -0.11682849 -0.63161176
#>  [64,]  0.07225190 -0.12949298  0.871374511 -0.98741976  0.83413569
#>  [65,]  0.63872517  0.43514948 -0.342683589  0.46002322  0.80397888
#>  [66,] -0.80560249  0.08754990 -0.767724942  0.37805704  0.76004081
#>  [67,] -0.79862145  0.43721831  0.193533169 -0.72456448  0.71169075
#>  [68,]  0.52039103 -0.14015298  0.525424648  0.52539560  0.51689944
#>  [69,] -0.69528849 -0.98837319 -0.869703069 -0.03004132 -0.13222921
#>  [70,] -0.27103367 -0.60015233  0.420372752 -0.31808290  0.55028193
#>  [71,]  0.01921681  0.94794630  0.478853474  0.82432681  0.35924864
#>  [72,]  0.26428142  0.22515489  0.833331713 -0.37226056 -0.78019626
#>  [73,]  0.32276616  0.62913756  0.464985257 -0.08996466  0.20653617
#>  [74,] -0.09930684  0.53406807  0.984546049  0.15904322  0.91597035
#>  [75,] -0.45466057 -0.36056150 -0.775228607 -0.89316957 -0.58647283
#>  [76,] -0.50723968 -0.27243421 -0.781222415 -0.17220593  0.11809305
#>  [77,]  0.07351974  0.81664796 -0.094247914 -0.41317068  0.09682741
#>  [78,]  0.01988341  0.61104241 -0.527685085 -0.21360567  0.50363748
#>  [79,]  0.10945618 -0.63475028 -0.652786429  0.82983070  0.33111083
#>  [80,]  0.96991871 -0.05074280 -0.622315697 -0.57281369 -0.77293887
#>  [81,] -0.57796782  0.52824559 -0.138690855 -0.51847184 -0.17537017
#>  [82,]  0.91739723  0.61691695  0.962849197 -0.86785158  0.94160375
#>  [83,] -0.47044772  0.30296728  0.893266642 -0.49544651 -0.90263637
#>  [84,]  0.72762902  0.34180927 -0.424331739  0.54096519  0.30259258
#>  [85,] -0.90791856  0.12621731  0.707266953  0.37287482  0.45356484
#>  [86,] -0.56155429  0.31725634  0.797930480  0.47963592 -0.44125934
#>  [87,] -0.05900099  0.31892481  0.748165501 -0.86392720 -0.42301977
#>  [88,] -0.70814577  0.19820770  0.677572050  0.25706231  0.80023050
#>  [89,]  0.01774626  0.46851918  0.288692706 -0.65946600 -0.25146592
#>  [90,] -0.39455361 -0.08509153 -0.235214368 -0.98296121 -0.54740224
#>  [91,] -0.79412106  0.91362558 -0.780764255 -0.36187945 -0.13401148
#>  [92,] -0.69827041  0.10271195  0.115168724  0.74310588  0.26962615
#>  [93,]  0.61912260  0.14738412  0.067082634  0.45540277 -0.49707163
#>  [94,]  0.23412742 -0.81687885  0.537734358  0.03366702  0.43655209
#>  [95,] -0.93420841 -0.21573690 -0.846656679  0.84544838 -0.14048942
#>  [96,]  0.88231526  0.60426307  0.631809099  0.93183940 -0.98067520
#>  [97,] -0.27277927 -0.23159463 -0.260564919 -0.70419519  0.31175640
#>  [98,] -0.50489746  0.57265640  0.209169126  0.31945088 -0.60078216
#>  [99,] -0.98759642 -0.84131144  0.598858617  0.13368037  0.52204801
#> [100,] -0.48391159 -0.96706603 -0.137568593  0.13051685 -0.35219751
#> 
#> $y
#>   [1] 0 0 0 1 0 0 1 1 0 0 1 0 0 1 0 1 1 1 1 1 0 1 1 1 0 1 0 0 1 1 0 1 1 1 1 0 1
#>  [38] 1 1 1 1 0 0 0 0 1 1 1 1 1 0 0 0 0 1 1 0 0 1 1 0 1 0 1 1 0 0 1 0 1 0 1 1 1
#>  [75] 1 1 0 1 1 0 1 0 1 0 0 1 1 1 1 1 0 1 1 0 0 0 1 1 0 0
#>