Simulate Data from Hyper-Sphere for Classification Problems

The classification problem Circle is described in the BARK paper (2008). Inputs are dim independent variables uniformly distributed on the interval \([-1,1]\), only the first 2 out of these dim are actually signals. Outputs are created according to the formula $$y = 1(x1^2+x2^2 \le 2/\pi)$$

Usage

sim_circle(n, dim = 5)

Arguments

n

number of data points to generate

dim

number of dimension of the problem, no less than 2

Value

Returns a list with components

x

input values (independent variables)

y

0/1 output values (dependent variable)

References

Ouyang, Zhi (2008) Bayesian Additive Regression Kernels. Duke University. PhD dissertation, Chapter 3.

See also

Examples

sim_circle(n=100, dim=5)
#> $x
#>                [,1]        [,2]         [,3]        [,4]        [,5]
#>   [1,] -0.633331520 -0.27497140 -0.634339766 -0.43404403 -0.97538451
#>   [2,]  0.128593748  0.31860613 -0.588902177 -0.40108627 -0.81375541
#>   [3,]  0.115347190  0.68115050 -0.463793562  0.84605702  0.27291677
#>   [4,] -0.195920723 -0.98922104 -0.042351491 -0.60514044 -0.24396511
#>   [5,]  0.227704688 -0.44016603  0.524748719 -0.56346139 -0.45525019
#>   [6,]  0.593480784  0.27642079  0.468217716  0.21201058  0.91582251
#>   [7,]  0.059240056  0.24836785 -0.939354430  0.50238008  0.64013566
#>   [8,]  0.259049730 -0.09862506  0.035409785  0.78683376 -0.23387881
#>   [9,] -0.189169566  0.45169397  0.006476560 -0.57234221 -0.87251432
#>  [10,]  0.521875794  0.94263818  0.845531838 -0.65550955 -0.60701030
#>  [11,] -0.186624643 -0.88019637 -0.021125837  0.42345817  0.40221590
#>  [12,]  0.270206193 -0.47519434  0.476901720  0.75595135  0.37477615
#>  [13,]  0.778698995  0.53351212 -0.085172086  0.73419562  0.31671515
#>  [14,] -0.381058561  0.55942479 -0.642338869 -0.05048466  0.67576598
#>  [15,] -0.991806238  0.74079636  0.984721646  0.42642439  0.67146814
#>  [16,] -0.384936198  0.56869965 -0.430203730  0.20836231  0.96266466
#>  [17,] -0.895659730 -0.35895352 -0.581922495  0.69441216  0.50911618
#>  [18,] -0.060795726 -0.98784303  0.323996676 -0.07897282 -0.15899808
#>  [19,]  0.926065792  0.95128303 -0.235110169  0.21736925 -0.15103778
#>  [20,] -0.370889691 -0.99579282 -0.587373880  0.08731083 -0.53162712
#>  [21,] -0.735162804 -0.57524974 -0.510013110 -0.03149834  0.57673671
#>  [22,]  0.078632311  0.63808516 -0.574153788  0.86188157  0.38809479
#>  [23,] -0.940862035  0.49794247  0.586265264  0.39494474 -0.37859190
#>  [24,]  0.677679727  0.34482198 -0.002896788  0.47492256  0.79030880
#>  [25,]  0.646095674 -0.23810416  0.227524244  0.74041122  0.30401707
#>  [26,]  0.750659580  0.74441218  0.041831801 -0.52440691 -0.77261611
#>  [27,]  0.887608064 -0.26211493  0.537840201 -0.76580828  0.88712054
#>  [28,]  0.198060588  0.10450891 -0.622340317 -0.03957169  0.31482720
#>  [29,]  0.725098789 -0.86810863 -0.044102530  0.54980678  0.56067128
#>  [30,]  0.573908086  0.96982657 -0.316431037 -0.03278978 -0.54807812
#>  [31,] -0.786733455 -0.38230274 -0.543442627 -0.66676302  0.57870558
#>  [32,] -0.813025144  0.07669469 -0.403389851 -0.97772908  0.49334833
#>  [33,]  0.560102595 -0.46787302 -0.798730068 -0.60835170 -0.06027769
#>  [34,] -0.318579926 -0.58426314  0.073261994 -0.12814038 -0.12744042
#>  [35,] -0.410143171  0.13145119  0.163757298  0.99171926 -0.07788017
#>  [36,] -0.938819175 -0.84985863  0.824251522  0.17295065  0.41089330
#>  [37,] -0.528356495  0.62535022  0.849950274 -0.42539625  0.05631634
#>  [38,] -0.546292884 -0.33397886  0.683584453  0.89222427 -0.42308000
#>  [39,] -0.576142997  0.31003165 -0.486585936  0.87597257  0.28806332
#>  [40,]  0.725725259  0.60962899 -0.272995782  0.46925319 -0.68593080
#>  [41,]  0.402553878 -0.67840799  0.664861889  0.79977579  0.73448045
#>  [42,] -0.698989445  0.15127592 -0.219409226 -0.29542327 -0.41599938
#>  [43,] -0.020778134  0.98489423 -0.239484914  0.70994293 -0.86740100
#>  [44,]  0.259078207 -0.97059851  0.556712150 -0.80580018 -0.38251866
#>  [45,] -0.200892930  0.47569309  0.114020912  0.57138344 -0.37580559
#>  [46,] -0.814094881  0.39609848  0.163656930 -0.32711008 -0.75637281
#>  [47,]  0.943401201 -0.73274489 -0.939258294  0.31260248  0.39225970
#>  [48,]  0.589645174  0.44313090  0.277237689 -0.53109943  0.27652378
#>  [49,] -0.561066949  0.38314334 -0.952332486  0.47459455 -0.92441351
#>  [50,]  0.439763052 -0.10445921  0.595648132  0.86151809  0.82965689
#>  [51,]  0.929951278  0.54801569  0.002681090 -0.98156509 -0.59345669
#>  [52,] -0.424122946 -0.37264873 -0.491641292  0.16286270  0.20471104
#>  [53,]  0.299176653 -0.95999713  0.866994080  0.97749855 -0.60895893
#>  [54,]  0.307453678  0.27418894 -0.865312417 -0.29035032  0.96856585
#>  [55,]  0.915728274 -0.08915785 -0.128612452 -0.49953622  0.14631663
#>  [56,]  0.138969509  0.57385129  0.477438050  0.47035174 -0.49500912
#>  [57,] -0.952529995 -0.73726878  0.491677154 -0.89506619 -0.20614318
#>  [58,] -0.685615230  0.74045735 -0.965248691  0.09489869  0.38877658
#>  [59,]  0.623208219 -0.43184488 -0.721656924  0.76735675  0.05686615
#>  [60,]  0.009743789 -0.80007393  0.894717193  0.58338885 -0.25316852
#>  [61,] -0.425328127 -0.99909006 -0.082107446  0.43960689  0.60831366
#>  [62,] -0.087571256  0.27274617 -0.470208414 -0.16467302  0.34439824
#>  [63,] -0.260070947  0.02870696  0.320415156 -0.84695516  0.37000155
#>  [64,]  0.089854046 -0.91221943  0.434038418 -0.52460034 -0.00207905
#>  [65,]  0.772171241  0.83098210 -0.398520507 -0.01655942  0.11812411
#>  [66,]  0.732740274  0.65444741  0.275701966 -0.74211773 -0.84085060
#>  [67,]  0.989174848 -0.79828835  0.055446018 -0.96575049  0.30516689
#>  [68,] -0.124210239 -0.86954699 -0.314460049 -0.56579372 -0.02648957
#>  [69,] -0.585543209 -0.15797173  0.287830100  0.32487513  0.24168456
#>  [70,]  0.909584220 -0.24721628  0.088534411  0.23703712  0.31628538
#>  [71,]  0.808047915  0.89444412 -0.636061653  0.90931138  0.46014041
#>  [72,]  0.615021407  0.42547789 -0.565387789  0.23635581 -0.31156075
#>  [73,] -0.162791940 -0.43227133  0.404626094  0.27813980  0.05240118
#>  [74,] -0.678960255 -0.36528872 -0.016354561  0.37987190  0.79957853
#>  [75,]  0.498961956 -0.04678336 -0.806126410 -0.66773830 -0.10136880
#>  [76,]  0.377011818  0.60039509 -0.995727379 -0.88905332  0.92511349
#>  [77,] -0.487453880 -0.30631256  0.133368673 -0.35351265 -0.83367181
#>  [78,]  0.494798507 -0.50590800 -0.747065849  0.43243703  0.88559898
#>  [79,]  0.816422363 -0.88194824  0.843349353  0.09782524  0.19598285
#>  [80,]  0.824952974  0.09767950  0.325641091 -0.61194529 -0.37234835
#>  [81,]  0.210247748 -0.33766467 -0.325891436  0.28086571 -0.60231719
#>  [82,] -0.469341823 -0.56357824  0.735419870  0.07829615 -0.50858227
#>  [83,] -0.142365145 -0.02967648  0.477799535 -0.31272478 -0.27075132
#>  [84,]  0.765675308 -0.99660903 -0.113807735  0.88066527  0.24616142
#>  [85,]  0.732930383 -0.20586474  0.073865024  0.07502360 -0.78351886
#>  [86,]  0.907588562  0.68973717  0.762770230  0.41526441  0.74588757
#>  [87,] -0.631678608 -0.09698204 -0.726942056  0.04657243 -0.56870148
#>  [88,]  0.710183995  0.27388920 -0.936832442 -0.26328517 -0.13756457
#>  [89,]  0.421330992 -0.81957996  0.316717296 -0.37067643  0.35301369
#>  [90,] -0.859692583  0.81234564 -0.988150407  0.20961623  0.13458787
#>  [91,]  0.245102538 -0.83566044 -0.403363832  0.29974538 -0.80168767
#>  [92,] -0.827280639  0.09086426  0.894263302  0.72863675  0.14262501
#>  [93,] -0.772792355 -0.35146833 -0.977657427  0.37012277  0.20461002
#>  [94,]  0.967433424  0.78483731 -0.299738396 -0.50327218  0.41269387
#>  [95,] -0.191870570  0.42594384  0.269955662  0.07296640 -0.99338869
#>  [96,]  0.489331944 -0.74672132 -0.198522171  0.08562716  0.23929109
#>  [97,] -0.676540194  0.71274615  0.049722551 -0.25971256 -0.89242136
#>  [98,]  0.512923948  0.51082557 -0.764083650  0.51778047 -0.41656189
#>  [99,]  0.110932868  0.35804638 -0.218894248  0.08800500 -0.06047573
#> [100,] -0.853987152  0.84077410 -0.708503340 -0.63078146 -0.70688269
#> 
#> $y
#>   [1] 1 1 1 0 1 1 1 1 1 0 0 1 0 1 0 1 0 0 0 0 0 1 0 1 1 0 0 1 0 0 0 0 1 1 1 0 0
#>  [38] 1 1 0 1 1 0 0 1 0 0 1 1 1 0 1 0 1 0 1 0 0 1 0 0 1 1 0 0 0 0 0 1 0 0 1 1 1
#>  [75] 1 1 1 1 0 0 1 1 1 0 1 0 1 1 0 0 0 0 0 0 1 0 0 1 1 0
#>