Simulate Data from Hyper-Sphere for Classification Problems

The classification problem Circle is described in the BARK paper (2008). Inputs are dim independent variables uniformly distributed on the interval \([-1,1]\), only the first 2 out of these dim are actually signals. Outputs are created according to the formula $$y = 1(x1^2+x2^2 \le 2/\pi)$$

sim_circle(n, dim = 5)

Arguments

n

number of data points to generate

dim

number of dimension of the problem, no less than 2

Value

Returns a list with components

x

input values (independent variables)

y

0/1 output values (dependent variable)

References

Ouyang, Zhi (2008) Bayesian Additive Regression Kernels. Duke University. PhD dissertation, Chapter 3.
https://stat.duke.edu/people/theses/OuyangZ.html

See also

Examples

sim_circle(n=100, dim=5)
#> $x
#>                [,1]         [,2]         [,3]         [,4]         [,5]
#>   [1,]  0.410567869  0.052099216 -0.125625792  0.572460813 -0.243558515
#>   [2,]  0.576177093  0.118559187 -0.211503902  0.029164153  0.736962467
#>   [3,]  0.415735733 -0.555143859  0.800469481  0.879698575  0.884927492
#>   [4,]  0.688707141 -0.369871981 -0.498464855 -0.303655834  0.314979286
#>   [5,]  0.203716539 -0.035221162 -0.239524393  0.370334897  0.979408158
#>   [6,]  0.366796191  0.490699069 -0.797281757  0.633619470  0.910324413
#>   [7,]  0.663387909 -0.249513804 -0.127385407  0.246226046 -0.262117160
#>   [8,]  0.408305049 -0.047027132  0.628470876  0.839378727 -0.488412879
#>   [9,] -0.040309193 -0.477417306  0.800706699 -0.837627582  0.143532112
#>  [10,] -0.717852117  0.664045825  0.048940086  0.007154909  0.035649318
#>  [11,] -0.205537995  0.087806860  0.638094450 -0.825086364  0.557938261
#>  [12,] -0.622390887 -0.980388824  0.801124535  0.586242855  0.956879310
#>  [13,]  0.320005738 -0.553407745  0.217945097 -0.823795131 -0.692228723
#>  [14,]  0.228796144  0.460810168 -0.436286130  0.974469696  0.228165983
#>  [15,] -0.359341633 -0.114305429  0.609448203 -0.937433927 -0.292757975
#>  [16,]  0.620864906  0.606666107 -0.922044274 -0.197205894  0.746807534
#>  [17,] -0.721693052  0.089470085  0.896343842 -0.185873750  0.835047412
#>  [18,] -0.182877370 -0.179093849  0.470381531  0.504238666 -0.645357286
#>  [19,]  0.635157610 -0.230908864  0.620979832 -0.651485483 -0.353795039
#>  [20,]  0.447134149 -0.493561376  0.320443629 -0.991724181 -0.631085800
#>  [21,]  0.112686208 -0.884510631 -0.577607162 -0.933562085  0.815875950
#>  [22,] -0.211818545  0.607503299 -0.977104459 -0.328097642  0.917250818
#>  [23,]  0.550385239  0.322150462  0.085586647 -0.612983572  0.771883026
#>  [24,] -0.796827441 -0.306104977  0.917583885  0.487862099  0.499465474
#>  [25,]  0.633417489 -0.440821784  0.729752067 -0.828020869  0.323977785
#>  [26,] -0.867568283  0.992938389  0.572830401  0.621571322 -0.809024244
#>  [27,]  0.492786432 -0.924756765  0.464631572 -0.691109225  0.452804112
#>  [28,]  0.741208563 -0.279940571 -0.096066866 -0.598372751 -0.805344727
#>  [29,] -0.893316087 -0.793300644 -0.014446267  0.500338950 -0.366526194
#>  [30,]  0.038533251 -0.584342463 -0.932420087  0.654287575 -0.995238912
#>  [31,] -0.486182236  0.709856014  0.769013622  0.029821281 -0.604779153
#>  [32,]  0.844184241  0.240305279  0.193489766  0.099700222 -0.315575613
#>  [33,]  0.811485847 -0.535745639 -0.010411776 -0.581188563  0.432140870
#>  [34,] -0.345057141 -0.308341246  0.935646103  0.581452332  0.847940181
#>  [35,]  0.582179922 -0.297947812 -0.092429403  0.418795897 -0.677891450
#>  [36,] -0.290714326  0.885081397  0.277263742 -0.230619130  0.287943518
#>  [37,] -0.877543636 -0.325288999  0.079364018  0.463451724 -0.586606634
#>  [38,]  0.532288770 -0.108816484 -0.523272861 -0.726285790 -0.009253968
#>  [39,]  0.860475331 -0.180289314 -0.920304622 -0.948185849 -0.025272181
#>  [40,]  0.555553436  0.965220918 -0.587855097  0.364422457 -0.602415468
#>  [41,]  0.305314985  0.224958394 -0.594641259  0.020655085  0.472178740
#>  [42,] -0.357254842  0.132038702  0.106102280 -0.352230787  0.060758879
#>  [43,]  0.088675068 -0.009827604  0.391533044 -0.037523813  0.735438534
#>  [44,] -0.993502508  0.773241473 -0.050381355 -0.285182138 -0.026310745
#>  [45,]  0.666474337  0.452089797 -0.282365667  0.438794646 -0.999834085
#>  [46,] -0.126282084  0.469556049 -0.262993437  0.808219284 -0.199262657
#>  [47,]  0.579432425  0.328491251  0.455034511 -0.439640159  0.114707990
#>  [48,]  0.043043973  0.818888232  0.042378514  0.170975058 -0.624796002
#>  [49,] -0.619138672 -0.214855865 -0.003251935 -0.586624513  0.661218140
#>  [50,] -0.595743706  0.036237823  0.473936577 -0.431505751 -0.325458978
#>  [51,] -0.929281728  0.136521352  0.988791243  0.305273534  0.904258305
#>  [52,] -0.941019465  0.790984966  0.371069201 -0.574865414 -0.562275039
#>  [53,]  0.676770827  0.627405011 -0.447144674  0.852161275  0.625448987
#>  [54,] -0.461227342 -0.198729493  0.754444123  0.560691785  0.384684511
#>  [55,]  0.830246343  0.591533033  0.790408238  0.293332828 -0.573197150
#>  [56,] -0.337336999 -0.076964410  0.892778864  0.532610787 -0.090574692
#>  [57,]  0.736210154  0.458253620 -0.157034603 -0.208347686  0.403288830
#>  [58,]  0.730739064  0.887898000 -0.033887885  0.934059133  0.238223367
#>  [59,] -0.948834640 -0.981406459 -0.851297153 -0.667613907  0.591840603
#>  [60,] -0.657133278 -0.585162494  0.368784382 -0.179008597 -0.484812081
#>  [61,] -0.793348948  0.701863714  0.570604459  0.743211843 -0.750920009
#>  [62,]  0.545795281  0.148474156 -0.942881425  0.702197364  0.622117686
#>  [63,]  0.884528985  0.063584880  0.522506776  0.738472287 -0.357747507
#>  [64,]  0.825700990 -0.674457967 -0.024048276  0.393518539 -0.375846556
#>  [65,]  0.081869271  0.840656233  0.037934448 -0.494163436  0.371046270
#>  [66,]  0.229837230  0.212015149 -0.701152088 -0.919859665  0.047522966
#>  [67,] -0.370777036 -0.369817391  0.941779454  0.356349200 -0.871501687
#>  [68,]  0.043283266  0.930647758 -0.439904090  0.885304469  0.281137770
#>  [69,] -0.637695048  0.991885830 -0.248578478  0.764462568  0.697720946
#>  [70,]  0.034138010  0.195959474 -0.644874163  0.097595381 -0.730996851
#>  [71,] -0.012352884  0.311734534  0.003557323  0.939845942 -0.403370667
#>  [72,]  0.350830373  0.419367831  0.967882601  0.100561392  0.457412317
#>  [73,] -0.007084301 -0.347638264  0.944273124  0.259969678  0.980211610
#>  [74,] -0.155593663 -0.472033508  0.921070002  0.240588231 -0.315071000
#>  [75,] -0.487959709 -0.182201720  0.932259741  0.960581115  0.678300702
#>  [76,]  0.628788330  0.732970904  0.047417525 -0.338688132 -0.306226726
#>  [77,]  0.423368534  0.061433327 -0.739619156 -0.432545885 -0.462211542
#>  [78,] -0.883387262  0.103729523 -0.608610414  0.086201106 -0.621087775
#>  [79,]  0.990515785  0.849088495  0.443669942  0.692733306  0.494243154
#>  [80,] -0.600028588  0.202851244  0.509156574 -0.861565290  0.231340921
#>  [81,] -0.998973958  0.589119678  0.965608191 -0.727710374 -0.196783810
#>  [82,]  0.711331704  0.702316036 -0.288734035 -0.169265434  0.741013917
#>  [83,] -0.652946891 -0.298406739 -0.381930009 -0.372034252 -0.635450379
#>  [84,] -0.014737375 -0.307886365 -0.858798784 -0.194877173  0.264851016
#>  [85,]  0.270201162  0.338458542 -0.130445336 -0.833376313 -0.360645384
#>  [86,] -0.725952469 -0.034563655 -0.459349753 -0.006579322  0.282921243
#>  [87,]  0.113233321 -0.972799199  0.691470821  0.585531649  0.516124182
#>  [88,]  0.241423106 -0.287924783 -0.316129967  0.099405582  0.410112462
#>  [89,]  0.230272321 -0.215868940  0.742045641 -0.210550244 -0.441128654
#>  [90,]  0.519825225 -0.886262249  0.852897504  0.150000584  0.549062094
#>  [91,]  0.950752043 -0.939848880  0.323159147 -0.488807230  0.790666130
#>  [92,] -0.678450472 -0.322671837 -0.335189553 -0.605888746 -0.227381582
#>  [93,]  0.367487821 -0.392293275 -0.560718625  0.478565922  0.194470705
#>  [94,]  0.056015624 -0.855303719  0.746858822 -0.763460238 -0.710762498
#>  [95,] -0.267801790  0.845852307  0.791912077  0.112336149 -0.173153820
#>  [96,] -0.044333974 -0.155911926 -0.300719639  0.284943567 -0.600013004
#>  [97,]  0.765608514  0.368846102  0.861573085  0.753819525 -0.564125708
#>  [98,] -0.438299327 -0.317553660  0.453167580  0.960838025 -0.627449931
#>  [99,] -0.643567102 -0.971396245 -0.964350261 -0.679408612  0.023600808
#> [100,] -0.702601249 -0.892718024 -0.240418093 -0.618125516 -0.197191921
#> 
#> $y
#>   [1] 1 1 1 1 1 1 1 1 1 0 1 0 1 1 1 0 1 1 1 1 0 1 1 0 1 0 0 1 0 1 0 0 0 1 1 0 0
#>  [38] 1 0 0 1 1 1 0 0 1 1 0 1 1 0 0 0 1 0 1 0 0 0 0 0 1 0 0 0 1 1 0 0 1 1 1 1 1
#>  [75] 1 0 1 0 0 1 0 0 1 1 1 1 0 1 1 0 0 1 1 0 0 1 0 1 0 0
#>