Simulate Data from Hyper-Sphere for Classification Problems

The classification problem Circle is described in the BARK paper (2008). Inputs are dim independent variables uniformly distributed on the interval \([-1,1]\), only the first 2 out of these dim are actually signals. Outputs are created according to the formula $$y = 1(x1^2+x2^2 \le 2/\pi)$$

sim_circle(n, dim = 5)

Arguments

n

number of data points to generate

dim

number of dimension of the problem, no less than 2

Value

Returns a list with components

x

input values (independent variables)

y

0/1 output values (dependent variable)

References

Ouyang, Zhi (2008) Bayesian Additive Regression Kernels. Duke University. PhD dissertation, Chapter 3.

See also

Examples

sim_circle(n=100, dim=5)
#> $x
#>                [,1]          [,2]         [,3]        [,4]          [,5]
#>   [1,] -0.513330844  0.0347121321 -0.708402940 -0.22045623  0.9455617946
#>   [2,] -0.450324621 -0.5285936724  0.259178786 -0.44272512 -0.2700564391
#>   [3,] -0.246414712 -0.4700010270 -0.989040732 -0.26977196 -0.2467003060
#>   [4,]  0.658320854  0.5001200647  0.858934754 -0.11474411  0.4869942041
#>   [5,] -0.452151998  0.3006539745  0.922253778  0.76774459 -0.2923890697
#>   [6,] -0.496467112  0.5477086101 -0.880481710  0.15750646 -0.0572079122
#>   [7,]  0.615344208 -0.0446285880 -0.066985326  0.53534589 -0.5160455392
#>   [8,] -0.335937551  0.2762336624 -0.051208884  0.99380933 -0.5913824383
#>   [9,] -0.954432982  0.5950903608  0.103489771  0.94680698  0.0321698100
#>  [10,]  0.290170346 -0.6673060036  0.201095785 -0.84792104 -0.4707641583
#>  [11,] -0.161253920  0.6779504740 -0.675526697  0.91144293  0.5090571600
#>  [12,]  0.227565395  0.4310090407  0.555435408  0.14873219 -0.0310504385
#>  [13,] -0.484981177 -0.5811755974  0.363624398 -0.37665375  0.9024051949
#>  [14,]  0.052080702 -0.6724122022  0.880152104  0.06588363  0.1875023558
#>  [15,]  0.927756296  0.0263554524 -0.602947718 -0.18000690 -0.5072117639
#>  [16,]  0.775058870 -0.4256500155 -0.796205176 -0.15349313  0.3947466477
#>  [17,] -0.659208719  0.3915507994  0.578665167  0.63631884 -0.2799952286
#>  [18,]  0.923806124 -0.8419282306 -0.114755407 -0.21058151 -0.6382830017
#>  [19,]  0.287526804  0.2047155672  0.191698565 -0.23302091  0.0413382109
#>  [20,]  0.887450768  0.2623762572  0.220478099 -0.14404106  0.4039080660
#>  [21,]  0.912727912  0.9030415141  0.411994560  0.76075551  0.7738901949
#>  [22,]  0.886505615  0.0003631325  0.330933072 -0.24294049  0.0562013276
#>  [23,] -0.105696179 -0.5322047044  0.404180675 -0.22348364  0.9392662928
#>  [24,]  0.506901070  0.8607276827 -0.380216492  0.62033320  0.4688644488
#>  [25,] -0.382934740  0.6090356051  0.678097897  0.67224118 -0.5783444387
#>  [26,]  0.701385522 -0.0788139063 -0.671313004  0.27241780 -0.2467128597
#>  [27,]  0.783948511  0.4792924840 -0.996831066  0.26473523 -0.4135890417
#>  [28,] -0.755872694  0.9504047087 -0.864427496  0.26441177  0.5615803353
#>  [29,]  0.628802594  0.5276301024  0.409680286  0.26598766  0.4045113618
#>  [30,] -0.645700352  0.3677721145  0.690398930  0.88775948 -0.0092341299
#>  [31,] -0.886440076 -0.2865475430  0.820706208  0.32871142  0.0077759321
#>  [32,]  0.507261872 -0.9033391103  0.514173910  0.97086596 -0.1542402520
#>  [33,] -0.985434794  0.9777833275  0.463928687 -0.44319588 -0.6860361500
#>  [34,]  0.312573523  0.2287495644 -0.959906478 -0.27480533 -0.7385089318
#>  [35,] -0.212862673 -0.3674408770  0.557673819 -0.77960289 -0.3785415762
#>  [36,] -0.895960655 -0.7118080365 -0.744660803 -0.82157170 -0.3757805447
#>  [37,] -0.481506674  0.6638314929 -0.328934768 -0.46282574  0.8991043037
#>  [38,]  0.819479055 -0.2835759809  0.217560056  0.03757458 -0.8791661384
#>  [39,] -0.727753434  0.1953452998  0.324626149  0.14845911 -0.0793124172
#>  [40,] -0.122669910  0.3523135087  0.272657993 -0.15278445 -0.0690209721
#>  [41,]  0.312688872  0.1014756351  0.805383063 -0.44823250 -0.8263423275
#>  [42,] -0.899438989 -0.4236641577 -0.284526824  0.08352355 -0.7213571141
#>  [43,] -0.175143373  0.4106805110  0.435164231  0.39381805 -0.0857848092
#>  [44,]  0.239122740  0.2523745629  0.500452395  0.78870070  0.7810041504
#>  [45,] -0.400482204  0.3565316005 -0.003739111  0.80515708  0.6076024845
#>  [46,]  0.193712650  0.7220697855 -0.854871443  0.52985418  0.6802800084
#>  [47,]  0.677979147  0.5083346963 -0.988135022  0.39614568 -0.0293375100
#>  [48,] -0.422093164 -0.3428209913 -0.120852395  0.51659684 -0.4659137991
#>  [49,] -0.735411922 -0.8258735724  0.443718412 -0.03309476 -0.5628153658
#>  [50,]  0.750859438  0.8965959749 -0.593862840  0.69460087 -0.9583157115
#>  [51,] -0.661120158 -0.1445841151 -0.887023403  0.28772599 -0.0464164955
#>  [52,] -0.154591995  0.8242314337  0.228048515  0.02609976  0.9485773025
#>  [53,]  0.654169479 -0.0437221858 -0.525561633 -0.15157127 -0.9210847742
#>  [54,] -0.941533361  0.5718308743 -0.845454998 -0.38395905  0.0003215885
#>  [55,]  0.533647983 -0.6133865444  0.283442370 -0.79608318 -0.4639913337
#>  [56,] -0.515599617  0.0144893741  0.721268762 -0.74548878 -0.6271009371
#>  [57,] -0.849785208  0.8873719228  0.275290293 -0.86831866 -0.9834342226
#>  [58,]  0.357604664  0.2565505276  0.473369519 -0.62667282  0.8581009661
#>  [59,]  0.990191916  0.9861752754 -0.388930365 -0.23551109  0.6639287313
#>  [60,]  0.597805535 -0.5923759346  0.948826418 -0.60948349  0.7905547810
#>  [61,]  0.853087278 -0.0807774588  0.601181164  0.74790707  0.1736276560
#>  [62,] -0.460899094 -0.5730453259  0.814458480  0.13855046 -0.3226881982
#>  [63,] -0.343643632  0.2790358406  0.433956266 -0.95291599  0.4541548449
#>  [64,]  0.666943686  0.1909600818  0.977730795  0.60614956 -0.2170683951
#>  [65,] -0.940803071 -0.1196411583  0.535955495  0.77678348 -0.7023443831
#>  [66,]  0.409432492  0.1132094208 -0.203659221 -0.45247839 -0.5397666479
#>  [67,] -0.802533937  0.9364350564 -0.423885759  0.47872716  0.3191420096
#>  [68,]  0.820374881 -0.8479157272 -0.308474167 -0.53010700 -0.1517388439
#>  [69,]  0.222972042  0.9032502724  0.267953104  0.91249539 -0.7473079800
#>  [70,]  0.514238915 -0.9617763017  0.995948958  0.57798115  0.4978998499
#>  [71,]  0.063315708 -0.6189640104  0.673096451 -0.11740502 -0.3392541492
#>  [72,] -0.951572489  0.5149484114  0.737209686 -0.50338344 -0.9272533306
#>  [73,] -0.361006859 -0.3208538145  0.830979179 -0.25346070  0.2832216523
#>  [74,]  0.825819044  0.8632487855  0.893751872  0.83222594  0.3917659903
#>  [75,] -0.497504496 -0.3677222445 -0.913203835  0.83364991  0.5497708162
#>  [76,]  0.518359928  0.2062247903 -0.699822303  0.18630701 -0.1394772409
#>  [77,] -0.075247922  0.9308349588 -0.867560467  0.53074269  0.4182933429
#>  [78,]  0.762541293 -0.3758682869  0.210194728  0.26291113 -0.5371340164
#>  [79,] -0.019434315 -0.3517865092  0.503651475  0.22989199  0.7007220183
#>  [80,]  0.601913569 -0.2058206876  0.383223708 -0.23154268 -0.0311642718
#>  [81,]  0.379890869 -0.0850230935 -0.010825584 -0.72459126 -0.7204725854
#>  [82,] -0.578931967  0.0412789122  0.844631089 -0.20306920 -0.0846991022
#>  [83,] -0.006456387 -0.8842177647 -0.240576327 -0.62152287 -0.2755946112
#>  [84,]  0.098063373  0.5418756539  0.821786333  0.88008445 -0.6493748850
#>  [85,] -0.659889644 -0.5391267072  0.303916147 -0.63723842 -0.3360063918
#>  [86,]  0.737437226  0.9626236041 -0.991848771  0.67384048  0.0266206535
#>  [87,] -0.113246356  0.7800164786 -0.032118186  0.28281852  0.9578103973
#>  [88,] -0.134404383  0.0754571366  0.926669979  0.90454191  0.8496257630
#>  [89,]  0.764185127  0.2156763724 -0.652923982  0.59475676  0.4640228273
#>  [90,]  0.392433156  0.5237569334  0.625745312  0.23355081  0.4167509116
#>  [91,]  0.914171441  0.9771909239  0.660342477 -0.34634398 -0.1025254093
#>  [92,]  0.194024157 -0.0728105162  0.227312783 -0.24347328 -0.2372503248
#>  [93,] -0.141577874  0.6770806299  0.410004762 -0.80674179 -0.1045974884
#>  [94,] -0.126719099 -0.9727869551  0.142972867 -0.51919780 -0.2636789870
#>  [95,] -0.494793155 -0.0238750107  0.785563709  0.16548099 -0.5411616438
#>  [96,] -0.048948053 -0.6984434989 -0.672576124  0.78752252 -0.0076418403
#>  [97,]  0.666635288  0.7299282444 -0.778055760  0.17513972  0.5946742622
#>  [98,] -0.509545166  0.4852729104 -0.077487923  0.38333844  0.2451785537
#>  [99,]  0.513743991  0.6607813961  0.750255598 -0.26325116 -0.7209468484
#> [100,] -0.039921342 -0.6075494173 -0.680724380 -0.93172347 -0.5876217014
#> 
#> $y
#>   [1] 1 1 1 0 1 1 1 1 0 1 1 1 1 1 0 0 1 0 1 0 0 0 1 0 1 1 0 0 0 1 0 0 0 1 1 0 0
#>  [38] 0 1 1 1 0 1 1 1 1 0 1 0 0 1 0 1 0 0 1 0 1 0 0 0 1 1 1 0 1 0 0 0 0 1 0 1 0
#>  [75] 1 1 0 0 1 1 1 1 0 1 0 0 1 1 1 1 0 1 1 0 1 1 0 1 0 1
#>