Simulate Data from Hyper-Sphere for Classification Problems

The classification problem Circle is described in the BARK paper (2008). Inputs are dim independent variables uniformly distributed on the interval \([-1,1]\), only the first 2 out of these dim are actually signals. Outputs are created according to the formula $$y = 1(x1^2+x2^2 \le 2/\pi)$$

Usage

sim_circle(n, dim = 5)

Arguments

n

number of data points to generate

dim

number of dimension of the problem, no less than 2

Value

Returns a list with components

x

input values (independent variables)

y

0/1 output values (dependent variable)

References

Ouyang, Zhi (2008) Bayesian Additive Regression Kernels. Duke University. PhD dissertation, Chapter 3.

See also

Examples

sim_circle(n=100, dim=5)
#> $x
#>                [,1]          [,2]        [,3]         [,4]         [,5]
#>   [1,] -0.926936613  4.868981e-01  0.06240194  0.538569878 -0.925652321
#>   [2,] -0.815784258 -6.396821e-01 -0.72379318  0.028146250  0.943498317
#>   [3,]  0.160725271  6.076963e-01 -0.11883692  0.147035654 -0.672695967
#>   [4,] -0.475725579  9.559185e-01 -0.56527115 -0.697350943  0.003467482
#>   [5,] -0.366433305  1.632198e-01 -0.05300840  0.194587056 -0.376832903
#>   [6,] -0.875650101  3.515178e-01  0.84388244  0.717186480 -0.229622333
#>   [7,] -0.829755022 -4.585261e-01 -0.50889416 -0.559266561 -0.349263296
#>   [8,] -0.143454231 -2.985881e-01 -0.45802623  0.707012234 -0.305619110
#>   [9,]  0.207467488  1.044298e-01  0.76579174 -0.949800869 -0.300471397
#>  [10,] -0.644776427  1.341855e-01 -0.65515357  0.632089979  0.953875842
#>  [11,]  0.213865324 -9.141177e-01 -0.51061385 -0.436249221  0.043004787
#>  [12,] -0.373156053 -5.928413e-01 -0.68064921  0.676617033 -0.361086737
#>  [13,] -0.536500596 -2.355142e-01  0.41696737  0.803346590  0.360347524
#>  [14,] -0.563014493  2.005708e-01 -0.99660049  0.504605203  0.006009315
#>  [15,]  0.413436932  9.541319e-01  0.64358418  0.058437754  0.706271638
#>  [16,] -0.489279788  3.175657e-01 -0.08544838  0.413759615 -0.361266509
#>  [17,] -0.243476676  1.448544e-01  0.53310306 -0.841788481  0.511326247
#>  [18,] -0.611047083  2.965811e-01 -0.84116259 -0.995152878  0.367265611
#>  [19,] -0.016661684 -1.743518e-01  0.09007129  0.541600808  0.296506995
#>  [20,] -0.135797421  4.212513e-01  0.94935576 -0.312724050  0.510988011
#>  [21,]  0.722246100 -5.325670e-01 -0.71784532 -0.855441840 -0.173579392
#>  [22,] -0.257795198  8.230388e-01 -0.05244776  0.800081309  0.895813816
#>  [23,] -0.226264150 -7.969984e-01 -0.18621726  0.368211215  0.356673773
#>  [24,] -0.277649775 -8.160520e-01  0.15235489  0.296479789 -0.442798910
#>  [25,] -0.996567922  3.448699e-01 -0.73850667  0.817383272  0.747632201
#>  [26,] -0.015616352 -1.787087e-01 -0.85715686  0.076092181 -0.504488626
#>  [27,]  0.812385136 -4.774834e-01  0.16062051 -0.976288727  0.193677773
#>  [28,] -0.449200912  1.571355e-01 -0.65064783  0.699628908 -0.445372010
#>  [29,] -0.670285263  4.725025e-01 -0.79141410  0.225663979  0.066122164
#>  [30,] -0.633148589 -8.868927e-01 -0.04415699  0.388104782 -0.155649482
#>  [31,] -0.725187732  5.113255e-01 -0.41513681  0.417894683  0.986510902
#>  [32,] -0.088255329 -2.143345e-01 -0.32950687 -0.787892686 -0.923002802
#>  [33,] -0.550749595 -4.596791e-01  0.37514181  0.135495777  0.245614705
#>  [34,]  0.387791618 -2.497310e-02 -0.84871485 -0.815806539 -0.684275818
#>  [35,]  0.917763927 -8.364316e-01 -0.92738558  0.020739669  0.389997893
#>  [36,]  0.707553728  4.556823e-01  0.35038186  0.146466645  0.390089473
#>  [37,]  0.451765270  3.660650e-01  0.28276124  0.441549306 -0.935850509
#>  [38,] -0.130679932 -5.938038e-01 -0.87232937 -0.923237034 -0.041352530
#>  [39,] -0.956714840  4.370081e-01 -0.03516490 -0.949612688  0.031387528
#>  [40,]  0.625065198  1.539332e-01  0.89717794  0.903769529  0.589613174
#>  [41,] -0.894079798 -7.203017e-01 -0.60212583 -0.194774958 -0.381036269
#>  [42,]  0.771780197  7.108577e-01  0.74740478 -0.152439958  0.913515028
#>  [43,]  0.533970415 -6.093921e-01  0.91536495 -0.064931794  0.191683974
#>  [44,]  0.465756294  1.350868e-01 -0.76757987 -0.150222678  0.844433842
#>  [45,]  0.336392303 -1.928956e-01 -0.48508986 -0.402312803 -0.959349934
#>  [46,] -0.367139529 -7.613255e-01  0.38165992 -0.287580565 -0.041267890
#>  [47,]  0.207193788 -1.704446e-01  0.39390818 -0.454747863  0.437190551
#>  [48,]  0.717525152 -4.443323e-01 -0.42492859  0.223606450  0.291142059
#>  [49,] -0.062657201 -2.511550e-01 -0.97149631 -0.748690796 -0.242768697
#>  [50,] -0.441011260 -5.042701e-01  0.80855423 -0.602166249 -0.705471619
#>  [51,]  0.462702146  3.251359e-01  0.04545159 -0.635331504 -0.406885140
#>  [52,] -0.663681757  4.495471e-01  0.50484374  0.296044619  0.526761096
#>  [53,] -0.808090948 -4.692006e-01  0.16850339  0.174878359 -0.813655409
#>  [54,]  0.948094852  6.804950e-01  0.50928378  0.290599475 -0.909197178
#>  [55,]  0.942367881 -7.665589e-01  0.79639452  0.538106448 -0.195557756
#>  [56,]  0.955489561  8.425481e-01 -0.90608470  0.411535819  0.063420178
#>  [57,] -0.785704508 -9.068311e-01 -0.67901972  0.664941584  0.403940575
#>  [58,]  0.211339640  7.353400e-01 -0.79509648  0.006737914  0.931382177
#>  [59,] -0.073506071 -3.872959e-01 -0.66576520 -0.848355900 -0.069864763
#>  [60,]  0.151561091  5.976965e-01  0.88724766  0.405889696  0.204106381
#>  [61,]  0.139454055 -9.427095e-01 -0.44166209  0.458050509  0.055943507
#>  [62,] -0.450317522 -7.305335e-01 -0.52617925  0.038643391  0.015591070
#>  [63,]  0.602963172  7.408432e-01  0.38918114  0.630800306 -0.079443532
#>  [64,] -0.142358897 -4.921941e-01 -0.57510939 -0.257120197  0.254528061
#>  [65,]  0.611432338  1.741615e-01  0.77055817  0.181699296 -0.074735376
#>  [66,]  0.491804657 -2.049527e-01  0.90704185 -0.345095154  0.398408620
#>  [67,] -0.785441375  6.837710e-01 -0.69726177 -0.235922676  0.574536445
#>  [68,]  0.198109614 -2.036802e-05  0.77986037 -0.903526732 -0.862904100
#>  [69,]  0.416768753  7.868125e-02 -0.29103964  0.107045629  0.219391379
#>  [70,]  0.951292020 -8.648730e-01 -0.85002124  0.847964863 -0.475352847
#>  [71,] -0.876244270  9.841604e-01 -0.29504141  0.880753960 -0.548559973
#>  [72,] -0.206847424  3.188446e-02 -0.80365001 -0.820444123  0.639003731
#>  [73,]  0.755695133  9.399585e-01  0.50437389  0.562545589  0.842154686
#>  [74,]  0.721228974 -9.184540e-01 -0.88404473 -0.578289792  0.288952840
#>  [75,] -0.857531405  5.844219e-01  0.08204884 -0.885729362  0.709295710
#>  [76,]  0.405067435 -6.660639e-01  0.95430823  0.155679871 -0.193813427
#>  [77,]  0.617943142  6.476538e-01  0.93259417  0.235106096  0.324087326
#>  [78,]  0.656915024 -4.028940e-01 -0.04515782  0.976130248  0.428346739
#>  [79,] -0.005323523  6.763360e-01  0.80653330 -0.932176241 -0.391795523
#>  [80,] -0.093652824  3.920815e-01  0.92023562  0.556130519  0.028513718
#>  [81,]  0.309998399  7.407754e-01 -0.56184698 -0.692598571  0.380030912
#>  [82,] -0.613769535 -4.484236e-01 -0.09944634  0.740949254 -0.896202601
#>  [83,] -0.834723376  2.478281e-01  0.18326878 -0.552223112  0.547758741
#>  [84,]  0.109113086 -9.225020e-01  0.79134068 -0.496789663 -0.617851967
#>  [85,] -0.657132444 -7.238244e-01 -0.96623806  0.119860371  0.154632259
#>  [86,]  0.208099179  3.112155e-01 -0.02882286  0.677166171  0.448293454
#>  [87,] -0.745515986  9.628473e-01 -0.17624002  0.371420437 -0.023829509
#>  [88,]  0.789288833 -9.766147e-01  0.80335674  0.070773901 -0.399350974
#>  [89,]  0.314792836  7.639958e-01  0.62093427  0.460313819 -0.184115388
#>  [90,]  0.627472007  4.145911e-01  0.30867240 -0.526824926 -0.671109695
#>  [91,]  0.344014773  6.734574e-01  0.25716340  0.955865520 -0.291407179
#>  [92,]  0.371060441  5.506264e-01 -0.76151997  0.744178816  0.953800193
#>  [93,] -0.826447315 -1.761435e-01  0.72232795  0.432296111  0.106292625
#>  [94,]  0.992508668 -7.330390e-01 -0.02941341 -0.919333437 -0.305313508
#>  [95,]  0.563209091 -9.087839e-01 -0.75678545  0.910711709  0.540370814
#>  [96,]  0.284720804 -1.592544e-01 -0.64816207  0.240437454  0.722272744
#>  [97,]  0.483315954 -9.137330e-01  0.89835774 -0.395254648  0.990717132
#>  [98,] -0.912901007 -2.526479e-01  0.11572593 -0.424378594  0.653718080
#>  [99,]  0.521250716  3.085797e-01  0.60767331  0.687448385  0.251642416
#> [100,] -0.441681569  7.302632e-01  0.39060130  0.587263802 -0.985424653
#> 
#> $y
#>   [1] 0 0 1 0 1 0 0 1 1 1 0 1 1 1 0 1 1 1 1 1 0 0 0 0 0 1 0 1 0 0 0 1 1 1 0 0 1
#>  [38] 1 0 1 0 0 0 1 1 0 1 0 1 1 1 0 0 0 0 0 0 1 1 1 0 0 0 1 1 1 0 1 1 0 0 1 0 0
#>  [75] 0 1 0 1 1 1 0 1 0 0 0 1 0 0 0 1 1 1 0 0 0 1 0 0 1 0
#>