Simulate Data from Hyper-Sphere for Classification Problems

The classification problem Circle is described in the BARK paper (2008). Inputs are dim independent variables uniformly distributed on the interval \([-1,1]\), only the first 2 out of these dim are actually signals. Outputs are created according to the formula $$y = 1(x1^2+x2^2 \le 2/\pi)$$

sim_circle(n, dim = 5)

Arguments

n

number of data points to generate

dim

number of dimension of the problem, no less than 2

Value

Returns a list with components

x

input values (independent variables)

y

0/1 output values (dependent variable)

References

Ouyang, Zhi (2008) Bayesian Additive Regression Kernels. Duke University. PhD dissertation, Chapter 3.

See also

Examples

sim_circle(n=100, dim=5)
#> $x
#>               [,1]        [,2]         [,3]        [,4]         [,5]
#>   [1,] -0.52120965 -0.41519631 -0.381940074  0.84100853 -0.127106186
#>   [2,]  0.77254518 -0.06251539 -0.605250022 -0.70793630  0.886443041
#>   [3,] -0.83887707  0.48293792 -0.330240815  0.77155397  0.393382689
#>   [4,] -0.79352448 -0.95595908 -0.964914919  0.37117266 -0.345582531
#>   [5,] -0.50748654  0.04706498 -0.719858691  0.01433458 -0.371495432
#>   [6,]  0.92224218 -0.38897787  0.827944593  0.04118527 -0.237584628
#>   [7,] -0.86848825 -0.67229376  0.661490068  0.97560308  0.172335534
#>   [8,]  0.89726401 -0.93530933 -0.420491288  0.99659984  0.864274935
#>   [9,] -0.36150056 -0.46094784  0.792624748 -0.32265597  0.531402593
#>  [10,] -0.36211942 -0.57076478  0.689356572  0.40962053  0.692425299
#>  [11,] -0.45157710  0.56220082  0.888302759 -0.97313531 -0.339945022
#>  [12,] -0.56432507 -0.89069984  0.154680254  0.21495321 -0.518413437
#>  [13,]  0.14975490 -0.06852047  0.074463962 -0.30231734  0.762715200
#>  [14,]  0.63290871 -0.67295389  0.986394865 -0.01563716 -0.900136060
#>  [15,] -0.21682820  0.56367084 -0.623814821 -0.10502236  0.034202878
#>  [16,] -0.12931929  0.28697443  0.433874731 -0.66728790 -0.347697377
#>  [17,] -0.92908701  0.66461252 -0.578278203  0.71378272  0.126499780
#>  [18,]  0.90170743 -0.78949826 -0.291573555  0.44162667  0.611127523
#>  [19,] -0.42007831 -0.62743993  0.951079948 -0.28014116 -0.216368413
#>  [20,] -0.75510426  0.23088501 -0.249659277  0.83323399 -0.872365573
#>  [21,] -0.86089167  0.59654251  0.875021223  0.07761345 -0.505594059
#>  [22,]  0.33406330 -0.54810476  0.359945157  0.04936522  0.849396916
#>  [23,] -0.17490879  0.06650519 -0.134817938 -0.60073099  0.039067542
#>  [24,] -0.69206155  0.81925680 -0.836521954  0.92465619  0.405333626
#>  [25,] -0.76634892  0.81157376 -0.824885389  0.87900838 -0.930236977
#>  [26,]  0.05169884 -0.11780529  0.289485561  0.87101288  0.955489791
#>  [27,] -0.03590490 -0.42384450  0.877203536  0.09185822 -0.266296832
#>  [28,]  0.67380798  0.75513664  0.757405158  0.93211343  0.616812949
#>  [29,] -0.77566682  0.70387169 -0.297453019  0.18715460  0.124862118
#>  [30,] -0.91861564  0.82393516 -0.573816939  0.06135737 -0.553277178
#>  [31,] -0.26684453  0.52191733  0.997323452 -0.57981985 -0.335485362
#>  [32,] -0.44565355 -0.94997102 -0.493140972 -0.83340984 -0.483518497
#>  [33,]  0.10923816  0.97459952  0.234245691  0.02769067 -0.047503151
#>  [34,] -0.33708167  0.53334383 -0.083946215  0.76560226 -0.077423585
#>  [35,] -0.33523071  0.59809128  0.263281248 -0.47440005 -0.632189394
#>  [36,]  0.42376146  0.13254699  0.621723148 -0.88007019 -0.404932960
#>  [37,]  0.52126053  0.04237893  0.957998110  0.76271246  0.630345057
#>  [38,]  0.75659666  0.23397084 -0.788142422 -0.80732602  0.677451078
#>  [39,]  0.36987419  0.20620379  0.047750330 -0.69161050 -0.073295914
#>  [40,] -0.12524709 -0.83468945  0.175917962  0.72234688  0.365124190
#>  [41,] -0.90803921  0.28989369  0.172265504 -0.45865322 -0.618423077
#>  [42,]  0.70014795 -0.55371993 -0.413383694  0.11231559  0.150638906
#>  [43,] -0.65476799  0.58600964 -0.029895002  0.13440452  0.516372420
#>  [44,]  0.98519428  0.55381676  0.881087120 -0.07662753 -0.200001986
#>  [45,]  0.21433872 -0.42863633  0.937235940 -0.90775835 -0.982852056
#>  [46,] -0.48220717 -0.63763519  0.473197082  0.95108531 -0.181278136
#>  [47,] -0.77760187 -0.53053847 -0.499370982  0.01923192 -0.654898626
#>  [48,] -0.32792013 -0.26029886 -0.088747136  0.02430235 -0.486796337
#>  [49,] -0.88435439 -0.81343581  0.972112933  0.75146824  0.411825559
#>  [50,] -0.12658675  0.83287985 -0.330311294 -0.42843683  0.320146466
#>  [51,]  0.34897079  0.75484549 -0.482551643 -0.87736018  0.799127145
#>  [52,] -0.79211408 -0.26923975 -0.093364737  0.99259875  0.300988248
#>  [53,]  0.99355103  0.55366765  0.439762146 -0.32428079  0.639532493
#>  [54,]  0.78499004 -0.46374873 -0.565397031  0.40912646  0.881353543
#>  [55,]  0.08448539 -0.74205178 -0.304676456  0.05236586 -0.267149224
#>  [56,] -0.11330842 -0.32193337 -0.477103514  0.69075331 -0.529578776
#>  [57,] -0.61974549 -0.28793188 -0.245949860  0.79757418 -0.501565408
#>  [58,] -0.13206173 -0.12199663  0.005347831 -0.42689196  0.102627600
#>  [59,]  0.23871297  0.21365972  0.894917246  0.62765725  0.128602660
#>  [60,] -0.73608583  0.94741282  0.112280731 -0.69243604  0.506702757
#>  [61,]  0.38263341 -0.97213102 -0.759552874  0.32352704  0.958817271
#>  [62,]  0.58545799 -0.29350042  0.489636739  0.91569737  0.028227618
#>  [63,] -0.62901169  0.14151494 -0.295557638 -0.88524154  0.653423217
#>  [64,] -0.91018008 -0.09955456 -0.299501122  0.93725125 -0.793506742
#>  [65,] -0.92641347  0.23909052 -0.342499745  0.16868092 -0.199739323
#>  [66,]  0.40652070 -0.90525671 -0.825462190  0.93735970 -0.924981080
#>  [67,] -0.89129695 -0.81906424  0.897024313  0.01327444 -0.735428482
#>  [68,]  0.69041323 -0.41647792  0.074492999 -0.36782791  0.042207933
#>  [69,] -0.35865072 -0.76343632  0.358809828 -0.80016405 -0.577829438
#>  [70,]  0.75496425  0.82075786  0.493813561 -0.08841535  0.092005138
#>  [71,]  0.41534868  0.38726857 -0.178541173 -0.25825765 -0.560352711
#>  [72,] -0.30335010  0.58053426 -0.984964255 -0.22576846 -0.327593390
#>  [73,] -0.87926512  0.59113283 -0.388536429  0.97036567  0.537591656
#>  [74,] -0.85273108 -0.64194598  0.608501793 -0.04869402  0.850081585
#>  [75,] -0.94644292 -0.71896128 -0.406320682  0.69832688  0.189667237
#>  [76,]  0.76403422  0.77808729  0.158878879  0.85948576  0.006741136
#>  [77,]  0.65667871 -0.08499517 -0.813656894 -0.91605344 -0.377158925
#>  [78,]  0.23488463 -0.19986735  0.792019265 -0.80526213  0.528994354
#>  [79,] -0.90669559  0.77915033 -0.790348377  0.80633374 -0.280765283
#>  [80,]  0.20347263 -0.68394763 -0.212957286 -0.96219982 -0.331687523
#>  [81,] -0.43007346  0.58852346  0.480784672  0.28112437  0.562814486
#>  [82,] -0.38922772  0.93292063  0.281969928 -0.11834398  0.076581818
#>  [83,]  0.51521050 -0.07104470  0.830584761  0.31429001 -0.603835541
#>  [84,] -0.40416981  0.11659562  0.374434726 -0.36436474  0.169629753
#>  [85,]  0.26540830 -0.06695599  0.100964487  0.14645548 -0.841914765
#>  [86,] -0.61727364 -0.69134125 -0.045638283 -0.20102585 -0.671894722
#>  [87,] -0.67171416 -0.43261004  0.255238189 -0.93062000  0.215933311
#>  [88,]  0.87244255  0.78741925  0.422742840 -0.87870348 -0.366878892
#>  [89,] -0.84448247 -0.85606188  0.415050903 -0.44217092 -0.785261265
#>  [90,] -0.60132035  0.33061130  0.496128724  0.17823228  0.472555630
#>  [91,] -0.01073910  0.31215759 -0.284288151  0.20520151  0.844260182
#>  [92,]  0.98935910 -0.41639348 -0.058066880  0.56092195 -0.148454663
#>  [93,]  0.91008778 -0.80898439  0.507815464 -0.92253345 -0.262222462
#>  [94,]  0.91986959 -0.32591460 -0.869192855 -0.48752848 -0.513653490
#>  [95,] -0.58693552  0.19609513  0.376439356 -0.31600294  0.654372876
#>  [96,]  0.88371442 -0.55313193 -0.823251448 -0.85264122 -0.773770289
#>  [97,] -0.26448064 -0.14989548  0.699820192 -0.61764194 -0.007094430
#>  [98,] -0.59866137 -0.11332309  0.843194585 -0.21168167  0.411539810
#>  [99,] -0.40185580 -0.02318086 -0.915283307 -0.73338650  0.850097592
#> [100,] -0.68793932 -0.70950648 -0.275431235  0.33362786  0.998280417
#> 
#> $y
#>   [1] 1 1 0 0 1 0 0 0 1 1 1 0 1 0 1 1 0 0 1 1 0 1 1 0 0 1 1 0 0 0 1 0 0 1 1 1 1
#>  [38] 1 1 0 0 0 0 0 1 0 0 1 0 0 0 0 0 0 1 1 1 1 1 0 0 1 1 0 0 0 0 0 0 0 1 1 0 0
#>  [75] 0 0 1 1 0 1 1 0 1 1 1 0 0 0 0 1 1 0 0 0 1 0 1 1 1 0
#>