Simulate Data from Hyper-Sphere for Classification Problems

The classification problem Circle is described in the BARK paper (2008). Inputs are dim independent variables uniformly distributed on the interval \([-1,1]\), only the first 2 out of these dim are actually signals. Outputs are created according to the formula $$y = 1(x1^2+x2^2 \le 2/\pi)$$

Usage

sim_circle(n, dim = 5)

Arguments

n

number of data points to generate

dim

number of dimension of the problem, no less than 2

Value

Returns a list with components

x

input values (independent variables)

y

0/1 output values (dependent variable)

References

Ouyang, Zhi (2008) Bayesian Additive Regression Kernels. Duke University. PhD dissertation, Chapter 3.

See also

Examples

sim_circle(n=100, dim=5)
#> $x
#>                [,1]        [,2]        [,3]         [,4]          [,5]
#>   [1,] -0.045076370 -0.68118630  0.10471526  0.201827928  0.3083168264
#>   [2,]  0.139403970 -0.13653291 -0.55102096 -0.614369143  0.6028405572
#>   [3,]  0.648284131  0.95350122  0.83555381  0.447077314  0.3710135659
#>   [4,]  0.817826527 -0.74294962 -0.33138544  0.744144526  0.0186200649
#>   [5,] -0.459438759 -0.30672700 -0.30591664 -0.415553018 -0.0249721524
#>   [6,] -0.705723077 -0.01990451 -0.12319788  0.466786212  0.9293080922
#>   [7,]  0.488937416 -0.42903581 -0.93443675  0.420429383 -0.0009305798
#>   [8,] -0.624882772 -0.47101319 -0.01970664 -0.416035492 -0.4551884430
#>   [9,]  0.904799684 -0.15935801  0.93313825  0.227709946 -0.3748269212
#>  [10,]  0.582576895  0.11939011 -0.75185296 -0.579133166 -0.2918300838
#>  [11,] -0.566383493 -0.08931084  0.07101133  0.064522511 -0.3527116221
#>  [12,] -0.075051495  0.50679127  0.04509113 -0.512396349  0.9726376152
#>  [13,] -0.676534033 -0.24202864  0.53580003  0.964837651  0.9750121436
#>  [14,] -0.299236301  0.86612517  0.82358241  0.837394386 -0.1952008270
#>  [15,] -0.688670816 -0.01404042  0.78151164  0.966851482 -0.0132023864
#>  [16,]  0.951512654  0.41175055 -0.17952269 -0.035611035  0.2316115689
#>  [17,]  0.762763189 -0.77044768 -0.59506191 -0.749701213 -0.1810925310
#>  [18,] -0.736212989 -0.49128269  0.91450010  0.025019161  0.8723693630
#>  [19,]  0.279700362 -0.56753579  0.93556319  0.476107847 -0.7055782694
#>  [20,]  0.959284332 -0.84508581 -0.02435083  0.819854254 -0.6076167305
#>  [21,]  0.978087798 -0.62372376  0.15843878  0.032309660  0.1377385152
#>  [22,] -0.154382891 -0.30918540 -0.41915510 -0.567352849 -0.2539595822
#>  [23,] -0.165672504  0.39672447  0.27867556  0.527591066  0.6862535751
#>  [24,]  0.045479070  0.77163618  0.08284971  0.506136716 -0.6101388871
#>  [25,] -0.007753041  0.37626556  0.78297345 -0.262531349 -0.5126172747
#>  [26,]  0.132849327  0.49539039 -0.97211926 -0.258583460 -0.8584846598
#>  [27,]  0.775318543  0.84622092  0.28741877 -0.143207887  0.7949783397
#>  [28,]  0.390863524 -0.29754331  0.72466793 -0.848480230  0.7562052696
#>  [29,]  0.363701284  0.25746412 -0.90445356  0.464467711 -0.1316237007
#>  [30,] -0.265290954 -0.27729399  0.03232968 -0.385038026 -0.5369408950
#>  [31,] -0.708801694 -0.31707174  0.30702345 -0.355138947 -0.6118206908
#>  [32,] -0.179929670 -0.88647109 -0.62743400  0.186046983 -0.7698311172
#>  [33,]  0.808907443  0.43500275  0.39412941  0.130322614 -0.5019406462
#>  [34,]  0.787511662 -0.38371742 -0.70555697 -0.254038457  0.2929720758
#>  [35,] -0.764042744  0.60064541 -0.76385809  0.211272982  0.3717373814
#>  [36,]  0.014602822  0.58651150 -0.52813404 -0.960940556  0.0463448893
#>  [37,]  0.542031004  0.24152191 -0.35216155 -0.795625668 -0.7715271320
#>  [38,]  0.574800622 -0.31408341 -0.71653677  0.717139191 -0.9477684172
#>  [39,] -0.781193791 -0.69967846 -0.75055885 -0.427124183  0.4381843521
#>  [40,] -0.246578235 -0.95601023 -0.98132485 -0.092503111  0.6690642401
#>  [41,] -0.220267169 -0.53103362 -0.74678594 -0.555867381  0.6173051465
#>  [42,] -0.864238749 -0.51056803  0.25952309 -0.377040273  0.6680709212
#>  [43,]  0.452181006  0.59157301  0.43918478  0.983514359  0.1156005268
#>  [44,]  0.162121535 -0.13841002  0.41202979 -0.171514851  0.2401627167
#>  [45,] -0.223916439  0.01010642  0.36407102  0.254008282 -0.7113821269
#>  [46,] -0.578772687 -0.35145335 -0.42942763 -0.760208116 -0.0321337064
#>  [47,]  0.212765609 -0.84566515  0.35726136 -0.973742934 -0.6814084477
#>  [48,]  0.853059042 -0.21077113 -0.32904528 -0.711871207 -0.5492174197
#>  [49,]  0.132331828 -0.23457221 -0.11450667  0.857858353 -0.4040162466
#>  [50,] -0.857079056  0.69050799 -0.01529084 -0.734512532 -0.0219763243
#>  [51,]  0.799967214 -0.18247061 -0.77589051 -0.636883000 -0.2486596243
#>  [52,]  0.446409951  0.08625934 -0.03555117  0.902241969 -0.5565143060
#>  [53,] -0.303708823  0.33502513 -0.79194308  0.581916659 -0.3062621271
#>  [54,] -0.838512520 -0.62076462  0.67680970 -0.773140273  0.2247655564
#>  [55,] -0.286071435 -0.20062576  0.89502798 -0.539475939  0.7971775327
#>  [56,] -0.076859794  0.41987791 -0.84311521 -0.076356101  0.0773519110
#>  [57,]  0.143588383 -0.94721506 -0.86747616  0.441733572  0.4385245866
#>  [58,]  0.567441156  0.06103718  0.79281821 -0.006953374  0.8113150019
#>  [59,]  0.248754478 -0.02859318  0.98118931  0.057039106  0.4949618885
#>  [60,] -0.720140689  0.35609889 -0.28266117 -0.759165006 -0.3490833971
#>  [61,]  0.138416523 -0.31068952 -0.75527465 -0.500692352 -0.0501748985
#>  [62,]  0.345466631  0.43778759  0.91670777  0.362535254  0.1647140910
#>  [63,] -0.757224281  0.29356343  0.47154659 -0.822025124 -0.6269342690
#>  [64,] -0.921338694 -0.12370199 -0.12470060  0.759572734 -0.8577302489
#>  [65,]  0.608388949  0.75638912  0.02718654  0.675378457 -0.4068336254
#>  [66,] -0.722881701 -0.22643764  0.72632499 -0.817773128  0.3805862577
#>  [67,] -0.005221523  0.32576434 -0.49200136  0.217493842 -0.3225743836
#>  [68,] -0.354985584  0.87813446  0.57370192 -0.079616587  0.3773302040
#>  [69,] -0.025773288  0.27836905 -0.80691095 -0.782096236 -0.6342728785
#>  [70,] -0.085009378  0.58040774  0.40985233  0.588849026  0.1127640111
#>  [71,]  0.432105693  0.84391925  0.23408736 -0.659798861 -0.9619614566
#>  [72,]  0.641671191  0.41437462 -0.14574106 -0.615984557  0.2725682654
#>  [73,] -0.531012428 -0.78278791 -0.77203750  0.629165563 -0.6019884329
#>  [74,]  0.625110461  0.14132239 -0.70217868 -0.589566687 -0.1920061638
#>  [75,]  0.313471495 -0.19387955  0.93185208 -0.979356461 -0.1320591513
#>  [76,]  0.211213487  0.64086862 -0.06035971  0.098491513  0.4612145848
#>  [77,] -0.050576555  0.92116825  0.73693672 -0.804062090  0.0654765270
#>  [78,] -0.690838750 -0.13409836  0.78573117 -0.586740995 -0.5993628036
#>  [79,] -0.345864434  0.32217981  0.26182381  0.815194648  0.3367649685
#>  [80,]  0.921220724 -0.86648233  0.80088701  0.529816009  0.7558845738
#>  [81,] -0.134488717 -0.47195591  0.77794032  0.417021875 -0.1934165452
#>  [82,]  0.591037106  0.19694815 -0.72036532  0.998423921  0.0497318758
#>  [83,]  0.017042239 -0.91797266 -0.69236660 -0.566781064  0.0150697674
#>  [84,]  0.562868370  0.48447645  0.59004889  0.384661217 -0.4018195230
#>  [85,]  0.706283731  0.88210460  0.67459039  0.693395302  0.0606838022
#>  [86,] -0.072384743 -0.40636709  0.72560503  0.503447220  0.3742074566
#>  [87,] -0.775061658  0.80934032 -0.01212825 -0.663411022 -0.6725152326
#>  [88,]  0.926610061 -0.82346244 -0.65000180  0.022065924 -0.3764282214
#>  [89,]  0.129743394 -0.57447570  0.90234087 -0.886366971  0.4552401281
#>  [90,]  0.930574034  0.71931705 -0.69158578 -0.877227644  0.1116130324
#>  [91,]  0.649262730  0.56950834  0.22592989 -0.795453528 -0.5697091715
#>  [92,]  0.627276164  0.75041606  0.85601412  0.623990409  0.3853781288
#>  [93,] -0.649182844 -0.94224222 -0.24460894  0.284735678 -0.9127532230
#>  [94,] -0.073401790  0.97309481  0.56969554 -0.068523235 -0.9609755748
#>  [95,]  0.538301418 -0.79179010 -0.31868919  0.178749397  0.8536587418
#>  [96,]  0.314421746  0.77458167 -0.77816342  0.182001961 -0.8126594205
#>  [97,]  0.254396850  0.80189454  0.17050218 -0.133424436 -0.1501455246
#>  [98,] -0.158822752  0.90835495  0.52976733 -0.255135227  0.7191061173
#>  [99,]  0.390700540 -0.36683565 -0.17054132 -0.559966323 -0.4339231988
#> [100,] -0.870376568  0.88434117 -0.02827984  0.888733468 -0.7513325121
#> 
#> $y
#>   [1] 1 1 0 0 1 1 1 1 0 1 1 1 1 0 1 0 0 0 1 0 0 1 1 1 1 1 0 1 1 1 1 0 0 0 0 1 1
#>  [38] 1 0 0 1 0 1 1 1 1 0 0 1 0 0 1 1 0 1 1 0 1 1 0 1 1 0 0 0 1 1 0 1 1 0 1 0 1
#>  [75] 1 1 0 1 1 0 1 1 0 1 0 1 0 0 1 0 0 0 0 0 0 0 0 0 1 0
#>