Simulate Data from Hyper-Sphere for Classification Problems

The classification problem Circle is described in the BARK paper (2008). Inputs are dim independent variables uniformly distributed on the interval \([-1,1]\), only the first 2 out of these dim are actually signals. Outputs are created according to the formula $$y = 1(x1^2+x2^2 \le 2/\pi)$$

sim_circle(n, dim = 5)

Arguments

n

number of data points to generate

dim

number of dimension of the problem, no less than 2

Value

Returns a list with components

x

input values (independent variables)

y

0/1 output values (dependent variable)

References

Ouyang, Zhi (2008) Bayesian Additive Regression Kernels. Duke University. PhD dissertation, Chapter 3.
http://stat.duke.edu/people/theses/OuyangZ.html

See also

Examples

sim_circle(n=100, dim=5)
#> $x
#>                [,1]        [,2]         [,3]        [,4]         [,5]
#>   [1,]  0.416962515 -0.71995035 -0.350551062  0.74730891 -0.651176546
#>   [2,]  0.206704699  0.48839614  0.982100766  0.55560344 -0.411478526
#>   [3,] -0.616245791  0.36352284  0.562776258  0.91575655  0.888630335
#>   [4,] -0.148163605 -0.87361396 -0.426851310  0.42952625  0.185002941
#>   [5,] -0.350376744  0.36845115 -0.077326682 -0.74552838  0.994987664
#>   [6,]  0.003298651  0.38249354  0.508762441 -0.67645825 -0.323577953
#>   [7,]  0.068985352  0.49744750  0.292894217  0.48687486 -0.276983859
#>   [8,] -0.401170760  0.65499242  0.424825520 -0.84171193 -0.120330657
#>   [9,] -0.221903275  0.18277287  0.708531696  0.25244343 -0.817523441
#>  [10,] -0.567776907 -0.85552519  0.996860689  0.78129082  0.812949305
#>  [11,] -0.357294266  0.28855313  0.197753186  0.62564322  0.405028633
#>  [12,]  0.815391262  0.30051380 -0.895491873 -0.26710170  0.366544885
#>  [13,]  0.874380277 -0.03547445  0.895830981  0.26634128 -0.389209255
#>  [14,]  0.155980473 -0.51038988  0.448800151  0.22619766 -0.072037117
#>  [15,] -0.313153112 -0.79060669  0.089245890  0.10178502 -0.358402794
#>  [16,] -0.108056773 -0.79661605  0.116687103 -0.37414641  0.581541092
#>  [17,]  0.439646640  0.22316893 -0.189921256  0.35098195 -0.785168561
#>  [18,]  0.574115814  0.04471953  0.078483960 -0.47270355 -0.545430151
#>  [19,] -0.176920334  0.60294952  0.005622508  0.27518833 -0.339143672
#>  [20,] -0.113629369 -0.38184384  0.588818476  0.87878844  0.611943173
#>  [21,] -0.837358322 -0.81793012 -0.705748925  0.51862030 -0.027002182
#>  [22,]  0.911302963 -0.52045527 -0.491634478 -0.43245360 -0.808260168
#>  [23,] -0.122551021  0.26986621 -0.948474194  0.58266373  0.763713500
#>  [24,]  0.176410838  0.95821721 -0.833552707 -0.74657044 -0.275768464
#>  [25,] -0.509467185 -0.26344555 -0.901499364 -0.82644275 -0.491603584
#>  [26,] -0.826991415  0.28535669  0.866524202  0.18349467  0.075270200
#>  [27,]  0.257439815 -0.13270560  0.847815233 -0.56724843 -0.529653913
#>  [28,] -0.873142648 -0.37936922  0.015681735 -0.08254534  0.911287124
#>  [29,]  0.285283078  0.23551725 -0.357196767 -0.86788859  0.602663179
#>  [30,] -0.878634847  0.19525998  0.362965243  0.08835703  0.486780044
#>  [31,] -0.257749314  0.15209596 -0.660744216 -0.56707087  0.169696821
#>  [32,] -0.492421140 -0.01707712 -0.652872665 -0.22421500  0.166905595
#>  [33,]  0.433281192  0.69908462  0.574022635  0.94584254  0.601057197
#>  [34,] -0.726382494  0.82415259 -0.610039084  0.94454735 -0.616606898
#>  [35,] -0.981898715 -0.06234443  0.949180785  0.64222263 -0.326555840
#>  [36,]  0.979681117  0.41412730 -0.025166593 -0.31497749  0.015501889
#>  [37,]  0.608404313  0.53958594 -0.968742727  0.42165752  0.147819663
#>  [38,]  0.996577387  0.66504728  0.285762901 -0.17638033  0.553440670
#>  [39,]  0.833229040  0.38561740  0.801783910 -0.84365577  0.886849879
#>  [40,]  0.403533414  0.67660544 -0.185534599 -0.72985677 -0.328059184
#>  [41,]  0.379764992  0.08396421  0.297600291 -0.11841345  0.435977256
#>  [42,]  0.998976266  0.12036367 -0.863061735 -0.31921890 -0.080883315
#>  [43,]  0.386281364  0.87899369 -0.287882554 -0.88180816  0.011964420
#>  [44,]  0.572633910  0.23227522 -0.099744292  0.03211692 -0.674328974
#>  [45,] -0.648746360 -0.56156331 -0.042800515  0.96889526  0.696699383
#>  [46,] -0.449722351 -0.70420331 -0.071537856 -0.23409923 -0.887159444
#>  [47,]  0.752374266 -0.56994010 -0.086340160  0.44868213 -0.476490574
#>  [48,]  0.473487473  0.34070615  0.076056625  0.03597435 -0.127969248
#>  [49,]  0.054190251  0.03260504  0.744540666  0.80568394 -0.510974021
#>  [50,]  0.514413161 -0.71689555 -0.296912426  0.47983467 -0.837196067
#>  [51,] -0.843819519 -0.87229020  0.445463817 -0.95628687 -0.836780798
#>  [52,] -0.921486167 -0.32100560  0.749055125  0.87823910  0.661567308
#>  [53,] -0.804448314  0.57089789  0.401602952  0.26556510 -0.111477456
#>  [54,] -0.231771980  0.63197983  0.272741138 -0.54896925  0.289932283
#>  [55,] -0.769471726  0.75999121  0.715406273 -0.81704263 -0.840915816
#>  [56,]  0.122097945  0.83534365 -0.150956050  0.88075218 -0.493761459
#>  [57,] -0.843567050  0.34818975 -0.871806828 -0.17065257 -0.354382540
#>  [58,]  0.508976060 -0.85585374  0.225443473 -0.69575550  0.889877810
#>  [59,] -0.733860841  0.05870039  0.912666399 -0.44663431  0.256001213
#>  [60,] -0.107282052  0.49299334  0.447137076 -0.02710358  0.087566436
#>  [61,] -0.888625515 -0.04356661 -0.121902135 -0.75089587  0.239290623
#>  [62,]  0.854196080 -0.14433019 -0.427790859 -0.43094438  0.647856602
#>  [63,]  0.037295562  0.22257027  0.757257909 -0.34155489 -0.614365927
#>  [64,]  0.567697817  0.36552877 -0.020955528  0.03185759 -0.959979988
#>  [65,] -0.865505786 -0.22352920 -0.710082171 -0.55022163  0.128704889
#>  [66,] -0.714541959  0.32206195 -0.375870715 -0.61516799  0.036782042
#>  [67,] -0.127770398  0.66371174 -0.037853210  0.87820566  0.980843325
#>  [68,]  0.791449883 -0.80815354  0.962987536  0.94520104 -0.809745495
#>  [69,]  0.391464918  0.40704045  0.468299820  0.29902504 -0.783777498
#>  [70,] -0.102020932 -0.88935161 -0.645272952 -0.33960677 -0.206235921
#>  [71,] -0.815854078 -0.01361295 -0.211322762  0.31512508 -0.291730913
#>  [72,]  0.738935045 -0.87770972  0.988122254 -0.21868376  0.154298879
#>  [73,] -0.274550475  0.88751886  0.070108575 -0.44912260 -0.676262536
#>  [74,] -0.997649766 -0.50379351 -0.277063013  0.38529356 -0.919709082
#>  [75,]  0.097627855  0.64306374 -0.906717573  0.71452738  0.814830319
#>  [76,]  0.008157993 -0.00806155  0.172889426  0.50269038  0.697954883
#>  [77,] -0.331697938  0.94997712 -0.715638204  0.91703699 -0.288325137
#>  [78,] -0.339684105 -0.77888937  0.803312273  0.76627809 -0.799426199
#>  [79,] -0.560927341  0.43819159 -0.021090037 -0.79931772  0.678241550
#>  [80,]  0.077663061 -0.98294887 -0.270742966  0.37360638 -0.660293893
#>  [81,]  0.069644106 -0.75857325 -0.605523704 -0.88434507  0.098029147
#>  [82,] -0.716886823  0.15546155  0.612759197  0.47022459  0.353422320
#>  [83,]  0.453938613 -0.42685615  0.873732415 -0.91704268  0.074555188
#>  [84,]  0.483381450  0.44639899  0.141978193  0.39951551 -0.459287030
#>  [85,] -0.735497769 -0.20957261  0.609163251 -0.59067185  0.549731025
#>  [86,]  0.988807555 -0.64972290  0.464907227 -0.15693498 -0.871564424
#>  [87,]  0.648276225  0.73262866 -0.080965985 -0.32032742 -0.006503727
#>  [88,] -0.481309325  0.91466171 -0.464638133 -0.77570954  0.437935882
#>  [89,]  0.871513181 -0.44850698 -0.062958460 -0.60415522 -0.485823414
#>  [90,] -0.922973256  0.77629746 -0.950001038 -0.13306623  0.704588165
#>  [91,]  0.864364198  0.24013901 -0.755186243 -0.20347416  0.817258204
#>  [92,]  0.422312114  0.12355172 -0.597639879  0.27024570  0.129351417
#>  [93,]  0.722723241 -0.30927876  0.431288494 -0.18415001  0.116508544
#>  [94,] -0.741821819 -0.28015502 -0.150520348  0.69523932  0.468582486
#>  [95,] -0.864807383  0.93055765  0.655869307  0.55880835 -0.655408176
#>  [96,]  0.223847340  0.44702303 -0.793439173  0.10610016 -0.471971138
#>  [97,] -0.626106559 -0.61301277  0.138962148  0.26670707 -0.732360614
#>  [98,]  0.128787008  0.34067248  0.011576809 -0.56032138 -0.269850350
#>  [99,] -0.129756793 -0.90308901 -0.349482662 -0.39115533  0.785062102
#> [100,]  0.094827486  0.26647680 -0.756820186  0.27306879  0.307641072
#> 
#> $y
#>   [1] 0 1 1 0 1 1 1 1 1 0 1 0 0 1 0 0 1 1 1 1 0 0 1 0 1 0 1 0 1 0 1 1 0 0 0 0 0
#>  [38] 0 0 1 1 0 0 1 0 0 0 1 1 0 0 0 0 1 0 0 0 0 1 1 0 0 1 1 0 1 1 0 1 0 0 0 0 0
#>  [75] 1 1 0 0 1 0 1 1 1 1 1 0 0 0 0 0 0 1 1 1 0 1 0 1 0 1
#>