Simulate Data from Hyper-Sphere for Classification Problems

The classification problem Circle is described in the BARK paper (2008). Inputs are dim independent variables uniformly distributed on the interval \([-1,1]\), only the first 2 out of these dim are actually signals. Outputs are created according to the formula $$y = 1(x1^2+x2^2 \le 2/\pi)$$

Usage

sim_circle(n, dim = 5)

Arguments

n

number of data points to generate

dim

number of dimension of the problem, no less than 2

Value

Returns a list with components

x

input values (independent variables)

y

0/1 output values (dependent variable)

References

Ouyang, Zhi (2008) Bayesian Additive Regression Kernels. Duke University. PhD dissertation, Chapter 3.

See also

Examples

sim_circle(n=100, dim=5)
#> $x
#>               [,1]        [,2]         [,3]         [,4]         [,5]
#>   [1,]  0.38279226  0.86871552 -0.842680607  0.704930563 -0.404477402
#>   [2,] -0.82239438  0.09503145  0.234690725 -0.410272280 -0.489137190
#>   [3,] -0.58548578 -0.10057729  0.763620373 -0.169523523 -0.702580940
#>   [4,] -0.07477130 -0.73831747  0.806184879 -0.365907134  0.122448187
#>   [5,]  0.36785633  0.73310771 -0.757958184  0.111413319 -0.333203219
#>   [6,]  0.28287313 -0.59255893 -0.715582789  0.727584659  0.572151009
#>   [7,]  0.64775145 -0.38243208 -0.915812019 -0.621287634  0.115559516
#>   [8,] -0.40154662 -0.93274566  0.167945427  0.469626425  0.603528371
#>   [9,] -0.38544514  0.78758027  0.107952416  0.531310497 -0.813125629
#>  [10,]  0.49153243 -0.14705057 -0.363518306  0.318185432  0.914106108
#>  [11,]  0.52908371  0.34569494 -0.680966224 -0.250658513  0.062523986
#>  [12,]  0.83047993  0.48817590  0.513204060  0.753223198  0.644331725
#>  [13,]  0.47568508 -0.67283342  0.580045357 -0.276790414 -0.970113880
#>  [14,] -0.68989019 -0.23358668 -0.435617403  0.250662189  0.047645422
#>  [15,]  0.46834367  0.78610370  0.811971627  0.452607684 -0.459185656
#>  [16,] -0.29756090  0.30864217 -0.804453278  0.732426051  0.703482054
#>  [17,] -0.55348030  0.73023510  0.341479434  0.477794634 -0.120716874
#>  [18,] -0.54381571  0.52132189  0.483152550 -0.661824837 -0.239330549
#>  [19,] -0.28940742  0.14474678  0.049365691 -0.134050044 -0.090223135
#>  [20,]  0.44260466  0.11399441  0.482364172 -0.703840610 -0.921120107
#>  [21,] -0.24808587 -0.13694759 -0.212876427 -0.001067765 -0.181012956
#>  [22,] -0.32147547  0.83149374 -0.703748371  0.294531116 -0.635997490
#>  [23,]  0.81329880  0.90206826  0.518974110  0.728875459  0.601239847
#>  [24,] -0.54948166  0.94463623 -0.808061382  0.536655808  0.325342301
#>  [25,]  0.01758963 -0.15110231  0.241447935  0.956790371  0.507588373
#>  [26,] -0.31102632  0.99729613  0.928718193 -0.054112363 -0.356157290
#>  [27,] -0.74693345 -0.91358827 -0.549043835  0.823021336 -0.921674357
#>  [28,] -0.22079021  0.37247116 -0.570980871  0.718304041 -0.645545994
#>  [29,]  0.55366378  0.95468464 -0.690440667 -0.164636460 -0.595228835
#>  [30,]  0.22165207 -0.96051061 -0.970473672  0.092041497  0.645403526
#>  [31,]  0.62644331 -0.69862246 -0.431139931  0.124662417  0.790923849
#>  [32,]  0.64108543 -0.27582572 -0.069317169  0.134737802  0.979649261
#>  [33,]  0.78497732 -0.78234719 -0.647058889  0.227646191 -0.830111018
#>  [34,]  0.62051200  0.57798152  0.482750719 -0.346614088  0.748329624
#>  [35,]  0.80304272  0.47518470  0.184215184 -0.554747877  0.661181986
#>  [36,]  0.94694932  0.88157120  0.662821646  0.957518719  0.393490318
#>  [37,] -0.58374898 -0.53934153 -0.431611538  0.615458333  0.470382080
#>  [38,]  0.42515381  0.81164404 -0.893503262  0.911205534 -0.395458730
#>  [39,]  0.84524907  0.12190718 -0.293250685 -0.353674189 -0.281967727
#>  [40,] -0.11868240 -0.97576793  0.373994121  0.557348021  0.055400991
#>  [41,]  0.38956624 -0.67807883 -0.596728969  0.326447765  0.073859647
#>  [42,]  0.58768531  0.65883053 -0.375992499 -0.593576787 -0.516746859
#>  [43,]  0.48981778  0.49251639 -0.744209893 -0.284802034 -0.019314218
#>  [44,] -0.62026645 -0.92922503  0.188428275 -0.865400224  0.316712773
#>  [45,] -0.07925485  0.88425600 -0.042278791  0.869739749 -0.186589116
#>  [46,] -0.83795795 -0.62401667 -0.135587594  0.747049674 -0.089693706
#>  [47,]  0.47469579 -0.76809113 -0.017257750  0.465120529 -0.174496933
#>  [48,]  0.17594406 -0.33420200  0.705624940 -0.991642608 -0.009755365
#>  [49,]  0.41914449  0.88672002 -0.997877407  0.434337099  0.883195884
#>  [50,] -0.34952002  0.17060836  0.454808481 -0.030632037  0.478200379
#>  [51,] -0.33175550  0.30571272  0.801553827 -0.642411530  0.079015340
#>  [52,]  0.83269751  0.12992675 -0.786071913  0.893712833  0.513234421
#>  [53,] -0.59837443 -0.17560563 -0.451354246  0.762977958  0.718621000
#>  [54,]  0.99383840  0.71572730  0.422792622 -0.644395388  0.167311896
#>  [55,]  0.64808407 -0.09888788 -0.321912858 -0.832904312 -0.234654199
#>  [56,] -0.11296619 -0.59504504 -0.869844662  0.748956028 -0.987806276
#>  [57,] -0.86849479 -0.43496939  0.679320069  0.203317449  0.906567876
#>  [58,] -0.20875983 -0.54877364  0.987801490  0.996376100 -0.887989248
#>  [59,]  0.15110548 -0.54648357 -0.006421861  0.314765127 -0.291751819
#>  [60,]  0.40989334  0.53866393 -0.528791321  0.356805818  0.939092972
#>  [61,]  0.07438878  0.47142602 -0.344059080  0.560913352 -0.398543457
#>  [62,]  0.19162589  0.86684521  0.289795382 -0.084501877 -0.140474639
#>  [63,]  0.94250627  0.58191479 -0.568511036 -0.824710350 -0.766734512
#>  [64,]  0.26443540 -0.06870153 -0.053921442 -0.178706968  0.409217968
#>  [65,] -0.19670445  0.61903319  0.204291179 -0.645277421 -0.841507817
#>  [66,] -0.44062718 -0.24868153  0.223693865 -0.089494015 -0.027033591
#>  [67,]  0.08740728 -0.03835941  0.118837201  0.434968639 -0.524726574
#>  [68,] -0.64754719  0.22884559 -0.560324402  0.344817221  0.578125360
#>  [69,] -0.70563622  0.06334892 -0.147760267  0.379737134  0.520984597
#>  [70,]  0.46567592 -0.27612664 -0.093391393 -0.932511558  0.462998490
#>  [71,]  0.49054235 -0.18178725 -0.679498214  0.170722981 -0.951463385
#>  [72,] -0.47995585 -0.47264691 -0.998116069  0.229116258 -0.180116415
#>  [73,] -0.37669472 -0.47855463  0.921998152  0.973417575  0.119744090
#>  [74,] -0.66871168 -0.62378494  0.240575384  0.571221800 -0.505184412
#>  [75,]  0.34103343  0.56660584 -0.100078664  0.695415269 -0.450095571
#>  [76,]  0.91508643  0.91341471 -0.956430017 -0.227768858  0.723968033
#>  [77,]  0.68633871  0.77008354  0.603010916 -0.746416832 -0.303906065
#>  [78,] -0.55131172  0.08816426  0.546249630 -0.512961439 -0.122378732
#>  [79,] -0.33958981  0.83076064  0.113732560  0.449434252 -0.563233388
#>  [80,] -0.13560211  0.76599170  0.306067762 -0.616231697  0.874006910
#>  [81,]  0.49584742  0.13417223 -0.854378098  0.186358040 -0.883678451
#>  [82,] -0.61905159 -0.88898005 -0.333363533 -0.477338166  0.393071451
#>  [83,] -0.98604290  0.26601675 -0.519151373  0.076336753  0.048859317
#>  [84,] -0.83382709 -0.68384753 -0.814125250 -0.167366761 -0.685373489
#>  [85,]  0.63201644  0.27984993  0.958823893 -0.670232005 -0.800895299
#>  [86,]  0.78011937 -0.61716222  0.877778190  0.393014587 -0.478737778
#>  [87,]  0.45334753 -0.71226066 -0.305356937  0.223095294  0.359931200
#>  [88,]  0.32064246 -0.54617957  0.166178725 -0.866202174  0.206393631
#>  [89,]  0.68094867 -0.20967137 -0.279931466  0.724301172  0.387602347
#>  [90,]  0.15201413 -0.27498119  0.963046256  0.577179814  0.168407385
#>  [91,]  0.51131835 -0.74758157  0.981405172 -0.904913491 -0.658414393
#>  [92,]  0.67765174 -0.36163378  0.522101577  0.786806174  0.532964840
#>  [93,] -0.65917862 -0.62950047  0.771226501 -0.603302329  0.074176067
#>  [94,]  0.59214002 -0.53334780  0.930598691 -0.827866677 -0.350039226
#>  [95,]  0.21658691  0.49293546  0.029525213  0.049326519  0.045448649
#>  [96,]  0.41206555  0.70529153  0.964047970  0.125338366 -0.602228787
#>  [97,] -0.61193309 -0.21932200 -0.200315889  0.840519086  0.214630657
#>  [98,] -0.40097240  0.01447452 -0.060149664  0.864183826  0.367001666
#>  [99,]  0.17147531  0.55580422 -0.455653105  0.074263861 -0.199179318
#> [100,]  0.35867910  0.92481479  0.930726755  0.992410550  0.597602407
#> 
#> $y
#>   [1] 0 0 1 1 0 1 1 0 0 1 1 0 0 1 0 1 0 1 1 1 1 0 0 0 1 0 0 1 0 0 0 1 0 0 0 0 1
#>  [38] 0 0 0 1 0 1 0 0 0 0 1 0 1 1 0 1 0 1 1 0 1 1 1 1 0 0 1 1 1 1 1 1 1 1 1 1 0
#>  [75] 1 0 0 1 0 1 1 0 0 0 1 0 0 1 1 1 0 1 0 1 1 0 1 1 1 0
#>