Simulate Data from Hyper-Sphere for Classification Problems

The classification problem Circle is described in the BARK paper (2008). Inputs are dim independent variables uniformly distributed on the interval \([-1,1]\), only the first 2 out of these dim are actually signals. Outputs are created according to the formula $$y = 1(x1^2+x2^2 \le 2/\pi)$$

Usage

sim_circle(n, dim = 5)

Arguments

n

number of data points to generate

dim

number of dimension of the problem, no less than 2

Value

Returns a list with components

x

input values (independent variables)

y

0/1 output values (dependent variable)

References

Ouyang, Zhi (2008) Bayesian Additive Regression Kernels. Duke University. PhD dissertation, Chapter 3.

See also

Examples

sim_circle(n=100, dim=5)
#> $x
#>                [,1]         [,2]         [,3]         [,4]         [,5]
#>   [1,]  0.834088827 -0.422625575 -0.491143414 -0.145405294 -0.618291964
#>   [2,]  0.178737875  0.293921790  0.151437100 -0.300391954  0.726512420
#>   [3,]  0.507854985 -0.247603711 -0.601266975 -0.004605016 -0.964619416
#>   [4,]  0.257021860  0.931450448  0.344723723 -0.872965217  0.628618587
#>   [5,] -0.479732614 -0.789862562 -0.767541810  0.061232629 -0.956873548
#>   [6,]  0.825563333 -0.029104395 -0.249130331 -0.215320547 -0.290948018
#>   [7,] -0.732268545 -0.840233208 -0.158075092  0.244822264  0.418190561
#>   [8,] -0.716351434 -0.950725403  0.405079752  0.763970151  0.955873900
#>   [9,] -0.769541547  0.697633721  0.820558861  0.340817499  0.022274264
#>  [10,]  0.064335423 -0.975345477 -0.119893106 -0.954644722 -0.696634024
#>  [11,]  0.115188732 -0.268857233 -0.795426331 -0.122303325  0.473162131
#>  [12,] -0.175964485 -0.544468769 -0.808317896 -0.616184340 -0.701205129
#>  [13,]  0.415074203  0.454430372  0.901445494  0.545191466  0.698766392
#>  [14,]  0.658818290 -0.990534398  0.068335393 -0.291865615  0.738560876
#>  [15,] -0.302295169  0.332040833 -0.927141084  0.605612800 -0.894494907
#>  [16,] -0.648537665  0.114978086 -0.841376520 -0.403775177  0.469195140
#>  [17,]  0.492255586 -0.743740310  0.360394779  0.674419091  0.012885520
#>  [18,] -0.438191531 -0.400068514  0.133810441  0.882156134  0.650014001
#>  [19,] -0.813631344 -0.448216280  0.680732921 -0.638490522 -0.353810521
#>  [20,]  0.250795812 -0.480004398  0.922039451  0.988877336 -0.315859681
#>  [21,] -0.752685202 -0.262468385 -0.588331211 -0.501001329 -0.801324344
#>  [22,]  0.847367515 -0.486796025  0.951059600  0.524562832  0.985570113
#>  [23,]  0.044724306  0.849776591 -0.101867483 -0.501549994 -0.568107212
#>  [24,] -0.408534277 -0.570349193  0.553573372 -0.002646740  0.584531804
#>  [25,]  0.284197423 -0.158174712  0.951791850 -0.925327820  0.266980541
#>  [26,] -0.729375979 -0.283245125 -0.320991035 -0.781454122  0.758677585
#>  [27,]  0.428988704 -0.395461316 -0.411286580  0.063266896 -0.539362035
#>  [28,]  0.907930362 -0.935663819  0.238488226  0.446645170 -0.464412040
#>  [29,]  0.867959033 -0.809142332 -0.358578863  0.805860524 -0.997274352
#>  [30,]  0.921025416  0.165709610 -0.064451084  0.251111832 -0.315980878
#>  [31,] -0.676303450 -0.590874696 -0.381231324  0.127748435 -0.073251043
#>  [32,]  0.064345766  0.264395672 -0.552487504 -0.669942646 -0.703456820
#>  [33,] -0.309135068 -0.170237028 -0.402276070 -0.221126796  0.136741863
#>  [34,] -0.275634889  0.853650881  0.469479437  0.153353938 -0.990054128
#>  [35,]  0.315169975  0.070719501  0.728214822  0.148183459 -0.937613728
#>  [36,]  0.303654020 -0.078948329  0.774547465 -0.299834063 -0.762674526
#>  [37,] -0.088580256 -0.950848056 -0.950381186 -0.721486115 -0.731187897
#>  [38,]  0.258863227 -0.186059400  0.655262562  0.500671920  0.451932141
#>  [39,] -0.526018444  0.825510149 -0.605629778 -0.205363097 -0.476039038
#>  [40,]  0.804360875 -0.221173699  0.438590693 -0.153248843 -0.856190345
#>  [41,] -0.938519810 -0.671446525 -0.227898550 -0.859617127 -0.826931388
#>  [42,] -0.780811114 -0.202856196 -0.534696271  0.939360166  0.040444002
#>  [43,] -0.071360846  0.049135951 -0.091711835  0.319580564  0.760923631
#>  [44,]  0.809497324  0.103413089  0.537399263  0.126739890  0.357249203
#>  [45,] -0.476874264  0.024697304  0.661012569  0.647822992  0.172138175
#>  [46,]  0.177071912  0.624469296 -0.433412509 -0.270098504  0.951838471
#>  [47,] -0.177973767 -0.202691547  0.368762450 -0.183564553  0.128902468
#>  [48,]  0.166219615 -0.189112583  0.002122128  0.299697543  0.278962545
#>  [49,]  0.070001551 -0.288182802 -0.290899941 -0.613628616  0.987448384
#>  [50,]  0.862046800  0.403396342 -0.450790172  0.616579110 -0.738514728
#>  [51,]  0.061374868  0.045367972  0.151056323 -0.488667213 -0.089524786
#>  [52,]  0.651037491 -0.536597255 -0.779048376 -0.700956768  0.478289476
#>  [53,] -0.466119688 -0.242311836 -0.890044073  0.308510662  0.513946282
#>  [54,] -0.341559234  0.883073826 -0.048009533 -0.210857484 -0.864010608
#>  [55,] -0.110823215  0.812028570  0.863147295  0.802369169  0.555897794
#>  [56,]  0.108917250 -0.288041397 -0.839092384  0.188820805 -0.207182557
#>  [57,]  0.897709533  0.942940511 -0.858644889  0.744054489 -0.899091176
#>  [58,] -0.544898314  0.530334125  0.559534238  0.997065606 -0.394707606
#>  [59,] -0.362069128 -0.330197502 -0.792777492 -0.410256360 -0.794937345
#>  [60,] -0.926656447  0.177793088  0.309657449  0.926794770 -0.385590212
#>  [61,]  0.173355629 -0.145667844 -0.997353788 -0.372473592  0.372843147
#>  [62,] -0.576492107  0.173547483  0.883038917 -0.092392523 -0.567662485
#>  [63,] -0.435483345  0.017159202 -0.122966369 -0.890770548 -0.980080572
#>  [64,]  0.708323388 -0.389581924  0.527026218 -0.432934863  0.037202792
#>  [65,] -0.004503821 -0.760380685  0.760739182  0.021497575  0.055672982
#>  [66,]  0.831380236  0.382294029  0.437300004  0.016684154  0.593958896
#>  [67,] -0.247106769  0.914338377  0.290301346  0.401580393  0.426850025
#>  [68,]  0.166396115 -0.830448569 -0.187592873 -0.486259711  0.684145391
#>  [69,] -0.289252492 -0.278453644 -0.023919073  0.725056985 -0.593751335
#>  [70,] -0.338312972  0.977532113  0.580016046 -0.884623926  0.770657027
#>  [71,]  0.567982615 -0.020281395  0.230339371  0.793549803  0.807000092
#>  [72,] -0.466493316 -0.028374138 -0.400849533 -0.538087754 -0.568846047
#>  [73,]  0.492421229 -0.951302896  0.889831878 -0.695138874 -0.790690979
#>  [74,] -0.617143861 -0.815927291 -0.198754139 -0.815471948 -0.807361494
#>  [75,]  0.405253767  0.004790220 -0.460751888  0.302615365 -0.769815052
#>  [76,] -0.647008631  0.788239046  0.342484768  0.648260341  0.286195262
#>  [77,] -0.329766736 -0.765534858 -0.791288497  0.415780538  0.489825952
#>  [78,]  0.624218123  0.652741302 -0.627402956 -0.879768206 -0.039836674
#>  [79,]  0.239149053  0.609186310  0.704926069  0.872574080 -0.860098454
#>  [80,]  0.215742430 -0.591858914  0.751229815  0.874812311 -0.537822202
#>  [81,] -0.685552908 -0.652495911 -0.964260558  0.306006914 -0.725585163
#>  [82,] -0.128054809 -0.908416001 -0.783412351  0.635513659 -0.089319460
#>  [83,]  0.048453538 -0.492295771 -0.094077793  0.141093385  0.483446880
#>  [84,] -0.951958071  0.003147763  0.018370495 -0.941120743  0.734019888
#>  [85,] -0.949032217 -0.355108836  0.430513858  0.546772698  0.357123305
#>  [86,] -0.130434530  0.932712407  0.294804367 -0.107810375 -0.341211973
#>  [87,]  0.487816415  0.090440806  0.748927028  0.895172982 -0.434663653
#>  [88,] -0.094115626 -0.566619937 -0.532893030 -0.406397377 -0.114755223
#>  [89,]  0.316638127 -0.587599495 -0.980142577 -0.179181396 -0.590828704
#>  [90,] -0.167203919  0.260707964 -0.819250717 -0.031413915 -0.413734561
#>  [91,] -0.821911341  0.571723937  0.777516590  0.671513030  0.220167052
#>  [92,]  0.281338519 -0.122795728  0.676544886 -0.238701950  0.293007900
#>  [93,]  0.478543652 -0.760865134 -0.332063382 -0.677868982  0.006202261
#>  [94,] -0.475454221  0.876050211  0.733433464 -0.126666989  0.143237322
#>  [95,]  0.107515982  0.048041387  0.442635021  0.145671448  0.615533483
#>  [96,]  0.537398213 -0.030322486  0.036371687  0.177476218  0.206413265
#>  [97,]  0.333150879  0.717413382  0.683922816  0.354673515 -0.215615370
#>  [98,]  0.238532809  0.002430805 -0.916577231  0.346160441  0.289945825
#>  [99,] -0.131876506  0.713128655 -0.847926980 -0.033053859 -0.836858914
#> [100,]  0.207406908  0.930778522 -0.754589512 -0.045045198 -0.351887094
#> 
#> $y
#>   [1] 0 1 1 0 0 0 0 0 0 0 1 1 1 0 1 1 0 1 0 1 1 0 0 1 1 1 1 0 0 0 0 1 1 0 1 1 0
#>  [38] 1 0 0 0 0 1 0 1 1 1 1 1 0 1 0 1 0 0 1 0 1 1 0 1 1 1 0 1 0 0 0 1 0 1 1 0 0
#>  [75] 1 0 0 0 1 1 0 0 1 0 0 0 1 1 1 1 0 1 0 0 1 1 1 1 1 0
#>