Simulate Data from Hyper-Sphere for Classification Problems

The classification problem Circle is described in the BARK paper (2008). Inputs are dim independent variables uniformly distributed on the interval \([-1,1]\), only the first 2 out of these dim are actually signals. Outputs are created according to the formula $$y = 1(x1^2+x2^2 \le 2/\pi)$$

sim_circle(n, dim = 5)

Arguments

n

number of data points to generate

dim

number of dimension of the problem, no less than 2

Value

Returns a list with components

x

input values (independent variables)

y

0/1 output values (dependent variable)

References

Ouyang, Zhi (2008) Bayesian Additive Regression Kernels. Duke University. PhD dissertation, Chapter 3.

See also

Examples

sim_circle(n=100, dim=5)
#> $x
#>               [,1]         [,2]        [,3]        [,4]         [,5]
#>   [1,]  0.20655827 -0.809304062  0.36259397 -0.78362707  0.833283722
#>   [2,]  0.16963232 -0.481872038  0.08839748  0.16813487  0.486994134
#>   [3,] -0.15428054 -0.556323513 -0.71104955  0.54198529  0.331304285
#>   [4,]  0.63225624  0.992989989  0.46127697  0.94774678 -0.918698668
#>   [5,]  0.77605504 -0.214245168 -0.64115348  0.40400550  0.855671868
#>   [6,]  0.50561943 -0.115816683  0.37784573 -0.68512288 -0.458331807
#>   [7,] -0.67238337  0.453598457 -0.90884587  0.90218294 -0.796034834
#>   [8,] -0.20391111  0.972393058  0.34778432 -0.87653408  0.175736369
#>   [9,] -0.01021507 -0.506565221 -0.92177596 -0.41313641 -0.131476295
#>  [10,] -0.90931931 -0.201513167  0.65819305 -0.76697999  0.684624177
#>  [11,]  0.53193746  0.149799779 -0.25826012 -0.20636270  0.443497037
#>  [12,] -0.76835695  0.335480851 -0.58159088 -0.37474893 -0.622744148
#>  [13,] -0.16064273  0.379081211  0.37383046 -0.71337390  0.647657634
#>  [14,]  0.76652362 -0.451186389  0.82339683  0.83103800 -0.558051235
#>  [15,]  0.92906039  0.445385017  0.72457277 -0.96173101 -0.411752959
#>  [16,]  0.87460841  0.494043122 -0.29117848 -0.80687058 -0.990803379
#>  [17,]  0.03795588  0.216129296 -0.17664855  0.72062538 -0.877665767
#>  [18,]  0.16661084 -0.751423698  0.64637916  0.09015451 -0.721052422
#>  [19,] -0.11883976 -0.990729748 -0.41569609 -0.28919175 -0.323924897
#>  [20,]  0.57224887 -0.963489316 -0.36328799  0.47483036 -0.965187814
#>  [21,] -0.90491301  0.416600854  0.67720282 -0.77493990 -0.691123475
#>  [22,] -0.92058699  0.729897229  0.54225202  0.43697741 -0.559319431
#>  [23,] -0.02605062  0.704415535 -0.32588498  0.25120713 -0.662653751
#>  [24,] -0.30871215  0.147467141 -0.65498267 -0.49103570  0.884184274
#>  [25,]  0.74716175 -0.165393327 -0.89951428 -0.52187453  0.826321591
#>  [26,]  0.95694759  0.951699507  0.43004464  0.12631158  0.492124540
#>  [27,] -0.86518598  0.513072457 -0.47309815 -0.59837281 -0.560393518
#>  [28,]  0.30092445  0.577473100 -0.75701107 -0.92866329  0.637371559
#>  [29,] -0.41814825 -0.911746462  0.04496545 -0.99998831  0.989632581
#>  [30,]  0.80761614  0.046100540  0.53242119 -0.90199578  0.792515404
#>  [31,] -0.90335110  0.551171584 -0.05342501  0.50399242 -0.204131209
#>  [32,] -0.16304044 -0.149646533 -0.98636009  0.40044481  0.913210714
#>  [33,]  0.13319782  0.887379060  0.86245854  0.20367012 -0.606498700
#>  [34,]  0.74033547 -0.141166816 -0.68725923  0.98847978  0.721742958
#>  [35,]  0.63096532  0.769920700  0.95806447 -0.94372825  0.924345897
#>  [36,] -0.16099662  0.197621139  0.50479157  0.63063427  0.658738025
#>  [37,]  0.39457930  0.890291910 -0.88679544 -0.41457145 -0.151768567
#>  [38,] -0.09021290  0.275679173 -0.77576685 -0.12805876 -0.454227662
#>  [39,] -0.67192766 -0.550700891  0.84127907  0.28917353  0.092151203
#>  [40,] -0.75159669 -0.884293761 -0.87727537 -0.61451058 -0.869110306
#>  [41,]  0.15452866 -0.477597415 -0.94402720  0.48103204  0.846996864
#>  [42,] -0.62319037 -0.654015705  0.32900565 -0.35056025  0.384038965
#>  [43,] -0.01294768  0.885328581  0.38255625 -0.83072016 -0.095426281
#>  [44,] -0.71331927 -0.672034614  0.11500145 -0.34070512 -0.562683583
#>  [45,]  0.36813190 -0.673012225 -0.73159862 -0.97938404  0.012153460
#>  [46,] -0.40840939 -0.863942494 -0.88248219  0.59639258 -0.481013787
#>  [47,] -0.19233235 -0.407688600 -0.84837806 -0.19238276  0.724207071
#>  [48,]  0.29583610 -0.389036468 -0.91504056 -0.46042836 -0.887339013
#>  [49,]  0.03460620 -0.134899613 -0.49527479  0.48003955 -0.003395842
#>  [50,] -0.90795132 -0.486099234 -0.25413999 -0.98812384 -0.593095832
#>  [51,] -0.41219962 -0.119553308  0.50846716 -0.78922468 -0.217293181
#>  [52,] -0.14390054 -0.777575004 -0.15513483 -0.16688421 -0.844815885
#>  [53,] -0.86744762 -0.948685835  0.94626635 -0.38747465 -0.289718753
#>  [54,]  0.56091387  0.675785912  0.11430629  0.68399583  0.252194718
#>  [55,] -0.19262218 -0.131583556  0.82738817 -0.82484377  0.959852444
#>  [56,]  0.25520441  0.124881305 -0.51545187 -0.09207554 -0.689542291
#>  [57,]  0.94948100  0.966754449  0.23338250 -0.15667887  0.446348169
#>  [58,]  0.24483994 -0.213825864  0.08232720 -0.41320829 -0.261604920
#>  [59,] -0.21953279  0.171467371  0.28726071 -0.36023737  0.079063610
#>  [60,]  0.74204899 -0.132534525  0.01620233 -0.22393808  0.070346773
#>  [61,]  0.82633312  0.773197586  0.07381992  0.01253657 -0.999508972
#>  [62,]  0.40800898  0.222916986  0.92739263  0.60300469  0.266195251
#>  [63,] -0.46720819 -0.609089101 -0.64520409 -0.83745592  0.129144384
#>  [64,] -0.49453472  0.189687492 -0.90973914 -0.22579346 -0.650038566
#>  [65,] -0.37695054  0.562817466 -0.57575630 -0.67005357 -0.213486498
#>  [66,]  0.69554806  0.301967957 -0.95092595 -0.97221146 -0.695345269
#>  [67,]  0.22085758  0.411169111  0.67595022 -0.76497788  0.294882701
#>  [68,] -0.52016027  0.014950002  0.91383153 -0.81435715 -0.207661550
#>  [69,] -0.66164552  0.293502545  0.17613210  0.14359709 -0.642715329
#>  [70,] -0.67939680  0.959749458  0.57964276  0.31302568 -0.275382629
#>  [71,]  0.33031273  0.121415238  0.74038024  0.37755879 -0.587842960
#>  [72,]  0.53911740  0.072905905  0.38684192 -0.64376114 -0.722239600
#>  [73,]  0.82154277 -0.952068932  0.70133941 -0.68464354 -0.929055607
#>  [74,] -0.08690230 -0.738536630 -0.45054581  0.27980717  0.700245704
#>  [75,] -0.92009539  0.426507063 -0.17379640  0.83939142 -0.966287405
#>  [76,]  0.58303714 -0.279763693  0.54289711  0.90230579  0.911263499
#>  [77,] -0.44493250  0.389335877  0.80317729 -0.29037738 -0.708361478
#>  [78,] -0.69046927  0.030070907  0.73652359 -0.98553044 -0.912418718
#>  [79,] -0.52376530 -0.807316449 -0.50991774  0.61580967  0.067124767
#>  [80,]  0.91753858 -0.222520882 -0.14904195  0.83719945 -0.884518157
#>  [81,]  0.66365342  0.480084269 -0.38017526  0.05100401 -0.868810348
#>  [82,]  0.19177105  0.389037974  0.54686387 -0.01088570 -0.910191026
#>  [83,]  0.84083654 -0.006037541  0.65456934 -0.47326025  0.116097155
#>  [84,] -0.31431486 -0.782954627  0.78605959 -0.75997449 -0.246245069
#>  [85,] -0.80672982  0.210701721 -0.02429784  0.56235725 -0.217047533
#>  [86,] -0.31611113 -0.809569650 -0.69817548  0.22062109 -0.610025438
#>  [87,]  0.12752579 -0.861862106  0.16981181 -0.66629028  0.979191023
#>  [88,]  0.48707174  0.449645503 -0.26124820  0.46207917  0.300448421
#>  [89,]  0.80228376  0.508955237 -0.59230924 -0.57055365 -0.694640417
#>  [90,] -0.54475701  0.858617612  0.27006476 -0.68830439 -0.517523348
#>  [91,]  0.06175415 -0.956221423 -0.61651091 -0.63630088 -0.855487830
#>  [92,]  0.21319757 -0.678171123 -0.90813797  0.55589091 -0.526377605
#>  [93,]  0.08642763  0.679072986 -0.80085708 -0.73839507 -0.200151096
#>  [94,]  0.12777198 -0.033016263 -0.48951639  0.63661452 -0.268896192
#>  [95,] -0.77805584 -0.504170062 -0.78001639 -0.53673079  0.479317747
#>  [96,] -0.36455633 -0.793282303  0.79096536  0.90793497 -0.524307650
#>  [97,] -0.08274367  0.018728351  0.60781774 -0.59370250  0.560053762
#>  [98,]  0.52573303 -0.750252216 -0.51988918  0.61171052  0.713728840
#>  [99,] -0.38073120 -0.880670582  0.96171042  0.88162898  0.760910629
#> [100,] -0.05595572 -0.049109794 -0.61360101  0.12306525  0.160712000
#> 
#> $y
#>   [1] 0 1 1 0 0 1 0 0 1 0 1 0 1 0 0 0 1 1 0 0 0 0 1 1 1 0 0 1 0 0 0 1 0 1 0 1 0
#>  [38] 1 0 0 1 0 0 0 1 0 1 1 1 0 1 1 0 0 1 1 0 1 1 1 0 1 1 1 1 1 1 1 1 0 1 1 0 1
#>  [75] 0 1 1 1 0 0 0 1 0 0 0 0 0 1 0 0 0 1 1 1 0 0 1 0 0 1
#>