Simulate Data from Hyper-Sphere for Classification Problems

The classification problem Circle is described in the BARK paper (2008). Inputs are dim independent variables uniformly distributed on the interval \([-1,1]\), only the first 2 out of these dim are actually signals. Outputs are created according to the formula $$y = 1(x1^2+x2^2 \le 2/\pi)$$

Usage

sim_circle(n, dim = 5)

Arguments

n

number of data points to generate

dim

number of dimension of the problem, no less than 2

Value

Returns a list with components

x

input values (independent variables)

y

0/1 output values (dependent variable)

References

Ouyang, Zhi (2008) Bayesian Additive Regression Kernels. Duke University. PhD dissertation, Chapter 3.

See also

Examples

sim_circle(n=100, dim=5)
#> $x
#>               [,1]        [,2]        [,3]        [,4]        [,5]
#>   [1,] -0.36961198 -0.53348013  0.04390449  0.53084967 -0.84585412
#>   [2,]  0.82964434 -0.23174421 -0.29801345 -0.17209116 -0.22384221
#>   [3,] -0.92141013 -0.39454730  0.17861588  0.66871383 -0.68623210
#>   [4,] -0.06626801 -0.17188535  0.70436912  0.48804964 -0.66654205
#>   [5,]  0.80359605 -0.64222080 -0.70373485  0.06619135  0.44846647
#>   [6,] -0.11024488  0.11410438  0.37040031  0.09731654 -0.25534388
#>   [7,] -0.95625644 -0.81458024 -0.47864998  0.81475285 -0.35820908
#>   [8,] -0.07179352  0.67931465  0.15075838  0.56166319  0.59252698
#>   [9,]  0.62808270  0.02934675  0.86989885  0.45354173  0.64932134
#>  [10,] -0.07412217  0.89272891 -0.52999655  0.94402577  0.97412275
#>  [11,]  0.82528140  0.61933879 -0.36285407 -0.37548485 -0.94712019
#>  [12,]  0.80115492  0.60975061  0.44724173 -0.19959865  0.10988136
#>  [13,] -0.59144936  0.44328569  0.49657631  0.92033281 -0.90330469
#>  [14,] -0.55136479  0.77806676 -0.78385368 -0.34679444  0.51185101
#>  [15,] -0.18442912  0.98401468  0.72880953  0.38528684 -0.48358955
#>  [16,] -0.05426707 -0.13593452  0.61552741 -0.45584759 -0.59024986
#>  [17,] -0.58750967 -0.79191208 -0.74556536 -0.16550657 -0.08797216
#>  [18,] -0.54359153 -0.84941683  0.15118085 -0.39175108 -0.01123266
#>  [19,] -0.02823852 -0.52783802  0.10834073  0.69009227  0.96140826
#>  [20,]  0.91528785 -0.57547413  0.37817514 -0.14237067  0.05151019
#>  [21,]  0.11466551  0.15264296  0.93721438  0.67283522  0.57926398
#>  [22,] -0.32224221  0.03722554  0.15664547 -0.74790581 -0.62995177
#>  [23,] -0.73332488  0.36186884  0.55322371  0.23826265  0.37911116
#>  [24,] -0.02436449  0.77995966  0.67104638  0.01600144  0.04964982
#>  [25,] -0.91311342 -0.25161412  0.63887258  0.19521453 -0.22375276
#>  [26,] -0.62915086  0.91008674 -0.17758918 -0.88392334  0.46320176
#>  [27,]  0.94443252 -0.01771911 -0.50826908 -0.07530125 -0.56941425
#>  [28,]  0.08077083  0.33242528 -0.39843471  0.56060325  0.94031181
#>  [29,]  0.92934212 -0.43089951 -0.58888358 -0.92339267 -0.85944036
#>  [30,]  0.62319353  0.03334553 -0.66302019  0.91710810  0.56872052
#>  [31,] -0.87448470  0.55884241 -0.41163263 -0.24722293  0.23734149
#>  [32,]  0.41825747 -0.56991957  0.39878037  0.43941041 -0.50697949
#>  [33,]  0.31504968 -0.31696958  0.86262213  0.23133925  0.08401318
#>  [34,]  0.98727221  0.12554945 -0.51931287  0.53837871  0.16840804
#>  [35,] -0.73374889  0.67663428  0.86946093 -0.60265122  0.26092658
#>  [36,]  0.80292693 -0.11099082  0.58018861  0.96082586 -0.84091230
#>  [37,] -0.50578878  0.61792415 -0.13524501  0.80050434  0.32069923
#>  [38,] -0.79403157 -0.26851671  0.92295879 -0.36142231  0.17578127
#>  [39,]  0.10717852  0.66752737 -0.71085527 -0.75074139 -0.50173431
#>  [40,]  0.81629818  0.84037349  0.91908353 -0.23941515 -0.20414892
#>  [41,] -0.81868678  0.71266347 -0.43317657  0.45063259  0.10540077
#>  [42,]  0.88411588  0.70984433 -0.82053764 -0.89635099  0.78570298
#>  [43,] -0.34465542  0.79900022 -0.62288665  0.13264575  0.08634110
#>  [44,]  0.97286408  0.76855422  0.43589038 -0.06112932  0.30560193
#>  [45,]  0.63407596  0.17983233  0.30302288  0.53540235  0.43225389
#>  [46,] -0.04600450 -0.67661826  0.71790500  0.41951485 -0.55324058
#>  [47,]  0.69029972 -0.63661438  0.39647756  0.37730017  0.46119620
#>  [48,]  0.39305345  0.07216508  0.64500360  0.65272928  0.46929901
#>  [49,] -0.86621717 -0.98429686 -0.11691203 -0.98987072 -0.48742647
#>  [50,]  0.02480789  0.65305788  0.39149669  0.64580090 -0.92784601
#>  [51,] -0.49044731  0.92149947 -0.69762126  0.20276055  0.36794836
#>  [52,] -0.29269878 -0.18300029 -0.27541838  0.04815876 -0.42504129
#>  [53,]  0.71638929  0.22168632  0.94671965 -0.49321917  0.05771053
#>  [54,]  0.42265712 -0.52352252 -0.51503595  0.78728143 -0.36840420
#>  [55,] -0.89734210 -0.35427580  0.23565921  0.38772930  0.78292476
#>  [56,] -0.69882658  0.85124572  0.44309480  0.86697364  0.16618402
#>  [57,]  0.73616982 -0.38431926  0.97331868  0.12656805  0.52608562
#>  [58,] -0.41918796  0.69859097  0.19536706 -0.81198306 -0.45062721
#>  [59,] -0.16759983 -0.23791182  0.14831036  0.61018136 -0.59235834
#>  [60,]  0.08487235  0.02499823  0.63517138  0.75701340 -0.87007909
#>  [61,] -0.97555374  0.54164358 -0.51956329 -0.78892429  0.45979337
#>  [62,] -0.05514069 -0.77307401  0.53763548  0.19841082 -0.50857366
#>  [63,]  0.10373489 -0.93135654 -0.27044365  0.89085365 -0.06480414
#>  [64,] -0.92429238 -0.69134562  0.06527183  0.60836127 -0.09648126
#>  [65,] -0.23224809 -0.94752681 -0.25352630  0.20297225 -0.65532145
#>  [66,] -0.19211284 -0.35056738  0.30430361  0.13038208  0.51877676
#>  [67,]  0.52911815 -0.75651087  0.50709842 -0.66040122  0.28506865
#>  [68,] -0.54805598  0.29026653  0.12332680 -0.53120167  0.11444038
#>  [69,]  0.49692436  0.51403481 -0.67611687  0.20341731  0.11050175
#>  [70,]  0.02327904  0.97387789  0.87905768 -0.73520857 -0.68221160
#>  [71,] -0.71453885 -0.28321520  0.88696009 -0.80871235  0.23614779
#>  [72,] -0.10971497  0.06321188  0.09388102  0.86343888  0.99198339
#>  [73,] -0.56782512  0.43512688 -0.98105764  0.56122805 -0.02549075
#>  [74,]  0.28911727 -0.02233556  0.13289654 -0.01473868 -0.23086325
#>  [75,] -0.94382693 -0.64647167 -0.07146918  0.54071923  0.73286827
#>  [76,]  0.37997347  0.20705025  0.10645463 -0.96989629  0.17497950
#>  [77,] -0.04218068 -0.78670278 -0.54732089 -0.90237957 -0.57008933
#>  [78,] -0.78674664  0.30851265  0.49924585  0.85160767  0.14351375
#>  [79,] -0.78466406 -0.09345593 -0.96954802 -0.51369824  0.74147924
#>  [80,] -0.41576293  0.13997156  0.43736281 -0.18030593 -0.61079430
#>  [81,]  0.86891625 -0.12030529  0.45447576  0.73029222  0.97601645
#>  [82,]  0.53870123  0.57615886  0.24647343  0.36051260  0.54284631
#>  [83,]  0.33092877  0.82540548  0.45382370 -0.72270106 -0.81094408
#>  [84,] -0.49273679 -0.05466931 -0.04017881 -0.40784876 -0.66125518
#>  [85,]  0.42836817 -0.75282815 -0.48963421  0.80396179  0.57931024
#>  [86,]  0.66025149  0.28479953 -0.30010050  0.88071495  0.48955030
#>  [87,] -0.84112590  0.32008329 -0.78408991 -0.81924587  0.13520493
#>  [88,]  0.57479691  0.80366592  0.83493361  0.23125854 -0.61130132
#>  [89,]  0.37963606 -0.03604875 -0.73911989 -0.05926663  0.96113859
#>  [90,]  0.54673579 -0.39087189 -0.73236645 -0.38397421  0.33357703
#>  [91,] -0.92409050 -0.74566926  0.04236730  0.87340353 -0.83092210
#>  [92,] -0.09493089  0.61825326  0.69592583 -0.59006902 -0.65829869
#>  [93,] -0.07009381  0.57038223 -0.04734799  0.06997909  0.41398911
#>  [94,]  0.09480195 -0.54707200 -0.02859985  0.43106884 -0.65387274
#>  [95,] -0.87063440 -0.10544866  0.19943193  0.49219876  0.49956800
#>  [96,] -0.82131646  0.15196260  0.72559079 -0.09205552  0.09506335
#>  [97,]  0.04699362 -0.91690812 -0.52904719  0.07393405 -0.41008476
#>  [98,]  0.29104869  0.68458533 -0.80825967 -0.37735868  0.57975089
#>  [99,] -0.86126464 -0.04841631 -0.86086917  0.39919964  0.23084096
#> [100,] -0.63726407 -0.33521192  0.86789920  0.47980161 -0.16847763
#> 
#> $y
#>   [1] 1 0 0 1 0 1 0 1 1 0 0 0 1 0 0 1 0 0 1 0 1 1 0 1 0 0 0 1 0 1 0 1 1 0 0 0 0
#>  [38] 0 1 0 0 0 0 0 1 1 0 1 0 1 0 1 1 1 0 0 0 0 1 1 0 1 0 0 0 1 0 1 1 0 1 1 1 1
#>  [75] 0 1 1 0 1 1 0 1 0 1 0 1 0 0 1 1 0 1 1 1 0 0 0 1 0 1
#>