Simulate Data from Hyper-Sphere for Classification Problems

The classification problem Circle is described in the BARK paper (2008). Inputs are dim independent variables uniformly distributed on the interval \([-1,1]\), only the first 2 out of these dim are actually signals. Outputs are created according to the formula $$y = 1(x1^2+x2^2 \le 2/\pi)$$

Usage

sim_circle(n, dim = 5)

Arguments

n

number of data points to generate

dim

number of dimension of the problem, no less than 2

Value

Returns a list with components

x

input values (independent variables)

y

0/1 output values (dependent variable)

References

Ouyang, Zhi (2008) Bayesian Additive Regression Kernels. Duke University. PhD dissertation, Chapter 3.

See also

Examples

sim_circle(n=100, dim=5)
#> $x
#>                [,1]        [,2]        [,3]          [,4]         [,5]
#>   [1,] -0.922705572 -0.33949439 -0.67236355 -0.1207115408  0.907020786
#>   [2,] -0.105862543  0.52546916 -0.07562261 -0.5003707418 -0.883904464
#>   [3,]  0.521245366  0.33042093 -0.24715225 -0.0005052378  0.932293639
#>   [4,] -0.184292371  0.67413336 -0.45559650  0.4900322398 -0.311358638
#>   [5,] -0.206081090  0.65704607  0.47060086 -0.9074192634  0.820610977
#>   [6,] -0.168400654  0.94272939 -0.54328542 -0.8253477374 -0.501387551
#>   [7,]  0.658186051 -0.46346776  0.76350679  0.5439806040  0.910689742
#>   [8,]  0.988009538  0.63358962 -0.99580475 -0.3469208232  0.061938917
#>   [9,] -0.439672930 -0.18067978 -0.29918557  0.2497163024 -0.079591818
#>  [10,]  0.193643512  0.41120589  0.20334926  0.4435057412  0.317656141
#>  [11,]  0.186208250 -0.76562737 -0.13581111 -0.3167984267  0.631614678
#>  [12,] -0.544021678  0.64818473  0.58483607  0.5111880638  0.898236665
#>  [13,] -0.740061636  0.11348139  0.08339322 -0.6533205546  0.410699623
#>  [14,]  0.930253952 -0.88087058  0.66607221  0.7592366147  0.274900085
#>  [15,] -0.657810538  0.32773126  0.53188213  0.8453421569 -0.519307625
#>  [16,]  0.512513124 -0.60611053  0.23512191  0.5234085219 -0.411476448
#>  [17,] -0.563985623  0.43932868 -0.12731126  0.6283732392  0.919672074
#>  [18,]  0.487794264  0.28020902 -0.60074938 -0.8204991911 -0.677385450
#>  [19,] -0.470303279 -0.84398132  0.47458801 -0.1526912078 -0.673175843
#>  [20,] -0.788972804 -0.09764133  0.85090573  0.6289250059  0.048056435
#>  [21,]  0.763241435  0.07048872  0.25898786 -0.3395032613 -0.116101393
#>  [22,]  0.440072240 -0.67017083  0.28451341 -0.5685911924  0.329119064
#>  [23,] -0.567148837 -0.30094573  0.30229620  0.7156065763  0.014452184
#>  [24,] -0.178912734 -0.82781399  0.83283228 -0.5569018498 -0.950919647
#>  [25,]  0.816697075 -0.35001843 -0.58902270 -0.0820254288 -0.604644429
#>  [26,] -0.713204132 -0.91645825 -0.46848178 -0.3761167512  0.079535743
#>  [27,] -0.083726816 -0.41987059  0.71440408 -0.0268794727 -0.898441747
#>  [28,] -0.132537093 -0.56131445 -0.07122965 -0.8091082005 -0.595427060
#>  [29,]  0.723165624  0.20140816 -0.85555238  0.4891259754 -0.910547239
#>  [30,]  0.624158724 -0.67258202 -0.08388794 -0.0083297924 -0.739590751
#>  [31,] -0.333140376 -0.43562414  0.76024427  0.6360339299 -0.558678259
#>  [32,]  0.808201869 -0.89251064  0.48958992  0.0897581829 -0.993086326
#>  [33,]  0.922998041  0.94326039  0.62701664 -0.5237929057  0.780784333
#>  [34,]  0.764758992 -0.04577912 -0.25686452  0.3821652499 -0.038166156
#>  [35,] -0.504459695 -0.26084495 -0.58891126  0.9311341373 -0.222524789
#>  [36,] -0.553263923  0.86556840  0.87485493  0.7747790781 -0.912134802
#>  [37,]  0.274702598  0.42313183  0.81499489  0.2961993804 -0.478268185
#>  [38,] -0.421789986 -0.62911268  0.81379544 -0.8711690968 -0.226891773
#>  [39,]  0.352704344  0.83686844 -0.37038533 -0.8663691599 -0.270739783
#>  [40,]  0.159321548  0.64282276  0.73036099  0.0123381317 -0.577942310
#>  [41,]  0.915917942 -0.58543422 -0.93359987 -0.2453246573  0.723552312
#>  [42,] -0.237274439  0.41971548 -0.65459200 -0.6125955246  0.740228304
#>  [43,] -0.604973067 -0.41762661 -0.80229301 -0.5909527624  0.888797799
#>  [44,] -0.679474667 -0.08734063  0.21459518 -0.5196816232  0.125004541
#>  [45,]  0.628693268  0.53272255  0.99781041 -0.2038816782 -0.736107583
#>  [46,]  0.947030011 -0.73642525 -0.10402713  0.3839044473 -0.719026028
#>  [47,]  0.558734620  0.77237434  0.88551522 -0.3703166498  0.297813404
#>  [48,]  0.853001433 -0.22528070 -0.21422105 -0.2380604972 -0.851296702
#>  [49,]  0.006423069  0.99234240  0.67268207 -0.9847366232  0.687627021
#>  [50,] -0.927636954  0.18557039 -0.05160703 -0.2722156942 -0.201513335
#>  [51,] -0.285414641  0.11652632 -0.60312346  0.3745570304  0.207961183
#>  [52,]  0.980814717 -0.54408771 -0.43563117 -0.5227741832  0.757846808
#>  [53,]  0.124942500  0.77169596 -0.15037731  0.3388036713 -0.561968418
#>  [54,] -0.844140113 -0.61467966 -0.27697321 -0.8702403610 -0.225128242
#>  [55,] -0.842804945  0.42609766 -0.91468098  0.6937690419  0.606063273
#>  [56,]  0.318557793 -0.20072426 -0.54685649 -0.0518101794 -0.914586109
#>  [57,] -0.556824011 -0.37701370  0.51351616 -0.9384516762 -0.315838297
#>  [58,]  0.054630800 -0.53445420 -0.29506596 -0.4073668211  0.562620681
#>  [59,] -0.094160649  0.69801619  0.57240427  0.3631676016  0.791192039
#>  [60,]  0.586675051  0.66103296 -0.19663340  0.2737808833  0.948172972
#>  [61,]  0.788896036  0.43372758  0.49294705  0.2939539738 -0.324391763
#>  [62,]  0.910478222  0.86459258 -0.83440061  0.0081135156 -0.103396287
#>  [63,]  0.504669366  0.10663690 -0.58737810  0.3222319675  0.578231929
#>  [64,]  0.501259044  0.96378827  0.81594815  0.9888811354  0.156272600
#>  [65,]  0.518877043 -0.09543636  0.85743740  0.8103069663 -0.675384617
#>  [66,]  0.299298430  0.91341449  0.55373175 -0.1015834371 -0.412373227
#>  [67,]  0.215978099 -0.51381455  0.31812234  0.9430914391 -0.537687819
#>  [68,] -0.911193994  0.14900590  0.51531212  0.1552196615  0.315408383
#>  [69,] -0.116798147  0.22933359  0.25208987  0.5425015008  0.008005199
#>  [70,]  0.575039891  0.16838001  0.05025878  0.0938607031 -0.602847161
#>  [71,] -0.294080518 -0.43666547 -0.81074912 -0.0428400487  0.275163529
#>  [72,]  0.868172152 -0.39787619  0.34701494  0.5520150200 -0.021295067
#>  [73,]  0.005230767  0.50787758  0.13496158 -0.7677032882 -0.974949529
#>  [74,]  0.740614317  0.24121974 -0.94596235 -0.9408823764  0.650688116
#>  [75,] -0.298960260 -0.72191028  0.77177902 -0.2958479729  0.362364846
#>  [76,]  0.281953095  0.56995306 -0.71270465  0.8733807341 -0.138447308
#>  [77,] -0.674925890 -0.99083335 -0.99712376 -0.7849738719  0.972190086
#>  [78,]  0.759641618  0.77405693  0.04127983 -0.9204769642 -0.188190614
#>  [79,]  0.725319879 -0.52157589  0.34923305 -0.9766173121  0.650306636
#>  [80,] -0.024399933 -0.91496247  0.40037425  0.1500139777  0.714405511
#>  [81,] -0.722479094 -0.50868626 -0.97061737 -0.6002373141 -0.229303482
#>  [82,] -0.741181963 -0.10752499  0.65895783 -0.7272343123  0.672692781
#>  [83,]  0.906144617 -0.25035425 -0.31428966 -0.2452820255 -0.578259377
#>  [84,] -0.974572545  0.92180614 -0.39868866  0.9698736318 -0.153636160
#>  [85,] -0.266614658  0.43851929  0.93191268 -0.8740886701  0.510852191
#>  [86,] -0.627486601 -0.85924411  0.56645455 -0.2853609123 -0.065190018
#>  [87,] -0.474586241  0.55834274 -0.78974271 -0.7089121710  0.499027778
#>  [88,]  0.855425111  0.36815996 -0.93758116 -0.4206495769 -0.555410022
#>  [89,]  0.652365780 -0.57735744 -0.33900062  0.0649004993 -0.633320131
#>  [90,] -0.765565674  0.58445018 -0.74506046 -0.9751714752  0.419719521
#>  [91,]  0.025560718 -0.31331196  0.17551880 -0.8395175659  0.087316395
#>  [92,] -0.455415922  0.47349978  0.73225056  0.9872847605 -0.197527684
#>  [93,] -0.798814856  0.91861133  0.82157233 -0.5003304179 -0.249690580
#>  [94,]  0.993977424 -0.41943616 -0.64214462  0.3692583162  0.726695394
#>  [95,]  0.658402566 -0.47547577 -0.32418627  0.7867110479  0.096564428
#>  [96,] -0.034199640 -0.60108407  0.76454819  0.5126632652 -0.491359941
#>  [97,]  0.790242287  0.08666090  0.15956693  0.1843926827 -0.116463218
#>  [98,]  0.369550381  0.77627297 -0.48524520  0.3313704133 -0.665274890
#>  [99,] -0.092680453  0.95697376 -0.19591013 -0.4431673698  0.599541126
#> [100,]  0.635538164 -0.11810252  0.58540184 -0.0801674789  0.200488472
#> 
#> $y
#>   [1] 0 1 1 1 1 0 0 0 1 1 1 0 1 0 1 1 1 1 0 1 1 0 1 0 0 0 1 1 1 0 1 0 0 1 1 0 1
#>  [38] 1 0 1 0 1 1 1 0 0 0 0 0 0 1 0 1 0 0 1 1 1 1 0 0 0 1 0 1 0 1 0 1 1 1 0 1 1
#>  [75] 1 1 0 0 0 0 0 1 0 0 1 0 1 0 0 0 1 1 0 0 0 1 1 0 0 1
#>