Simulate Data from Hyper-Sphere for Classification Problems

The classification problem Circle is described in the BARK paper (2008). Inputs are dim independent variables uniformly distributed on the interval \([-1,1]\), only the first 2 out of these dim are actually signals. Outputs are created according to the formula $$y = 1(x1^2+x2^2 \le 2/\pi)$$

sim_circle(n, dim = 5)

Arguments

n

number of data points to generate

dim

number of dimension of the problem, no less than 2

Value

Returns a list with components

x

input values (independent variables)

y

0/1 output values (dependent variable)

References

Ouyang, Zhi (2008) Bayesian Additive Regression Kernels. Duke University. PhD dissertation, Chapter 3.

See also

Examples

sim_circle(n=100, dim=5)
#> $x
#>               [,1]         [,2]          [,3]         [,4]         [,5]
#>   [1,]  0.58956495 -0.194215273  0.7631992716  0.288923502 -0.015430235
#>   [2,] -0.04989597  0.216330010 -0.9298618198 -0.226036156  0.853645553
#>   [3,]  0.78037155  0.204914447  0.6816840298  0.386799742  0.983891519
#>   [4,]  0.27029353 -0.783419763 -0.4499793299  0.153662314  0.404738063
#>   [5,]  0.55486696 -0.311921736 -0.8559122873 -0.291489471 -0.383811623
#>   [6,] -0.05540496 -0.885989401 -0.5847821254 -0.306924518  0.426148113
#>   [7,] -0.37506526 -0.745762888  0.3113077604  0.534130407  0.645562414
#>   [8,] -0.43908753 -0.790239217  0.8249577228 -0.071095295  0.707676777
#>   [9,]  0.59093884 -0.967474074 -0.4999014623  0.985947602 -0.105383683
#>  [10,] -0.34184318  0.965894196 -0.4491909989  0.821973999 -0.111423798
#>  [11,] -0.75484408 -0.404835700  0.2628695020  0.666275187 -0.958392303
#>  [12,]  0.44525378 -0.951492472  0.6753254235 -0.520179029  0.472969099
#>  [13,]  0.84506242  0.761348014 -0.9512383752 -0.232240437  0.826853234
#>  [14,]  0.24172490 -0.334998317  0.0003451058  0.449001123 -0.815141578
#>  [15,]  0.29341761  0.750785782 -0.5206568576  0.057512707 -0.822216578
#>  [16,]  0.65603496  0.022121523 -0.9368217560  0.575326316  0.009363447
#>  [17,] -0.66103499 -0.677405052 -0.2882837746 -0.707096669  0.059449575
#>  [18,] -0.21469271 -0.090982446  0.4043025714 -0.882481674  0.384694726
#>  [19,] -0.78748830 -0.543055107  0.4519232623 -0.534229225  0.359440585
#>  [20,]  0.63845013 -0.040165861  0.1641142387 -0.656352584 -0.252920571
#>  [21,] -0.31166119  0.516829659 -0.6296939943 -0.208127011 -0.627353412
#>  [22,]  0.39658011 -0.209507697  0.0979535426  0.076736432  0.087174437
#>  [23,] -0.18223210  0.698017274 -0.3491819180 -0.942367760 -0.164746976
#>  [24,]  0.34896059 -0.070295651 -0.7953566071  0.878903755 -0.896637375
#>  [25,] -0.05002809 -0.580039711 -0.9339125711 -0.545937156  0.842813787
#>  [26,] -0.99788494  0.740498856 -0.4228378288 -0.996669845  0.253271793
#>  [27,] -0.97836161  0.551151549 -0.7608849015  0.638128925 -0.236952624
#>  [28,] -0.27246689 -0.563057814  0.9677081848  0.874512599 -0.403828626
#>  [29,]  0.61151677  0.582661012  0.1060360270 -0.721762106  0.399187683
#>  [30,] -0.28150835 -0.676846360  0.6137586134  0.614614405 -0.107371730
#>  [31,] -0.48606960  0.210043554 -0.7160251462  0.560652733  0.279134457
#>  [32,]  0.88185164 -0.816972928 -0.9267397756  0.139473378 -0.225497982
#>  [33,]  0.27604392 -0.696052569  0.4869594821 -0.526248956  0.167667062
#>  [34,]  0.95322560 -0.913264520  0.0661959993  0.750180599 -0.177344713
#>  [35,] -0.64941839 -0.484703874  0.2087606904 -0.311891055 -0.526470698
#>  [36,] -0.56167336 -0.063007244 -0.7765678535  0.563853744  0.683596411
#>  [37,]  0.33174674 -0.518259994 -0.0646815589 -0.892046073 -0.102797962
#>  [38,] -0.97963696  0.585051741  0.8546970943  0.173188066  0.526440952
#>  [39,] -0.31428702 -0.125472682 -0.5588966738  0.975164760  0.964803486
#>  [40,] -0.49720693  0.532679174  0.0213612569 -0.549511336 -0.827775393
#>  [41,] -0.63964457  0.420116078  0.4684261549  0.344693456  0.436272394
#>  [42,] -0.32793939  0.039567193 -0.7850722116 -0.892036333 -0.922674287
#>  [43,] -0.63146851 -0.352877618 -0.3063560450  0.417906187 -0.772257828
#>  [44,]  0.59054081  0.649826976  0.9030059911  0.001888991 -0.131109956
#>  [45,]  0.76011406  0.561303596  0.3142132051  0.077983713 -0.197053347
#>  [46,]  0.18723508  0.571746779 -0.2280284739  0.221706186 -0.913300382
#>  [47,] -0.78846971  0.196394280 -0.8381156311  0.921341576  0.089729667
#>  [48,]  0.30459642 -0.607187734 -0.5888003376 -0.696027604  0.665719439
#>  [49,]  0.75353994 -0.679117705  0.8255887427  0.956410366 -0.423067770
#>  [50,] -0.77524079  0.427109625  0.1821436249 -0.850912448  0.506766452
#>  [51,]  0.65073833 -0.401992528 -0.2880053408  0.939383974 -0.044225703
#>  [52,] -0.60006778  0.606974210  0.8618460242 -0.540873433  0.473017093
#>  [53,]  0.08332738 -0.588679604 -0.6816459098  0.776412818  0.303828141
#>  [54,]  0.95166627  0.940687001  0.1696354616 -0.844278200 -0.650326513
#>  [55,] -0.75033377  0.298691557 -0.6375604551 -0.163539376  0.650855982
#>  [56,]  0.28890083  0.680034241 -0.9253746397  0.919233362 -0.278544316
#>  [57,]  0.51675080 -0.670423432 -0.5629164493  0.959080386  0.892877199
#>  [58,]  0.36328411  0.511662868 -0.7939211191 -0.748009827  0.013615833
#>  [59,] -0.71527758 -0.841341082 -0.2858081716 -0.469656161 -0.798792338
#>  [60,]  0.32753445 -0.342929562  0.3576429151  0.379954961 -0.374948744
#>  [61,]  0.33264093  0.476500808  0.7302850690 -0.603322281 -0.770471334
#>  [62,] -0.53703792  0.841634818  0.0699391207  0.750340676  0.550493969
#>  [63,] -0.31519385  0.341967962 -0.7130586077  0.859969628  0.208887897
#>  [64,]  0.31181873  0.050510806  0.6923106103 -0.292879968  0.204337703
#>  [65,] -0.08311333 -0.281222158  0.1183991889  0.309023359 -0.991730030
#>  [66,]  0.49722449  0.473583477  0.3733933708  0.064846847 -0.165422834
#>  [67,]  0.39540114  0.091732624  0.2509176014 -0.950056499  0.131963193
#>  [68,] -0.38372455 -0.084391213  0.0066649327 -0.420342292 -0.187500149
#>  [69,] -0.37650033  0.534855488  0.7223546095 -0.205400409  0.408375085
#>  [70,] -0.04953507 -0.112872228 -0.3362212027 -0.846795911 -0.974607301
#>  [71,]  0.60374441  0.898443974  0.5995636107  0.238686909 -0.710934136
#>  [72,] -0.93644527  0.166900335  0.2204033458 -0.060231119  0.234285940
#>  [73,]  0.40859020  0.695788917 -0.0779921846 -0.499636920  0.925482750
#>  [74,]  0.66723120  0.987945654 -0.6382864090 -0.894533996 -0.028311562
#>  [75,] -0.81639528  0.751587228 -0.7210160601 -0.419168125 -0.620236554
#>  [76,] -0.47310815  0.668821977 -0.8419353426 -0.395724856 -0.640745413
#>  [77,] -0.58691840 -0.157096475 -0.6086425562 -0.285881007 -0.074442522
#>  [78,] -0.04508207 -0.306570884  0.0032549161 -0.178204264 -0.437157703
#>  [79,]  0.57600783 -0.057005372  0.1547572180  0.211827026  0.598567483
#>  [80,]  0.88034099 -0.134576085  0.8275812902 -0.254995701 -0.183977070
#>  [81,] -0.78466187 -0.492186763  0.3295325153 -0.228211317 -0.523388124
#>  [82,]  0.60731226 -0.977477278  0.9018765064 -0.141517981  0.747912525
#>  [83,]  0.80034479  0.312840661 -0.5223902799 -0.693226022 -0.696972127
#>  [84,] -0.86144110 -0.769747741 -0.7099428750  0.713813376  0.357102331
#>  [85,]  0.81505097  0.375593992 -0.0329657015  0.135037574 -0.307020193
#>  [86,]  0.10748770  0.170885619 -0.0420939564  0.788100320 -0.355578008
#>  [87,] -0.26205865  0.001282372 -0.0180904553 -0.887148253  0.290560659
#>  [88,] -0.72201707 -0.866361600  0.4426542330 -0.440894061 -0.563639106
#>  [89,]  0.27030229  0.062947415 -0.1565691680  0.693677461 -0.710299273
#>  [90,] -0.76030153  0.970829138 -0.3259652532  0.535048407  0.092065158
#>  [91,]  0.24568151 -0.626938689  0.7971648648  0.299265414  0.319030149
#>  [92,] -0.73352292  0.876135339  0.9901817953 -0.230197167 -0.996188703
#>  [93,] -0.64434697 -0.773665956  0.1147483303  0.770184610 -0.517349281
#>  [94,]  0.85179146 -0.021921838 -0.5841782563  0.875011167  0.114505538
#>  [95,] -0.45683451 -0.419605738 -0.7959583676  0.502992543 -0.998470625
#>  [96,] -0.26548629  0.610673508  0.4932367723 -0.656175084  0.022544647
#>  [97,] -0.06945582 -0.628386649 -0.9196738042 -0.061953278  0.424108676
#>  [98,] -0.89496685  0.625080413  0.8138785898  0.997172973 -0.836399701
#>  [99,] -0.60766121 -0.534864285  0.2476455411 -0.293067105 -0.515410341
#> [100,]  0.11546590 -0.376731568  0.7246271507 -0.174797953 -0.262913107
#> 
#> $y
#>   [1] 1 1 0 0 1 0 0 0 0 0 0 0 0 1 0 1 0 1 0 1 1 1 1 1 1 0 0 1 0 1 1 0 1 0 0 1 1
#>  [38] 0 1 1 1 1 1 0 0 1 0 1 0 0 1 0 1 0 0 1 0 1 0 1 1 0 1 1 1 1 1 1 1 1 0 0 0 0
#>  [75] 0 0 1 1 1 0 0 0 0 0 0 1 1 0 1 0 1 0 0 0 1 1 1 0 0 1
#>