Simulate Data from Hyper-Sphere for Classification Problems

The classification problem Circle is described in the BARK paper (2008). Inputs are dim independent variables uniformly distributed on the interval \([-1,1]\), only the first 2 out of these dim are actually signals. Outputs are created according to the formula $$y = 1(x1^2+x2^2 \le 2/\pi)$$

Usage

sim_circle(n, dim = 5)

Arguments

n

number of data points to generate

dim

number of dimension of the problem, no less than 2

Value

Returns a list with components

x

input values (independent variables)

y

0/1 output values (dependent variable)

References

Ouyang, Zhi (2008) Bayesian Additive Regression Kernels. Duke University. PhD dissertation, Chapter 3.

See also

Examples

sim_circle(n=100, dim=5)
#> $x
#>               [,1]        [,2]        [,3]         [,4]         [,5]
#>   [1,]  0.91843343 -0.17549004  0.26775240 -0.560909710 -0.615281755
#>   [2,] -0.72927028  0.99364273 -0.96918784  0.435361659 -0.583179305
#>   [3,] -0.42428061 -0.66576054  0.80498427 -0.578158492  0.635914512
#>   [4,]  0.88329991 -0.64993553  0.34949959  0.979426938 -0.416681993
#>   [5,] -0.20904030  0.45941593  0.22598097 -0.952268171  0.224367901
#>   [6,] -0.49333808 -0.37263470  0.04795585 -0.798653341 -0.833513015
#>   [7,] -0.64140235  0.27286072 -0.66153821 -0.212970133 -0.678411244
#>   [8,]  0.23785492  0.45759954  0.35452514 -0.756130554 -0.067541667
#>   [9,]  0.04410471 -0.93191180 -0.65569945  0.987522995 -0.065193992
#>  [10,] -0.21702943  0.40188162  0.26092876  0.232949946 -0.051217161
#>  [11,]  0.19454488 -0.53577959  0.73975340 -0.177840598 -0.179879099
#>  [12,]  0.29384949  0.57144356 -0.40363420 -0.498411116  0.793095689
#>  [13,] -0.75485053 -0.88770641 -0.56218032 -0.007206569 -0.701283623
#>  [14,] -0.60985227 -0.65395053  0.55619537  0.724597048 -0.112579977
#>  [15,] -0.49626589 -0.97613022  0.10607340 -0.240367529  0.954185539
#>  [16,]  0.29986806  0.29584910  0.18449408 -0.034186450  0.321453730
#>  [17,] -0.92383312  0.79566543  0.83296681  0.381372971  0.764473893
#>  [18,]  0.85255176 -0.20470079  0.05322193  0.978168361 -0.025226593
#>  [19,]  0.47573814  0.51187419 -0.59117991 -0.615000834 -0.987515695
#>  [20,]  0.13639372  0.88299826  0.52296525 -0.791172389 -0.080411196
#>  [21,]  0.11788781 -0.06986068 -0.74938793  0.401848310 -0.395151611
#>  [22,] -0.38750414  0.59753547  0.24821988  0.353915580  0.196869440
#>  [23,] -0.46642473 -0.59451300  0.36552345 -0.417871337 -0.405256490
#>  [24,] -0.25494325  0.20964672  0.49781216  0.331566128 -0.354040692
#>  [25,] -0.03306799  0.18964303  0.94833039  0.144519388  0.554182288
#>  [26,]  0.63218821  0.01258494 -0.83137147  0.882098619 -0.813216541
#>  [27,]  0.48638747  0.84681806  0.72668050  0.555770670 -0.314936464
#>  [28,]  0.80009791 -0.87256609 -0.70121538  0.748136757  0.512586236
#>  [29,]  0.90735516  0.48842143  0.78152553  0.709820488  0.675227544
#>  [30,]  0.19366551 -0.07809131  0.83485002  0.482336243  0.004250372
#>  [31,]  0.62266408 -0.87154463  0.82161038  0.189176497  0.431541667
#>  [32,] -0.53971787 -0.68602581 -0.25016921 -0.382764361  0.527761141
#>  [33,]  0.84860019  0.59201190  0.20939974 -0.240695367  0.370952792
#>  [34,] -0.75013191 -0.33236499 -0.95938409  0.045534790 -0.679283059
#>  [35,] -0.60396573  0.34126034  0.52042874  0.672349569 -0.063468072
#>  [36,] -0.53379221 -0.73698684  0.14553512 -0.152989676 -0.821273172
#>  [37,]  0.63159061 -0.90020104 -0.63712166 -0.054143530 -0.667788294
#>  [38,] -0.02697658  0.23911898 -0.50250939  0.507715065  0.861504390
#>  [39,]  0.33228841  0.28403639 -0.88726767 -0.822031033  0.998818747
#>  [40,] -0.21176129  0.63376316 -0.53295711 -0.390947135 -0.449675115
#>  [41,] -0.89136163  0.70919377  0.17098622  0.962575166  0.574769084
#>  [42,] -0.91717428  0.11793194  0.30223664 -0.204668147  0.476441576
#>  [43,]  0.60636038 -0.79205036  0.85891717  0.868807842  0.614483239
#>  [44,] -0.23481972  0.75205641  0.61888362  0.081925420  0.139553453
#>  [45,] -0.46630953 -0.61942676 -0.36920837  0.791231044 -0.326853756
#>  [46,]  0.89039326 -0.05245484  0.66458771  0.446631633  0.393915524
#>  [47,] -0.14880477 -0.15483603 -0.33706378 -0.553289720  0.389135973
#>  [48,]  0.70872281  0.92451241 -0.79339498  0.180277649 -0.573446885
#>  [49,] -0.08827067  0.42044886 -0.24098786  0.301288081 -0.509392892
#>  [50,] -0.18143647  0.87816841  0.42974339 -0.264522378  0.944579952
#>  [51,] -0.22181165  0.53653680 -0.71664487  0.095208941 -0.703519573
#>  [52,] -0.45520612 -0.52371997 -0.03248673 -0.602395777  0.032673484
#>  [53,]  0.01624503 -0.18557867 -0.74968076 -0.892174683 -0.312280835
#>  [54,]  0.18754702  0.50498905  0.60009677  0.198983842  0.384549576
#>  [55,] -0.99108805  0.22248421  0.96521816 -0.689542463 -0.755573341
#>  [56,]  0.19369146  0.99765686 -0.07620308 -0.306729440 -0.537373300
#>  [57,]  0.61545379 -0.37067416 -0.46312240 -0.670789543  0.921121430
#>  [58,]  0.27668887  0.32108986 -0.16305071 -0.231537400  0.780487311
#>  [59,]  0.98628229 -0.67705043 -0.66820225 -0.787395400  0.814257351
#>  [60,]  0.60725061  0.16057533  0.46619022 -0.835427243 -0.348971325
#>  [61,] -0.50095151 -0.98792496 -0.99304195  0.557289187 -0.594719839
#>  [62,]  0.28294308  0.30765679  0.59522502  0.723182773  0.963736458
#>  [63,] -0.71176825 -0.43436723 -0.75289999  0.576941528 -0.711758057
#>  [64,] -0.80804539 -0.91298884  0.16180755  0.728828456  0.812068255
#>  [65,] -0.59060514 -0.18014195 -0.27343164 -0.387662541  0.291236022
#>  [66,]  0.93548509 -0.73521790  0.77362933  0.060126062 -0.626460283
#>  [67,]  0.74429310 -0.90197142 -0.08501866 -0.838971830  0.006283067
#>  [68,]  0.90174266  0.83152170  0.81604280  0.375198704 -0.254047028
#>  [69,] -0.52592292 -0.23736254  0.09059860 -0.616159401 -0.667980512
#>  [70,] -0.85153310  0.15455320 -0.83687932 -0.297880578  0.881348351
#>  [71,] -0.15079389  0.68252765 -0.38505239 -0.769397425  0.836856402
#>  [72,]  0.73276438 -0.68474494  0.40693771 -0.934106356 -0.794557399
#>  [73,]  0.76209235 -0.75466460  0.02571994  0.373483558  0.544459018
#>  [74,] -0.90629702 -0.17543503  0.41859139 -0.010346361 -0.410562651
#>  [75,] -0.69788679 -0.26291698  0.34912952 -0.267882247  0.827549649
#>  [76,] -0.10225260 -0.48353885 -0.03654442  0.958524922 -0.813998032
#>  [77,] -0.58520465  0.34443194  0.61267725 -0.238503996  0.252386494
#>  [78,] -0.60845259 -0.51728839  0.69038300 -0.359544673 -0.216097713
#>  [79,]  0.70634924  0.86102518  0.22986060 -0.215115678  0.959811726
#>  [80,]  0.67433895 -0.19349271 -0.30718515 -0.517594507 -0.132428420
#>  [81,] -0.42155759  0.45857430 -0.47650076  0.119470086  0.014628698
#>  [82,]  0.32037903 -0.37101947  0.33801463  0.283240896  0.870283147
#>  [83,] -0.95243226  0.11978339  0.97495780  0.634144238  0.262107789
#>  [84,]  0.84890243  0.38469971 -0.42914363 -0.250287994  0.434003980
#>  [85,]  0.45815999  0.70406155  0.70204578 -0.104171268 -0.239569434
#>  [86,]  0.07234656  0.87208093  0.57525641 -0.411198253 -0.376479591
#>  [87,] -0.53468794  0.50930676  0.76426563 -0.808053702  0.139195002
#>  [88,]  0.01644966 -0.34322894 -0.96941017  0.271473038  0.834050563
#>  [89,] -0.95993225 -0.71939200  0.17480699  0.968473809 -0.438298128
#>  [90,] -0.67233531  0.96844219 -0.54737444 -0.998911223 -0.983490356
#>  [91,]  0.47474110  0.93786408  0.57518601 -0.810417015 -0.417695958
#>  [92,] -0.04364886 -0.56577853 -0.77847610 -0.135194524 -0.997163476
#>  [93,]  0.44387487 -0.95897845 -0.55608378 -0.570268367 -0.478019142
#>  [94,]  0.45041346  0.44601926  0.79598870  0.836598378 -0.123308545
#>  [95,] -0.39978251  0.78501735 -0.24770705  0.516770183 -0.846402213
#>  [96,] -0.41122364 -0.21943333 -0.56990402 -0.149013077 -0.370326575
#>  [97,] -0.44183604  0.89949844 -0.84628855  0.077946524  0.535609904
#>  [98,]  0.45716547 -0.27257741 -0.98957948 -0.534120802  0.425844240
#>  [99,]  0.32404397 -0.97222565  0.40259518  0.209305640  0.050419342
#> [100,] -0.14252757 -0.27862392  0.15914796  0.358037928  0.965537463
#> 
#> $y
#>   [1] 0 0 1 0 1 1 1 1 0 1 1 1 0 0 0 1 0 0 1 0 1 1 1 1 1 1 0 0 0 1 0 0 0 0 1 0 0
#>  [38] 1 1 1 0 0 0 1 1 0 1 0 1 0 1 1 1 1 0 0 1 1 0 1 0 1 0 0 1 0 0 0 1 0 1 0 0 0
#>  [75] 1 1 1 0 0 1 1 1 0 0 0 0 1 1 0 0 0 1 0 1 0 1 0 1 0 1
#>