Simulate Data from Hyper-Sphere for Classification Problems

The classification problem Circle is described in the BARK paper (2008). Inputs are dim independent variables uniformly distributed on the interval \([-1,1]\), only the first 2 out of these dim are actually signals. Outputs are created according to the formula $$y = 1(x1^2+x2^2 \le 2/\pi)$$

sim_circle(n, dim = 5)

Arguments

n

number of data points to generate

dim

number of dimension of the problem, no less than 2

Value

Returns a list with components

x

input values (independent variables)

y

0/1 output values (dependent variable)

References

Ouyang, Zhi (2008) Bayesian Additive Regression Kernels. Duke University. PhD dissertation, Chapter 3.
https://stat.duke.edu/people/theses/OuyangZ.html

See also

Examples

sim_circle(n=100, dim=5)
#> $x
#>                [,1]         [,2]         [,3]         [,4]        [,5]
#>   [1,]  0.709108983  0.528359063 -0.528958700 -0.491278590  0.63282197
#>   [2,] -0.124496409 -0.516410785 -0.077847349  0.262434544 -0.77839377
#>   [3,] -0.592336299 -0.263793047  0.391728938  0.720276494  0.64110722
#>   [4,] -0.515512578  0.305278086  0.157190769 -0.679606960  0.14428568
#>   [5,]  0.535129243 -0.733150472 -0.804848907 -0.458113952  0.68962540
#>   [6,] -0.982496300  0.395241055 -0.203896594  0.259432773 -0.65121274
#>   [7,]  0.236304810  0.224052361 -0.914837660 -0.418874547  0.70397057
#>   [8,] -0.498015387  0.114336621 -0.933822666  0.722910172  0.01436877
#>   [9,] -0.532467959 -0.469545361 -0.131755403  0.806999755 -0.34296350
#>  [10,]  0.300113371 -0.632665376 -0.549821792 -0.017596494  0.24646982
#>  [11,] -0.475924509 -0.720357546  0.844850751 -0.004336456  0.13881741
#>  [12,] -0.573925892  0.930637914 -0.590233310  0.483751698 -0.20869100
#>  [13,]  0.296173754 -0.586077010 -0.390979146  0.121736721 -0.62550047
#>  [14,]  0.084365949 -0.908992577  0.201689814  0.001313308 -0.33841365
#>  [15,]  0.113875701  0.989840410  0.312034921  0.073101444  0.18130527
#>  [16,]  0.706717419  0.306043977 -0.006886898  0.845125657  0.21175574
#>  [17,]  0.545941826 -0.250474286 -0.722498061  0.528333958  0.05682295
#>  [18,] -0.524163820  0.321306188 -0.493068400  0.059044397  0.93978967
#>  [19,] -0.565935505  0.796650705 -0.156790704 -0.153531739  0.44656478
#>  [20,] -0.890572167  0.008384283  0.386426094 -0.961848303  0.04928801
#>  [21,]  0.933782930 -0.007569815 -0.230876384 -0.326629230  0.72044909
#>  [22,] -0.354851425  0.220037262  0.123782092 -0.664803885  0.05318884
#>  [23,]  0.652668890 -0.742676416 -0.510773839 -0.928884957  0.58529657
#>  [24,] -0.751803599  0.341870583 -0.696233054  0.894952048  0.78896517
#>  [25,] -0.085224856 -0.679328342 -0.124826557 -0.013191528 -0.81361882
#>  [26,] -0.339420574  0.937366067 -0.755006354  0.413283302  0.91582277
#>  [27,]  0.098285379  0.318430649 -0.228610185  0.691285318  0.40245193
#>  [28,] -0.489370493  0.317157235  0.005216324  0.239179641 -0.33737477
#>  [29,]  0.633778670  0.165733353 -0.953905716  0.832049305  0.55410626
#>  [30,] -0.471289170  0.130206056  0.290779161  0.257917221 -0.85668775
#>  [31,]  0.022236709 -0.083195080  0.919752718  0.730122351  0.12541755
#>  [32,] -0.131214191  0.660273582 -0.503712991  0.781497665 -0.59506696
#>  [33,] -0.593079040 -0.116245442  0.353849808  0.752145742 -0.52673440
#>  [34,] -0.971323277 -0.970224486  0.373533084 -0.529635480 -0.51633215
#>  [35,] -0.912014178  0.374381750  0.268545240 -0.295658290 -0.92240179
#>  [36,]  0.267372699  0.021885564  0.484628900 -0.757601345 -0.36634673
#>  [37,]  0.172838842 -0.612373803  0.888209231  0.946126842 -0.67378776
#>  [38,]  0.911252038 -0.854439110  0.141104727 -0.076833547 -0.84428571
#>  [39,]  0.371220693  0.163685067  0.549853806 -0.047150352 -0.84219860
#>  [40,] -0.474831055  0.518150160  0.570822565  0.253137339  0.36137421
#>  [41,]  0.739517108  0.056402801 -0.857411757 -0.251495757 -0.10841926
#>  [42,] -0.643930359  0.273860545  0.601917119 -0.668852315  0.90832054
#>  [43,] -0.562242912  0.804574385  0.120167125  0.819164629 -0.68419889
#>  [44,]  0.904679820  0.037413958  0.662392046  0.518821430 -0.07719229
#>  [45,]  0.331598936 -0.320249566  0.339688612  0.914235070 -0.86304811
#>  [46,] -0.865444913 -0.238196983  0.260426021 -0.531195723  0.76315144
#>  [47,] -0.509482407 -0.261026792  0.659571801 -0.898036315  0.59506693
#>  [48,] -0.999561674 -0.394076830 -0.633799382  0.383108695  0.67544312
#>  [49,] -0.406913373 -0.796418267 -0.860079826  0.045196965  0.24721874
#>  [50,]  0.858300189  0.795736135 -0.417030129  0.413268816 -0.53575944
#>  [51,]  0.318483374  0.945257943 -0.377015528 -0.620638883  0.16511963
#>  [52,]  0.713454787  0.173025564  0.475781822  0.456066797  0.22817949
#>  [53,] -0.251178298 -0.368227585 -0.050694361  0.689693273  0.83528042
#>  [54,] -0.970620147 -0.521729957  0.485663266 -0.249315628  0.58037967
#>  [55,]  0.841252324 -0.879696418 -0.038506515 -0.982243821  0.19595149
#>  [56,]  0.457341136  0.049007189 -0.659792157 -0.660642650  0.42625777
#>  [57,]  0.569011316  0.228334425  0.948542508 -0.967750764  0.58578367
#>  [58,] -0.823965773  0.315058777 -0.522359233  0.572609677  0.63777969
#>  [59,] -0.289079416  0.739325922 -0.895277038 -0.385195455  0.10620018
#>  [60,]  0.878510357  0.772502535  0.156912095 -0.837056931  0.54073940
#>  [61,]  0.478511470  0.086716947 -0.102536961 -0.585036261  0.12940548
#>  [62,] -0.500856379 -0.678099441 -0.557393325  0.569596824  0.86477733
#>  [63,]  0.414133816  0.657391029 -0.284397472 -0.782207700  0.86476389
#>  [64,] -0.123322928 -0.520145114  0.939540134  0.050789317 -0.15145423
#>  [65,] -0.605290989  0.863141073  0.259638674 -0.259475309 -0.53364654
#>  [66,]  0.840267144 -0.098821212 -0.170249245  0.583377372 -0.71917158
#>  [67,]  0.629132197  0.713818643 -0.313476222 -0.838352480  0.31952388
#>  [68,] -0.603987171 -0.200211613 -0.814346561 -0.031887752 -0.59830562
#>  [69,] -0.276961290  0.964121838  0.816443102  0.407585024 -0.92296755
#>  [70,] -0.174444090 -0.483569821 -0.439592089  0.552512757 -0.11492134
#>  [71,] -0.473530764 -0.876243760  0.907077206 -0.194704912  0.75221768
#>  [72,]  0.795708500 -0.308230324 -0.186262196 -0.110539275  0.89376276
#>  [73,] -0.139232850 -0.825177206 -0.834912325 -0.498267013  0.92849657
#>  [74,] -0.608009275 -0.663627521 -0.180999560  0.675287893 -0.47721875
#>  [75,] -0.843651821  0.662232046  0.188488544  0.015597606  0.93311995
#>  [76,] -0.794244312 -0.209867094  0.465195732 -0.707699496  0.15572920
#>  [77,]  0.695915972  0.244090765  0.627529853  0.127767100  0.58372053
#>  [78,] -0.615508923 -0.533804022  0.876031701  0.374587732  0.53086713
#>  [79,]  0.378942075  0.580531001 -0.738985009  0.377628199 -0.42467538
#>  [80,] -0.009089844  0.581940623 -0.301687279  0.079532959 -0.21600586
#>  [81,] -0.790962445 -0.992902216 -0.273369516  0.916384315  0.13595223
#>  [82,] -0.571636997  0.953909034  0.087430425 -0.613485543 -0.41321461
#>  [83,] -0.661675855 -0.282255393  0.571068922  0.608206205  0.56940519
#>  [84,]  0.490866931  0.537734149  0.679808095  0.788212139 -0.07572422
#>  [85,]  0.518527422 -0.573800957 -0.294462517 -0.989185933  0.66448463
#>  [86,]  0.854423640  0.668275305  0.593134791  0.379348488  0.27347973
#>  [87,] -0.368039584 -0.209724353  0.845505851  0.180719859 -0.36614139
#>  [88,] -0.574443029  0.741254216  0.148553077 -0.221155625  0.92172841
#>  [89,]  0.155209798  0.424655241 -0.640815846 -0.910671226 -0.44067308
#>  [90,] -0.349771371 -0.880968447  0.880440131 -0.154732093  0.62307051
#>  [91,]  0.038372400 -0.732455461 -0.182366149  0.937067854 -0.88337777
#>  [92,] -0.485947359  0.980514207  0.050736799 -0.634451908 -0.49329701
#>  [93,] -0.297514388 -0.377304536 -0.255029332 -0.126715049 -0.88196875
#>  [94,]  0.300278547 -0.854136241  0.207356831 -0.078719169  0.52734181
#>  [95,] -0.240325488  0.102803659 -0.550729049  0.770978951  0.76993444
#>  [96,]  0.358652317  0.679175789 -0.504708796 -0.999510945  0.13996206
#>  [97,]  0.912193300 -0.286439096 -0.665640881  0.010089048  0.25100653
#>  [98,]  0.084827906  0.004542241  0.892429793  0.610725570 -0.38840755
#>  [99,]  0.096250471  0.571301910  0.398657411  0.852671869 -0.92394371
#> [100,] -0.441608179  0.813023239 -0.985834030 -0.492888881 -0.23385880
#> 
#> $y
#>   [1] 0 1 1 1 0 0 1 1 1 1 0 0 1 0 0 1 1 1 0 0 0 1 0 0 1 0 1 1 1 1 1 1 1 0 0 1 1
#>  [38] 0 1 1 1 1 0 0 1 0 1 0 0 0 0 1 1 0 0 1 1 0 1 0 1 0 1 1 0 0 0 1 0 1 0 0 0 0
#>  [75] 0 0 1 0 1 1 0 0 1 1 1 0 1 0 1 0 1 0 1 0 1 1 0 1 1 0
#>