Simulate Data from Hyper-Sphere for Classification Problems

The classification problem Circle is described in the BARK paper (2008). Inputs are dim independent variables uniformly distributed on the interval \([-1,1]\), only the first 2 out of these dim are actually signals. Outputs are created according to the formula $$y = 1(x1^2+x2^2 \le 2/\pi)$$

sim_circle(n, dim = 5)

Arguments

n

number of data points to generate

dim

number of dimension of the problem, no less than 2

Value

Returns a list with components

x

input values (independent variables)

y

0/1 output values (dependent variable)

References

Ouyang, Zhi (2008) Bayesian Additive Regression Kernels. Duke University. PhD dissertation, Chapter 3.
http://stat.duke.edu/people/theses/OuyangZ.html

See also

Examples

sim_circle(n=100, dim=5)
#> $x
#>                [,1]        [,2]         [,3]         [,4]        [,5]
#>   [1,] -0.903870597 -0.72488441  0.041408977  0.638233029 -0.35999122
#>   [2,] -0.473902013  0.06471180 -0.762146213  0.522262662 -0.30247015
#>   [3,] -0.672134912  0.92828923 -0.616455573  0.870242299  0.08951725
#>   [4,] -0.898478482  0.44228511 -0.004999617 -0.666564291 -0.38275109
#>   [5,] -0.043445881 -0.51402154 -0.888817935  0.213163124  0.36425546
#>   [6,] -0.771767595  0.87757967  0.599534723 -0.535555031 -0.23231771
#>   [7,]  0.670691049 -0.28288767  0.651939374 -0.762995093 -0.76690431
#>   [8,]  0.817648483  0.11923913  0.121631943 -0.528126076  0.40521198
#>   [9,] -0.629469682  0.74805145  0.174939814  0.064901395 -0.15465494
#>  [10,]  0.061758144  0.84694074 -0.926411747  0.305820588  0.35579633
#>  [11,]  0.213785547 -0.37198032  0.565229349 -0.495145053  0.70655005
#>  [12,] -0.857410147  0.31585160 -0.834709110  0.299471642 -0.61617973
#>  [13,]  0.010947396 -0.46375101  0.747583429 -0.833298628  0.70091605
#>  [14,]  0.357063535 -0.34003575  0.220240597  0.344982656  0.39381389
#>  [15,] -0.635023053  0.82082002  0.395837619  0.358133085 -0.80965490
#>  [16,]  0.099700513 -0.22723357  0.604476007 -0.018370025  0.07421492
#>  [17,] -0.326553353 -0.70527903 -0.952935945 -0.852313682  0.02687623
#>  [18,] -0.509730099  0.12209859  0.502462027 -0.628681134 -0.96444393
#>  [19,] -0.434085196  0.52256945  0.197778788 -0.773721431  0.37614290
#>  [20,] -0.662016027  0.65039414  0.385878406  0.511040718  0.69043165
#>  [21,]  0.031523955  0.69356625 -0.065433982 -0.670678257  0.53570974
#>  [22,] -0.937628406  0.02656579 -0.579710620 -0.577389482  0.09069199
#>  [23,]  0.704659882 -0.44012245 -0.861503182  0.805634686  0.35973609
#>  [24,] -0.118490362  0.37456188  0.882530201 -0.120328997 -0.71764314
#>  [25,]  0.577334344 -0.79682990  0.555230932  0.339668775 -0.97720848
#>  [26,] -0.214551936  0.52828129 -0.366593353 -0.386684969  0.96142122
#>  [27,]  0.730924046 -0.77867530 -0.303509100  0.044052642 -0.62766152
#>  [28,] -0.861861397 -0.48967448 -0.846437872 -0.049679411  0.85650643
#>  [29,]  0.503225290 -0.82920747 -0.025460097 -0.430349616 -0.22920811
#>  [30,] -0.368132742 -0.18383277  0.333839968  0.863981910 -0.66173012
#>  [31,]  0.212462661  0.98071632 -0.979555027 -0.226694132  0.40305942
#>  [32,]  0.974526442  0.03106338  0.463602304  0.671876498  0.32946438
#>  [33,]  0.950172679  0.19220103  0.223896216  0.964527773 -0.19130647
#>  [34,] -0.326254882  0.84961319 -0.749826503  0.850598817  0.79083645
#>  [35,] -0.898915823  0.45207141  0.725032802 -0.736677036 -0.69302542
#>  [36,] -0.289778149 -0.17281658 -0.314958762 -0.042814091 -0.72743613
#>  [37,] -0.755524349  0.38745604  0.036673758  0.338493717  0.73140759
#>  [38,] -0.663812301  0.27235812 -0.677178168  0.883869660 -0.22221333
#>  [39,] -0.004536393 -0.03370005 -0.869423793  0.016357271 -0.67811587
#>  [40,] -0.146272324  0.24649056 -0.214359227  0.010370262 -0.52420971
#>  [41,] -0.266240053 -0.47813319  0.487701132 -0.260792591 -0.50387002
#>  [42,]  0.879047120 -0.35475574 -0.759883752  0.756269480 -0.75317375
#>  [43,] -0.451438389  0.70988367  0.398717322 -0.540523802  0.74131344
#>  [44,] -0.960570610  0.89128125  0.809733575 -0.113241795 -0.84746002
#>  [45,] -0.488581460 -0.64444387  0.994562359  0.534939606  0.87158315
#>  [46,] -0.102124879 -0.16382383 -0.235145590 -0.140994507  0.34406563
#>  [47,] -0.003213482  0.14241760  0.550632026 -0.682871437 -0.14807646
#>  [48,]  0.486864377  0.70544703  0.565528404  0.581148663 -0.06600971
#>  [49,]  0.289719689  0.58722526  0.729645211 -0.109571196  0.90771154
#>  [50,] -0.959923729  0.70735725 -0.456990481  0.623641514 -0.32476661
#>  [51,]  0.430952919 -0.43241362 -0.898672713  0.308735570  0.95466866
#>  [52,]  0.438112156  0.93479927  0.058838190 -0.261565840  0.18445964
#>  [53,]  0.751384119 -0.32693869  0.472818248  0.470202074 -0.17567139
#>  [54,]  0.022691421 -0.15516680  0.251168752  0.476977660 -0.68206684
#>  [55,] -0.926540437 -0.53194582 -0.085220897 -0.949579691  0.85876971
#>  [56,]  0.136738020 -0.17993442 -0.867127966 -0.258846846 -0.02536199
#>  [57,] -0.523779177 -0.85702943  0.525529621 -0.617954853  0.17145691
#>  [58,]  0.777264111  0.41234284 -0.316665999  0.687162339 -0.27592816
#>  [59,]  0.601235780  0.95900029  0.719611909 -0.260925828 -0.10791023
#>  [60,] -0.092705993  0.07363533 -0.477262944  0.001826332  0.37481406
#>  [61,] -0.641185321  0.34821089  0.632722662 -0.605504131  0.26875657
#>  [62,] -0.629553910  0.69360772  0.140805549 -0.968517512 -0.47241319
#>  [63,] -0.306998966 -0.79268265 -0.871978824  0.666457097  0.91219646
#>  [64,]  0.550686546 -0.81029852  0.563780858 -0.482845777 -0.70277330
#>  [65,]  0.618228326  0.22457045  0.164019385  0.539599616  0.72440727
#>  [66,]  0.668095050 -0.59282293 -0.890989188  0.953966234  0.73135371
#>  [67,] -0.584644268 -0.84306212  0.713462716  0.577332997 -0.62140399
#>  [68,] -0.490868205 -0.29969165  0.291496088  0.447407471  0.30954184
#>  [69,] -0.927545932 -0.48641361 -0.585466784 -0.736143069 -0.20172694
#>  [70,]  0.121424472 -0.44279596  0.772066517 -0.688282604 -0.86246127
#>  [71,]  0.752410963  0.08755480 -0.374303687 -0.628272498 -0.84383787
#>  [72,] -0.912923768 -0.03500527 -0.076265545 -0.198720692 -0.06149833
#>  [73,]  0.533838875  0.05521995  0.248433907 -0.532927996  0.19919495
#>  [74,] -0.178090484  0.29393706 -0.633923119  0.142063430  0.98078141
#>  [75,] -0.776409845  0.26543627 -0.765467293  0.657750306 -0.34948909
#>  [76,]  0.166771561 -0.08333687  0.010734920 -0.149403007  0.60310191
#>  [77,]  0.761675897 -0.68853717  0.420215094 -0.998526062  0.11845757
#>  [78,] -0.428526201 -0.79137910 -0.513766340  0.681999426  0.80946055
#>  [79,] -0.265618580  0.10394141  0.148974385 -0.001388798  0.73730848
#>  [80,]  0.127304579 -0.93975003 -0.820842525 -0.998957516 -0.63965831
#>  [81,]  0.676222386  0.46321794 -0.825306715 -0.670430459 -0.66638157
#>  [82,] -0.428670902  0.57526939 -0.984652677  0.247797964 -0.20277537
#>  [83,]  0.546774988  0.92178846  0.759947413 -0.775119854 -0.98428934
#>  [84,]  0.964466599 -0.53385116  0.196317563 -0.083957595  0.15165284
#>  [85,]  0.885625782 -0.55652241 -0.809101896  0.025900683  0.37281974
#>  [86,] -0.409258903 -0.47566331  0.984760307  0.704808705  0.66121895
#>  [87,]  0.144392678 -0.65312349  0.565894277 -0.912879884  0.87022342
#>  [88,]  0.606522700 -0.08492576  0.987403929  0.152580771 -0.83172681
#>  [89,] -0.953671606  0.41556816 -0.446708231  0.551910896  0.16515349
#>  [90,]  0.891859935  0.14501650  0.138346116 -0.106719050 -0.86599399
#>  [91,] -0.145684208 -0.26854517  0.142606307 -0.748677915 -0.21877647
#>  [92,]  0.198876648 -0.13840165 -0.905313753  0.734178809 -0.34278282
#>  [93,] -0.621395519  0.75954925 -0.127113791 -0.312519046 -0.67971435
#>  [94,]  0.322796046 -0.73295742 -0.614159197 -0.787297537  0.15160834
#>  [95,]  0.744491238 -0.61472873  0.063120827  0.093701935  0.38376280
#>  [96,]  0.776742505 -0.99031154  0.755816804  0.905697287  0.74243246
#>  [97,] -0.089017581 -0.91752500 -0.104016402  0.197420830 -0.51637509
#>  [98,]  0.093621624 -0.30648797  0.493139737  0.725929912 -0.16033538
#>  [99,] -0.158269907 -0.06038185 -0.454422828 -0.788889301  0.09924278
#> [100,]  0.285359382 -0.87712642  0.226794700  0.073373365 -0.19557275
#> 
#> $y
#>   [1] 0 1 0 0 1 0 1 0 0 0 1 0 1 1 0 1 1 1 1 0 1 0 0 1 0 1 0 0 0 1 0 0 0 0 0 1 0
#>  [38] 1 1 1 1 0 0 0 0 1 1 0 1 0 1 0 0 1 0 1 0 0 0 1 1 0 0 0 1 0 0 1 0 1 1 0 1 1
#>  [75] 0 1 0 0 1 0 0 1 0 0 0 1 1 1 0 0 1 1 0 0 0 0 0 1 1 0
#>