Simulate Data from Hyper-Sphere for Classification Problems

The classification problem Circle is described in the BARK paper (2008). Inputs are dim independent variables uniformly distributed on the interval \([-1,1]\), only the first 2 out of these dim are actually signals. Outputs are created according to the formula $$y = 1(x1^2+x2^2 \le 2/\pi)$$

Usage

sim_circle(n, dim = 5)

Arguments

n

number of data points to generate

dim

number of dimension of the problem, no less than 2

Value

Returns a list with components

x

input values (independent variables)

y

0/1 output values (dependent variable)

References

Ouyang, Zhi (2008) Bayesian Additive Regression Kernels. Duke University. PhD dissertation, Chapter 3.

See also

Examples

sim_circle(n=100, dim=5)
#> $x
#>               [,1]         [,2]         [,3]        [,4]         [,5]
#>   [1,] -0.62974289  0.839709125  0.485366106 -0.12485203  0.610535549
#>   [2,] -0.81385731 -0.174827380 -0.275302324 -0.88444633 -0.005455901
#>   [3,] -0.98770709  0.193278589 -0.083845105 -0.39242465 -0.910141401
#>   [4,] -0.95197990  0.457060360 -0.693583555  0.63782449  0.762102404
#>   [5,] -0.08786214 -0.701273591  0.742700766 -0.10859597 -0.478392066
#>   [6,]  0.06250959  0.682013776 -0.122681328  0.42875542 -0.992243251
#>   [7,] -0.85024256  0.776700180  0.592577400  0.69275672  0.541956111
#>   [8,] -0.36803559  0.203425012  0.254864569 -0.30420190  0.076207181
#>   [9,]  0.04567539 -0.796954918 -0.705548527 -0.31025712 -0.926347923
#>  [10,]  0.42751067  0.277585288 -0.197612560  0.03150315  0.069468368
#>  [11,]  0.34924394 -0.635811255 -0.311338808 -0.94723345 -0.088767354
#>  [12,]  0.37931466 -0.260980940 -0.995424050  0.83614122 -0.960105072
#>  [13,]  0.59178328  0.204192775 -0.265220952  0.64568407 -0.855394950
#>  [14,]  0.99265368 -0.833104846 -0.100258630  0.70066621 -0.849895662
#>  [15,] -0.96030673 -0.931495276  0.763454776  0.89968986  0.031503583
#>  [16,] -0.74314150  0.264486513 -0.033757094 -0.72050369  0.581052647
#>  [17,]  0.29759985 -0.452704745 -0.240179487  0.94171656  0.654019829
#>  [18,]  0.96668510  0.133600387 -0.103741409  0.85764336  0.872514458
#>  [19,] -0.18481836  0.175822335 -0.020835382  0.86149034 -0.404871368
#>  [20,]  0.97228041 -0.479736045 -0.079944479 -0.84137577  0.242634831
#>  [21,]  0.08275798  0.850725217 -0.121270020  0.52413569  0.010298313
#>  [22,] -0.01369214 -0.661336840  0.341289192 -0.57063119 -0.154289692
#>  [23,]  0.25748800  0.576070780 -0.754461209 -0.74402447 -0.561855047
#>  [24,] -0.07023769  0.316642346 -0.198690111  0.75716178  0.721200577
#>  [25,] -0.50567146 -0.536144511 -0.877849812 -0.10208263  0.155346901
#>  [26,] -0.66613765  0.877577759 -0.950408851 -0.51816871 -0.720582974
#>  [27,]  0.15797801  0.358466540 -0.464749931  0.69770996 -0.741721117
#>  [28,] -0.30923730  0.615478652  0.022610596 -0.90680890  0.112751968
#>  [29,]  0.41165692  0.769626840 -0.447988150  0.45093328  0.371069355
#>  [30,] -0.96975683  0.930973032 -0.597522093  0.70298526 -0.144844473
#>  [31,] -0.64183378  0.140875044  0.689228225  0.64954875 -0.696268000
#>  [32,] -0.50378435 -0.050845818  0.813880720 -0.18676259 -0.683351557
#>  [33,]  0.53382187 -0.681417520  0.727965984  0.71361653 -0.034070543
#>  [34,]  0.77857999 -0.088849121 -0.233312835 -0.50805713 -0.044626863
#>  [35,] -0.22972040  0.603849848 -0.046689169 -0.73422469 -0.627032796
#>  [36,] -0.26912069 -0.124262758 -0.361054407 -0.32739950  0.616690119
#>  [37,] -0.71288197  0.343232382 -0.071232668  0.57363936  0.479177514
#>  [38,]  0.23669322  0.810807073  0.562935361 -0.96767933  0.665481686
#>  [39,]  0.18655053  0.724314221  0.340282714 -0.12618202  0.091977316
#>  [40,]  0.61231225 -0.998209632 -0.556177504 -0.56209572  0.461847899
#>  [41,]  0.40373139  0.209199035 -0.558868290  0.82578656  0.957233978
#>  [42,]  0.33169834  0.307006319  0.213865289 -0.68723168 -0.588809504
#>  [43,] -0.98735214  0.240994317  0.070551187  0.12088791  0.171700661
#>  [44,]  0.35131962  0.814910217 -0.814662708 -0.73325414  0.151715473
#>  [45,]  0.76564826  0.150292968 -0.314347748  0.96296569 -0.951578665
#>  [46,] -0.42119773  0.451056371  0.089278239 -0.48039537  0.262959658
#>  [47,]  0.13696709  0.908398069 -0.842307774 -0.77018331  0.939405364
#>  [48,] -0.57277380 -0.027881179 -0.156622567  0.37146334 -0.441321979
#>  [49,]  0.30921239 -0.864707051  0.395924889 -0.38777132  0.075669416
#>  [50,] -0.77891482  0.273968785  0.874174422 -0.03421574 -0.123357223
#>  [51,]  0.63849069 -0.507747547 -0.325336792 -0.17328455  0.919707502
#>  [52,] -0.48584144 -0.526462440  0.592639287 -0.82497281 -0.176455619
#>  [53,] -0.29610636  0.792259904 -0.973917347 -0.18059703 -0.775552228
#>  [54,]  0.17421050 -0.541984594  0.349031153 -0.25531750  0.194777892
#>  [55,] -0.59751593 -0.780261507 -0.936290644  0.19942287 -0.878696942
#>  [56,] -0.67564673 -0.933567394  0.943799764  0.70393709 -0.700246094
#>  [57,]  0.87152199 -0.817680441 -0.335356324  0.17655597  0.320223785
#>  [58,] -0.55169836 -0.337524319 -0.090818990  0.93870404 -0.175045648
#>  [59,] -0.54092081  0.007895201  0.253040491  0.60740611 -0.297154509
#>  [60,] -0.10033498  0.030433664  0.004900395  0.49656095  0.771075789
#>  [61,]  0.76690446  0.226920359  0.629389443 -0.13722735 -0.936046037
#>  [62,] -0.39142121 -0.150747675 -0.552920992  0.78549098  0.527569361
#>  [63,] -0.10283267 -0.214398370 -0.400592410  0.40168930 -0.677189683
#>  [64,]  0.42709494 -0.806759843  0.872527303 -0.53293247  0.069955911
#>  [65,]  0.19859451 -0.258227176  0.824532697  0.98595974 -0.853634530
#>  [66,]  0.56570198 -0.466170395  0.127437934  0.22434444  0.737411663
#>  [67,]  0.46645252 -0.457089806  0.039196650 -0.50540248 -0.396173842
#>  [68,]  0.94465352  0.893977619  0.836265787  0.88127680 -0.641236924
#>  [69,] -0.96107567  0.413204513 -0.070086716  0.96275642  0.088000800
#>  [70,]  0.67528165 -0.632940803  0.778990086 -0.07258992  0.593218019
#>  [71,] -0.81087179 -0.373715855  0.989290269  0.37842411  0.918470428
#>  [72,] -0.79827902 -0.503447562  0.961732334  0.18201316  0.509539756
#>  [73,] -0.31364149  0.084871570  0.233721538  0.23910962  0.058443343
#>  [74,] -0.73349673 -0.665796032  0.477040344 -0.90844079 -0.737785576
#>  [75,]  0.07574419 -0.678011052  0.186785190  0.19996487  0.528306572
#>  [76,] -0.66676523 -0.359313678 -0.158085553 -0.86145870 -0.089240209
#>  [77,]  0.99326641  0.513249762  0.206243265  0.05323107  0.644145465
#>  [78,]  0.29870678 -0.349893036 -0.889779348 -0.68446233  0.622981095
#>  [79,]  0.35227634  0.142648593  0.514572645  0.88826240 -0.747403037
#>  [80,] -0.45589976  0.829981723 -0.349334075 -0.05149288 -0.969650117
#>  [81,]  0.18340353  0.168573635 -0.419832860 -0.20702228 -0.095574983
#>  [82,] -0.80663888  0.203457522 -0.170769186 -0.43217228  0.619488304
#>  [83,] -0.84224349 -0.109728847  0.158973991 -0.02209394  0.773512138
#>  [84,] -0.39433699 -0.973439962  0.767111997 -0.58418411 -0.209770980
#>  [85,] -0.38783019 -0.132359123 -0.078636039  0.57382970  0.874312620
#>  [86,] -0.61777082 -0.228024343  0.915474379  0.14682610 -0.950625489
#>  [87,]  0.21925623  0.835815859  0.799034241 -0.65049660 -0.512518945
#>  [88,] -0.54575323  0.955166246 -0.283962147  0.50967925 -0.589569835
#>  [89,] -0.56338105  0.050107228 -0.826276056 -0.86611281 -0.200446121
#>  [90,] -0.29916897 -0.664895960  0.908381213 -0.28026783  0.924620904
#>  [91,] -0.19337438  0.659893210  0.021306751 -0.64118378 -0.274338606
#>  [92,] -0.58163691 -0.944630770  0.538702064 -0.43841474  0.763810948
#>  [93,]  0.84102762  0.746183752  0.379103097  0.78194124 -0.674094711
#>  [94,]  0.97408587 -0.789328655 -0.387145699  0.35850895  0.615140502
#>  [95,] -0.70224576 -0.356423426 -0.647051501 -0.69092220 -0.738256041
#>  [96,] -0.43126630  0.500639454 -0.816888056  0.55699310  0.867089106
#>  [97,]  0.74511339 -0.018596968  0.995559858 -0.23029934 -0.345681732
#>  [98,] -0.75927843  0.002399330 -0.976606191 -0.62770748 -0.857551451
#>  [99,] -0.48390691  0.555060189  0.220232523 -0.61511545  0.083549005
#> [100,]  0.29796853  0.532118300 -0.580259369  0.30489602  0.668541685
#> 
#> $y
#>   [1] 0 0 0 0 1 1 0 1 0 1 1 1 1 0 0 1 1 0 1 0 0 1 1 1 1 0 1 1 0 0 1 1 0 1 1 1 1
#>  [38] 0 1 0 1 1 0 0 1 1 0 1 0 0 0 1 0 1 0 0 0 1 1 1 0 1 1 0 1 1 1 0 0 0 0 0 1 0
#>  [75] 1 1 0 1 1 0 1 0 0 0 1 1 0 0 1 1 1 0 0 0 1 1 1 1 1 1
#>