Simulate Data from Hyper-Sphere for Classification Problems

The classification problem Circle is described in the BARK paper (2008). Inputs are dim independent variables uniformly distributed on the interval \([-1,1]\), only the first 2 out of these dim are actually signals. Outputs are created according to the formula $$y = 1(x1^2+x2^2 \le 2/\pi)$$

sim_circle(n, dim = 5)

Arguments

n

number of data points to generate

dim

number of dimension of the problem, no less than 2

Value

Returns a list with components

x

input values (independent variables)

y

0/1 output values (dependent variable)

References

Ouyang, Zhi (2008) Bayesian Additive Regression Kernels. Duke University. PhD dissertation, Chapter 3.

See also

Examples

sim_circle(n=100, dim=5)
#> $x
#>               [,1]         [,2]         [,3]        [,4]         [,5]
#>   [1,]  0.45790124 -0.041006661 -0.704508819  0.94511027  0.103863655
#>   [2,]  0.42439154 -0.040011051 -0.345436670  0.49518253  0.922222517
#>   [3,] -0.64953142  0.773072657 -0.779050143  0.77966229 -0.697436168
#>   [4,] -0.12143620  0.346462571  0.322645006  0.79326152  0.948182279
#>   [5,]  0.19581114  0.194438265 -0.624311152  0.47508105 -0.735151283
#>   [6,] -0.63380455 -0.361993464 -0.710805582 -0.06341305 -0.043154649
#>   [7,]  0.05556974 -0.697681143  0.036296567  0.53682589 -0.976985714
#>   [8,] -0.89769957 -0.944575166  0.059697818  0.80820331  0.801131314
#>   [9,]  0.64868960 -0.786756504  0.153491343  0.60558441 -0.587771412
#>  [10,] -0.23061924 -0.621140957 -0.566562475  0.98422106 -0.199628480
#>  [11,]  0.40916149 -0.237354514 -0.657967584 -0.51440054 -0.061198667
#>  [12,] -0.27179587  0.283241979  0.602656173 -0.41387108  0.273598825
#>  [13,] -0.78588618  0.821262104 -0.355870201 -0.21613202 -0.873823878
#>  [14,] -0.44027266  0.736223221 -0.727832458 -0.95601028  0.615710615
#>  [15,]  0.61678393 -0.695942804  0.599401946 -0.90566977  0.634265710
#>  [16,]  0.73596854 -0.969001874 -0.821549120  0.48233973 -0.950858849
#>  [17,] -0.79258621  0.405955452  0.347610304 -0.52520420  0.259900762
#>  [18,] -0.84060350  0.426830814  0.885139534  0.52769999 -0.537526869
#>  [19,]  0.58271658 -0.805455640  0.965622306 -0.98520209  0.693860200
#>  [20,] -0.74195913  0.001812189  0.436437542 -0.98153392  0.173351295
#>  [21,]  0.06046274  0.714380782 -0.058057667  0.06083727 -0.868143844
#>  [22,]  0.83749036 -0.811790028 -0.333098306  0.78093435 -0.114986463
#>  [23,] -0.81350256 -0.601448536  0.677343324 -0.99417555  0.665778582
#>  [24,]  0.06594439  0.356587064  0.762281151 -0.73095536  0.592306384
#>  [25,] -0.09356189 -0.586530094  0.909170222 -0.42354667 -0.130520545
#>  [26,]  0.63059577 -0.180075229 -0.337733916 -0.43129720  0.524299500
#>  [27,] -0.53040128  0.724251455 -0.164613409 -0.97801990  0.691065069
#>  [28,]  0.84852211  0.060168268  0.657485214 -0.23907069 -0.098083072
#>  [29,]  0.16373543 -0.709325965 -0.030903056 -0.74115350  0.643259649
#>  [30,]  0.47235545  0.077081047 -0.289381102 -0.12964933  0.400690796
#>  [31,]  0.10618180  0.635851955  0.872466646 -0.35784080  0.071949384
#>  [32,] -0.59553350 -0.676115073 -0.047070135  0.05678277 -0.188393715
#>  [33,] -0.77508928 -0.327648971  0.646629808  0.55875094 -0.487567599
#>  [34,]  0.83542307 -0.443739799  0.205834084  0.05766118  0.277647041
#>  [35,] -0.85567409  0.733981892 -0.055165261 -0.06712142  0.005523638
#>  [36,]  0.81033010 -0.457575620  0.042733068 -0.11210467 -0.684853533
#>  [37,]  0.48404047  0.726942215 -0.944572360  0.23430021 -0.471797050
#>  [38,]  0.60719779 -0.248088006 -0.610362378 -0.38792977 -0.615076383
#>  [39,] -0.61455672 -0.004610938  0.831185159  0.79553768 -0.096475522
#>  [40,] -0.13801951  0.698888526  0.989991834 -0.48008378  0.636905012
#>  [41,]  0.30941831  0.390482784  0.276242930  0.85117887 -0.558029789
#>  [42,]  0.66810435 -0.819897412  0.771889012 -0.76345527  0.240145324
#>  [43,]  0.23802685 -0.260504918 -0.839473224 -0.59483869  0.107006400
#>  [44,] -0.21840038  0.190971226  0.178676992  0.34146092  0.240909815
#>  [45,] -0.02119299 -0.149508261  0.838190319  0.38663954  0.289510638
#>  [46,] -0.67655401 -0.449829424  0.814862004 -0.20861490  0.882644041
#>  [47,]  0.16255347  0.553121928 -0.596309239  0.55450169  0.459438622
#>  [48,] -0.48120534 -0.627807387  0.403001033  0.61254748 -0.465524865
#>  [49,]  0.08784952  0.824569828  0.153188693  0.66498723 -0.345020283
#>  [50,] -0.65561289 -0.942809318  0.505156009  0.10202545 -0.894225044
#>  [51,]  0.40274944 -0.946000227 -0.471726127 -0.28724499  0.581207148
#>  [52,] -0.48949519  0.042892990  0.998168541  0.67837735  0.906788979
#>  [53,]  0.65137495  0.889559411  0.199079652  0.26847959 -0.263294761
#>  [54,] -0.82781065  0.706335371 -0.459774219  0.35916422 -0.093755770
#>  [55,] -0.87605637 -0.527017281 -0.276203457  0.62642583  0.273546475
#>  [56,]  0.96911837  0.438180623  0.515358230  0.98493783  0.268234115
#>  [57,] -0.28909485 -0.105471083 -0.005279188 -0.70874956 -0.966684764
#>  [58,]  0.08104898  0.136011642 -0.542509572  0.05246266  0.830035141
#>  [59,] -0.19495144  0.012850413  0.097500016 -0.86717169  0.586192001
#>  [60,] -0.88945045  0.701883990 -0.139498596  0.03934339 -0.685740059
#>  [61,]  0.38197769 -0.647567597  0.091927141 -0.25537365  0.323747135
#>  [62,]  0.07802104  0.810875732 -0.904956381  0.24418009  0.603748293
#>  [63,] -0.62341714 -0.836534573 -0.148688847 -0.95331397  0.503919424
#>  [64,] -0.49806270 -0.682889883 -0.211800429  0.39224687 -0.777294651
#>  [65,]  0.45591532  0.219806776 -0.356522854  0.15321998 -0.388399735
#>  [66,] -0.87548657 -0.782002914  0.839929127  0.35064398 -0.302450734
#>  [67,]  0.78168722 -0.685577849  0.438448243  0.24722926 -0.425643390
#>  [68,] -0.55375232  0.648732443  0.960092192  0.20861717 -0.133381738
#>  [69,] -0.43279590  0.439930885  0.918296968  0.36450314  0.467897008
#>  [70,] -0.71055566  0.888643576 -0.357101923 -0.52451723  0.209553719
#>  [71,]  0.01912864  0.641937428 -0.640548898  0.34477693 -0.354420139
#>  [72,]  0.17635314 -0.319626394  0.412476931 -0.70847791  0.910212322
#>  [73,] -0.44202869 -0.918399037  0.543946317 -0.61704013  0.507200674
#>  [74,] -0.51754766  0.078119887  0.821570569 -0.52050915 -0.961291751
#>  [75,]  0.47892510  0.450764562  0.520129210 -0.69440252 -0.127890116
#>  [76,]  0.69238772  0.816724632 -0.977803567 -0.46812231  0.384946119
#>  [77,] -0.21502081  0.776942095  0.282971335  0.37861968 -0.790319878
#>  [78,]  0.47855195 -0.853158823 -0.389092103 -0.34311283 -0.001873090
#>  [79,]  0.39207411 -0.920255725  0.451721717  0.70925886  0.154087605
#>  [80,] -0.75861908  0.128644379  0.589080539 -0.32973674 -0.205633843
#>  [81,]  0.41213559  0.916483549 -0.585689077  0.39627368 -0.071597058
#>  [82,]  0.65816962 -0.169044305 -0.907806275  0.34096437  0.795324368
#>  [83,]  0.86268729  0.527843575 -0.851368560  0.05384420  0.477925725
#>  [84,]  0.01052166  0.480156044 -0.952267546  0.86238640 -0.828069228
#>  [85,] -0.79866588  0.002366647 -0.109033877  0.58094928  0.269715638
#>  [86,] -0.65664301 -0.623308696  0.577517948  0.92917638  0.358240398
#>  [87,]  0.85879256 -0.498713049 -0.690055100  0.13786112 -0.950602105
#>  [88,] -0.23363103 -0.396005319 -0.116166412  0.17540120 -0.030019222
#>  [89,] -0.16728430  0.236106018 -0.706545587 -0.43556234 -0.946885107
#>  [90,] -0.59026037 -0.687462566  0.179368632 -0.96812814  0.276424326
#>  [91,] -0.00774873 -0.704849316 -0.316305391 -0.61522872  0.876082656
#>  [92,]  0.17647126 -0.824288104  0.418928920 -0.95577880 -0.496383858
#>  [93,] -0.39008804 -0.092502658  0.259187419 -0.30751478  0.686312813
#>  [94,] -0.08651732  0.102598142  0.723231953 -0.78234048 -0.153445784
#>  [95,] -0.64230085 -0.190640238 -0.988195831 -0.48649324  0.713685370
#>  [96,]  0.04381104  0.567518035 -0.968084093 -0.27978957 -0.110426360
#>  [97,]  0.28598527 -0.037786377 -0.114000142  0.83033343  0.774656351
#>  [98,] -0.30098923 -0.955885017  0.664727254 -0.26562516  0.962346629
#>  [99,]  0.34903232  0.063680279 -0.767376965  0.99691582 -0.699344761
#> [100,] -0.41173767 -0.596120346  0.288535961  0.45368269  0.455435126
#> 
#> $y
#>   [1] 1 1 0 1 1 1 1 0 0 1 1 1 0 0 0 0 0 0 0 1 1 0 0 1 1 1 0 0 1 1 1 0 0 0 0 0 0
#>  [38] 1 1 1 1 0 1 1 1 0 1 1 0 0 0 1 0 0 0 0 1 1 1 0 1 0 0 0 1 0 0 0 1 0 1 1 0 1
#>  [75] 1 0 0 0 0 1 0 1 0 1 0 0 0 1 1 0 1 0 1 1 1 1 1 0 1 1
#>