Simulate Data from Hyper-Sphere for Classification Problems

The classification problem Circle is described in the BARK paper (2008). Inputs are dim independent variables uniformly distributed on the interval \([-1,1]\), only the first 2 out of these dim are actually signals. Outputs are created according to the formula $$y = 1(x1^2+x2^2 \le 2/\pi)$$

sim_circle(n, dim = 5)

Arguments

n

number of data points to generate

dim

number of dimension of the problem, no less than 2

Value

Returns a list with components

x

input values (independent variables)

y

0/1 output values (dependent variable)

References

Ouyang, Zhi (2008) Bayesian Additive Regression Kernels. Duke University. PhD dissertation, Chapter 3.

See also

Examples

sim_circle(n=100, dim=5)
#> $x
#>                [,1]         [,2]        [,3]         [,4]          [,5]
#>   [1,] -0.183489182 -0.051446059 -0.01388655  0.294589194  0.4101732485
#>   [2,]  0.155085319  0.922392267  0.16071989 -0.716725611 -0.7244708557
#>   [3,]  0.233135888 -0.554508213  0.79139476  0.865384481  0.4494651179
#>   [4,] -0.938891927  0.179407042  0.80561903  0.404011616  0.3238284136
#>   [5,]  0.968084020  0.131871247 -0.74110906  0.920414158 -0.7278327951
#>   [6,] -0.469826681  0.573517684 -0.55933439  0.977020486  0.7282252866
#>   [7,] -0.272745868 -0.502596669 -0.43099785 -0.515641894  0.8128875340
#>   [8,] -0.536912769  0.292558099  0.48466980  0.745534881 -0.3985595684
#>   [9,] -0.964861910 -0.040510344 -0.54535379  0.383479547 -0.0503169256
#>  [10,]  0.990972156 -0.076154747  0.34695087 -0.900412790 -0.2167056832
#>  [11,] -0.255834234  0.296903145  0.14779443 -0.326830249 -0.0323963361
#>  [12,] -0.924833995 -0.345062702  0.45994947 -0.649909866 -0.4388850336
#>  [13,]  0.467909854  0.035759624 -0.53784779  0.320965229  0.7494747369
#>  [14,] -0.237019576 -0.527485968  0.84591974  0.204243665  0.6535321693
#>  [15,]  0.634999310 -0.860938415 -0.50044725 -0.305334331  0.1715838541
#>  [16,]  0.928958235  0.638970660  0.91375450 -0.013153349  0.5095559503
#>  [17,] -0.300597758  0.931198484  0.97467268 -0.006395333  0.3230917077
#>  [18,] -0.232724541 -0.057017387  0.77372576 -0.882070453  0.8209822206
#>  [19,] -0.074890034  0.066772964 -0.78344198 -0.370304096 -0.2548901322
#>  [20,]  0.090976391 -0.610403965 -0.34615771 -0.437501391 -0.1250077942
#>  [21,]  0.921269131  0.299036175  0.23276404 -0.472293427 -0.9795738268
#>  [22,]  0.795975343 -0.409797774 -0.70248068  0.588608397 -0.1422244352
#>  [23,]  0.695754568  0.682860780 -0.95197055 -0.522112860 -0.1286582239
#>  [24,] -0.540601584  0.470980639 -0.91383245 -0.337667768 -0.3079166054
#>  [25,]  0.435986576  0.639104954 -0.26062456 -0.941351796  0.4797319975
#>  [26,] -0.619460952  0.304843570  0.66169783  0.203798685 -0.2467972273
#>  [27,]  0.429834514 -0.500798136  0.08899727 -0.015755706 -0.1487802058
#>  [28,]  0.812851950  0.850511587 -0.60506054 -0.539665947 -0.5099226311
#>  [29,]  0.374874217  0.004377981  0.67256481  0.366938449 -0.0368992658
#>  [30,]  0.220757099  0.960876969 -0.48215070  0.820351437  0.0472718556
#>  [31,]  0.806822565  0.550473786  0.03107543  0.311029567  0.9488073573
#>  [32,]  0.477696747  0.571960406 -0.45997375 -0.840918080  0.3820769587
#>  [33,] -0.952145277 -0.947919403  0.77023806  0.607577133  0.2874987549
#>  [34,]  0.381652494 -0.386326591 -0.95384109 -0.485785872 -0.9988299101
#>  [35,]  0.969168403  0.688483565 -0.44877257 -0.311948463  0.2475778251
#>  [36,] -0.681140257  0.900054346  0.94155789  0.901054543  0.0002421648
#>  [37,] -0.627015401 -0.310753290 -0.16054884  0.709128422  0.0050292932
#>  [38,]  0.371295243  0.484237496 -0.27303209  0.049777530 -0.2916512997
#>  [39,]  0.706080902 -0.306444741 -0.20158712  0.989393483  0.1572820670
#>  [40,] -0.285130671 -0.061165311  0.80461060  0.323719167  0.1948352898
#>  [41,]  0.811263552 -0.116301451  0.21630871  0.452315240 -0.9630291751
#>  [42,] -0.555856684  0.428056403  0.59728459  0.522634645  0.0696551870
#>  [43,]  0.653965118  0.267282471 -0.75547415  0.836683786  0.8869790202
#>  [44,]  0.086444249 -0.788699106  0.57776216  0.871319996 -0.5812361804
#>  [45,]  0.663381157  0.601496621 -0.85100635 -0.003123410 -0.1088977028
#>  [46,] -0.980468198 -0.666012672  0.15711292 -0.895862431 -0.8492612108
#>  [47,] -0.124118196  0.875707874 -0.16363770  0.539775864 -0.1958329831
#>  [48,] -0.700244603 -0.415161636 -0.84566039  0.073043972  0.1786265615
#>  [49,]  0.235419207  0.978664711  0.48388076  0.031116661 -0.3653886244
#>  [50,] -0.638501646 -0.844003089 -0.99324941  0.056799219  0.3313310323
#>  [51,]  0.637706954 -0.865060945 -0.42067763 -0.321525532  0.8312324658
#>  [52,]  0.006267907 -0.095053819  0.40794379  0.605531483  0.8154096277
#>  [53,]  0.848251140 -0.739159337 -0.54690670 -0.938504604 -0.8706488586
#>  [54,] -0.395492959 -0.782463919 -0.25925379  0.441437116 -0.7105290219
#>  [55,] -0.647023939 -0.687568265 -0.01700171  0.817698133  0.3248761687
#>  [56,] -0.048230436  0.843700658 -0.74038438  0.300417639  0.9320110227
#>  [57,] -0.460144788 -0.733646572 -0.61338269 -0.045847991 -0.4006550135
#>  [58,]  0.101627992 -0.242245379  0.43301539  0.557265942 -0.0785271749
#>  [59,] -0.880877476  0.845716469 -0.16324442 -0.135892693 -0.8239629068
#>  [60,]  0.528304247  0.223646313 -0.34530824 -0.542006833 -0.9478631536
#>  [61,] -0.541625594 -0.714110360 -0.34054388 -0.251774666 -0.2416956294
#>  [62,]  0.917807763  0.552567863  0.49874704  0.588317220  0.3690443370
#>  [63,]  0.955194449 -0.018079197  0.98354276  0.097570440 -0.7589233415
#>  [64,]  0.555712283  0.038623151 -0.52509212  0.111168006  0.1980199036
#>  [65,] -0.493508035  0.189281868 -0.45098171 -0.468084570 -0.4707230842
#>  [66,]  0.855891996 -0.324431425 -0.86117509  0.190975107 -0.4762471681
#>  [67,] -0.202288194 -0.200805820 -0.57313323  0.144648514  0.7692801277
#>  [68,] -0.839464491  0.323223902 -0.62909121  0.242981080  0.3158409907
#>  [69,]  0.999106079  0.996622253 -0.57035664 -0.466886123 -0.9437727295
#>  [70,] -0.286460584  0.586778277  0.74687519 -0.396177104 -0.3451698800
#>  [71,]  0.344789997  0.876713415  0.47375636  0.405567717 -0.5038692467
#>  [72,] -0.673324789  0.040656474 -0.35870552 -0.646641757 -0.4495747751
#>  [73,]  0.332778738 -0.955052837  0.94651123 -0.698298967  0.3023367766
#>  [74,] -0.034731036  0.854724429  0.54043024  0.599377248 -0.7739889589
#>  [75,]  0.229849757 -0.531269369  0.45006630  0.795093162 -0.6080826656
#>  [76,] -0.908004130  0.771048750  0.81265191 -0.190228419 -0.4752168194
#>  [77,] -0.087683779 -0.434079132 -0.24820634  0.586620734 -0.4099588455
#>  [78,] -0.962622886  0.352074612  0.32846633  0.153360145  0.6907569291
#>  [79,] -0.115076740  0.970010314  0.37523008 -0.889443242 -0.3061860162
#>  [80,]  0.259324036  0.320208111  0.49747282  0.475349062  0.2565846737
#>  [81,]  0.468192119 -0.737183537  0.58406904 -0.437297349 -0.7916537318
#>  [82,] -0.854392358 -0.588622228 -0.58788868 -0.373635715 -0.7613072298
#>  [83,] -0.401843718 -0.076599305 -0.25592486  0.669612214  0.7327608690
#>  [84,]  0.407862851  0.066437861 -0.26468765 -0.808302740 -0.8526642751
#>  [85,] -0.105354167 -0.393500914 -0.90425213 -0.176868693  0.1446999861
#>  [86,]  0.845989393  0.851848651  0.48668694 -0.386127131 -0.8416146105
#>  [87,] -0.762754108  0.007006152 -0.28555976  0.946232449 -0.1214396819
#>  [88,]  0.686512512 -0.485115062  0.17700078 -0.983430422 -0.8295024820
#>  [89,]  0.931422168  0.963017926 -0.39462739 -0.541347785 -0.5476660160
#>  [90,] -0.696809503  0.471320388  0.75695167  0.298910915 -0.1713887565
#>  [91,] -0.412501911 -0.254368923 -0.79529229  0.528611358  0.8592739669
#>  [92,] -0.006670525 -0.223597958  0.01077205 -0.506325188 -0.9907933818
#>  [93,] -0.785603124  0.276215769 -0.55158799  0.738111287 -0.3634821638
#>  [94,]  0.582179237 -0.295229879  0.34356369 -0.871069819 -0.0738179982
#>  [95,]  0.253077613 -0.979511008  0.71936404  0.644581447 -0.3404363887
#>  [96,] -0.880436287 -0.644864189 -0.46551646  0.112692439 -0.7841526465
#>  [97,]  0.722456385  0.101416716 -0.41297147  0.032763372 -0.2502157628
#>  [98,]  0.297160414 -0.368933466 -0.86466867 -0.688472794  0.5984199559
#>  [99,]  0.561174863  0.066122889  0.66517786 -0.814755445  0.8445884576
#> [100,]  0.985560865 -0.120044462 -0.72375458  0.435702620  0.9596966370
#> 
#> $y
#>   [1] 1 0 1 0 0 1 1 1 0 0 1 0 1 1 0 0 0 1 1 1 0 0 0 1 1 1 1 0 1 0 0 1 0 1 0 0 1
#>  [38] 1 1 1 0 1 1 1 0 0 0 0 0 0 0 1 0 0 0 0 0 1 0 1 0 0 0 1 1 0 1 0 0 1 0 1 0 0
#>  [75] 1 0 1 0 0 1 0 0 1 1 1 0 1 0 0 0 1 1 0 1 0 0 1 1 1 0
#>