Simulate Data from Hyper-Sphere for Classification Problems

The classification problem Circle is described in the BARK paper (2008). Inputs are dim independent variables uniformly distributed on the interval \([-1,1]\), only the first 2 out of these dim are actually signals. Outputs are created according to the formula $$y = 1(x1^2+x2^2 \le 2/\pi)$$

sim_circle(n, dim = 5)

Arguments

n

number of data points to generate

dim

number of dimension of the problem, no less than 2

Value

Returns a list with components

x

input values (independent variables)

y

0/1 output values (dependent variable)

References

Ouyang, Zhi (2008) Bayesian Additive Regression Kernels. Duke University. PhD dissertation, Chapter 3.

See also

Examples

sim_circle(n=100, dim=5)
#> $x
#>                [,1]         [,2]         [,3]         [,4]          [,5]
#>   [1,]  0.459473452 -0.909803256  0.886884761  0.492472039  0.4239336145
#>   [2,]  0.444140630  0.872883487 -0.096765317  0.847352914 -0.6389284115
#>   [3,] -0.637209850  0.868553501  0.135618907 -0.107156876  0.8191812551
#>   [4,]  0.621960800  0.051500751  0.742801464  0.099766741  0.3742706613
#>   [5,]  0.964889346  0.131717379  0.019235507  0.591558627 -0.1879171883
#>   [6,]  0.196572532  0.802030178  0.297791361 -0.150334830  0.6508415220
#>   [7,] -0.442512228  0.665538377  0.559995078  0.700185252  0.1810575305
#>   [8,]  0.141070825  0.275295639 -0.665662389 -0.334560534  0.0613080435
#>   [9,] -0.024002068 -0.665022645 -0.512223895 -0.746634838 -0.6141776242
#>  [10,] -0.788537280 -0.903025433  0.890046382 -0.331549726  0.2165416502
#>  [11,]  0.185813589  0.563358271 -0.211704342 -0.890206252  0.5535208457
#>  [12,]  0.721195836  0.658892913 -0.632934101 -0.407864042  0.7707299544
#>  [13,] -0.742323503  0.231224650  0.921669005 -0.806345021 -0.4611251028
#>  [14,] -0.731292953  0.648990747  0.774018008 -0.445237855 -0.2822505892
#>  [15,] -0.015731809  0.027250720 -0.779698653 -0.232749071  0.2944359188
#>  [16,]  0.441651102  0.593107654  0.974047113 -0.733366111  0.4280580613
#>  [17,]  0.265105443  0.026500424  0.895722019  0.735315325 -0.1515704109
#>  [18,] -0.197641090  0.432601342 -0.927354040 -0.201518742  0.8653116943
#>  [19,] -0.140924369  0.306280259 -0.353909592 -0.301113896 -0.5825379426
#>  [20,] -0.838642455 -0.007641925 -0.773494171 -0.013768257  0.9713152163
#>  [21,] -0.723073558  0.954944722 -0.667922571  0.996881362 -0.5815124647
#>  [22,] -0.263225026  0.879693495 -0.046703096  0.337998150 -0.5525745125
#>  [23,]  0.268180169 -0.939650917 -0.546390162  0.301335521 -0.9715520241
#>  [24,]  0.655167841 -0.810122940  0.092220140 -0.657743289 -0.0024247253
#>  [25,]  0.410893563  0.590805147 -0.936721068 -0.410250836  0.1282534916
#>  [26,]  0.211949058  0.008706426  0.327828705 -0.593552633 -0.7604087973
#>  [27,] -0.770172381 -0.725784708  0.996042948  0.831732504 -0.7060608333
#>  [28,]  0.040743869  0.649717188 -0.314177341  0.760139129  0.6136044241
#>  [29,]  0.645876743  0.920080491 -0.842277394 -0.731025605  0.4353655111
#>  [30,]  0.971810461  0.200546337 -0.447467459 -0.595287976  0.7750298609
#>  [31,] -0.772529860 -0.230995519  0.386401352 -0.261758875 -0.6195207429
#>  [32,] -0.659274210 -0.479841947 -0.299137821 -0.891210739 -0.9190596794
#>  [33,] -0.229572110 -0.888945124 -0.578022256  0.802230762 -0.0005766912
#>  [34,]  0.944495648 -0.201066714  0.268487715  0.707704920  0.2579198815
#>  [35,] -0.219870638  0.838072781 -0.271080955  0.167225395  0.7802544632
#>  [36,] -0.405687326  0.154670720  0.351392319  0.982848018  0.4559889869
#>  [37,]  0.158011226 -0.291031458 -0.245347264  0.452315701  0.2362543545
#>  [38,] -0.076574551  0.755573527 -0.508046608 -0.278661687  0.3711353540
#>  [39,] -0.919970381  0.330206861  0.501359698 -0.916775114 -0.3968518507
#>  [40,]  0.600567290 -0.833145707  0.699276488  0.011433343  0.0361060426
#>  [41,] -0.582619840 -0.972177880 -0.537080253 -0.645784033  0.2529323827
#>  [42,]  0.014577003  0.777174551  0.180936200  0.215654334  0.2849415317
#>  [43,]  0.215677046  0.439598210  0.747093996 -0.348705449  0.6945540993
#>  [44,] -0.760434926 -0.013579375  0.845519983 -0.948008745  0.8152591921
#>  [45,]  0.742773154 -0.653791217  0.766990772 -0.781389632  0.0024573533
#>  [46,] -0.529791467  0.281474205  0.242514294  0.935241333 -0.1521903854
#>  [47,]  0.609714086 -0.631555240 -0.508321351 -0.609581108 -0.9460391044
#>  [48,] -0.443470870  0.327962064 -0.490646778  0.230651164 -0.1338367499
#>  [49,]  0.566638142 -0.361624957 -0.168374755 -0.473027959  0.5815922054
#>  [50,] -0.862468474  0.028271612  0.675547202 -0.567765234  0.6506839292
#>  [51,] -0.371849330 -0.951229462 -0.072617705  0.860690778 -0.7675908566
#>  [52,] -0.304073204 -0.288629027 -0.836697312  0.421217203  0.8708597757
#>  [53,]  0.610026054 -0.650917578 -0.061913039  0.987089653  0.0462782136
#>  [54,] -0.516616356 -0.414923259 -0.173873753  0.009695477  0.4355057273
#>  [55,]  0.789808247 -0.239561422 -0.860859558  0.877601916 -0.1752944239
#>  [56,] -0.981845162  0.159749789  0.475589762 -0.541884960 -0.0783602688
#>  [57,]  0.391358334 -0.163039682 -0.814328081 -0.124278897 -0.2010609112
#>  [58,] -0.237678708  0.108444529  0.950661181 -0.931304641 -0.8549384619
#>  [59,]  0.096348871  0.212621147  0.252836031 -0.307540477 -0.3189067715
#>  [60,]  0.077971971  0.940586730  0.008353714 -0.217246976 -0.8081101361
#>  [61,]  0.107051137 -0.100726871 -0.077033892  0.271749505  0.0989018348
#>  [62,]  0.005515776 -0.614331979 -0.834626681  0.678077129  0.6403566296
#>  [63,]  0.466202260  0.679677194 -0.694175515 -0.423326056 -0.2026361413
#>  [64,] -0.026221371  0.141583984 -0.198080900  0.328370830  0.9721181197
#>  [65,] -0.163290366  0.263179640  0.795991347  0.776428067  0.5236809775
#>  [66,] -0.204781280  0.641676946 -0.849104172 -0.088129267  0.1681874688
#>  [67,] -0.590065241 -0.501383448 -0.250755884 -0.470351993  0.6622790298
#>  [68,] -0.421282164  0.919283343 -0.786045792 -0.183238594 -0.8815463893
#>  [69,]  0.892436079 -0.284013656  0.243353760  0.916667943  0.5964496080
#>  [70,] -0.181780394 -0.774538274 -0.274044410  0.397410379 -0.2069488643
#>  [71,]  0.131208434 -0.558291873  0.026895176  0.383953976  0.3630510904
#>  [72,] -0.529820526  0.261270768  0.463220171  0.842815554  0.1684894809
#>  [73,] -0.286668944 -0.454735516  0.608998584 -0.338040465 -0.4669726985
#>  [74,] -0.088413731  0.639650891 -0.730226743 -0.873920363  0.9805134484
#>  [75,]  0.612018109  0.556018204 -0.549369763 -0.481640255 -0.6281857267
#>  [76,]  0.174105796  0.288106195 -0.820236437 -0.836112733  0.1453790474
#>  [77,]  0.885684839 -0.273228833 -0.181816631 -0.174771534  0.3375367830
#>  [78,]  0.716198242  0.664957683  0.731708352  0.181310188 -0.0746108517
#>  [79,] -0.546702603 -0.807178682  0.586236825  0.466915473  0.2655170942
#>  [80,]  0.638430244 -0.223107943  0.416189872 -0.572095431  0.4179951870
#>  [81,] -0.220559549 -0.030788848  0.394058363 -0.370243604 -0.6833053729
#>  [82,]  0.755076481  0.294851670 -0.848698687  0.336902579  0.8027433893
#>  [83,]  0.109892691 -0.577162297  0.029058500  0.696529904  0.2616637708
#>  [84,] -0.362373815 -0.334010032 -0.617670218 -0.770198367  0.7352158553
#>  [85,] -0.520564973  0.770308561  0.269739700 -0.806838600 -0.9322311818
#>  [86,] -0.159470466  0.380715339  0.898794170  0.102079913  0.2518250155
#>  [87,]  0.800318310 -0.965814565 -0.654894983  0.551800526  0.7008834672
#>  [88,] -0.487043423  0.103033726  0.909304988 -0.898839401 -0.2982370001
#>  [89,]  0.190581376  0.056438265 -0.762810936 -0.278395128 -0.0340700457
#>  [90,] -0.469656474  0.129062376  0.031024586  0.108250427  0.8255015435
#>  [91,]  0.924088633 -0.914404549  0.410200253 -0.056802753 -0.2100900966
#>  [92,] -0.181294184  0.784793743  0.382157503 -0.042508522  0.1738744699
#>  [93,]  0.318542060  0.521976491 -0.722450171  0.636543016  0.7666145028
#>  [94,]  0.004714357  0.819504138  0.826694583 -0.236453061 -0.0602955953
#>  [95,]  0.421937136 -0.061752204  0.191640834  0.844478822  0.4120611874
#>  [96,] -0.883121987 -0.097150899 -0.020925426 -0.959477646  0.7846398517
#>  [97,] -0.900376962 -0.419542796 -0.462935164 -0.700500908  0.2075414192
#>  [98,] -0.428266394  0.430066928 -0.374980334 -0.937806435  0.6928219702
#>  [99,]  0.572165845 -0.701365268  0.697806816 -0.874871064  0.2320488486
#> [100,]  0.040435739  0.624063774 -0.423337011 -0.005926531  0.2053513220
#> 
#> $y
#>   [1] 0 0 0 1 0 0 0 1 1 0 1 0 1 0 1 1 1 1 1 0 0 0 0 0 1 1 0 1 0 0 0 0 0 0 0 1 1
#>  [38] 1 0 0 0 1 1 1 0 1 0 1 1 0 0 1 0 1 0 0 1 1 1 0 1 1 0 1 1 1 1 0 0 1 1 1 1 1
#>  [75] 0 1 0 0 0 1 1 0 1 1 0 1 0 1 1 1 0 0 1 0 1 0 0 1 0 1
#>