Simulate Data from Hyper-Sphere for Classification Problems

The classification problem Circle is described in the BARK paper (2008). Inputs are dim independent variables uniformly distributed on the interval \([-1,1]\), only the first 2 out of these dim are actually signals. Outputs are created according to the formula $$y = 1(x1^2+x2^2 \le 2/\pi)$$

sim_circle(n, dim = 5)

Arguments

n

number of data points to generate

dim

number of dimension of the problem, no less than 2

Value

Returns a list with components

x

input values (independent variables)

y

0/1 output values (dependent variable)

References

Ouyang, Zhi (2008) Bayesian Additive Regression Kernels. Duke University. PhD dissertation, Chapter 3.

See also

Examples

sim_circle(n=100, dim=5)
#> $x
#>               [,1]         [,2]        [,3]        [,4]         [,5]
#>   [1,] -0.92554743 -0.320972760  0.97346520  0.21743800  0.307904867
#>   [2,] -0.01097902 -0.453177987 -0.56194879  0.86332593 -0.326067196
#>   [3,] -0.52249486  0.502320482 -0.45125357 -0.60422637 -0.793144905
#>   [4,] -0.98421495 -0.763693952  0.02949360 -0.15229405  0.049016926
#>   [5,] -0.60943432 -0.561651534 -0.55712291 -0.44892160 -0.444272808
#>   [6,]  0.50882552  0.959987469  0.64994396  0.15921400 -0.758078777
#>   [7,]  0.15964980 -0.166688737 -0.38969335  0.67852349  0.488916845
#>   [8,] -0.74594665  0.754322718  0.72199124  0.59390882  0.375736813
#>   [9,] -0.76125079 -0.488796830 -0.27947839  0.97641300 -0.049303634
#>  [10,]  0.19330295 -0.378469741  0.26519236  0.54103511 -0.671459401
#>  [11,] -0.05094176 -0.593674105  0.85835649  0.26887212  0.453229526
#>  [12,] -0.76193542 -0.024887202 -0.33155810 -0.93999958 -0.561345315
#>  [13,]  0.03734588 -0.483867606 -0.71332660  0.62024774  0.657226657
#>  [14,]  0.21118529  0.506991698 -0.67742027 -0.35445257  0.244000539
#>  [15,] -0.39606086 -0.888949502 -0.01334185 -0.26139269 -0.609460890
#>  [16,] -0.18848442 -0.930185725 -0.45306488  0.23437133  0.874808735
#>  [17,] -0.40639418 -0.783317565 -0.20523594 -0.42063565 -0.108563913
#>  [18,] -0.18694969  0.173921090  0.75439595  0.54628964 -0.566966491
#>  [19,]  0.74274686 -0.768599134  0.88364596  0.33706805 -0.589764020
#>  [20,] -0.17599742  0.081977846 -0.27517326 -0.11125273  0.729813545
#>  [21,]  0.57926169  0.098433050  0.63258474  0.77053204  0.579456208
#>  [22,]  0.13796069  0.004240714  0.78985134 -0.96173470  0.085327778
#>  [23,] -0.73857498  0.541848812  0.01187490  0.12719856  0.341344481
#>  [24,]  0.79472290  0.216359329  0.91030632  0.61763543  0.449670556
#>  [25,] -0.73653598 -0.011421838  0.77397055  0.76610623  0.524255598
#>  [26,] -0.97588118  0.627373947 -0.15602742 -0.98794699  0.738253741
#>  [27,]  0.75245107 -0.691422197 -0.85414395  0.44875075  0.693491023
#>  [28,]  0.31415228  0.205587320 -0.99955431 -0.84254403 -0.389325618
#>  [29,] -0.40489452  0.154166943 -0.36010334 -0.31605801  0.082249809
#>  [30,] -0.38117234  0.869025788  0.79460942  0.82841493  0.150859176
#>  [31,]  0.35971359  0.978233827  0.94655731 -0.38892639  0.482750843
#>  [32,]  0.59594692 -0.795335823 -0.75104268 -0.22959118  0.794540520
#>  [33,]  0.58798105 -0.249757819  0.80133497  0.38591135  0.139253183
#>  [34,] -0.10193498  0.056686106 -0.31682570 -0.30853564 -0.794749255
#>  [35,]  0.97807306 -0.621871942  0.36499539  0.26794627  0.547613469
#>  [36,]  0.82817801  0.629329760 -0.89204879 -0.01451777  0.064209593
#>  [37,] -0.70642191  0.463471733  0.89833396  0.32658040  0.377479312
#>  [38,] -0.39815198  0.911310125  0.97495485  0.01164870  0.915534181
#>  [39,]  0.07068783  0.595633870 -0.15945947 -0.86186259  0.908347494
#>  [40,] -0.05132893  0.153493586 -0.26748226 -0.05886955 -0.024571237
#>  [41,] -0.82653361 -0.550982353  0.20562535  0.81501346  0.587111105
#>  [42,] -0.40798746 -0.118186063 -0.43872771  0.55236598  0.473806092
#>  [43,] -0.34959169  0.394269137 -0.30805206 -0.26589031  0.449595153
#>  [44,]  0.13937157  0.129896579 -0.82718011 -0.65885646  0.660929834
#>  [45,] -0.98035380 -0.135692432 -0.81547977  0.50582256  0.167975250
#>  [46,]  0.03210597 -0.502558414  0.67746267  0.30843411 -0.818038573
#>  [47,]  0.73134794  0.112009455 -0.69609975  0.72007732 -0.642141979
#>  [48,] -0.72547796 -0.643619829 -0.47121809 -0.59336003 -0.647662895
#>  [49,]  0.79909846  0.248083867 -0.48282874  0.75397337 -0.007260238
#>  [50,] -0.65954367 -0.116557148 -0.59941985 -0.35863104 -0.965435060
#>  [51,]  0.98534987 -0.473721262 -0.99455392  0.84157117 -0.470851810
#>  [52,]  0.82199545 -0.141355950 -0.85394076  0.03002378 -0.951967402
#>  [53,]  0.22572399 -0.641204671 -0.41999503  0.46937569 -0.291155519
#>  [54,] -0.27208193  0.273917036  0.18999183 -0.12665523 -0.460118175
#>  [55,] -0.58173234 -0.297675089  0.13267328 -0.07124054 -0.440296580
#>  [56,] -0.16583522  0.279114991  0.93060165 -0.70210275  0.850104041
#>  [57,]  0.29596013  0.480503241  0.38886671  0.31252404 -0.688549419
#>  [58,] -0.01653420  0.534181083  0.25217582 -0.33558395 -0.805681480
#>  [59,]  0.18571241  0.896131311 -0.52845645  0.05171251  0.547385701
#>  [60,] -0.62012638  0.008922105  0.11591629  0.64548791 -0.860009694
#>  [61,]  0.93347401  0.157923946  0.12023794 -0.49703527 -0.961607977
#>  [62,] -0.41483417 -0.975439980 -0.15161277 -0.28202141  0.987752489
#>  [63,]  0.20960316  0.472332572  0.74213228 -0.47017345  0.888947922
#>  [64,]  0.12966838 -0.753978770  0.21850702  0.87439274 -0.201224612
#>  [65,]  0.05187802 -0.349603555  0.46511113  0.11480497 -0.382936873
#>  [66,]  0.68815304 -0.456521443  0.88245356 -0.18003665 -0.235078689
#>  [67,]  0.05424845  0.634551074 -0.82624026  0.25477834 -0.517251781
#>  [68,] -0.73668664  0.608831116  0.77804794 -0.54096011 -0.766041270
#>  [69,]  0.11515132 -0.278022703  0.07763963 -0.81310342  0.859188271
#>  [70,] -0.77642794  0.062878698  0.63874039 -0.36616585  0.755906608
#>  [71,] -0.38853626 -0.583280270  0.76504051 -0.82212052  0.227763775
#>  [72,]  0.91399911  0.991161779 -0.16411371 -0.83857347 -0.629753253
#>  [73,]  0.40800145  0.184623122  0.81159784 -0.27572479  0.202327119
#>  [74,] -0.32744960 -0.749679683  0.17794921 -0.68366633 -0.687076348
#>  [75,] -0.96996820 -0.890149531  0.42193732 -0.52499540  0.479815363
#>  [76,] -0.64259716  0.835151470 -0.54909699 -0.30130735  0.647275921
#>  [77,]  0.70086754  0.326577080  0.24965749  0.07622889 -0.761986402
#>  [78,] -0.45762226 -0.361433665 -0.25360708  0.18971131 -0.123612499
#>  [79,]  0.89424850  0.939458180  0.19669315 -0.11440576  0.274734580
#>  [80,] -0.19860135 -0.208411229 -0.55305098  0.67199111  0.105843338
#>  [81,] -0.20925561 -0.605906327 -0.98337025  0.23229949  0.010387918
#>  [82,]  0.03013579  0.345125467 -0.22469790 -0.85712415  0.555296454
#>  [83,]  0.24228402  0.294348287 -0.69851978  0.02141094 -0.190514452
#>  [84,]  0.26170889  0.432117807 -0.09488498  0.14339494 -0.391150160
#>  [85,]  0.12082585 -0.361200251 -0.09436958 -0.31568239 -0.295762318
#>  [86,]  0.95871686  0.571308088 -0.33459196  0.60512662  0.467638740
#>  [87,] -0.88606544  0.882105809 -0.81026494 -0.47461031  0.175075325
#>  [88,]  0.72780204  0.740520588  0.84108982  0.95446323  0.466832039
#>  [89,] -0.41106997  0.117033883  0.68431387 -0.05152502 -0.832032856
#>  [90,] -0.75609541 -0.119277676  0.32640289  0.24654200  0.926486300
#>  [91,]  0.45756158 -0.974019724  0.86432065  0.27464100 -0.113634520
#>  [92,]  0.28447194 -0.112813842 -0.61710065 -0.50750882  0.980635454
#>  [93,]  0.01939090  0.502341954  0.34209503 -0.62209575  0.534275061
#>  [94,] -0.13326699 -0.161368303 -0.27784534 -0.13757265  0.692352611
#>  [95,] -0.06547778  0.180553947  0.45150766  0.24549100 -0.628416957
#>  [96,]  0.53906310 -0.647855264 -0.58640333  0.52661512 -0.448912663
#>  [97,]  0.18680887 -0.047221510  0.50699019 -0.69035509 -0.549695588
#>  [98,] -0.10254131  0.413923846  0.27269755  0.58672244 -0.597906943
#>  [99,]  0.05262888  0.746692428  0.95247276 -0.61208996 -0.514752577
#> [100,]  0.98097097 -0.684452459  0.79059040 -0.91858152 -0.528549174
#> 
#> $y
#>   [1] 0 1 1 0 0 0 1 0 0 1 1 1 1 1 0 0 0 1 0 1 1 1 0 0 1 0 0 1 1 0 0 0 1 1 0 0 0
#>  [38] 0 1 1 0 1 1 1 0 1 1 0 0 1 0 0 1 1 1 1 1 1 0 1 0 0 1 1 1 0 1 0 1 1 1 0 1 0
#>  [75] 0 0 1 1 0 1 1 1 1 1 1 0 0 0 1 1 0 1 1 1 1 0 1 1 1 0
#>