Simulate Data from Hyper-Sphere for Classification Problems

The classification problem Circle is described in the BARK paper (2008). Inputs are dim independent variables uniformly distributed on the interval \([-1,1]\), only the first 2 out of these dim are actually signals. Outputs are created according to the formula $$y = 1(x1^2+x2^2 \le 2/\pi)$$

Usage

sim_circle(n, dim = 5)

Arguments

n

number of data points to generate

dim

number of dimension of the problem, no less than 2

Value

Returns a list with components

x

input values (independent variables)

y

0/1 output values (dependent variable)

References

Ouyang, Zhi (2008) Bayesian Additive Regression Kernels. Duke University. PhD dissertation, Chapter 3.

See also

Examples

sim_circle(n=100, dim=5)
#> $x
#>                [,1]        [,2]          [,3]        [,4]         [,5]
#>   [1,]  0.603803983  0.51598413  0.5165420175  0.82309335 -0.826834050
#>   [2,]  0.649597460 -0.18120169 -0.3877700646 -0.72597825  0.180360097
#>   [3,] -0.144748102  0.73022047  0.7753272713 -0.52575106 -0.359836655
#>   [4,]  0.578793676  0.38919585  0.0607321472  0.96195553 -0.190793194
#>   [5,]  0.744654286 -0.61777445  0.7669381844  0.41784049  0.355251723
#>   [6,]  0.796482386 -0.18317174 -0.4091845187 -0.18522533  0.455627887
#>   [7,] -0.560462697 -0.11337931  0.7963541611 -0.83896812  0.192330712
#>   [8,] -0.446586767  0.74601431  0.0001533935 -0.35001454  0.325299944
#>   [9,]  0.294896920  0.58881779  0.9065097766 -0.61309433  0.632226132
#>  [10,] -0.119146903  0.79235961 -0.2600648263  0.24710309 -0.431979855
#>  [11,]  0.070979518 -0.20796414  0.2536203219 -0.68737401  0.016243496
#>  [12,]  0.083478951 -0.44527443 -0.9771939088 -0.43780229 -0.806763396
#>  [13,] -0.752327464 -0.15391312 -0.3290051157  0.94344104  0.208698611
#>  [14,]  0.761527544  0.53535841 -0.9419507370  0.58860780 -0.535803664
#>  [15,] -0.765762178 -0.28031721 -0.4734760448 -0.88853995 -0.187649691
#>  [16,]  0.398644646  0.99257472 -0.7541470719 -0.99710454  0.473227397
#>  [17,]  0.378096816 -0.85279754  0.9881815985  0.73794069  0.187640341
#>  [18,] -0.106805665 -0.51159367  0.2943018111 -0.54901277  0.429104568
#>  [19,]  0.432135299 -0.30209506  0.4801610461  0.94860039 -0.275213599
#>  [20,] -0.586935149 -0.79276850 -0.0241506128 -0.99010420  0.630087629
#>  [21,]  0.554584994 -0.08903971  0.5121229929  0.49294015 -0.187961532
#>  [22,]  0.945050875  0.87272161  0.5691874065 -0.69338379  0.575089929
#>  [23,] -0.724317821 -0.32169145 -0.9561125874  0.82299398 -0.749707354
#>  [24,] -0.007153759 -0.98205471 -0.3202286894  0.06740977 -0.671306774
#>  [25,]  0.564737328  0.20993735 -0.5576963099  0.20053399  0.926735312
#>  [26,]  0.029408092 -0.25497161  0.0389446816  0.72722107  0.195294088
#>  [27,] -0.161241148  0.02924497 -0.4846746805 -0.41423433 -0.249358887
#>  [28,]  0.135181407 -0.09343255  0.3781648623 -0.45302973 -0.497162410
#>  [29,] -0.066771793  0.61173876  0.9948295429  0.55930628  0.763445364
#>  [30,] -0.814221033  0.50405372  0.4754849244 -0.39223771 -0.503411503
#>  [31,] -0.106435807  0.22592831 -0.6866424629  0.77181032 -0.875259166
#>  [32,] -0.589812454  0.91814576  0.9156628619  0.77366045  0.975492493
#>  [33,]  0.159470799 -0.13466048  0.1345776096  0.37079049 -0.506580246
#>  [34,] -0.454777718 -0.91011482  0.8188648522  0.32966959  0.060278674
#>  [35,] -0.534588383  0.37547012  0.1964050150 -0.60391863  0.390982223
#>  [36,] -0.597698569  0.16996201  0.3079525935  0.01265765 -0.178134988
#>  [37,]  0.529571980 -0.96665010 -0.8060296644  0.23341150 -0.648828247
#>  [38,] -0.408479597 -0.37497674 -0.9012419535  0.47614467  0.513430214
#>  [39,]  0.782698043 -0.39633884  0.6822522935 -0.32972914 -0.995862300
#>  [40,]  0.779787127  0.13664108 -0.8745528851 -0.27846301  0.701640854
#>  [41,]  0.122276999 -0.90159477  0.5500012622  0.67676706  0.577828215
#>  [42,] -0.344488805  0.24319356 -0.2428048993 -0.21989736 -0.875273210
#>  [43,] -0.025350842  0.53457041 -0.0999329221  0.45879362 -0.744819450
#>  [44,] -0.925690786  0.03326080  0.1512872693  0.11214753  0.394044190
#>  [45,]  0.266865969  0.84946323 -0.7999199135  0.62238041 -0.872379500
#>  [46,]  0.696429065  0.09565507  0.6937101074  0.86909756  0.686937402
#>  [47,] -0.707317149 -0.89555789  0.4867579876 -0.08034297  0.658083618
#>  [48,]  0.292680351  0.13319095 -0.5936265932 -0.68008169  0.501477310
#>  [49,] -0.853657329  0.55927096  0.2718088157 -0.99920899  0.881666255
#>  [50,] -0.319359550  0.68118110 -0.0459535667  0.21409893  0.088180205
#>  [51,]  0.507326426  0.87969817 -0.1157721570  0.19009047 -0.667323240
#>  [52,] -0.357306877 -0.89306448 -0.8293086546 -0.82945184  0.515553909
#>  [53,]  0.374597348  0.25071215 -0.4849084383  0.68069506  0.286432285
#>  [54,]  0.938010017 -0.32062929 -0.0421226784  0.08355844  0.028516180
#>  [55,]  0.031032496  0.82486408 -0.0772395041 -0.51225233 -0.154074418
#>  [56,] -0.559384621 -0.55124261 -0.9281570581  0.25481532 -0.594448160
#>  [57,]  0.886051279  0.54701198 -0.3901936393  0.65846511  0.019009146
#>  [58,]  0.206934753 -0.34501325 -0.2509513739 -0.99681326  0.469014553
#>  [59,]  0.076476781 -0.11765105  0.7579414481 -0.91260095  0.294883163
#>  [60,]  0.737879700 -0.65427313  0.5708559081 -0.36646646 -0.385679189
#>  [61,] -0.450240369  0.60753037 -0.3600179087  0.69048181 -0.006705693
#>  [62,] -0.420523868  0.84381400  0.2844357830 -0.21262920  0.875563021
#>  [63,] -0.409799158 -0.69199520 -0.4778245324  0.75450657  0.482908423
#>  [64,]  0.451516325  0.86456811 -0.0361468489  0.92408313 -0.420224497
#>  [65,] -0.066089926 -0.77714364 -0.1899487255 -0.85681652  0.529081778
#>  [66,] -0.542736155 -0.17070355  0.6264345446 -0.43470454 -0.109461754
#>  [67,]  0.796444453 -0.21357345 -0.9411309250 -0.55637349 -0.508889960
#>  [68,] -0.451626841  0.79438107  0.5525397365 -0.77975381  0.317109802
#>  [69,]  0.702588263  0.32310878  0.6953316550 -0.45299556  0.498926678
#>  [70,] -0.788089846 -0.30254569  0.1838965048 -0.13957862  0.577667388
#>  [71,] -0.150787815 -0.30201781 -0.3742812406  0.60190580 -0.728514828
#>  [72,]  0.640907198  0.71650875  0.8139619441  0.83905872 -0.038427682
#>  [73,] -0.239392732 -0.24130895 -0.0319127459  0.07458205  0.982057068
#>  [74,]  0.981504905 -0.46075785  0.7907182984  0.80100739 -0.881289719
#>  [75,] -0.812199186 -0.04380314 -0.4827202796 -0.32900927 -0.028628635
#>  [76,]  0.619895905  0.03578304  0.9916337626  0.84782459 -0.855957855
#>  [77,] -0.652065342 -0.90079693  0.1927346243 -0.48889182 -0.352075779
#>  [78,] -0.107652649 -0.31403041 -0.1626685779  0.07240673  0.538876759
#>  [79,] -0.556576473  0.98518892  0.1317231935 -0.76218444 -0.524282063
#>  [80,]  0.758020252  0.88212353 -0.4194313958 -0.67911014 -0.259542520
#>  [81,] -0.930538537 -0.87166254 -0.6432007710 -0.47193915  0.886383621
#>  [82,]  0.322460376  0.67595796  0.2498811288 -0.72802339  0.856750609
#>  [83,] -0.177355672  0.15130606  0.1763583943 -0.40935787 -0.965040746
#>  [84,]  0.914863256  0.30815068  0.7472780827 -0.36892374  0.592982535
#>  [85,] -0.785128660 -0.95926620 -0.3823565631  0.67605386 -0.575540610
#>  [86,] -0.911096226 -0.91417758  0.5830970313 -0.37548251  0.977572633
#>  [87,] -0.225434748  0.05977408 -0.7059207093  0.11241643 -0.208749690
#>  [88,]  0.349540808 -0.69403834 -0.5144953202  0.45481180 -0.873750064
#>  [89,]  0.591712384  0.38454345 -0.6463577962 -0.53573501  0.604172985
#>  [90,]  0.552357278  0.02133330 -0.7139752256 -0.12294166  0.489155516
#>  [91,]  0.713354964  0.37014641 -0.8011993985  0.84707315 -0.506601379
#>  [92,]  0.819876740 -0.01534878  0.7695220769  0.24858479 -0.238844763
#>  [93,]  0.893792155  0.83308205  0.8383915732 -0.69699267 -0.304277983
#>  [94,]  0.352662611 -0.77910790  0.7182268109 -0.73313493  0.432250825
#>  [95,] -0.923421431  0.44563786 -0.5350125427  0.51246908  0.267469427
#>  [96,] -0.830097177 -0.36102428 -0.8459713953 -0.37450134 -0.738998873
#>  [97,] -0.342164832 -0.24737237 -0.9545007995  0.45935304  0.950024148
#>  [98,]  0.966636308 -0.75779617 -0.1353460271 -0.17728707 -0.478835781
#>  [99,] -0.330832329 -0.37667596  0.5617789989 -0.63520108  0.364016967
#> [100,] -0.005560056 -0.80286043 -0.6223511267  0.04905408  0.235101381
#> 
#> $y
#>   [1] 1 1 1 1 0 0 1 0 1 0 1 1 1 0 0 0 0 1 1 0 1 0 1 0 1 1 1 1 1 0 1 0 1 0 1 1 0
#>  [38] 1 0 1 0 1 1 0 0 1 0 1 0 1 0 0 1 0 0 1 0 1 1 0 1 0 0 0 1 1 0 0 1 0 1 0 1 0
#>  [75] 0 1 0 1 0 0 0 1 1 0 0 0 1 1 1 1 0 0 0 0 0 0 1 0 1 0
#>