Simulate Data from Hyper-Sphere for Classification Problems

The classification problem Circle is described in the BARK paper (2008). Inputs are dim independent variables uniformly distributed on the interval \([-1,1]\), only the first 2 out of these dim are actually signals. Outputs are created according to the formula $$y = 1(x1^2+x2^2 \le 2/\pi)$$

sim_circle(n, dim = 5)

Arguments

n

number of data points to generate

dim

number of dimension of the problem, no less than 2

Value

Returns a list with components

x

input values (independent variables)

y

0/1 output values (dependent variable)

References

Ouyang, Zhi (2008) Bayesian Additive Regression Kernels. Duke University. PhD dissertation, Chapter 3.

See also

Examples

sim_circle(n=100, dim=5)
#> $x
#>               [,1]          [,2]         [,3]         [,4]         [,5]
#>   [1,] -0.76288562  0.2674209010 -0.301787330 -0.216409440 -0.712870021
#>   [2,]  0.17152627 -0.8753532115 -0.953948974 -0.628540567  0.692276471
#>   [3,] -0.21600448  0.0861132396 -0.326916351 -0.316016945 -0.941180947
#>   [4,]  0.30917099  0.1557034077 -0.893063821  0.974425788 -0.892006441
#>   [5,] -0.45719074 -0.3082204252 -0.044000640  0.359558937 -0.084995562
#>   [6,]  0.46443604  0.8358907118 -0.472850580 -0.777638967 -0.006629480
#>   [7,] -0.23175809  0.1061011455 -0.961948598  0.852696532 -0.425441716
#>   [8,]  0.96029660  0.8728614007 -0.235870857  0.716622238  0.037077314
#>   [9,] -0.42195188  0.9440388507  0.466737916  0.968746698  0.513433797
#>  [10,] -0.95342382 -0.5320166918 -0.043143421  0.888265340  0.439473791
#>  [11,] -0.62141895  0.8805003362 -0.367332160 -0.489231955  0.021234964
#>  [12,] -0.89859075  0.4638865651 -0.952335970 -0.754907248  0.083466928
#>  [13,] -0.87652188 -0.2604730325 -0.585677769 -0.750573830  0.360854770
#>  [14,] -0.06260556  0.2654584977  0.591998278  0.592551178  0.700004509
#>  [15,]  0.40421908  0.9595697452  0.097486967 -0.924320666 -0.672214229
#>  [16,]  0.18027853  0.0723092295 -0.612983407  0.445924614  0.752235170
#>  [17,] -0.28074290 -0.1752879741 -0.844882182  0.556453863  0.784459305
#>  [18,] -0.56927167  0.1274130833 -0.344852951  0.159304161 -0.127081546
#>  [19,]  0.77095552  0.7435505921  0.402254902 -0.038757103  0.944336240
#>  [20,]  0.25391008  0.5647894065 -0.001281957 -0.050249935 -0.504955045
#>  [21,]  0.19153639  0.0109205050 -0.870424679  0.809442557  0.278123193
#>  [22,] -0.37972519 -0.3782603405  0.680673924 -0.607645073  0.606029909
#>  [23,] -0.38174334  0.4999630586 -0.778931384  0.773230509 -0.482975272
#>  [24,] -0.65032089 -0.8709959211  0.796399008 -0.478453749  0.915731451
#>  [25,] -0.50273451 -0.1242251405 -0.585394575  0.841660595 -0.364083041
#>  [26,]  0.43305637  0.4216345963  0.216041153  0.192933600  0.258689466
#>  [27,] -0.32247489 -0.6753405281 -0.557269119  0.879797137 -0.411529414
#>  [28,]  0.51976199 -0.7832031166 -0.377498426 -0.197484265 -0.198606313
#>  [29,] -0.66885201  0.3842183338  0.395059528 -0.559692109  0.213289978
#>  [30,]  0.96315695 -0.5107576237 -0.477030245 -0.144231640 -0.178666662
#>  [31,]  0.36615125  0.9597836300 -0.111038193  0.551038177 -0.495304814
#>  [32,] -0.36895563 -0.6400587615 -0.682785095 -0.627675856  0.503640300
#>  [33,]  0.93087139  0.5910173408  0.717830047 -0.885955136  0.824991266
#>  [34,]  0.92730481 -0.7743409886  0.134883195 -0.748337788 -0.091079552
#>  [35,] -0.74909859 -0.6491818372  0.036285441  0.549859087  0.160563742
#>  [36,]  0.55788415  0.5472946879  0.212532315 -0.431717131 -0.095692276
#>  [37,]  0.12400996  0.8100748090  0.686219310 -0.731224534  0.068878128
#>  [38,]  0.03930957  0.0002695345 -0.022109448 -0.707012572 -0.505484183
#>  [39,] -0.99655609  0.4601626676 -0.031722527  0.834318678 -0.423048392
#>  [40,] -0.34689797 -0.8156432947  0.450206783 -0.911161975  0.069326755
#>  [41,] -0.73721102  0.9954150449 -0.721446180 -0.200858979 -0.136590349
#>  [42,] -0.08913654 -0.3099550102  0.774231936  0.518611136 -0.319093809
#>  [43,]  0.89275853 -0.5605812240 -0.902450170 -0.066925815 -0.580537377
#>  [44,]  0.22851243  0.7062702272  0.230041142  0.736331012  0.715225364
#>  [45,] -0.22182138  0.2072091671 -0.251515953  0.615730271 -0.884501771
#>  [46,] -0.83048052 -0.4517139220 -0.583929018 -0.267176256 -0.930223916
#>  [47,]  0.57395076  0.3749279603  0.338654542 -0.362909873  0.858745260
#>  [48,]  0.10756558  0.4595176154  0.985975509  0.630406510  0.856106295
#>  [49,] -0.70851669 -0.9406268834  0.830957298  0.466161118 -0.972573539
#>  [50,]  0.37858941 -0.1110229208  0.183626871  0.326378394  0.234822711
#>  [51,]  0.56402552  0.6008475600 -0.951662587 -0.198726781 -0.699805136
#>  [52,] -0.05014761  0.6096346327  0.634901635  0.445104340 -0.099342834
#>  [53,] -0.38584871  0.3339427807  0.111883967  0.541723111  0.738563702
#>  [54,]  0.95512430 -0.9652678017 -0.919904689 -0.997927580  0.786878633
#>  [55,] -0.77274291 -0.8751597954  0.295734480 -0.161003639  0.055846788
#>  [56,] -0.41475127 -0.3678289414 -0.340365854 -0.635656349 -0.246060870
#>  [57,]  0.14223657  0.6614328818  0.407166489 -0.982724957  0.513355896
#>  [58,]  0.16848341  0.3758461741 -0.483638406 -0.305629830  0.174086525
#>  [59,]  0.09971171  0.2053674832 -0.713676965  0.389685885  0.595007550
#>  [60,] -0.64462551 -0.6005245289 -0.102884436 -0.536373592 -0.915039526
#>  [61,]  0.06000530 -0.6148331668 -0.649591944  0.892154194 -0.391397481
#>  [62,] -0.26662001  0.3034775527 -0.546340141 -0.942944000 -0.674935373
#>  [63,]  0.47772692 -0.2336917478  0.288259826  0.034923263 -0.716263851
#>  [64,] -0.18610603 -0.4265176072  0.268010418  0.004975517  0.007716966
#>  [65,] -0.33258468 -0.7559048352  0.206502092 -0.088632853  0.353433901
#>  [66,]  0.29753267  0.1775280097  0.076753572 -0.688300417  0.011146421
#>  [67,]  0.67296401  0.0018080217 -0.018428593  0.931628119 -0.700795805
#>  [68,]  0.59045860  0.7369122826 -0.413886104 -0.298932064  0.529626513
#>  [69,] -0.98957385 -0.2707074173  0.087787441  0.299966357 -0.922754471
#>  [70,] -0.90546362 -0.6929099704 -0.477370291 -0.777118285  0.387609933
#>  [71,]  0.88526760  0.6982072792 -0.658314335 -0.666801866  0.773526150
#>  [72,]  0.16596305 -0.0957070319 -0.311848431  0.066153814  0.702264820
#>  [73,]  0.28350550  0.8885302749  0.566363586 -0.802793518  0.862243030
#>  [74,]  0.59823980  0.6580497739 -0.114813627  0.021226079  0.678890062
#>  [75,] -0.18983520 -0.5657964079  0.608117152 -0.446417684  0.073240134
#>  [76,] -0.56566627  0.6872398411 -0.426202063 -0.339183256 -0.493142758
#>  [77,]  0.42236159  0.5882115504  0.269707256  0.632644660 -0.497576451
#>  [78,]  0.97732127 -0.3499365658  0.131759572  0.271531601 -0.863092463
#>  [79,] -0.76596110 -0.7294887574 -0.457509706  0.627067422 -0.851450947
#>  [80,]  0.13355515 -0.4832436885  0.976006071 -0.836714661  0.551058542
#>  [81,] -0.33478602  0.7284630532 -0.482558321 -0.953508908 -0.446687458
#>  [82,] -0.31293029 -0.6589852087  0.160835415 -0.295697773 -0.722328526
#>  [83,]  0.52375773  0.5355139030 -0.840796527 -0.665580580  0.624922669
#>  [84,] -0.08668255 -0.6430170862  0.795067824 -0.717719333 -0.258498224
#>  [85,] -0.64430985 -0.1103189196 -0.287230909 -0.185575242  0.812919369
#>  [86,] -0.85066197 -0.2095707520  0.458419065  0.366414234 -0.646445145
#>  [87,] -0.54740838  0.7317233966 -0.016084402 -0.913396468  0.821033539
#>  [88,] -0.79675362  0.8306879117  0.937237419  0.973386722  0.150313936
#>  [89,]  0.25083737  0.7339889440  0.084782737 -0.418668354 -0.432745170
#>  [90,]  0.39943265  0.8527618293  0.229695146  0.090820526  0.011581767
#>  [91,] -0.45937387 -0.2279601442  0.992535870 -0.518059995 -0.187661030
#>  [92,]  0.28967094 -0.6098117451  0.729684711 -0.630034520 -0.973258061
#>  [93,] -0.11167393  0.2308219080 -0.978469219  0.415319021  0.737444734
#>  [94,]  0.44779387  0.4308621259  0.078044743 -0.903563724 -0.613465682
#>  [95,]  0.14049860  0.4390274361  0.374815587 -0.664202030  0.984168896
#>  [96,]  0.39289357  0.1680014338  0.100974615  0.751726185 -0.677646065
#>  [97,]  0.59440090  0.6372264018 -0.745262336 -0.860283767  0.005265999
#>  [98,] -0.84702534  0.4139877674 -0.360868810  0.128031216 -0.948919415
#>  [99,] -0.86505875  0.4452454150  0.758877346 -0.937063123 -0.711921081
#> [100,]  0.52711106  0.0885044606 -0.506982447 -0.204230643 -0.356625457
#> 
#> $y
#>   [1] 0 0 1 1 1 0 1 0 0 0 0 0 0 1 0 1 1 1 0 1 1 1 1 0 1 1 1 0 1 0 0 1 0 0 0 1 0
#>  [38] 1 0 0 0 1 0 1 1 0 1 1 0 1 0 1 1 0 0 1 1 1 1 0 1 1 1 1 0 1 1 0 0 0 0 1 0 0
#>  [75] 1 0 1 0 0 1 0 1 1 1 1 0 0 0 1 0 1 1 1 1 1 1 0 0 0 1
#>