Simulate Data from Hyper-Sphere for Classification Problems

The classification problem Circle is described in the BARK paper (2008). Inputs are dim independent variables uniformly distributed on the interval \([-1,1]\), only the first 2 out of these dim are actually signals. Outputs are created according to the formula $$y = 1(x1^2+x2^2 \le 2/\pi)$$

Usage

sim_circle(n, dim = 5)

Arguments

n

number of data points to generate

dim

number of dimension of the problem, no less than 2

Value

Returns a list with components

x

input values (independent variables)

y

0/1 output values (dependent variable)

References

Ouyang, Zhi (2008) Bayesian Additive Regression Kernels. Duke University. PhD dissertation, Chapter 3.

See also

Examples

sim_circle(n=100, dim=5)
#> $x
#>                [,1]        [,2]         [,3]        [,4]         [,5]
#>   [1,]  0.706993004 -0.37062222 -0.763979014 -0.42901460 -0.275916380
#>   [2,] -0.302650794  0.80879516 -0.928646183 -0.81409803 -0.151954745
#>   [3,]  0.969259051  0.53560777 -0.726040389 -0.75997814  0.430898484
#>   [4,] -0.447498299 -0.07598874  0.188552552  0.72486328 -0.959618382
#>   [5,] -0.379181049 -0.48802297  0.486539122  0.40308740  0.168590438
#>   [6,] -0.858079384 -0.55440789 -0.317194259  0.06805306  0.473846294
#>   [7,]  0.129324029 -0.27611437 -0.011012945 -0.50242764 -0.552619446
#>   [8,] -0.772296632  0.94045191  0.142316854 -0.16405555 -0.425474505
#>   [9,]  0.162575835 -0.20831596 -0.563249358  0.01143979  0.909117903
#>  [10,]  0.426438319 -0.37173786 -0.053284835  0.26078937  0.655659260
#>  [11,] -0.163482659 -0.53324122  0.393005719  0.03742102 -0.444453209
#>  [12,] -0.965751703 -0.30524205 -0.878543495  0.27331476 -0.651998334
#>  [13,] -0.708235602  0.48159092 -0.023498777  0.75961355  0.534337526
#>  [14,]  0.963869208  0.68550145  0.238104871  0.07321570 -0.353572968
#>  [15,] -0.124630868  0.36605820 -0.224460335  0.41768535  0.137570543
#>  [16,]  0.684394660 -0.39953451  0.337369736 -0.90150866 -0.110127120
#>  [17,]  0.515748670 -0.95772891 -0.368160862 -0.51984688 -0.939362644
#>  [18,]  0.459577614  0.69681014  0.237974981  0.38088867 -0.959197678
#>  [19,]  0.658264526 -0.20419411 -0.554101247 -0.92679310  0.985729286
#>  [20,] -0.117883774 -0.27940310 -0.513137409  0.68554834 -0.644532310
#>  [21,] -0.669482240 -0.96986578  0.310055974  0.95686780 -0.421715481
#>  [22,]  0.170180148  0.79697048  0.367255961  0.24981836  0.360601402
#>  [23,] -0.843411409 -0.29537113 -0.822167638 -0.01178988 -0.090633214
#>  [24,]  0.190194819 -0.42575592 -0.037395412 -0.64897819 -0.325905283
#>  [25,] -0.582845452  0.84231836 -0.081708990  0.37135228 -0.456746208
#>  [26,] -0.505740366 -0.73894956  0.981740207 -0.26945029  0.868116139
#>  [27,]  0.992619972  0.69484372  0.130720952 -0.03743892  0.820618868
#>  [28,]  0.167243291 -0.34288562 -0.487511191  0.32630491  0.221552849
#>  [29,]  0.512084180  0.28761226  0.439498413  0.04592795 -0.016418055
#>  [30,]  0.849254566 -0.27631643 -0.632279612 -0.07458396  0.758131688
#>  [31,] -0.626824392  0.98082695  0.222010286 -0.46120037 -0.086246508
#>  [32,]  0.655279195 -0.40451062 -0.473755036 -0.38410801  0.890165184
#>  [33,]  0.430711330 -0.75096960 -0.381174469  0.34618749 -0.887263902
#>  [34,] -0.729363055  0.02936725 -0.179492805  0.06732276 -0.677174714
#>  [35,]  0.710236865 -0.93218305  0.274512768 -0.82868124 -0.506500026
#>  [36,] -0.882956111 -0.90704264 -0.592386776  0.59073880  0.893238119
#>  [37,] -0.441281084  0.88121723  0.683682574  0.27974114  0.413646934
#>  [38,]  0.523792207 -0.30829446  0.237953197  0.86490232  0.478130619
#>  [39,] -0.351485149  0.97255073 -0.118057127 -0.16536941  0.282325194
#>  [40,] -0.612794026 -0.07717452 -0.184016302  0.35705045 -0.069857080
#>  [41,] -0.007328899 -0.31058322 -0.073335321  0.63683488  0.498725534
#>  [42,]  0.801576470  0.70916515 -0.007057341 -0.38615822  0.396875991
#>  [43,]  0.577513370  0.71905006  0.966155312 -0.47191814 -0.429188205
#>  [44,] -0.350568956  0.40576453  0.817885035 -0.91091404  0.009843513
#>  [45,] -0.755160375 -0.20718226 -0.233032524  0.45736915 -0.034996593
#>  [46,] -0.929700451 -0.54250899  0.303625233  0.08595099  0.407898404
#>  [47,]  0.688735549 -0.33991882  0.616436413 -0.72004597 -0.761329390
#>  [48,] -0.845720414 -0.05174603 -0.007514405 -0.01303011 -0.296915799
#>  [49,]  0.119977989 -0.71495206 -0.817611629 -0.66890352 -0.035909394
#>  [50,] -0.047669323 -0.98526947  0.451306363  0.22571417  0.318350208
#>  [51,]  0.541759068  0.90481278 -0.630130443 -0.41636559 -0.039202035
#>  [52,]  0.507480665  0.60143834  0.761958044 -0.18629366 -0.490893065
#>  [53,]  0.784350336 -0.30401833  0.914949915 -0.33016979  0.337871511
#>  [54,] -0.490403750  0.38658859 -0.410743251 -0.12158065 -0.607750161
#>  [55,] -0.782133851 -0.50893729 -0.942488151 -0.71063483  0.804443695
#>  [56,]  0.534958303  0.03387989  0.893210828 -0.88405489  0.406188048
#>  [57,] -0.299664643  0.84439336 -0.633251014  0.69269817  0.082283795
#>  [58,] -0.875138725 -0.69380625  0.334999512  0.09727243 -0.089847583
#>  [59,] -0.470399750  0.79915965  0.784817260 -0.98049509 -0.847508142
#>  [60,] -0.959896920  0.65465244  0.750787359  0.49260598  0.943701119
#>  [61,] -0.089777099  0.38393457 -0.303868199  0.88801001  0.203519344
#>  [62,]  0.237912958 -0.42043777  0.330472389  0.98659700  0.638291024
#>  [63,]  0.678863142 -0.99199997 -0.502359018  0.07526385 -0.949505619
#>  [64,] -0.495328160 -0.67785269  0.949954337  0.32689406 -0.537787156
#>  [65,] -0.337737189  0.70241657 -0.683887345  0.72851239 -0.102314050
#>  [66,]  0.523295840  0.26801204 -0.272705668 -0.91019343  0.290459691
#>  [67,]  0.025575547  0.08892597  0.998830001  0.01586426 -0.911093807
#>  [68,]  0.420097091  0.40407950 -0.865513552  0.22695279  0.989645719
#>  [69,] -0.779058672  0.76157472 -0.650091762  0.68516734 -0.864279497
#>  [70,]  0.363946366  0.09073205  0.646854450  0.69124584 -0.961051485
#>  [71,]  0.897132003  0.85190074  0.541853731 -0.42896046 -0.143846073
#>  [72,] -0.253199531  0.61605510 -0.461565665  0.90980247  0.475920978
#>  [73,] -0.326805288 -0.84906697 -0.127748404  0.79432802 -0.068160273
#>  [74,] -0.424105049  0.86419532  0.989631599 -0.83023990 -0.055725069
#>  [75,]  0.458646578  0.72488667 -0.648518366  0.78636871  0.243914231
#>  [76,]  0.174344580  0.53954266 -0.201642044  0.29876581 -0.184395330
#>  [77,] -0.272115953  0.62879307 -0.641431658 -0.87358435  0.691117218
#>  [78,]  0.817754670 -0.16170394  0.778160394 -0.86013644 -0.936527818
#>  [79,]  0.993880052  0.96773448 -0.922433479  0.79819445 -0.966230243
#>  [80,] -0.359810202 -0.68403846  0.127510761  0.16749386 -0.597810464
#>  [81,]  0.773616097  0.62664577 -0.233008737  0.81190825  0.426380794
#>  [82,]  0.822821118  0.28302241 -0.371584885 -0.63534091  0.307019234
#>  [83,]  0.707224625 -0.64105034 -0.569879910 -0.63150716  0.101325897
#>  [84,]  0.619946183 -0.85352471  0.774030133 -0.24553919  0.146542310
#>  [85,]  0.787664169  0.73285217 -0.721714623  0.42050538 -0.941954815
#>  [86,] -0.119036787 -0.01355405  0.774847896  0.69568837  0.564972604
#>  [87,] -0.728489595 -0.28321747  0.799673169  0.18383698 -0.340182085
#>  [88,]  0.723076505 -0.77728159  0.385762767 -0.56105877  0.394676098
#>  [89,] -0.429759785 -0.13555184 -0.256636357 -0.62696507  0.936411588
#>  [90,] -0.458990531  0.80532855 -0.809452094 -0.12445940  0.011596636
#>  [91,] -0.597713349 -0.61588257 -0.743555203 -0.32096931  0.325178888
#>  [92,]  0.702151226 -0.90791640  0.692803625  0.49422609 -0.248175260
#>  [93,] -0.220919174 -0.91837534 -0.687654909  0.97741043 -0.859226687
#>  [94,] -0.062624124  0.82839357 -0.877108746  0.49875566  0.577630738
#>  [95,] -0.234606890 -0.02334265 -0.835461885 -0.82903100 -0.370704219
#>  [96,]  0.694371874  0.37885927 -0.476652687 -0.06593732 -0.029676312
#>  [97,]  0.664089027  0.04614045  0.867700519 -0.37550976  0.073394978
#>  [98,] -0.286974705  0.20443514  0.043752163 -0.67043389 -0.786086577
#>  [99,] -0.229841419  0.71661748  0.624758616 -0.97509737 -0.477621094
#> [100,] -0.564981489  0.60426995  0.769791985  0.04660063  0.854390671
#> 
#> $y
#>   [1] 0 0 0 1 1 0 1 0 1 1 1 0 0 0 1 1 0 0 1 1 0 0 0 1 0 0 0 1 1 0 0 1 0 1 0 0 0
#>  [38] 1 0 1 1 0 0 1 1 0 1 0 1 0 0 1 0 1 0 1 0 0 0 0 1 1 0 0 1 1 1 1 0 1 0 1 0 0
#>  [75] 0 1 1 0 0 1 0 0 0 0 0 1 1 0 1 0 0 0 0 0 1 1 1 1 1 0
#>