Simulate Data from Hyper-Sphere for Classification Problems

The classification problem Circle is described in the BARK paper (2008). Inputs are dim independent variables uniformly distributed on the interval \([-1,1]\), only the first 2 out of these dim are actually signals. Outputs are created according to the formula $$y = 1(x1^2+x2^2 \le 2/\pi)$$

Usage

sim_circle(n, dim = 5)

Arguments

n

number of data points to generate

dim

number of dimension of the problem, no less than 2

Value

Returns a list with components

x

input values (independent variables)

y

0/1 output values (dependent variable)

References

Ouyang, Zhi (2008) Bayesian Additive Regression Kernels. Duke University. PhD dissertation, Chapter 3.

See also

Examples

sim_circle(n=100, dim=5)
#> $x
#>               [,1]          [,2]          [,3]        [,4]         [,5]
#>   [1,] -0.64646597 -0.1218805527 -0.3761649276 -0.40462760  0.617908903
#>   [2,] -0.92512329  0.9199783113 -0.9999162452  0.34993908  0.169254814
#>   [3,]  0.05908748 -0.4821120417  0.0842181514 -0.70990095 -0.505089656
#>   [4,] -0.95617515 -0.7344336007  0.7615781971  0.82570670  0.027558585
#>   [5,]  0.77755973 -0.9797849250  0.6919989665 -0.80241053 -0.619294124
#>   [6,]  0.18464929  0.5732608316 -0.9641770041  0.33215308 -0.515076360
#>   [7,] -0.42688751  0.8467093101 -0.6964784493  0.82938302 -0.497608200
#>   [8,] -0.59886936  0.6469515190 -0.5812163888  0.03797823 -0.108869311
#>   [9,]  0.36824604 -0.8488324676  0.7775064609 -0.23679413 -0.062773762
#>  [10,] -0.96108534  0.2968224920 -0.5678637782 -0.83376431  0.939888543
#>  [11,] -0.30544945  0.2446442237  0.5857034801  0.12819720  0.684814536
#>  [12,]  0.78248155 -0.5550338435  0.5582565484  0.65374012  0.883853769
#>  [13,] -0.13721330  0.6759943091  0.6151215006 -0.33139050  0.932480467
#>  [14,]  0.51727906 -0.3755456819 -0.1715758387 -0.97288908 -0.176595825
#>  [15,] -0.89029877  0.4416036387  0.5265245861 -0.91787565 -0.355950424
#>  [16,]  0.21839340 -0.1642554495  0.5582239050 -0.25124768  0.813869068
#>  [17,]  0.28874163  0.2705249661 -0.5319394125  0.62186320  0.193618661
#>  [18,] -0.71687344  0.1045900644 -0.5976631488 -0.67121367  0.634357674
#>  [19,]  0.62308953  0.3784432467  0.4436272387 -0.13284910 -0.405380875
#>  [20,] -0.84432983 -0.2642665175  0.0062562632  0.73175589  0.037496447
#>  [21,]  0.35644376 -0.7291814461  0.4903817358 -0.43304253  0.833784860
#>  [22,] -0.96016816 -0.4321136461  0.7963317730  0.80394773 -0.031972771
#>  [23,]  0.86327696 -0.5245793480  0.2872158936 -0.72679447 -0.541141267
#>  [24,] -0.78300554  0.8883582023  0.2228412083 -0.44353536  0.987783797
#>  [25,] -0.41869685  0.8031505658 -0.4683701112 -0.05173309  0.912009407
#>  [26,] -0.43880075 -0.3460550127  0.7138828631  0.63978240  0.453791587
#>  [27,] -0.04081195 -0.0734232864 -0.4914738284  0.93452500  0.686147240
#>  [28,] -0.60763516 -0.0046633286  0.0859583002  0.71667482 -0.496184003
#>  [29,] -0.53436104  0.0125577874  0.1487983959 -0.66974615  0.254407763
#>  [30,] -0.74259869 -0.9858435122  0.7024643864  0.08875098  0.745061174
#>  [31,] -0.14954404 -0.6876916802  0.4022655711 -0.56335172 -0.234120800
#>  [32,] -0.26595915  0.2446547002  0.5665411181 -0.74964604 -0.850030370
#>  [33,]  0.53071407  0.8239147291 -0.6197296949 -0.34998827  0.322276758
#>  [34,] -0.97995464  0.0764964130 -0.9909667885 -0.16243641  0.662357068
#>  [35,]  0.03634505  0.7331221350 -0.9649157519  0.40826082  0.190492382
#>  [36,] -0.55836371 -0.2706733067  0.6633575959  0.95767449  0.095853986
#>  [37,] -0.69976578  0.2518586917 -0.3366315290 -0.16624367 -0.763884953
#>  [38,]  0.79681558  0.2112932578 -0.4565859879  0.83069301  0.930617006
#>  [39,]  0.50153255 -0.9276192882 -0.1175305517  0.51164965  0.212252331
#>  [40,] -0.19469949 -0.2969390266  0.7412364767  0.73605845  0.532564599
#>  [41,]  0.05942734  0.4577946537  0.3338580625 -0.82570507  0.208643559
#>  [42,] -0.72376148 -0.9009305206 -0.4648987404  0.87893183 -0.510694237
#>  [43,]  0.25207274  0.5392810516  0.7784802527  0.63082252 -0.818687554
#>  [44,]  0.65171447 -0.6771856784  0.6648594900 -0.42437505  0.195732372
#>  [45,] -0.73230440 -0.8333396818 -0.3199064438 -0.73599933 -0.012311910
#>  [46,]  0.11282491 -0.3355734437 -0.0261488496 -0.77897857  0.670537080
#>  [47,]  0.29748233  0.8855838021 -0.4328304017 -0.81097688 -0.864383795
#>  [48,] -0.69554251  0.0002415627 -0.4753690888 -0.56734990 -0.679150190
#>  [49,] -0.73170241  0.5589209790 -0.8664896102  0.32367015 -0.552499267
#>  [50,] -0.17617558  0.0077213463 -0.2532781512  0.01665198 -0.818960930
#>  [51,] -0.05679216 -0.7553515378 -0.9364774586  0.95238903 -0.730860764
#>  [52,]  0.96191996  0.7509560832  0.1725718738  0.05456511  0.964857278
#>  [53,] -0.86414680 -0.6332036802 -0.0007224004 -0.97471781 -0.587854605
#>  [54,] -0.93187083 -0.7018237165  0.6530017015 -0.15499416 -0.826719541
#>  [55,]  0.94605041  0.9482241971  0.6122228024  0.31412296 -0.712347487
#>  [56,]  0.11618323  0.1061324165  0.0864977837  0.87663312  0.169816552
#>  [57,] -0.06199439 -0.4431765703  0.6586440001 -0.80646491 -0.918685583
#>  [58,] -0.46298283  0.8352341177  0.7651325269 -0.42601749 -0.567777698
#>  [59,] -0.87574798 -0.0154366111 -0.5436901315  0.44078512  0.442722287
#>  [60,] -0.67167707 -0.8640224775  0.9659848213  0.62130348  0.981687733
#>  [61,]  0.27328859 -0.5326425731  0.5077716857  0.57005892  0.673952119
#>  [62,] -0.65401814 -0.8071737713  0.1814859509  0.38359628 -0.689667091
#>  [63,]  0.52719655 -0.1417960324 -0.3781236531 -0.92917940  0.693223270
#>  [64,] -0.66576782 -0.0176433846 -0.2175687496 -0.90133503 -0.188736940
#>  [65,]  0.55164076 -0.1526833284 -0.3047628901  0.90475786 -0.651008775
#>  [66,] -0.07795059  0.3525639367 -0.3053013091  0.18123164  0.869983019
#>  [67,] -0.36323338 -0.1157016689  0.6700100373  0.28883925 -0.320021130
#>  [68,]  0.20528980 -0.9387572394 -0.8178959666 -0.91273051  0.503601422
#>  [69,]  0.70641665  0.8938016556  0.7113263435 -0.89265241 -0.727371186
#>  [70,] -0.88621286 -0.7259473056 -0.4164266633 -0.52445256  0.630720873
#>  [71,]  0.87527766 -0.7169674989 -0.0793138286  0.31553796  0.434582475
#>  [72,] -0.19008869  0.0422454798  0.7172407075  0.81652875 -0.399823739
#>  [73,] -0.90963178  0.5631051632  0.8954591532  0.27182286  0.413096312
#>  [74,]  0.99302735  0.4258571402 -0.5535986810 -0.51674362  0.344702499
#>  [75,]  0.03698706  0.7854943438  0.5342364437 -0.25862407  0.636782449
#>  [76,] -0.95801041 -0.9532389492 -0.6136733224 -0.96915263  0.004788013
#>  [77,] -0.65356225 -0.0417159805  0.0492215306  0.60620953 -0.298030014
#>  [78,]  0.65528804  0.2859381218 -0.9493275755 -0.05727343  0.448091556
#>  [79,]  0.11466904 -0.1401724508 -0.0459033423  0.63934788  0.191718793
#>  [80,] -0.17070808 -0.3909232235 -0.9831605242  0.81518727 -0.806326443
#>  [81,]  0.78630566 -0.4710201346  0.4802657547  0.59167222 -0.380784326
#>  [82,]  0.61208702 -0.8557303115 -0.6132925665 -0.15387644 -0.450050555
#>  [83,]  0.37293881  0.4537608321  0.2083499813  0.99414832  0.047011273
#>  [84,] -0.63207257 -0.1813035146 -0.8843296850 -0.28641188 -0.838076024
#>  [85,]  0.54460391  0.7975512864  0.3215072444  0.53386178 -0.227782871
#>  [86,]  0.28862937 -0.4882946075 -0.1959956959  0.74664938  0.710450771
#>  [87,]  0.01308774 -0.4264553268 -0.1606867188  0.69675256 -0.359778775
#>  [88,] -0.61011344 -0.6721279882  0.1890944974 -0.24902249  0.642762824
#>  [89,] -0.48276451 -0.1333716093  0.0917266440  0.76879458  0.156498127
#>  [90,] -0.61015822  0.4508568314  0.0640394879  0.82072535  0.682476670
#>  [91,]  0.88381108  0.6746103941  0.0541540636  0.81788187 -0.862563440
#>  [92,]  0.34918884 -0.3282319950 -0.9894649838  0.37464547  0.252953317
#>  [93,] -0.34949739  0.0992768742  0.3566258084 -0.04809342 -0.168443371
#>  [94,] -0.76301424 -0.3468540725  0.9369434156 -0.22792077 -0.207723090
#>  [95,]  0.64049207  0.6558934841  0.0180143421  0.57623876  0.622176878
#>  [96,]  0.09168134  0.1657288461 -0.3637729948  0.06911845  0.265909242
#>  [97,] -0.86208149 -0.6786781377  0.3812019206 -0.58256150  0.271733194
#>  [98,]  0.02674301 -0.1903021680  0.1342048789 -0.25610108  0.231514443
#>  [99,] -0.73461461 -0.2834774205 -0.7665267289 -0.84413894  0.746424411
#> [100,] -0.82136792  0.4365724083 -0.0515799639 -0.31993138 -0.378743351
#> 
#> $y
#>   [1] 1 0 1 0 0 1 0 0 0 0 1 0 1 1 0 1 1 1 1 0 0 0 0 0 0 1 1 1 1 0 1 1 0 0 1 1 1
#>  [38] 0 0 1 1 0 1 0 0 1 0 1 0 1 1 0 0 0 0 1 1 0 0 0 1 0 1 1 1 1 1 0 0 0 0 1 0 0
#>  [75] 1 0 1 1 1 1 0 0 1 1 0 1 1 0 1 1 0 1 1 0 0 1 0 1 1 0
#>