Simulate Data from Hyper-Sphere for Classification Problems

The classification problem Circle is described in the BARK paper (2008). Inputs are dim independent variables uniformly distributed on the interval \([-1,1]\), only the first 2 out of these dim are actually signals. Outputs are created according to the formula $$y = 1(x1^2+x2^2 \le 2/\pi)$$

sim_circle(n, dim = 5)

Arguments

n

number of data points to generate

dim

number of dimension of the problem, no less than 2

Value

Returns a list with components

x

input values (independent variables)

y

0/1 output values (dependent variable)

References

Ouyang, Zhi (2008) Bayesian Additive Regression Kernels. Duke University. PhD dissertation, Chapter 3.

See also

Examples

sim_circle(n=100, dim=5)
#> $x
#>                [,1]        [,2]         [,3]        [,4]        [,5]
#>   [1,]  0.501449167 -0.24971700  0.538276922 -0.71839675 -0.98460701
#>   [2,] -0.719805090  0.03518443 -0.340628827 -0.59378249 -0.92236711
#>   [3,]  0.487585030  0.04889534  0.777288577  0.06136816  0.54425774
#>   [4,] -0.484644183  0.09481878  0.092557827 -0.21733938 -0.27280968
#>   [5,] -0.425699984 -0.30948825  0.471934224  0.52601111 -0.95272096
#>   [6,]  0.790502063 -0.02022642  0.836467744  0.78320713 -0.40456905
#>   [7,]  0.815765382  0.53329890 -0.461682706 -0.56280600 -0.80511933
#>   [8,]  0.533351295 -0.54228319 -0.038453999  0.57274059  0.27967179
#>   [9,]  0.295425307  0.36734073  0.201546034  0.72184666  0.68748059
#>  [10,] -0.152071024 -0.33624532 -0.265370203 -0.39249368 -0.91980019
#>  [11,] -0.653419504  0.86093548 -0.929579723 -0.40213247 -0.23142824
#>  [12,]  0.608792847  0.31155969  0.319778957  0.18284307 -0.31036679
#>  [13,] -0.215300669  0.75350091  0.689653680 -0.72940521  0.92876097
#>  [14,] -0.879356619  0.31248414 -0.969091595 -0.29932308  0.05545003
#>  [15,] -0.858467439  0.33306558  0.335941257  0.53331800  0.18872337
#>  [16,]  0.609073974 -0.63055540 -0.419642702  0.94062958  0.78859360
#>  [17,]  0.079514064 -0.48318183 -0.344302811 -0.82225767 -0.74543525
#>  [18,] -0.381819504 -0.40192215 -0.167117254  0.32761150  0.46474546
#>  [19,] -0.914143792 -0.75790729  0.968858126 -0.89212815  0.43324688
#>  [20,] -0.813140885 -0.55569792  0.141678536 -0.16519523 -0.09924734
#>  [21,]  0.632370429 -0.07394727 -0.809108187  0.54838818 -0.89521094
#>  [22,]  0.516845292 -0.36075188  0.976496579 -0.82330557  0.97626401
#>  [23,]  0.320011576  0.68401774 -0.288219974  0.27216384 -0.83145203
#>  [24,] -0.702702314  0.94503941 -0.709304440  0.77198476  0.67714612
#>  [25,]  0.632045872  0.35821695  0.984566475  0.24706628  0.60073612
#>  [26,] -0.515365498  0.50482910 -0.983537257 -0.91712329 -0.64190753
#>  [27,] -0.540475430  0.42158447  0.267636275  0.18619739  0.32043366
#>  [28,]  0.604418446 -0.48698091  0.433107428 -0.04745331  0.29114031
#>  [29,] -0.108543896 -0.26224995 -0.244250754  0.65584201 -0.84969916
#>  [30,] -0.833095359  0.19279284  0.827846386  0.48833722  0.24795954
#>  [31,] -0.161057951 -0.04763941 -0.006738462  0.57923019 -0.97002574
#>  [32,] -0.002066765 -0.57992110  0.885084601  0.62645423 -0.99515594
#>  [33,]  0.500056264 -0.79301694 -0.523892625  0.21356313  0.21130740
#>  [34,]  0.574002977  0.61170244 -0.884210318  0.34663844  0.41292195
#>  [35,]  0.816074642  0.27197050 -0.753997163  0.78008662 -0.83269943
#>  [36,]  0.102193840  0.66757847 -0.239204434 -0.44532883  0.95778686
#>  [37,] -0.792734260 -0.73132070 -0.189930227  0.07914464  0.82907902
#>  [38,]  0.287768141  0.10061511 -0.564443226  0.37389209 -0.99068045
#>  [39,]  0.179187591  0.39893080 -0.069645613 -0.24902525 -0.02231952
#>  [40,]  0.633555984 -0.87398179  0.631862960 -0.50763600  0.26967782
#>  [41,] -0.093698323  0.09409854 -0.576130465 -0.75786959  0.82069312
#>  [42,]  0.040284728 -0.10738026  0.422071946  0.14230343  0.52169926
#>  [43,] -0.643759266  0.06639947  0.691788220  0.10620876  0.42418427
#>  [44,]  0.368894118  0.17151260  0.030474694 -0.35560606 -0.71850561
#>  [45,] -0.814730392 -0.57545221 -0.133637981  0.12436679 -0.57721337
#>  [46,]  0.643242721  0.47719848  0.659953326  0.33230525 -0.97723697
#>  [47,] -0.839137099 -0.95247752 -0.496456588 -0.54035422 -0.19568657
#>  [48,]  0.052312908 -0.62517609  0.014573934 -0.65292270 -0.72641631
#>  [49,]  0.134096615  0.31787211  0.666204698 -0.66360003 -0.82570263
#>  [50,]  0.185050010 -0.84628660 -0.628961137  0.46140686 -0.53749460
#>  [51,]  0.715948124 -0.66130461 -0.991211763 -0.60248250  0.79464069
#>  [52,] -0.796590735  0.98709151  0.127762475  0.69379396  0.04230779
#>  [53,]  0.108991105  0.07846944  0.359203276  0.88119667  0.24015828
#>  [54,]  0.410779111  0.70608500 -0.071735966  0.50092594  0.17509016
#>  [55,]  0.898671359 -0.09716108  0.167690508 -0.39582776  0.83258160
#>  [56,]  0.012790883 -0.37270995 -0.029315773  0.61682483  0.04190958
#>  [57,]  0.192522680 -0.91235926 -0.057739440  0.51631909 -0.42662506
#>  [58,] -0.858346303  0.12882457  0.608747286 -0.98061427 -0.71510979
#>  [59,] -0.121930776 -0.66198270 -0.378728546  0.93302144 -0.11704436
#>  [60,] -0.203560875  0.32528333 -0.420159431 -0.92958473  0.88921048
#>  [61,] -0.389540272  0.18899980  0.907998663 -0.87726348 -0.60811313
#>  [62,] -0.580660191 -0.98119380  0.987860666  0.30956975  0.32466119
#>  [63,] -0.051889537 -0.29425229 -0.955805311  0.92944107 -0.18880110
#>  [64,] -0.366668982 -0.73509354 -0.636364849 -0.94934269  0.65244002
#>  [65,]  0.297407616 -0.18273223 -0.460617561 -0.65244204  0.45119418
#>  [66,] -0.516421826  0.58491696  0.344947484  0.31373746  0.73676756
#>  [67,] -0.022674046  0.89608334  0.677949744 -0.71746973 -0.67946311
#>  [68,]  0.862101067 -0.02489986  0.504519729 -0.60012739  0.24192618
#>  [69,] -0.364282807 -0.61545312  0.058674928 -0.30412057 -0.15397182
#>  [70,]  0.941289379 -0.96145347 -0.205161681  0.73251420 -0.45566637
#>  [71,]  0.233466047  0.02264707 -0.171695773  0.96633605  0.68419863
#>  [72,] -0.446290485  0.32448265  0.196103354 -0.22512976  0.66737635
#>  [73,]  0.565185077 -0.39115634 -0.819034111 -0.56171301 -0.83116102
#>  [74,] -0.543253996 -0.69082481  0.617770472 -0.77379970  0.67359087
#>  [75,] -0.580877572  0.13223196  0.580330006 -0.97856067  0.31019530
#>  [76,]  0.947085502 -0.37691456 -0.251352350 -0.37469010  0.38381565
#>  [77,] -0.953454677 -0.34000532  0.240102973  0.03772633 -0.61929627
#>  [78,] -0.563034865  0.07337226  0.996492502 -0.80224029  0.80024270
#>  [79,] -0.181081802  0.42501785  0.712947429 -0.89504723 -0.38352451
#>  [80,] -0.537850168 -0.05779684  0.783020337 -0.71657233  0.77753415
#>  [81,] -0.026924951  0.65495196 -0.041063872 -0.99324313  0.68188547
#>  [82,]  0.292215668 -0.21463472  0.623240253  0.93470210 -0.35042810
#>  [83,]  0.325070756  0.76173028 -0.512349484  0.30656316  0.45901640
#>  [84,] -0.019665588 -0.58466501  0.006088303  0.84826244 -0.15290988
#>  [85,]  0.184338556 -0.85564540 -0.119604976  0.88735232 -0.44018640
#>  [86,] -0.407928800  0.94316768  0.035392735 -0.49736132 -0.12743506
#>  [87,] -0.215276748  0.07141055  0.468206453 -0.92570702  0.65822850
#>  [88,] -0.546770881 -0.18163600 -0.881794231 -0.36040525  0.30972143
#>  [89,]  0.931241412  0.27711067  0.702157580  0.20456547  0.08807019
#>  [90,]  0.736036167  0.19293054  0.915998050 -0.53374061  0.65969097
#>  [91,]  0.199881350  0.98082490 -0.889195351  0.44471001 -0.21935779
#>  [92,] -0.316596724 -0.64677627 -0.302686473 -0.29864682 -0.38360741
#>  [93,] -0.914580604  0.51143017  0.734486763 -0.10818125  0.92128625
#>  [94,]  0.726617440  0.34506982  0.533864390  0.35865496  0.24576415
#>  [95,] -0.636861805 -0.21867009  0.870267057  0.29234739 -0.58701064
#>  [96,] -0.814594842  0.68207514  0.677080659 -0.67094989 -0.45498865
#>  [97,] -0.534370970  0.03431731 -0.850289090  0.44572057 -0.04700683
#>  [98,] -0.973289995  0.86858203 -0.252331949 -0.73551284  0.91155465
#>  [99,]  0.006844700  0.25695464 -0.103581466 -0.56156720 -0.64083962
#> [100,]  0.223868383  0.84598693 -0.616222030  0.14765012  0.80579311
#> 
#> $y
#>   [1] 1 1 1 1 1 1 0 1 1 1 0 1 1 0 0 0 1 1 0 0 1 1 1 0 1 1 1 1 1 0 1 1 0 0 0 1 0
#>  [38] 1 1 0 1 1 1 1 0 0 0 1 1 0 0 0 1 0 0 1 0 0 1 1 1 0 1 0 1 1 0 0 1 0 1 1 1 0
#>  [75] 1 0 0 1 1 1 1 1 0 1 0 0 1 1 0 1 0 1 0 0 1 0 1 0 1 0
#>