Simulate Data from Hyper-Sphere for Classification Problems

The classification problem Circle is described in the BARK paper (2008). Inputs are dim independent variables uniformly distributed on the interval \([-1,1]\), only the first 2 out of these dim are actually signals. Outputs are created according to the formula $$y = 1(x1^2+x2^2 \le 2/\pi)$$

Usage

sim_circle(n, dim = 5)

Arguments

n

number of data points to generate

dim

number of dimension of the problem, no less than 2

Value

Returns a list with components

x

input values (independent variables)

y

0/1 output values (dependent variable)

References

Ouyang, Zhi (2008) Bayesian Additive Regression Kernels. Duke University. PhD dissertation, Chapter 3.

See also

Examples

sim_circle(n=100, dim=5)
#> $x
#>                [,1]         [,2]         [,3]         [,4]        [,5]
#>   [1,] -0.352974321 -0.711610991 -0.698420495  0.278021929  0.77258019
#>   [2,]  0.055838876  0.499371284  0.388372199 -0.968929235 -0.22324955
#>   [3,]  0.760787531 -0.984619496 -0.703016418 -0.878496889  0.55287946
#>   [4,]  0.427730148  0.795631275 -0.729253448  0.126944686  0.63386504
#>   [5,]  0.619700197 -0.704968907 -0.498850465  0.525559936 -0.66485614
#>   [6,] -0.352168766 -0.149599686 -0.128850019  0.161909894 -0.37657538
#>   [7,]  0.578386599  0.454685360  0.467046685  0.577470888  0.66589411
#>   [8,] -0.085011397 -0.654590170 -0.213323608  0.629110321  0.67502939
#>   [9,]  0.980463146  0.013583009  0.623330344  0.708865152 -0.26546473
#>  [10,]  0.478339882  0.343903025  0.925702715  0.351742866 -0.87531142
#>  [11,] -0.928569378  0.559081445 -0.902318215  0.477630701  0.81937364
#>  [12,]  0.386757046  0.828393775  0.848118590 -0.732713481  0.65754474
#>  [13,]  0.895264454 -0.391628572  0.803085731 -0.841141298  0.79220676
#>  [14,]  0.711752126 -0.609223538 -0.169842359 -0.517672851  0.30941973
#>  [15,]  0.306804288 -0.168061148 -0.278424815  0.409838630  0.58368170
#>  [16,] -0.725063353  0.241544285 -0.034038405 -0.710698980  0.10827352
#>  [17,]  0.501767884  0.499719221 -0.749489189  0.691669873  0.64788480
#>  [18,] -0.290225171 -0.773005101  0.593931065  0.378455446  0.42186041
#>  [19,]  0.336200420  0.257293168 -0.403116536 -0.725751389  0.83910176
#>  [20,] -0.061849216  0.947082655  0.910079082 -0.965692321  0.68178380
#>  [21,] -0.764273370  0.070799974 -0.213817177 -0.806930731 -0.20400977
#>  [22,] -0.482837912  0.630181134 -0.083155113  0.162491208 -0.75017951
#>  [23,]  0.971218350  0.573956672 -0.194958382  0.492309546  0.66240052
#>  [24,]  0.846321775 -0.444773891 -0.446435508 -0.741999594 -0.47937668
#>  [25,]  0.857953531 -0.950533285 -0.279769917  0.125939867  0.36830244
#>  [26,]  0.201621373 -0.398338426  0.569866842 -0.340236027 -0.55079611
#>  [27,] -0.011255330  0.311748089  0.893579669 -0.117069748 -0.14265431
#>  [28,]  0.676737643  0.979089309 -0.922533013  0.085497496  0.28796786
#>  [29,]  0.513163293 -0.345652227  0.907265358  0.073642749 -0.57695791
#>  [30,]  0.780968619  0.387663935 -0.724473434 -0.279327132 -0.35118567
#>  [31,]  0.550974440 -0.005336067 -0.392543603 -0.026888455 -0.21057096
#>  [32,] -0.800927727 -0.114321876 -0.736522387  0.697534204  0.87356296
#>  [33,] -0.571741745 -0.522583198 -0.059191026 -0.966957088  0.92408449
#>  [34,]  0.893259271 -0.398800692 -0.095468502  0.646142294 -0.39789354
#>  [35,] -0.138622961 -0.444281001  0.361119708  0.871885919  0.48304066
#>  [36,] -0.133776439  0.227146899  0.412801870  0.198577572  0.42127916
#>  [37,]  0.646676733 -0.577343957  0.190188542  0.001424045 -0.74405193
#>  [38,]  0.566241926  0.213471137  0.948677962  0.931013864 -0.94356810
#>  [39,]  0.145904924  0.198903080  0.523760527 -0.692673913 -0.93215592
#>  [40,] -0.138887615  0.920911469  0.439828384 -0.177312222 -0.46087041
#>  [41,]  0.206092326  0.276917411 -0.279852421 -0.183786387 -0.50457395
#>  [42,]  0.143828552  0.656262537  0.109919498  0.165231866  0.11183812
#>  [43,] -0.456664942 -0.298037316 -0.849818817  0.297018748  0.66701876
#>  [44,] -0.035077605 -0.974219132  0.823406543 -0.501128052 -0.30308165
#>  [45,]  0.033101975  0.969145328  0.658266599  0.239732390 -0.60313122
#>  [46,]  0.023681640  0.639937339  0.979482711  0.601036558 -0.51997791
#>  [47,] -0.681211712  0.133628312  0.141830927  0.126339460 -0.52794314
#>  [48,] -0.168337882  0.500618001 -0.776888964  0.968351035 -0.19606757
#>  [49,]  0.245305747 -0.219949563 -0.332024754 -0.307300481  0.77533328
#>  [50,]  0.005963032 -0.869688824  0.935963638  0.109371751  0.87343919
#>  [51,]  0.885106892 -0.882336781 -0.947085072  0.011479336  0.57289170
#>  [52,]  0.255904828 -0.852160634  0.231850347  0.228477562 -0.38298611
#>  [53,]  0.951079261  0.851286208 -0.449915650 -0.568490366 -0.05016633
#>  [54,] -0.757116079 -0.444078530 -0.422541961  0.115795807  0.21912717
#>  [55,] -0.887731348  0.620936065 -0.389404894 -0.435105427  0.18002378
#>  [56,]  0.401194914 -0.536850682  0.045920546 -0.286209244 -0.53342324
#>  [57,]  0.039726666 -0.133621190 -0.763012497  0.980417629  0.90313421
#>  [58,] -0.417750971  0.047690721  0.667665719  0.057117397  0.53262090
#>  [59,] -0.232909947 -0.663930712  0.120656772 -0.713622187  0.02694739
#>  [60,] -0.276657749 -0.005439890  0.010071191  0.804032077 -0.51096366
#>  [61,]  0.899303574 -0.162769879  0.759946962  0.695565650 -0.81284500
#>  [62,]  0.302729480  0.786873348 -0.784975864 -0.736130753  0.96884915
#>  [63,]  0.059922957  0.965047651 -0.094359742  0.577320467  0.64727099
#>  [64,]  0.828035067  0.831339467 -0.181466121 -0.183056990  0.82107375
#>  [65,] -0.991312300  0.878148895 -0.379444978  0.948288461  0.97654810
#>  [66,]  0.167704987 -0.876976418 -0.154981720 -0.354442406 -0.73313520
#>  [67,] -0.903623092 -0.557962771  0.870309726 -0.702038080  0.24324331
#>  [68,] -0.556879438  0.510474560  0.791047634 -0.590370311 -0.28184388
#>  [69,] -0.344679767 -0.981154863 -0.364596903 -0.972320324  0.65544418
#>  [70,]  0.156838991 -0.316305848  0.336715298 -0.999904350  0.86239190
#>  [71,] -0.572359192  0.492510788  0.673645185 -0.053543823 -0.22100115
#>  [72,]  0.594821876 -0.297857841  0.937747485  0.719659221 -0.75142782
#>  [73,]  0.957408299 -0.526235798 -0.181672267  0.945309596 -0.09118611
#>  [74,]  0.606940054 -0.275507024  0.173201929 -0.485722492 -0.94284728
#>  [75,]  0.299804711  0.682377012 -0.894655713  0.962045027 -0.42502368
#>  [76,]  0.146357492  0.733229953  0.887064874  0.866670881 -0.10771398
#>  [77,] -0.327339920 -0.309103663  0.618493257 -0.602756824  0.67501316
#>  [78,] -0.985104427  0.957245295 -0.453438491 -0.715632644  0.32431925
#>  [79,]  0.002663887 -0.614940837 -0.371552153 -0.972935784 -0.92388547
#>  [80,] -0.100810189  0.394965428 -0.646410220  0.670143455  0.43792671
#>  [81,] -0.269473146  0.746485689 -0.475480231 -0.562528326  0.31757351
#>  [82,]  0.331487339 -0.678627741  0.300046452  0.216213854  0.86672192
#>  [83,] -0.365304571  0.533822385 -0.922105654  0.005540617 -0.08854372
#>  [84,]  0.330116090  0.490353451  0.147846723  0.909719925 -0.09019551
#>  [85,] -0.882692622 -0.907200960 -0.101213305  0.960081626  0.67614105
#>  [86,] -0.893775233  0.801585127 -0.001334836 -0.385859107  0.93371359
#>  [87,]  0.739554170  0.317020575 -0.718816583  0.227962258 -0.29772517
#>  [88,] -0.011578127  0.689949079 -0.407200293 -0.572084447  0.93662793
#>  [89,]  0.921349005  0.688356094 -0.957129108 -0.237979825 -0.92541414
#>  [90,] -0.352873011 -0.941355621  0.995022517 -0.187379707 -0.14359835
#>  [91,] -0.567188596 -0.471559841 -0.685468667 -0.827368671  0.56893335
#>  [92,] -0.821049535  0.759625512  0.619916372 -0.862192272  0.67396026
#>  [93,] -0.558253764  0.340804594  0.116017314 -0.005192131 -0.82822288
#>  [94,]  0.296655966  0.582843004 -0.370492995  0.919572617 -0.31910352
#>  [95,]  0.527980676 -0.903564297  0.667855447 -0.677049610  0.07493216
#>  [96,] -0.912210106  0.009721672 -0.257322215  0.176962269 -0.64212401
#>  [97,] -0.001528773  0.182845413  0.063451027 -0.692061102 -0.46698058
#>  [98,]  0.136365665 -0.881790068 -0.774156684 -0.195199919  0.97362667
#>  [99,] -0.123322636 -0.529367649 -0.708776109 -0.346843130  0.79002311
#> [100,]  0.564824431 -0.148866439  0.561331617 -0.739708447  0.58254456
#> 
#> $y
#>   [1] 1 1 0 0 0 1 1 1 0 1 0 0 0 0 1 1 1 0 1 0 1 1 0 0 0 1 1 0 1 0 1 0 1 0 1 1 0
#>  [38] 1 1 0 1 1 1 0 0 1 1 1 1 0 0 0 0 0 0 1 1 1 1 1 0 0 0 0 0 0 0 1 0 1 1 1 0 1
#>  [75] 1 1 1 0 1 1 1 1 1 1 0 0 0 1 0 0 1 0 1 1 0 0 1 0 1 1
#>