Simulate Data from Hyper-Sphere for Classification Problems

The classification problem Circle is described in the BARK paper (2008). Inputs are dim independent variables uniformly distributed on the interval \([-1,1]\), only the first 2 out of these dim are actually signals. Outputs are created according to the formula $$y = 1(x1^2+x2^2 \le 2/\pi)$$

Usage

sim_circle(n, dim = 5)

Arguments

n

number of data points to generate

dim

number of dimension of the problem, no less than 2

Value

Returns a list with components

x

input values (independent variables)

y

0/1 output values (dependent variable)

References

Ouyang, Zhi (2008) Bayesian Additive Regression Kernels. Duke University. PhD dissertation, Chapter 3.

See also

Examples

sim_circle(n=100, dim=5)
#> $x
#>                [,1]         [,2]         [,3]          [,4]        [,5]
#>   [1,]  0.414519012 -0.275718892 -0.614319277 -0.7106174664 -0.55491085
#>   [2,]  0.859606497  0.410578476  0.737352552 -0.3236148679  0.17290557
#>   [3,]  0.683713114 -0.356301605 -0.825221025 -0.6182013117 -0.14481444
#>   [4,]  0.689242663  0.964152752 -0.194894444 -0.9070302239  0.23088807
#>   [5,]  0.228120845  0.060160209  0.210781631 -0.0615045796 -0.68987282
#>   [6,]  0.801160565 -0.748625625  0.459254549 -0.5589625984  0.26632730
#>   [7,] -0.697706859 -0.786599881 -0.537077027 -0.6159874545 -0.74202619
#>   [8,]  0.129444794 -0.212159057 -0.640396000 -0.6590448809 -0.92099355
#>   [9,]  0.547162758  0.539413481 -0.519600366  0.0497388551  0.29340905
#>  [10,]  0.967179054 -0.519575452  0.417296572  0.8323067972  0.98768850
#>  [11,]  0.540259407  0.374082611 -0.080765238  0.6845134264  0.77272432
#>  [12,]  0.160655620  0.590820493 -0.954406624 -0.5592903928  0.13808725
#>  [13,] -0.371314870 -0.929493726  0.044020887 -0.9313845257 -0.65861604
#>  [14,] -0.704409421  0.823909868 -0.897287091  0.0663600024  0.59511342
#>  [15,]  0.308422099 -0.581775633 -0.130738055 -0.2371019283 -0.88202902
#>  [16,]  0.018960869  0.876075987 -0.504388064 -0.5412956127 -0.88212711
#>  [17,] -0.539958678  0.361625562 -0.048823855 -0.0706361397  0.67919424
#>  [18,]  0.617471191 -0.403457950 -0.611059804 -0.9495159574  0.87240942
#>  [19,]  0.703198973 -0.805913961  0.118424721 -0.6821453520  0.50090430
#>  [20,] -0.467346746  0.969870150 -0.024961573 -0.0122595360 -0.12204445
#>  [21,] -0.942591042  0.253040283  0.365914736 -0.3277307726  0.76109110
#>  [22,]  0.458029442  0.569698633 -0.079042123 -0.3470137557 -0.71075716
#>  [23,]  0.846044797  0.407107586  0.378155733 -0.0989794182  0.69815608
#>  [24,] -0.267179757 -0.013664800  0.061473862 -0.5776381237  0.46719001
#>  [25,] -0.652968678  0.562628634 -0.803832633  0.5063589057  0.31304647
#>  [26,] -0.463931375  0.092338063 -0.328237218 -0.9959512656 -0.87304508
#>  [27,] -0.721700556 -0.159271325  0.139916032  0.2991091013 -0.49829781
#>  [28,]  0.260766146  0.211515248 -0.434187038 -0.3818032895  0.95068164
#>  [29,]  0.060423969 -0.002413395  0.373926560 -0.1893377635 -0.32713608
#>  [30,]  0.109949652  0.766748853 -0.451999359  0.5977904797  0.07468641
#>  [31,] -0.759341593 -0.120431865 -0.796918896  0.6955921459 -0.97755840
#>  [32,]  0.774526765 -0.303012628  0.700662387  0.6744861864 -0.79954701
#>  [33,] -0.543271499  0.028997723  0.429451623  0.7772248746 -0.69216502
#>  [34,] -0.888736215 -0.314324311 -0.984935091 -0.6475986568 -0.48381881
#>  [35,]  0.140050402  0.370190168  0.436341156 -0.9178966735  0.81525311
#>  [36,] -0.687339805  0.668499141  0.133759269  0.0854557259 -0.09290382
#>  [37,]  0.126481246 -0.793302104 -0.045848290 -0.5829291916 -0.68981547
#>  [38,] -0.089261594 -0.365421179  0.777611165  0.5188205009 -0.45046641
#>  [39,] -0.776364387 -0.049253110 -0.576287332  0.3478918066 -0.59083008
#>  [40,] -0.941094392 -0.960137085 -0.608026531  0.4652204616  0.88035944
#>  [41,]  0.571880723  0.245090856  0.406352274 -0.9174191365  0.63313401
#>  [42,] -0.843002565 -0.414400391  0.929611418  0.9838057072  0.64054066
#>  [43,]  0.256499927  0.515299986 -0.585768457  0.2930315309  0.11749537
#>  [44,] -0.344226911  0.915052961 -0.363214061  0.4699167069  0.85851037
#>  [45,] -0.914580000  0.033638707  0.317668707  0.3878706051  0.61885003
#>  [46,]  0.920122927 -0.577048827  0.436311246 -0.2600937281  0.07713640
#>  [47,]  0.021564804  0.549950098  0.301702045 -0.5980046326  0.63110719
#>  [48,]  0.584919540  0.071607341  0.276137101 -0.4725949634  0.78228386
#>  [49,] -0.538840258 -0.638681370  0.708895315  0.9672334022  0.36314469
#>  [50,]  0.650248971 -0.798779104 -0.567281870 -0.9439332434  0.86399059
#>  [51,] -0.079130088  0.071174607 -0.496100031  0.3402267764  0.96052786
#>  [52,] -0.440367977 -0.753729624  0.677806250  0.0769242011 -0.25566302
#>  [53,] -0.947556580 -0.221014656  0.055877308  0.9550582613  0.77163302
#>  [54,]  0.463251709 -0.230976547  0.471008929  0.3266282966  0.68048213
#>  [55,] -0.818963197  0.181321383  0.985160987 -0.4482655623 -0.63729009
#>  [56,]  0.061521855  0.379540058  0.227604026 -0.5125150727 -0.34389009
#>  [57,] -0.482434375 -0.943472709  0.427166145 -0.1367577431  0.53418304
#>  [58,]  0.366578677  0.867570936  0.975379479 -0.3317498630  0.04973315
#>  [59,] -0.003778959  0.997774944  0.842435888  0.4415001990  0.91913935
#>  [60,] -0.844559710 -0.667091428 -0.933417240  0.0799557907 -0.56742697
#>  [61,]  0.803801762 -0.554038088  0.116526801  0.4279813790  0.64485853
#>  [62,] -0.915782742  0.329709083 -0.935087798 -0.5016861754  0.81249869
#>  [63,] -0.339968981  0.428498656 -0.315713502 -0.2915313700 -0.93849491
#>  [64,] -0.670526468 -0.471015289  0.229621980  0.3846473913  0.32846904
#>  [65,]  0.345937140 -0.198260924  0.606011887 -0.0889559695 -0.95602942
#>  [66,] -0.937370687  0.078908680  0.152746769 -0.2314836150  0.40496328
#>  [67,] -0.755529943 -0.755927266  0.093709344 -0.5947816214 -0.37029299
#>  [68,] -0.329013356 -0.712599999  0.605550742 -0.3734084242 -0.51950360
#>  [69,] -0.908018463  0.308873350 -0.542074665  0.6858079252 -0.63083061
#>  [70,] -0.071022648  0.395700615 -0.313429005  0.1022992963 -0.94319356
#>  [71,] -0.741619173 -0.731125911 -0.269515748  0.4893292920  0.42420629
#>  [72,]  0.699481181  0.558614918  0.112021744  0.6021359297  0.77918523
#>  [73,] -0.307831225 -0.560766971  0.244590359 -0.1826597997 -0.57250837
#>  [74,]  0.526977063 -0.460056894  0.900229487 -0.3182419920  0.77062954
#>  [75,] -0.908148494  0.484050676  0.519051737  0.7787112575 -0.39343994
#>  [76,]  0.456257924  0.507126486 -0.257738057  0.1935362117 -0.48669484
#>  [77,] -0.307093991 -0.756055602  0.893108991  0.8044773974  0.78259838
#>  [78,]  0.102257447 -0.528800495 -0.007594725  0.0858423207 -0.47899455
#>  [79,]  0.681912869  0.045349820 -0.865065005 -0.3364015031 -0.45034885
#>  [80,] -0.315161101  0.188587438 -0.295598523  0.5246876394 -0.11734612
#>  [81,] -0.952922297 -0.750055623  0.370664136 -0.5800035135 -0.07600669
#>  [82,] -0.236222711 -0.209905930  0.816062092  0.5089696976  0.80334454
#>  [83,]  0.685174853 -0.333104013  0.101518150  0.0203207484 -0.67870759
#>  [84,] -0.080082848 -0.911828911  0.897171333 -0.1249843976 -0.94412587
#>  [85,]  0.265342423 -0.493298737 -0.210815711  0.0223104437 -0.41034007
#>  [86,] -0.809024091  0.469981609 -0.517061653 -0.5735857836 -0.47875779
#>  [87,]  0.465491574 -0.811252221  0.850132280 -0.2569200550  0.47995119
#>  [88,]  0.106130742 -0.164025295 -0.436071162  0.2999167158  0.98727577
#>  [89,]  0.734488414 -0.151900387  0.275155783 -0.8299212568 -0.48495457
#>  [90,] -0.556419922  0.460171610 -0.315927140  0.3443619781  0.72778277
#>  [91,]  0.373111401 -0.076662773  0.136201203 -0.8092438718 -0.28677544
#>  [92,]  0.373531295  0.598825652 -0.499146804  0.3139700997 -0.82693584
#>  [93,] -0.217518733 -0.336425686  0.009707028 -0.0007044552  0.16343421
#>  [94,] -0.353256494 -0.851263850  0.999012191  0.5856752549  0.57828598
#>  [95,]  0.727556931 -0.890859405 -0.120729603  0.0692915963  0.61880498
#>  [96,]  0.939800965 -0.886509988 -0.893139858 -0.9750685813 -0.32793596
#>  [97,] -0.792390975  0.013471225 -0.861074526 -0.4819134874  0.43580055
#>  [98,]  0.490252362 -0.281527125 -0.452464832  0.3575887824  0.45613892
#>  [99,] -0.942432589  0.738675226  0.581069201  0.5517343394 -0.71724784
#> [100,] -0.474742211  0.018814272 -0.613668814 -0.9985567410 -0.09875754
#> 
#> $y
#>   [1] 1 0 1 0 1 0 0 1 1 0 1 1 0 0 1 0 1 1 0 0 0 1 0 1 0 1 1 1 1 1 1 0 1 0 1 0 0
#>  [38] 1 1 0 1 0 1 0 0 0 1 1 0 0 1 0 0 1 0 1 0 0 0 0 0 0 1 0 1 0 0 1 0 1 0 0 1 1
#>  [75] 0 1 0 1 1 1 0 1 1 0 1 0 0 1 1 1 1 1 1 0 0 0 1 1 0 1
#>