Simulate Data from Hyper-Sphere for Classification Problems

The classification problem Circle is described in the BARK paper (2008). Inputs are dim independent variables uniformly distributed on the interval \([-1,1]\), only the first 2 out of these dim are actually signals. Outputs are created according to the formula $$y = 1(x1^2+x2^2 \le 2/\pi)$$

sim_circle(n, dim = 5)

Arguments

n

number of data points to generate

dim

number of dimension of the problem, no less than 2

Value

Returns a list with components

x

input values (independent variables)

y

0/1 output values (dependent variable)

References

Ouyang, Zhi (2008) Bayesian Additive Regression Kernels. Duke University. PhD dissertation, Chapter 3.

See also

Examples

sim_circle(n=100, dim=5)
#> $x
#>                [,1]        [,2]         [,3]         [,4]        [,5]
#>   [1,] -0.795975273 -0.83455102 -0.914864898  0.873624728 -0.47556414
#>   [2,]  0.081210823  0.51165539  0.763688720 -0.469538142  0.68309259
#>   [3,]  0.249115758  0.26036336  0.700838565  0.288169152  0.70908332
#>   [4,] -0.469563099  0.49756453  0.004875764 -0.364620620  0.94451759
#>   [5,]  0.176379916 -0.15853506  0.959615911  0.759296890 -0.19959675
#>   [6,]  0.748112624  0.95386377  0.985519053  0.949629900 -0.32573770
#>   [7,] -0.069415701  0.02678022  0.492643568  0.217526568 -0.56246268
#>   [8,]  0.690255768 -0.42985440 -0.876071429  0.331876006  0.92125108
#>   [9,]  0.646109329  0.78309561  0.440468958 -0.676447733 -0.27202533
#>  [10,]  0.201391270  0.88577027  0.461581989 -0.252302285 -0.05572495
#>  [11,] -0.991922130  0.68951639 -0.294698823 -0.704539492 -0.42977883
#>  [12,] -0.023786280  0.80223931 -0.755023335  0.647659363  0.31443070
#>  [13,] -0.480432615  0.52521742 -0.445413795  0.570590508 -0.99789413
#>  [14,]  0.401718850 -0.35117066 -0.331714090 -0.021350177 -0.39362596
#>  [15,] -0.453289224 -0.02553331  0.685655444 -0.768984324  0.25057465
#>  [16,] -0.891372127 -0.67798824  0.994799955 -0.841057342  0.36079693
#>  [17,] -0.070534421 -0.81430469 -0.654697615  0.440606091  0.44646195
#>  [18,] -0.927980902 -0.00891479  0.365414780 -0.334290296  0.27580728
#>  [19,] -0.937541129 -0.97708689 -0.307277287 -0.849664729  0.33217740
#>  [20,] -0.648318090  0.92553901 -0.496736653  0.437134605 -0.33237458
#>  [21,] -0.637436693 -0.32777584 -0.217302158 -0.191467206 -0.18211447
#>  [22,] -0.449145110  0.53153245  0.634439101 -0.428146446  0.01980882
#>  [23,]  0.861185791 -0.05055535  0.534056944 -0.929138850  0.38184166
#>  [24,] -0.443710919 -0.45494007 -0.294530669 -0.828226193 -0.11449471
#>  [25,]  0.957914384 -0.44617433 -0.069956963 -0.702816917  0.47752449
#>  [26,]  0.212940574  0.95466399 -0.060589917  0.430067572  0.41103880
#>  [27,] -0.532678023 -0.87472960 -0.949694131  0.777513884  0.24149295
#>  [28,] -0.883460434 -0.95391082  0.669464679  0.802046152 -0.19497415
#>  [29,]  0.049279333  0.28079680  0.123664362 -0.388660627 -0.04612837
#>  [30,]  0.043145645  0.22934840  0.608201011  0.821488865  0.98767675
#>  [31,] -0.363816862 -0.35438178 -0.847659802 -0.035444295 -0.08238590
#>  [32,] -0.365615809  0.16517717  0.215022349  0.672440513  0.95160210
#>  [33,]  0.420516243 -0.66209217 -0.411388596 -0.117265943 -0.63760264
#>  [34,] -0.986477111  0.94860110 -0.135829072  0.384724849 -0.55830416
#>  [35,]  0.476988560  0.15019137 -0.990373385  0.830030049 -0.97808417
#>  [36,] -0.153369287 -0.99633736  0.925915155 -0.109662570 -0.86220786
#>  [37,] -0.881794927  0.25073892  0.120288496  0.923465356  0.58677350
#>  [38,]  0.259695479 -0.95748837  0.774071727 -0.173056623 -0.42708377
#>  [39,]  0.234532485 -0.45896226  0.221087706 -0.486115746 -0.99310599
#>  [40,]  0.027185909  0.28389631  0.381865428  0.280039405  0.05653796
#>  [41,] -0.074817555 -0.62335724 -0.456635609 -0.047206133  0.57013861
#>  [42,]  0.802843658  0.46820300 -0.095482707 -0.731363775  0.21181972
#>  [43,] -0.360395883  0.81816434  0.194578155 -0.004706042  0.97894286
#>  [44,]  0.410248759  0.29546613 -0.847054632 -0.858633015  0.97445230
#>  [45,] -0.098435692  0.48649131  0.205806858  0.331781368  0.95217308
#>  [46,] -0.744030107 -0.87856842 -0.229499539 -0.903931693 -0.04664221
#>  [47,]  0.894675323  0.01526409 -0.293585765  0.396847282  0.25557495
#>  [48,] -0.011070788  0.14393972  0.202935907 -0.526351907 -0.32936237
#>  [49,] -0.540905409 -0.25035331 -0.529702343  0.588294487 -0.08289310
#>  [50,] -0.023959782  0.64131083  0.775235277  0.624188561  0.66083546
#>  [51,] -0.791763383 -0.01167929 -0.195043277 -0.274582392 -0.17306646
#>  [52,] -0.813140837  0.51948498  0.019101237 -0.848720720  0.41938311
#>  [53,] -0.819898265  0.57402551  0.210771796 -0.320268235 -0.53622759
#>  [54,]  0.451247451 -0.71487385  0.285959177 -0.050261179  0.12357870
#>  [55,] -0.527542910  0.01464613 -0.978559740  0.551446185  0.50429490
#>  [56,] -0.543941884  0.78292167 -0.162286923  0.458336531  0.93056159
#>  [57,] -0.100556426 -0.27674354  0.656766812 -0.253343378  0.56926763
#>  [58,] -0.454294739  0.04282022 -0.542350779  0.001017674  0.75460522
#>  [59,] -0.420654347  0.63959000  0.208248305 -0.498078422 -0.03457007
#>  [60,]  0.293305566  0.70171558  0.197414205  0.071867025  0.19519502
#>  [61,] -0.323013858  0.43406172  0.990765588 -0.411368669  0.22914122
#>  [62,]  0.987689144  0.04376094 -0.612697345  0.324074811 -0.77760426
#>  [63,] -0.826283809 -0.98341083 -0.975247469  0.413434196  0.01489739
#>  [64,]  0.753996356  0.70642985 -0.899044023 -0.586461468 -0.85765362
#>  [65,] -0.971619916  0.29061492  0.939265234 -0.854219466  0.59194603
#>  [66,]  0.030625726  0.77854029  0.502488911 -0.991434353 -0.79429199
#>  [67,]  0.204073234  0.19004981 -0.752592520  0.182374807 -0.09469289
#>  [68,] -0.009870289  0.42256891 -0.769566838  0.091384026  0.61738025
#>  [69,] -0.228949737  0.42955260 -0.831390631  0.665774446 -0.64133030
#>  [70,] -0.821567208 -0.97604454  0.888681505 -0.768330878  0.67215778
#>  [71,]  0.356469278  0.42247767 -0.717938231  0.667480984 -0.27884486
#>  [72,]  0.573829051 -0.28137282 -0.793978713 -0.921883773  0.34725467
#>  [73,] -0.095687800 -0.84014597 -0.618921106 -0.811181072 -0.53425758
#>  [74,] -0.423062378  0.36152015 -0.329128369 -0.040430432  0.58734040
#>  [75,] -0.285015061 -0.48530589 -0.405294567  0.958004422 -0.27224567
#>  [76,]  0.559006950  0.34424528  0.013607128  0.310098084 -0.63889381
#>  [77,]  0.475267444 -0.24697532  0.820763554  0.337975327 -0.08366945
#>  [78,] -0.716846575  0.81896376 -0.138850746  0.405257079 -0.96227438
#>  [79,]  0.408910933 -0.58945966 -0.462485574  0.015975480 -0.74144012
#>  [80,] -0.862932643  0.17239271 -0.864615246  0.029837761 -0.56193220
#>  [81,] -0.681252359 -0.90555066  0.171142175  0.969384734  0.62231211
#>  [82,] -0.652475618  0.35594004  0.115152224  0.823087313  0.51258905
#>  [83,]  0.366993703  0.23672800  0.234239053  0.620499893 -0.39909190
#>  [84,] -0.541581716  0.99121853 -0.457942767  0.528115575 -0.41254382
#>  [85,] -0.914980320 -0.01240563 -0.253325147  0.564347492  0.04706565
#>  [86,]  0.617650693 -0.01755486  0.649756438  0.545728347 -0.96368707
#>  [87,]  0.933284083 -0.68700290 -0.493960904  0.589224077  0.90661770
#>  [88,] -0.016411877 -0.24838067 -0.478070620 -0.643559442  0.99628856
#>  [89,]  0.154287099 -0.67346500  0.223918072  0.154257270 -0.90885983
#>  [90,]  0.092244808  0.05868819 -0.214168075 -0.618781368 -0.29906718
#>  [91,] -0.561056321  0.27660650  0.282061718 -0.078478341  0.96245815
#>  [92,] -0.065929160  0.22068334 -0.170197215  0.233347332 -0.27609164
#>  [93,]  0.881699320  0.86412538 -0.709983258  0.039385695 -0.34198127
#>  [94,] -0.678215321  0.70773853  0.724328994  0.698003716 -0.63267495
#>  [95,] -0.384551952  0.96441409  0.673906578  0.103577241  0.05620962
#>  [96,]  0.807015618 -0.01937550  0.923365877 -0.090091426 -0.27364959
#>  [97,] -0.646007242 -0.11839728  0.257677791  0.019573426 -0.98927912
#>  [98,]  0.101473067  0.26338957 -0.789262441 -0.551057954 -0.97754765
#>  [99,]  0.023544332 -0.27408603  0.632296237 -0.905522750 -0.69298552
#> [100,]  0.622926122  0.77230233  0.377711754 -0.512926168 -0.53383652
#> 
#> $y
#>   [1] 0 1 1 1 1 0 1 0 0 0 0 0 1 1 1 0 0 0 0 0 1 1 0 1 0 0 0 0 1 1 1 1 1 0 1 0 0
#>  [38] 0 1 1 1 0 0 1 1 0 0 1 1 1 1 0 0 0 1 0 1 1 1 1 1 0 0 0 0 1 1 1 1 0 1 1 0 1
#>  [75] 1 1 1 0 1 0 0 1 1 0 0 1 0 1 1 1 1 1 0 0 0 0 1 1 1 0
#>