Simulate Data from Hyper-Sphere for Classification Problems

The classification problem Circle is described in the BARK paper (2008). Inputs are dim independent variables uniformly distributed on the interval \([-1,1]\), only the first 2 out of these dim are actually signals. Outputs are created according to the formula $$y = 1(x1^2+x2^2 \le 2/\pi)$$

sim_circle(n, dim = 5)

Arguments

n

number of data points to generate

dim

number of dimension of the problem, no less than 2

Value

Returns a list with components

x

input values (independent variables)

y

0/1 output values (dependent variable)

References

Ouyang, Zhi (2008) Bayesian Additive Regression Kernels. Duke University. PhD dissertation, Chapter 3.

See also

Examples

sim_circle(n=100, dim=5)
#> $x
#>                [,1]        [,2]         [,3]          [,4]         [,5]
#>   [1,] -0.617979584  0.31431889  0.456964524 -0.0189453652 -0.382357033
#>   [2,] -0.665078264 -0.88591259 -0.579160422 -0.1858333889  0.542663785
#>   [3,] -0.433883354  0.28220744  0.082839508  0.3813674473  0.384788857
#>   [4,] -0.141579187 -0.21564216 -0.686163487  0.9717800790  0.658157656
#>   [5,]  0.111720623  0.39234692 -0.726466281 -0.8774860417  0.142979445
#>   [6,]  0.006266833 -0.90649633 -0.192910168  0.9018851123  0.943023277
#>   [7,]  0.383941560 -0.57185238 -0.907578493 -0.0841055894 -0.906709122
#>   [8,]  0.242003210 -0.42344426 -0.741442617  0.3940299680  0.070648179
#>   [9,] -0.107898474  0.30556087 -0.729530584  0.8534216429 -0.664136103
#>  [10,]  0.758161348  0.67609513  0.655885838  0.1478921785  0.998998788
#>  [11,] -0.458651671 -0.34146550  0.300388240  0.4791960032 -0.427552856
#>  [12,]  0.026132283  0.49073512 -0.493267394  0.7622258267  0.157423999
#>  [13,]  0.740708933 -0.95412499  0.591103725 -0.3646454755 -0.065171037
#>  [14,] -0.363609985  0.25332906 -0.382622567 -0.2358058924 -0.973428531
#>  [15,] -0.426566401 -0.29702931 -0.977886933  0.6621500934  0.133545866
#>  [16,]  0.460007741  0.88098231  0.960179691  0.0568879237 -0.092573182
#>  [17,] -0.060053925 -0.64970406  0.255243482  0.8202578011  0.429184520
#>  [18,] -0.175684154  0.87639048 -0.178414024  0.9344309736  0.632964145
#>  [19,]  0.492435957  0.46096139  0.896305589  0.2656254023 -0.025916495
#>  [20,]  0.136156883 -0.64007370 -0.079857863 -0.5085274191 -0.265469148
#>  [21,]  0.081715344  0.13828137  0.500881036  0.4146878603 -0.733424749
#>  [22,]  0.500559527  0.40268506 -0.619291142 -0.0371899270 -0.547319769
#>  [23,]  0.707789629 -0.86709765  0.745648123 -0.0225014589 -0.008848295
#>  [24,] -0.966256395  0.01869281  0.929841931 -0.0693724141  0.439106387
#>  [25,]  0.860860940  0.45127359 -0.957790984  0.4953629924  0.310825267
#>  [26,]  0.536654865  0.09154909  0.888005822 -0.7244280945  0.496581003
#>  [27,]  0.982925357 -0.07809499  0.868275527 -0.5673407847 -0.973614314
#>  [28,] -0.166115464  0.65752313  0.001125265  0.0598473707  0.340337634
#>  [29,]  0.920268446 -0.83598766  0.731179055  0.8622587454  0.061635818
#>  [30,] -0.171069655 -0.32141441 -0.249106977  0.0305106989  0.894193069
#>  [31,]  0.224361198 -0.37907157 -0.774270607  0.2551321187 -0.571346602
#>  [32,]  0.890449528  0.60187156 -0.898673388 -0.6398813375  0.169132838
#>  [33,]  0.320530192  0.62985732 -0.209878609  0.9209003220 -0.102767495
#>  [34,]  0.136737897 -0.20036012  0.040138335  0.6033840547 -0.004961822
#>  [35,] -0.689103640  0.45056878  0.618703161 -0.0631450745 -0.083030478
#>  [36,]  0.541415120 -0.33332617  0.804112029  0.1512768283  0.201348173
#>  [37,]  0.724539495 -0.77703344 -0.071725726 -0.9074186878 -0.305935804
#>  [38,]  0.507547940 -0.21892779 -0.096760873 -0.5939139212 -0.341618364
#>  [39,] -0.145132361  0.01241820  0.475890657  0.0008829366 -0.970117288
#>  [40,] -0.451140173 -0.11164600  0.721648875  0.4890293768 -0.290129814
#>  [41,] -0.040060697 -0.33425279 -0.396508864 -0.1071456894  0.337703588
#>  [42,]  0.897007918 -0.25868689 -0.834696931 -0.3918714612  0.812284888
#>  [43,] -0.264689087 -0.69949291 -0.947693801  0.2810281175 -0.355203748
#>  [44,]  0.006329392  0.46847048 -0.331631343  0.6671861107  0.365723950
#>  [45,] -0.404625684 -0.43725497 -0.575410357 -0.3123051496  0.406538422
#>  [46,] -0.464501437  0.83499105  0.139801808  0.9577104412  0.839526854
#>  [47,] -0.661807799 -0.66837213  0.587502519 -0.4856464360 -0.798189758
#>  [48,] -0.656286775  0.99364714 -0.648413497 -0.7252927744  0.817720422
#>  [49,]  0.818781785 -0.84643256 -0.188384700 -0.3509607762  0.379298881
#>  [50,]  0.722792916 -0.15906600  0.678889764  0.6385483663  0.168121802
#>  [51,] -0.290847583 -0.28210982  0.520842386 -0.7628318393 -0.051721004
#>  [52,] -0.802962605  0.93940921 -0.748425727 -0.8651981121  0.383174091
#>  [53,]  0.615883067 -0.69541869  0.288271937 -0.2000099556  0.222982476
#>  [54,] -0.475338872  0.45745099  0.916033359  0.2275267360  0.984667883
#>  [55,]  0.026713578  0.10363409 -0.021112809 -0.6183218057 -0.057120652
#>  [56,] -0.319942363  0.76116603  0.784216639  0.7689294275  0.962169160
#>  [57,] -0.380059274 -0.52096460 -0.609589087  0.4054686553 -0.474556400
#>  [58,]  0.978091968 -0.84942085 -0.532570094 -0.1618796196  0.332228302
#>  [59,]  0.753041828 -0.43684292  0.462370370 -0.6080798153 -0.858252062
#>  [60,] -0.751331206  0.55666434 -0.462599103  0.6830374934 -0.928963236
#>  [61,] -0.646779247 -0.47284826  0.676056823  0.1535644275  0.978824293
#>  [62,]  0.648374725  0.06598777 -0.379911122 -0.0064823446 -0.100806519
#>  [63,] -0.620998628 -0.70883378 -0.124595627  0.3649256476  0.455316495
#>  [64,] -0.408813422  0.88008565  0.029318848 -0.0152609227 -0.598922547
#>  [65,] -0.649548807 -0.19089018 -0.303458145 -0.9526661308  0.949586226
#>  [66,] -0.068192326  0.37194970 -0.849384341 -0.8658189629 -0.857761942
#>  [67,] -0.640259513 -0.63576923  0.190252403  0.4332233681 -0.677656204
#>  [68,] -0.504283494  0.82183399 -0.370580385  0.8939338340 -0.509381176
#>  [69,]  0.752134457 -0.14377471 -0.369901722  0.1273722495  0.805111830
#>  [70,]  0.975780016  0.59804138 -0.195201346 -0.0579323587 -0.964931420
#>  [71,]  0.885619959 -0.11449870 -0.234424491 -0.2158022341 -0.544193971
#>  [72,] -0.204194496  0.29003190  0.805312406 -0.1418472012  0.055172176
#>  [73,]  0.080602939 -0.55388374 -0.751653275  0.9855005904 -0.238674765
#>  [74,] -0.477072771 -0.50107032 -0.741299150  0.9899212630  0.582928709
#>  [75,] -0.235045139 -0.05231660 -0.478107003 -0.9925582879 -0.706044112
#>  [76,]  0.454266556 -0.73080053 -0.828301417  0.7591999536 -0.477538816
#>  [77,] -0.190885912  0.56220632  0.974601790 -0.8091572537  0.893507042
#>  [78,]  0.179429373 -0.11843361  0.845607934  0.2326243725  0.486987317
#>  [79,] -0.305074360 -0.98501459 -0.425907933  0.3861990659  0.676361047
#>  [80,] -0.355846428  0.98704414  0.491837463 -0.7632737211  0.434142731
#>  [81,]  0.581469514 -0.44821060 -0.878530432 -0.7971241646 -0.328768752
#>  [82,] -0.004394213  0.34489566  0.259756619  0.8357255277  0.262217571
#>  [83,]  0.195126600  0.35412202 -0.556827054  0.1583119081  0.662897273
#>  [84,]  0.912218620 -0.13343955  0.675455642  0.8489132556  0.299517355
#>  [85,] -0.390783425 -0.09055182 -0.760246933  0.3373589101 -0.259627411
#>  [86,]  0.066618626  0.93504796  0.713387646  0.1634424180  0.468912714
#>  [87,]  0.495236285  0.73594004 -0.797127944 -0.1505006412 -0.702646202
#>  [88,]  0.331771893 -0.08434541  0.642229390  0.6036190102 -0.753003367
#>  [89,] -0.884451830  0.19671571  0.573723292  0.5792928911  0.448498145
#>  [90,]  0.293623548  0.18041603 -0.550457529  0.2302467637  0.799920442
#>  [91,] -0.976162401  0.25422381 -0.202626796 -0.4298293646 -0.298726811
#>  [92,] -0.345107874  0.99906535 -0.069104721  0.4158302555  0.448271159
#>  [93,]  0.469576119  0.11532636 -0.154759985  0.4184260410 -0.742786912
#>  [94,] -0.770916845 -0.70060612 -0.628577960  0.8062715880 -0.595184742
#>  [95,]  0.339816771 -0.35223333 -0.596017079  0.4307015459 -0.038210918
#>  [96,] -0.541503522  0.46838070  0.050140888 -0.5122011323 -0.022373986
#>  [97,] -0.953441836  0.51477494 -0.579649326 -0.9333283664  0.650378444
#>  [98,] -0.710671597  0.55883290 -0.240171602 -0.5499601928  0.610327428
#>  [99,] -0.012683205  0.60755646 -0.570393035  0.9795428449 -0.183209612
#> [100,] -0.409832732 -0.95234708 -0.460670904  0.7592347865 -0.740500703
#> 
#> $y
#>   [1] 1 0 1 1 1 0 1 1 1 0 1 1 0 1 1 0 1 0 1 1 1 1 0 0 0 1 0 1 0 1 1 0 1 1 0 1 0
#>  [38] 1 1 1 1 0 1 1 1 0 0 0 0 1 1 0 0 1 1 0 1 0 0 0 0 1 0 0 1 1 0 0 1 0 0 1 1 1
#>  [75] 1 0 1 1 0 0 1 1 1 0 1 0 0 1 0 1 0 0 1 0 1 1 0 0 1 0
#>