Simulate Data from Hyper-Sphere for Classification Problems

The classification problem Circle is described in the BARK paper (2008). Inputs are dim independent variables uniformly distributed on the interval \([-1,1]\), only the first 2 out of these dim are actually signals. Outputs are created according to the formula $$y = 1(x1^2+x2^2 \le 2/\pi)$$

Usage

sim_circle(n, dim = 5)

Arguments

n

number of data points to generate

dim

number of dimension of the problem, no less than 2

Value

Returns a list with components

x

input values (independent variables)

y

0/1 output values (dependent variable)

References

Ouyang, Zhi (2008) Bayesian Additive Regression Kernels. Duke University. PhD dissertation, Chapter 3.

See also

Examples

sim_circle(n=100, dim=5)
#> $x
#>                [,1]         [,2]         [,3]         [,4]         [,5]
#>   [1,]  0.376300708 -0.366337790  0.929037086 -0.575242431 -0.705777587
#>   [2,]  0.489695535  0.864995383  0.001679459  0.043047958 -0.530464707
#>   [3,] -0.327643610 -0.020278882 -0.101195769  0.977896225 -0.248324229
#>   [4,]  0.059955371 -0.995512553 -0.404203771 -0.434036790 -0.290379103
#>   [5,]  0.399644537 -0.736841928  0.280778196 -0.286522117  0.432040651
#>   [6,]  0.393203750  0.514706444 -0.322324668 -0.189452861  0.879563115
#>   [7,] -0.716349165  0.249907974  0.840511562 -0.985673724 -0.377963440
#>   [8,] -0.106316844 -0.651473281 -0.339282001 -0.368605467  0.861898919
#>   [9,] -0.425055359 -0.684819753 -0.635642386  0.107160329 -0.132388658
#>  [10,]  0.892764413 -0.253099991  0.267239043  0.105881029 -0.611878368
#>  [11,] -0.574475329  0.973572052  0.352745458 -0.263523584 -0.819553945
#>  [12,] -0.046622426  0.002329699 -0.009534826 -0.505048447 -0.261465744
#>  [13,]  0.932762297  0.897373639 -0.461515347 -0.596180890  0.770301019
#>  [14,] -0.566618634 -0.882599503  0.172850854  0.090025993 -0.465115413
#>  [15,]  0.690703383  0.454185748 -0.446599771  0.081928178  0.684159477
#>  [16,] -0.146431765 -0.306715456  0.169884752 -0.323358446 -0.540686549
#>  [17,]  0.201191506  0.154758142  0.391727097  0.040718475  0.466400577
#>  [18,] -0.001190073 -0.164937031 -0.101576382  0.973720129 -0.188585316
#>  [19,]  0.297059490 -0.373689370  0.489032134  0.437223854 -0.522296834
#>  [20,]  0.831732982  0.478319509 -0.681874838  0.907789209  0.510586289
#>  [21,]  0.299097566 -0.213912534  0.202430211 -0.024233567 -0.357877965
#>  [22,]  0.624191533 -0.378321683 -0.906843847 -0.929203875  0.720161838
#>  [23,]  0.346155233 -0.868214777 -0.102664449  0.874318746 -0.742339207
#>  [24,]  0.030460163  0.769227240  0.449926963 -0.483661157  0.121343574
#>  [25,]  0.560026796  0.529533730 -0.720521184  0.043994406  0.905011522
#>  [26,]  0.901457345 -0.697846017 -0.133563217 -0.427645852 -0.148789641
#>  [27,] -0.759629289 -0.127632868 -0.171794186 -0.587085084  0.549616565
#>  [28,] -0.470242072  0.099342825 -0.407108811 -0.045652358  0.591159037
#>  [29,]  0.266908477 -0.770214290  0.690051632 -0.502006044 -0.353223056
#>  [30,]  0.505434621 -0.562205089 -0.042641012 -0.535050948 -0.763829364
#>  [31,]  0.905059404  0.025566470 -0.465032595  0.814196342 -0.338987276
#>  [32,] -0.247645275 -0.769878102 -0.063606726 -0.031343411  0.786307001
#>  [33,] -0.797723887  0.785767633  0.772525438 -0.777748297 -0.835767294
#>  [34,]  0.978689888 -0.225257481  0.529849569  0.861933975 -0.865440650
#>  [35,] -0.213132541  0.787010747 -0.170553837 -0.131546323 -0.587537532
#>  [36,]  0.677315853  0.656248501 -0.880198458 -0.121060864 -0.867866736
#>  [37,] -0.067946091 -0.197774153 -0.561187588  0.413660973 -0.530534067
#>  [38,] -0.535720693 -0.937059456  0.516327501 -0.051853125  0.130533067
#>  [39,]  0.795266873  0.983650797  0.028149436 -0.041283725  0.701385892
#>  [40,] -0.350940926 -0.688213941  0.279437742  0.110436504 -0.648732779
#>  [41,]  0.753343936 -0.325861396 -0.593795555 -0.589922158 -0.538165668
#>  [42,] -0.877770426 -0.259802794 -0.649993797 -0.106726219 -0.852435827
#>  [43,]  0.594976769  0.918099328 -0.847597098 -0.848129524  0.443880741
#>  [44,]  0.292181324  0.027279785 -0.666079678  0.202682943 -0.889959469
#>  [45,]  0.848724538  0.587188737  0.633426707 -0.917733040 -0.007758301
#>  [46,] -0.341975335  0.220502476 -0.357149941  0.443250002  0.228831112
#>  [47,]  0.011176528 -0.456332619  0.585723884  0.159747692  0.021689492
#>  [48,]  0.889391905 -0.151546014 -0.234393901  0.004516363 -0.480954577
#>  [49,]  0.363520186 -0.691879177  0.970334575 -0.185533800  0.477261734
#>  [50,] -0.158094842 -0.429327735  0.925458586  0.493437185 -0.220575275
#>  [51,] -0.496345703 -0.403692731 -0.284384239  0.582490255 -0.610089288
#>  [52,]  0.018127366  0.592622839  0.373709470  0.862155838 -0.553060093
#>  [53,]  0.602560936  0.961348603 -0.478830493  0.809995793  0.230494261
#>  [54,] -0.393752168 -0.359400245 -0.179190722 -0.283118949  0.390068377
#>  [55,] -0.637012813 -0.613479879 -0.111492090 -0.090192887  0.433655994
#>  [56,]  0.251248559 -0.555610239  0.469357972  0.392225928 -0.295916534
#>  [57,] -0.430442289  0.978515543 -0.432011413 -0.017833250 -0.068427581
#>  [58,]  0.968083031  0.759341741  0.438924811 -0.271439946 -0.447060842
#>  [59,] -0.121910071 -0.381495652 -0.444537478  0.923602762 -0.052961592
#>  [60,] -0.731588813 -0.167502258 -0.656807402 -0.475420812  0.985513553
#>  [61,]  0.880187480  0.111477488  0.466887530  0.309541712 -0.557405146
#>  [62,]  0.913912198 -0.678747046  0.903794682  0.195703713 -0.475269868
#>  [63,]  0.260290232 -0.474314792 -0.066852092 -0.499562396 -0.793859320
#>  [64,] -0.945301219 -0.169486878 -0.418384125 -0.650034490  0.360310366
#>  [65,] -0.440926810  0.543596783  0.552269214  0.725832671  0.289802822
#>  [66,] -0.596332932  0.057284958 -0.694772988  0.124903323 -0.232516440
#>  [67,]  0.398422662  0.601099306 -0.176633002 -0.789891933  0.945263104
#>  [68,]  0.018073036 -0.986338920  0.231773594  0.380554545  0.319464440
#>  [69,]  0.894192221 -0.922169654  0.522556905 -0.589485144  0.572629142
#>  [70,]  0.311698051  0.247457730 -0.524892200 -0.260568694 -0.545962364
#>  [71,] -0.229067322 -0.536667252 -0.897103184  0.381929178  0.116541902
#>  [72,]  0.518757171 -0.205708823 -0.425619402 -0.203888678  0.974489169
#>  [73,] -0.851675229 -0.356560361 -0.159897841 -0.818102391 -0.128220132
#>  [74,]  0.623393448  0.673571463 -0.341380448 -0.463103087  0.608873382
#>  [75,] -0.208140023  0.080019966 -0.094135011  0.658894633 -0.114116682
#>  [76,] -0.958903181  0.595050981 -0.423984913 -0.186406501  0.500196230
#>  [77,]  0.554840246  0.416841546 -0.757822728  0.321633297  0.002888379
#>  [78,] -0.332413686  0.239203404  0.982907169  0.210075244 -0.943713816
#>  [79,] -0.749714420 -0.092923549 -0.665762809 -0.021992398 -0.247515233
#>  [80,] -0.537541788  0.098872435 -0.587211366 -0.100423011  0.255993820
#>  [81,]  0.107676555 -0.507932470 -0.473084818 -0.817275084  0.405844902
#>  [82,] -0.657613653 -0.780645129 -0.400675846 -0.734895004  0.991952240
#>  [83,]  0.339134435 -0.192550396  0.728116803  0.745732648  0.809888715
#>  [84,] -0.797604309  0.447398666  0.211967763 -0.710821009  0.828485292
#>  [85,] -0.442677356 -0.553153098 -0.582972231 -0.958449148 -0.879901842
#>  [86,]  0.821649252  0.613642295 -0.868879541 -0.204635637  0.819022280
#>  [87,]  0.356546666  0.227941143  0.257532148  0.032365493  0.288349409
#>  [88,] -0.184808573 -0.912235419  0.042184640  0.391937273  0.249633838
#>  [89,] -0.789488513 -0.256448552  0.138902360 -0.807935753  0.061791021
#>  [90,]  0.079043110 -0.036808189 -0.965994437  0.428392530 -0.452641924
#>  [91,]  0.478692116 -0.672072315  0.185508776  0.401760791  0.146251680
#>  [92,]  0.625345544 -0.292402237 -0.396265148  0.848732573 -0.060892299
#>  [93,] -0.188403154  0.079720030  0.379436080 -0.445281357 -0.917767540
#>  [94,] -0.179917211  0.048642912  0.068694043 -0.150022663  0.537657341
#>  [95,]  0.407921619  0.741931242  0.659755816 -0.478506725 -0.340825817
#>  [96,]  0.712705146 -0.499917753 -0.730070859  0.712389593 -0.011139353
#>  [97,]  0.652535570  0.941296644  0.131542565  0.233268286  0.196075676
#>  [98,] -0.402364439  0.722238487 -0.494545904 -0.436421926  0.175379945
#>  [99,]  0.645806623  0.785073507  0.717462549 -0.575350450 -0.051867198
#> [100,]  0.087269355  0.747594874 -0.435691131  0.628258454 -0.365249633
#> 
#> $y
#>   [1] 1 0 1 0 0 1 1 1 0 0 0 1 0 0 0 1 1 1 1 0 1 1 0 1 1 0 1 1 0 1 0 0 0 0 0 0 1
#>  [38] 0 0 1 0 0 0 1 0 1 1 0 1 1 1 1 0 1 0 1 0 0 1 1 0 0 1 0 1 1 1 0 0 1 1 1 0 0
#>  [75] 1 0 1 1 1 1 1 0 1 0 1 0 1 0 0 1 0 1 1 1 0 0 0 0 0 1
#>