Simulate Data from Hyper-Sphere for Classification Problems

The classification problem Circle is described in the BARK paper (2008). Inputs are dim independent variables uniformly distributed on the interval \([-1,1]\), only the first 2 out of these dim are actually signals. Outputs are created according to the formula $$y = 1(x1^2+x2^2 \le 2/\pi)$$

Usage

sim_circle(n, dim = 5)

Arguments

n

number of data points to generate

dim

number of dimension of the problem, no less than 2

Value

Returns a list with components

x

input values (independent variables)

y

0/1 output values (dependent variable)

References

Ouyang, Zhi (2008) Bayesian Additive Regression Kernels. Duke University. PhD dissertation, Chapter 3.

See also

Examples

sim_circle(n=100, dim=5)
#> $x
#>               [,1]          [,2]         [,3]        [,4]         [,5]
#>   [1,]  0.42929976 -0.5343703809  0.445483156 -0.35264547  0.204377216
#>   [2,]  0.37387942  0.0162284952 -0.099899372 -0.89781660  0.144926509
#>   [3,] -0.27632135  0.9808540223  0.496513173 -0.94011103  0.935093313
#>   [4,] -0.54447965  0.7835169914 -0.825883454  0.83392367  0.867495505
#>   [5,]  0.44746055  0.1977646048  0.624572120 -0.80076662  0.719000532
#>   [6,]  0.98564585 -0.7229769952 -0.030676269  0.48437629 -0.553884409
#>   [7,] -0.21186049  0.3386505493 -0.258442103  0.33465896  0.678240276
#>   [8,]  0.40457716 -0.4260876882  0.416939181 -0.07981853  0.819873589
#>   [9,]  0.06921948 -0.9364545001  0.026705541 -0.47962266 -0.945537649
#>  [10,] -0.83738538 -0.7524949773  0.408399962  0.39252722  0.174118174
#>  [11,] -0.60646119 -0.4810321517  0.353185615 -0.59292661 -0.229530734
#>  [12,]  0.56569441 -0.1432594080  0.715632159 -0.74978701 -0.834951763
#>  [13,]  0.64488656 -0.6477462817 -0.552456984  0.59071886  0.169460915
#>  [14,] -0.63861754  0.0440198891  0.630770730  0.79980566  0.860307567
#>  [15,] -0.65934230  0.0817580591  0.147233249 -0.34294534  0.980512647
#>  [16,] -0.38992141  0.6873332704 -0.663410255  0.48118032  0.777076131
#>  [17,]  0.54400474 -0.0218698015 -0.395147094  0.35710471 -0.597358144
#>  [18,] -0.02847036 -0.0152725307 -0.994488497 -0.38306372  0.995282607
#>  [19,]  0.03060068 -0.2471215604 -0.533134596  0.30822747  0.667137636
#>  [20,] -0.75420825  0.0105914320  0.818051341  0.49467413 -0.082594824
#>  [21,]  0.34254192  0.5601577736  0.343860698  0.88040838 -0.946077102
#>  [22,]  0.73923583  0.9157915576  0.228193982 -0.25532210  0.766840626
#>  [23,]  0.95012423 -0.8068329152  0.689566544  0.18741369  0.262079068
#>  [24,]  0.18428406  0.5547278072 -0.710966859 -0.31353286  0.484702243
#>  [25,]  0.34805838 -0.9962751190  0.842362633  0.29373824  0.834456318
#>  [26,]  0.17664575  0.3747463343  0.833710042 -0.90472249  0.449522905
#>  [27,]  0.80001868  0.5995486104 -0.249589403  0.07310224  0.725692115
#>  [28,] -0.02666458  0.5346616688 -0.815721348 -0.77566715  0.321950525
#>  [29,] -0.96748758 -0.9018364931 -0.659071241  0.35085713 -0.663953780
#>  [30,]  0.50519951 -0.2136472566 -0.519386015  0.35284601 -0.155129806
#>  [31,]  0.16674465  0.0179052018 -0.156693130 -0.87486600 -0.983438625
#>  [32,]  0.21600846 -0.1832653736 -0.292143299  0.72351997  0.732866027
#>  [33,]  0.71672931 -0.1546799256 -0.375984369 -0.75320778 -0.146640482
#>  [34,]  0.01352362  0.9514003284 -0.960768826 -0.54164429 -0.738872206
#>  [35,]  0.69142873 -0.9611414177  0.127786335  0.56863835 -0.842430202
#>  [36,]  0.30352853  0.1125565292  0.580330614 -0.37772216 -0.487497584
#>  [37,] -0.48785431 -0.1117596002 -0.576019792 -0.38198522 -0.208500043
#>  [38,]  0.13928642 -0.3686439428 -0.253236418 -0.72329590  0.532814708
#>  [39,] -0.39084478  0.1986586344 -0.137518992  0.68035733  0.173866306
#>  [40,]  0.34655777  0.5353487572  0.335052231  0.36726658 -0.783012402
#>  [41,]  0.68719054 -0.7212374993  0.645526137  0.29672982  0.814654290
#>  [42,]  0.62271887  0.6622189959  0.986744970  0.45231150 -0.334326623
#>  [43,]  0.03646576  0.6561822291  0.001888623  0.50309317 -0.529735826
#>  [44,] -0.99475952  0.9247705885 -0.404532869  0.57307219 -0.984491645
#>  [45,]  0.80675805  0.2270934321 -0.377454326 -0.40946741 -0.363291592
#>  [46,] -0.73226995  0.9590663332 -0.319286548  0.56225164  0.060042359
#>  [47,]  0.44104842 -0.2922363929 -0.410742507  0.93767995 -0.457620167
#>  [48,] -0.16470077  0.3617066508  0.577886799  0.01675785  0.144213990
#>  [49,] -0.84217955  0.6188937672 -0.003621990  0.76756816  0.264663044
#>  [50,]  0.54340471 -0.0007605352 -0.969948004  0.59778482  0.066871072
#>  [51,] -0.27952292  0.4754950129  0.216687965  0.87652991  0.081596313
#>  [52,] -0.01415279 -0.8121445389 -0.247125501 -0.31165934  0.272159025
#>  [53,]  0.05581382 -0.1261460660  0.745901899  0.55271568  0.962860653
#>  [54,] -0.94633080 -0.1789127532 -0.505489725  0.89069356 -0.890051192
#>  [55,]  0.10341583  0.0506213647 -0.255824702  0.16486121 -0.364434715
#>  [56,] -0.71357139 -0.9081108333  0.172604089  0.27183295  0.819443889
#>  [57,] -0.64165364  0.2761134929 -0.375336027 -0.83634127 -0.701027581
#>  [58,]  0.86057371 -0.4454884920 -0.030375510  0.28877195 -0.871512425
#>  [59,]  0.71758586  0.8085481259  0.263800581 -0.72127457 -0.737218535
#>  [60,] -0.41963359 -0.9186043045 -0.967785551  0.52080432  0.349023783
#>  [61,]  0.15135489 -0.8598418958  0.625755788 -0.32460792 -0.184092583
#>  [62,] -0.69388801 -0.9234449971  0.860589730  0.06543931 -0.516699118
#>  [63,]  0.43764619 -0.5723274136  0.333256798  0.71623378  0.737166697
#>  [64,] -0.03048344  0.6789583806  0.275408953  0.79115511 -0.310293249
#>  [65,]  0.25947137 -0.5153101734 -0.543884953  0.02145756  0.886571066
#>  [66,]  0.85330416  0.8635516791  0.740122060  0.61307263 -0.885871497
#>  [67,] -0.81424696  0.7221221365  0.696390290 -0.97785091  0.685843259
#>  [68,] -0.17435094  0.1471606744  0.880083230  0.77691692  0.443672561
#>  [69,] -0.27632836 -0.6857168018  0.141769188  0.60860485 -0.007351142
#>  [70,] -0.22295190  0.8615904003 -0.389941217  0.03087962 -0.397973119
#>  [71,] -0.54812515 -0.7170760096 -0.076959450  0.05825740 -0.474324377
#>  [72,]  0.78936075  0.8997853626 -0.370170841 -0.92348583 -0.838658284
#>  [73,] -0.14287930  0.0948695815 -0.707654172 -0.21700312  0.095501211
#>  [74,] -0.94481511  0.3944375338 -0.782966529  0.54096598 -0.127631717
#>  [75,]  0.58241349  0.1931995694 -0.749227121 -0.60375349  0.307795581
#>  [76,] -0.12721804 -0.6741769919 -0.788260975  0.22071951  0.069173310
#>  [77,] -0.76631365  0.3796149986  0.102358148 -0.58919358  0.764780256
#>  [78,]  0.97276945  0.3446836160  0.858573312 -0.39305828 -0.839715107
#>  [79,] -0.04703554 -0.6357294782 -0.953423558 -0.07184351 -0.422037827
#>  [80,]  0.62447639 -0.9514514897 -0.005534201 -0.24503293  0.955345042
#>  [81,] -0.62038317  0.0856661657  0.660704370 -0.28127951  0.282911127
#>  [82,] -0.47820558 -0.1794764725  0.497268207  0.39144021 -0.590714030
#>  [83,] -0.15577508 -0.2659423999 -0.831768107 -0.19211242 -0.944672027
#>  [84,] -0.89806373  0.4299176484 -0.881607911  0.28941388  0.369851803
#>  [85,]  0.01840725  0.4308130834  0.461221478 -0.26038432 -0.671280134
#>  [86,] -0.97656553 -0.6376289655 -0.774262512 -0.69662041  0.660318638
#>  [87,] -0.96276920  0.3992663887  0.733609864 -0.26994450  0.439213370
#>  [88,]  0.10225112 -0.5824279962  0.620872740 -0.05975391  0.963169476
#>  [89,] -0.47556909 -0.6920122001 -0.057262129 -0.47187997 -0.362015583
#>  [90,] -0.37351677 -0.3637713464 -0.403392979  0.95147488 -0.134218354
#>  [91,]  0.58341864  0.2265944495 -0.784025535  0.52672080  0.320891778
#>  [92,]  0.60918151 -0.6288983971  0.834296952  0.41839188  0.458187580
#>  [93,]  0.69074506 -0.8789285379  0.812398396 -0.06237327  0.470872690
#>  [94,]  0.02104480  0.5718956324  0.727371647 -0.96469087 -0.499909173
#>  [95,] -0.97853321  0.8454567427  0.826162087  0.79050256  0.607726646
#>  [96,]  0.50200664  0.7743570711  0.751347671  0.55673273 -0.352409925
#>  [97,] -0.36264974  0.8010840123 -0.314093613 -0.59693888  0.916020251
#>  [98,] -0.99693350 -0.6617436702  0.323967935  0.45467592  0.580915844
#>  [99,] -0.48623220 -0.4680796005 -0.289874326 -0.76799638 -0.969996213
#> [100,]  0.81502206 -0.4232050199  0.912748885 -0.83308842 -0.372835704
#> 
#> $y
#>   [1] 1 1 0 0 1 0 1 1 0 0 1 1 0 1 1 1 1 1 1 1 1 0 0 1 0 1 0 1 0 1 1 1 1 0 0 1 1
#>  [38] 1 1 1 0 0 1 0 0 0 1 1 0 1 1 0 1 0 1 0 1 0 0 0 0 0 1 1 1 0 0 1 1 0 0 0 1 0
#>  [75] 1 1 0 0 1 0 1 1 1 0 1 0 0 1 0 1 1 0 0 1 0 0 0 0 1 0
#>