Simulate Data from Hyper-Sphere for Classification Problems

The classification problem Circle is described in the BARK paper (2008). Inputs are dim independent variables uniformly distributed on the interval \([-1,1]\), only the first 2 out of these dim are actually signals. Outputs are created according to the formula $$y = 1(x1^2+x2^2 \le 2/\pi)$$

sim_circle(n, dim = 5)

Arguments

n

number of data points to generate

dim

number of dimension of the problem, no less than 2

Value

Returns a list with components

x

input values (independent variables)

y

0/1 output values (dependent variable)

References

Ouyang, Zhi (2008) Bayesian Additive Regression Kernels. Duke University. PhD dissertation, Chapter 3.

See also

Examples

sim_circle(n=100, dim=5)
#> $x
#>                [,1]         [,2]         [,3]        [,4]        [,5]
#>   [1,]  0.710595632  0.006022267  0.225708730 -0.24440735 -0.25645336
#>   [2,] -0.370995477  0.480786846  0.544371661 -0.41937779 -0.83206430
#>   [3,] -0.444245248  0.726609054 -0.822265323 -0.34109874  0.57976037
#>   [4,]  0.488970991 -0.329950136 -0.298746679  0.57625497 -0.48915531
#>   [5,] -0.787283950  0.018699083  0.314015443 -0.13137241 -0.17525343
#>   [6,] -0.606275063 -0.263293788  0.065915312  0.70199921  0.47831931
#>   [7,] -0.558607657  0.748384210  0.451333860 -0.47010820  0.05355341
#>   [8,] -0.380985504 -0.284667997 -0.257602000 -0.54489112 -0.86921495
#>   [9,]  0.468317122  0.141416264 -0.900754894  0.04450047 -0.94951617
#>  [10,] -0.347406814  0.875121193 -0.421644159  0.02162535 -0.04482827
#>  [11,] -0.496863575 -0.924903640  0.469747732 -0.35921220 -0.78539911
#>  [12,]  0.645406656 -0.989102072  0.715059307  0.48768099 -0.04226606
#>  [13,]  0.833893361  0.709552511  0.131011623  0.48581341  0.01097983
#>  [14,]  0.283585110  0.701355524 -0.879100425 -0.53760048 -0.93343511
#>  [15,]  0.912968904  0.717641463  0.330134525 -0.72444218  0.80116792
#>  [16,] -0.958175017  0.721155154 -0.746215991  0.06965216  0.04203878
#>  [17,] -0.280803683 -0.908336035  0.283928727 -0.75794548  0.57402935
#>  [18,] -0.668950729 -0.569115958 -0.592105773  0.51533931  0.19370475
#>  [19,] -0.381793161 -0.060668492  0.188936991 -0.06745612 -0.21443396
#>  [20,]  0.167786141 -0.846130843 -0.555999145  0.14889212 -0.96539147
#>  [21,] -0.008592401 -0.998699933  0.987045673 -0.12020955 -0.91575861
#>  [22,]  0.931538141  0.180937212 -0.145110060 -0.73809138  0.93972991
#>  [23,] -0.314138152  0.442215470  0.311875239 -0.32993603  0.90534848
#>  [24,]  0.115045300  0.751571714  0.726626956  0.53686978 -0.96174250
#>  [25,]  0.720188172 -0.784213924 -0.405454810  0.58986555 -0.20174721
#>  [26,]  0.991307292 -0.271117551 -0.316967008 -0.89171133 -0.10239268
#>  [27,]  0.249056643  0.281318335  0.818458586  0.89421980  0.84733380
#>  [28,] -0.479000472  0.974159105 -0.624640759 -0.55841681  0.87307689
#>  [29,]  0.713718576 -0.241712678 -0.195689075 -0.51970117  0.49300352
#>  [30,] -0.601008810 -0.402390380  0.579978554  0.14378100  0.51372829
#>  [31,] -0.175085955 -0.057033212 -0.208615219 -0.96463248 -0.70537260
#>  [32,] -0.818825619 -0.196157030 -0.931628965  0.82995752  0.98812547
#>  [33,]  0.591354476 -0.910298887 -0.313916966 -0.23761851 -0.07689879
#>  [34,]  0.768211529  0.567332901  0.669055145  0.70909220  0.60726715
#>  [35,] -0.906500349 -0.191827471  0.677776036 -0.72903341  0.40490921
#>  [36,]  0.309600872  0.568679382  0.893025932 -0.48779304  0.18567969
#>  [37,] -0.182102502 -0.207325232  0.448520640 -0.02632400 -0.34104656
#>  [38,]  0.247283342  0.751913866  0.294872877 -0.64208930  0.90210531
#>  [39,] -0.529591580  0.968260951 -0.636557689 -0.16313361  0.70472436
#>  [40,]  0.001951559 -0.848709761  0.207384121  0.99749638 -0.87293474
#>  [41,] -0.956838257  0.355082255 -0.947303631 -0.55047536 -0.33199527
#>  [42,]  0.300374553 -0.923725998  0.142078168 -0.99149885  0.85211814
#>  [43,] -0.955659304 -0.499873884 -0.427565597  0.66178434  0.41685853
#>  [44,]  0.316598054  0.631788769  0.940116900  0.44761267 -0.12208756
#>  [45,] -0.900508435  0.439298210 -0.082673743  0.98049023 -0.37693811
#>  [46,] -0.812805095  0.594126033 -0.812900490  0.63408617 -0.51937198
#>  [47,]  0.702930163 -0.206273296  0.650692634  0.53050015 -0.43323577
#>  [48,]  0.359038257 -0.827887936 -0.418186308 -0.90669931 -0.31713833
#>  [49,] -0.332511093 -0.243117770  0.924292410  0.43543985 -0.15194725
#>  [50,]  0.703072226 -0.087208461  0.474902247 -0.08973007  0.21754363
#>  [51,] -0.222809684  0.643289773 -0.984773214 -0.29625093 -0.58980451
#>  [52,] -0.228404490 -0.080827852 -0.624964440 -0.07916240 -0.67348348
#>  [53,]  0.973908854 -0.767439019  0.422935030 -0.84469911 -0.23303964
#>  [54,] -0.456819157 -0.645029960 -0.575575935 -0.33324238 -0.82648556
#>  [55,]  0.326548710  0.961589192  0.298774922  0.07334674  0.48312603
#>  [56,] -0.547791254  0.182362064  0.785759954  0.03657233  0.06512971
#>  [57,]  0.747468974  0.832087531  0.157999043  0.95583859  0.90384461
#>  [58,] -0.022453514  0.587109867  0.118960845 -0.76380294 -0.91771612
#>  [59,] -0.737835030  0.938029808 -0.007299344 -0.90981800 -0.91828173
#>  [60,] -0.970834410 -0.033905170  0.636595065  0.51367943 -0.54537913
#>  [61,]  0.853880055 -0.080701666  0.381717077  0.05511858 -0.95077696
#>  [62,] -0.969062294 -0.809124933 -0.810859088  0.98400928 -0.65326543
#>  [63,] -0.789151394  0.553223329  0.107911688  0.63336038 -0.12673211
#>  [64,] -0.297155497  0.793355893 -0.614624310  0.89768524 -0.54673182
#>  [65,] -0.288332304  0.402761440 -0.351750875  0.44584351 -0.94659722
#>  [66,] -0.138755476 -0.057089609  0.214149424  0.01947359  0.54326120
#>  [67,]  0.736884491  0.459806495 -0.653256969  0.10243081 -0.67122914
#>  [68,] -0.400007633  0.235428237  0.467900178 -0.50631951  0.93097906
#>  [69,] -0.170090757  0.511578759  0.407926512  0.41257276  0.42762856
#>  [70,] -0.440681957  0.700480565 -0.672701183  0.72193862  0.22780554
#>  [71,]  0.498335585  0.989220073 -0.988276503  0.37642061  0.39505757
#>  [72,]  0.567803095  0.339347667 -0.280658444  0.28656346 -0.19810853
#>  [73,]  0.186738417  0.363827160 -0.733682820 -0.92854869 -0.73090469
#>  [74,]  0.824361328 -0.724893617  0.557839600 -0.38077886 -0.10986128
#>  [75,] -0.880130320 -0.646368699  0.953953001 -0.96005776 -0.47759264
#>  [76,] -0.940585339  0.639782789  0.907271312 -0.15791436  0.29611078
#>  [77,]  0.231887188 -0.230599227  0.315337997 -0.84806227 -0.83042374
#>  [78,] -0.387927917  0.729508257  0.956794952 -0.51596405 -0.50747858
#>  [79,]  0.692472650 -0.749405997 -0.885148861 -0.42444347  0.05287377
#>  [80,]  0.952902158 -0.117981636  0.391280666 -0.45837618 -0.28443492
#>  [81,] -0.535254858 -0.332610810 -0.938702816 -0.87735992  0.85734979
#>  [82,]  0.674479686  0.976368143  0.968181206 -0.45162724 -0.99035420
#>  [83,] -0.593016460  0.106807454 -0.569623207 -0.18831269  0.67638046
#>  [84,] -0.550123405 -0.719849911 -0.393970906 -0.95837553 -0.79629427
#>  [85,]  0.776617406 -0.844405848  0.623329774  0.90474409  0.55518798
#>  [86,]  0.659619370  0.920412756 -0.500107264 -0.44901264  0.48421022
#>  [87,]  0.887606918  0.879982868  0.523824487 -0.76862321 -0.19313538
#>  [88,] -0.621453202  0.205355789  0.130863696  0.54572971 -0.59505756
#>  [89,] -0.219473151 -0.795777525 -0.519048473  0.72104411 -0.79231144
#>  [90,]  0.033819274  0.425102455 -0.799768227 -0.34587339  0.62587035
#>  [91,] -0.884021088  0.643394351 -0.844582492 -0.23131562 -0.96429220
#>  [92,] -0.147460815 -0.635536576  0.926758138  0.70063885  0.55866145
#>  [93,]  0.430481065  0.204534119 -0.986770133  0.44090540 -0.66180750
#>  [94,] -0.174356773 -0.278248323 -0.238477273  0.42807756  0.37076157
#>  [95,] -0.971115631  0.078524943  0.291461766 -0.52587489 -0.93499746
#>  [96,]  0.723225727 -0.662177898 -0.701814340 -0.84356891 -0.24241846
#>  [97,]  0.013246134  0.285664423  0.176811452  0.61404747  0.64984741
#>  [98,] -0.745150358  0.837042563  0.814975678  0.98778721  0.77287214
#>  [99,]  0.076426516  0.990894176  0.223969672  0.64370184  0.49376821
#> [100,]  0.469573688  0.767393362 -0.661593359 -0.08191723 -0.99818304
#> 
#> $y
#>   [1] 1 1 0 1 1 1 0 1 1 0 0 0 0 1 0 0 0 0 1 0 0 0 1 1 0 0 1 0 1 1 1 0 0 0 0 1 1
#>  [38] 1 0 0 0 0 0 1 0 0 1 0 1 1 1 1 0 1 0 1 0 1 0 0 0 0 0 0 1 1 0 1 1 0 0 1 1 0
#>  [75] 0 0 1 0 0 0 1 0 1 0 0 0 0 1 0 1 0 1 1 1 0 0 1 0 0 0
#>