Simulate Data from Hyper-Sphere for Classification Problems

The classification problem Circle is described in the BARK paper (2008). Inputs are dim independent variables uniformly distributed on the interval \([-1,1]\), only the first 2 out of these dim are actually signals. Outputs are created according to the formula $$y = 1(x1^2+x2^2 \le 2/\pi)$$

sim_circle(n, dim = 5)

Arguments

n

number of data points to generate

dim

number of dimension of the problem, no less than 2

Value

Returns a list with components

x

input values (independent variables)

y

0/1 output values (dependent variable)

References

Ouyang, Zhi (2008) Bayesian Additive Regression Kernels. Duke University. PhD dissertation, Chapter 3.

See also

Examples

sim_circle(n=100, dim=5)
#> $x
#>                [,1]         [,2]         [,3]          [,4]         [,5]
#>   [1,] -0.015124226  0.952241013  0.217044646  0.9502618713  0.782402296
#>   [2,]  0.630458492 -0.267948404 -0.904517164 -0.4977662768  0.815490304
#>   [3,]  0.741905452  0.647580113 -0.873521785  0.2305285111  0.870040059
#>   [4,]  0.210066395  0.120113392  0.584821715  0.3230662984  0.002811205
#>   [5,] -0.766812596 -0.883890067  0.730768404  0.6715312274  0.552660173
#>   [6,] -0.441998640 -0.110474015  0.834069518  0.7157234862 -0.083322093
#>   [7,] -0.411372514  0.923849765  0.428772501  0.8999211211 -0.503687920
#>   [8,] -0.167264909  0.665594723  0.782769188  0.3150646440  0.581594227
#>   [9,] -0.659989272  0.954041864  0.280017823 -0.7219589227 -0.097013153
#>  [10,]  0.792319596 -0.364258368  0.752816163  0.3732201322  0.119369917
#>  [11,] -0.017862429 -0.278329660  0.408289135 -0.1002999898 -0.456749074
#>  [12,]  0.616833090 -0.141105053  0.968068972 -0.6221758449 -0.951762810
#>  [13,] -0.675579580  0.296004411  0.109212086  0.7378856824 -0.792161967
#>  [14,]  0.177163392 -0.075473222 -0.171049091 -0.1918651187  0.754629310
#>  [15,]  0.140293209 -0.639115859  0.446048019  0.7836182569  0.840098622
#>  [16,]  0.249304725  0.083981516 -0.143910183 -0.2930240757 -0.931330732
#>  [17,]  0.375286678 -0.916134779 -0.949276209 -0.4421047322 -0.334216813
#>  [18,]  0.813339122  0.228817880  0.901161199  0.7010656027 -0.473473400
#>  [19,]  0.216335570 -0.386551886 -0.108140132  0.7895420818  0.803273088
#>  [20,] -0.631253029 -0.448684014  0.869555906  0.8389517376  0.475132535
#>  [21,] -0.323778770 -0.462562009 -0.437348607  0.4618076654 -0.138599621
#>  [22,]  0.038937372  0.980420406  0.778660676  0.5974924034 -0.173926091
#>  [23,]  0.982860280 -0.704149424  0.245235610 -0.4946218426  0.165890887
#>  [24,] -0.200306012  0.792713474 -0.544017474 -0.1510211220 -0.318686055
#>  [25,] -0.638258727  0.141712104 -0.288879635 -0.6285392158  0.672026251
#>  [26,] -0.944782313 -0.678010558  0.016486916 -0.3512221309  0.973449486
#>  [27,] -0.688496642 -0.675926450  0.567544297 -0.1044251523  0.210290984
#>  [28,]  0.460169969 -0.427488920 -0.299959281 -0.6434148075 -0.252516490
#>  [29,]  0.201295651  0.540286863 -0.387702341 -0.0614691558 -0.779247849
#>  [30,] -0.996129038  0.714713494 -0.858841610  0.1666432526  0.197428811
#>  [31,] -0.884271817 -0.217949165 -0.541463278  0.7128033754  0.297340728
#>  [32,] -0.499062222 -0.021570237  0.740715364 -0.9256216153 -0.878486788
#>  [33,] -0.370279809  0.451076749  0.511742269  0.2505826950  0.017321456
#>  [34,]  0.325171337  0.745898565 -0.146314236 -0.5568075762  0.055944777
#>  [35,] -0.535970669  0.624371190  0.241541365  0.2300860044 -0.944807075
#>  [36,]  0.263002544 -0.804769816  0.810982371  0.1584561639  0.710662761
#>  [37,]  0.437108444 -0.146535668  0.207479790 -0.4767149435 -0.150825755
#>  [38,]  0.856733481 -0.877621032 -0.925342509 -0.5228952509  0.145549352
#>  [39,] -0.715507188  0.606926257  0.558174332 -0.7424109885  0.807410323
#>  [40,]  0.577631887  0.449203483 -0.052510303 -0.1998824729  0.675658431
#>  [41,] -0.221376548 -0.590682444  0.005179343 -0.9480290958  0.665465953
#>  [42,]  0.771329154 -0.517986963 -0.935528572  0.4159896597 -0.045767864
#>  [43,] -0.785134636  0.425047531 -0.970987217 -0.4896078976 -0.823757101
#>  [44,]  0.934751252 -0.598073432 -0.049369739  0.7846134184 -0.479159812
#>  [45,]  0.883804264  0.120968803  0.727530629 -0.0562417991  0.296753340
#>  [46,]  0.370954591  0.377467561 -0.266438363  0.7227149368  0.948425155
#>  [47,] -0.140413106 -0.019329646 -0.924258765  0.4041043990 -0.734056685
#>  [48,]  0.629749767  0.061919506  0.082625894 -0.0510751237  0.435391721
#>  [49,] -0.632320459 -0.408936166  0.832258030  0.6085809013  0.558058427
#>  [50,]  0.470813483  0.810831388  0.744621018 -0.7645876282  0.726080651
#>  [51,]  0.117128223  0.551308977 -0.700925284  0.2129608458  0.253439423
#>  [52,]  0.576688240  0.165064848  0.148102054  0.3110787305  0.842286963
#>  [53,] -0.297258931  0.463946647  0.413886513  0.3610768281  0.029570035
#>  [54,]  0.270380497  0.257076153 -0.472171389 -0.8301324830  0.907509760
#>  [55,]  0.489914781 -0.654732995  0.538259406 -0.6013033832 -0.148664564
#>  [56,] -0.528999683 -0.813822865 -0.980000650  0.0754753430  0.323914103
#>  [57,]  0.184419295  0.707986773 -0.604069144 -0.8425095170  0.933012308
#>  [58,] -0.016328213 -0.393472601 -0.725425297 -0.8991566207 -0.709160002
#>  [59,]  0.765655734 -0.041031300  0.889742802 -0.7301739948 -0.158210807
#>  [60,] -0.814750798 -0.687246993 -0.171134019 -0.2491846867  0.258301680
#>  [61,] -0.202786501 -0.362501699  0.486124202  0.8240422285 -0.565453503
#>  [62,]  0.884924782 -0.526403985 -0.775444577  0.1940878853  0.282020903
#>  [63,]  0.215076087  0.435617631 -0.427811982  0.5173250856  0.652103206
#>  [64,] -0.745843740  0.224590481  0.357993728 -0.6555566448 -0.120730331
#>  [65,]  0.626702070 -0.562530307 -0.389560004  0.9013561266  0.112099272
#>  [66,] -0.025098343  0.293766297 -0.007725463 -0.2597871870  0.256639106
#>  [67,]  0.321600117  0.309725833  0.441361277  0.3748383676 -0.806008493
#>  [68,] -0.417023983 -0.056397505 -0.486294102  0.5299472609  0.537761068
#>  [69,] -0.908448114 -0.464505553 -0.964072862  0.5908731092 -0.810631780
#>  [70,]  0.878672282 -0.115036958 -0.413509832 -0.2711475077 -0.593144918
#>  [71,] -0.042785633  0.742258870  0.212825258 -0.4680704195  0.757207780
#>  [72,] -0.083165165  0.404510706 -0.702578241  0.1980246687  0.371619090
#>  [73,] -0.453835976  0.892743660 -0.475256495  0.9171368736  0.986477605
#>  [74,] -0.882693195  0.647503242 -0.258480966  0.2476400258  0.731535238
#>  [75,]  0.834017453 -0.884856530 -0.349610475  0.2615882210  0.825043238
#>  [76,]  0.599380475  0.112975871 -0.983123326  0.2636491205  0.040478780
#>  [77,]  0.118652210 -0.536460933 -0.248102611 -0.7033168073  0.999970474
#>  [78,]  0.996768983  0.004513782 -0.112090081  0.0009900816  0.012847564
#>  [79,] -0.942803722 -0.518203555 -0.551259981 -0.7635037205  0.036711136
#>  [80,] -0.689290093 -0.358839697 -0.426280312  0.1153091858  0.605128979
#>  [81,] -0.144251945  0.885400734  0.090092019  0.2693914552  0.941422591
#>  [82,] -0.449192685  0.508480695 -0.216819611  0.5338551695 -0.254800091
#>  [83,]  0.972031676  0.157871210  0.751995749  0.6845021565 -0.785480685
#>  [84,] -0.280929759  0.024873128  0.463985582 -0.3090054831 -0.194144097
#>  [85,]  0.736415663 -0.338746626 -0.905178996  0.0871969727  0.173655712
#>  [86,] -0.220721909  0.758793116 -0.230082471 -0.3703019191  0.852216940
#>  [87,]  0.892802650 -0.844677160  0.725192542  0.9307734575  0.428694849
#>  [88,] -0.754847182 -0.983613283 -0.218975649 -0.9853951153  0.310590523
#>  [89,] -0.313128467  0.749488706  0.640007537  0.6710980400 -0.307848714
#>  [90,] -0.239421489 -0.516657267 -0.781462476 -0.2216942441  0.894298807
#>  [91,]  0.488979812 -0.964140469  0.224328557  0.5662707291  0.301762400
#>  [92,] -0.049759849  0.406944667 -0.420126695 -0.3037034804  0.481348155
#>  [93,]  0.003711776  0.675359313  0.892081462 -0.1491571749 -0.900087154
#>  [94,]  0.793456132  0.215762950 -0.782118558 -0.4579707603  0.325644463
#>  [95,] -0.682448203 -0.525740399  0.379813636 -0.8072633133  0.830662892
#>  [96,]  0.901847127 -0.846638823  0.310579120 -0.9450103464 -0.264611872
#>  [97,] -0.732181721  0.442789481 -0.292995413  0.5805120440 -0.343705601
#>  [98,] -0.410751273 -0.506498780  0.222511008  0.7595793996 -0.774599330
#>  [99,] -0.679084733 -0.648428087 -0.279633014 -0.0474370401 -0.982936437
#> [100,] -0.802830985  0.940407734 -0.295050015  0.3494466613  0.110465333
#> 
#> $y
#>   [1] 0 1 0 1 0 1 0 1 0 0 1 1 1 1 1 1 0 0 1 1 1 0 0 0 1 0 0 1 1 0 0 1 1 0 0 0 1
#>  [38] 0 0 1 1 0 0 0 0 1 1 1 1 0 1 1 1 1 0 0 1 1 1 0 1 0 1 1 0 1 1 1 0 0 1 1 0 0
#>  [75] 0 1 1 0 0 1 0 1 0 1 0 1 0 0 0 1 0 1 1 0 0 0 0 1 0 0
#>