Simulate Data from Hyper-Sphere for Classification Problems

The classification problem Circle is described in the BARK paper (2008). Inputs are dim independent variables uniformly distributed on the interval \([-1,1]\), only the first 2 out of these dim are actually signals. Outputs are created according to the formula $$y = 1(x1^2+x2^2 \le 2/\pi)$$

sim_circle(n, dim = 5)

Arguments

n

number of data points to generate

dim

number of dimension of the problem, no less than 2

Value

Returns a list with components

x

input values (independent variables)

y

0/1 output values (dependent variable)

References

Ouyang, Zhi (2008) Bayesian Additive Regression Kernels. Duke University. PhD dissertation, Chapter 3.

See also

Examples

sim_circle(n=100, dim=5)
#> $x
#>                [,1]         [,2]          [,3]         [,4]        [,5]
#>   [1,]  0.814723942  0.547324289 -0.3692065924  0.814319252 -0.44282682
#>   [2,] -0.419218032 -0.169606103 -0.1669495432 -0.589827147 -0.96211369
#>   [3,]  0.244211387  0.796546875  0.7869638954  0.752906831 -0.51930784
#>   [4,]  0.270322108 -0.860006701 -0.6648622425  0.564197555 -0.49418387
#>   [5,]  0.788504674 -0.577031807 -0.5260691894 -0.632483163  0.79417226
#>   [6,] -0.768931552  0.357989999 -0.2415045188  0.647139935  0.68655840
#>   [7,] -0.986823643 -0.464628943  0.9493719600 -0.986812834 -0.63704722
#>   [8,]  0.609585847 -0.461029470 -0.1674100226  0.576283341  0.47584856
#>   [9,] -0.950508194 -0.346419230 -0.5536865019  0.842442250 -0.55108280
#>  [10,]  0.496423183  0.046318247 -0.1852084659 -0.639603735 -0.71155004
#>  [11,] -0.480736152 -0.044604452  0.5427047345  0.391700028 -0.26798517
#>  [12,]  0.703332101  0.156646938  0.3784106551  0.315273931  0.75380209
#>  [13,]  0.750180318 -0.201463670  0.7508008010  0.109143943  0.19459006
#>  [14,] -0.051217788 -0.966967948  0.3630469274 -0.998032650  0.62729083
#>  [15,] -0.946918431 -0.098797833 -0.2317057564  0.884166930 -0.43936992
#>  [16,] -0.304779604 -0.403921519  0.3885981711 -0.473652118 -0.17301586
#>  [17,]  0.973962829  0.552647742 -0.5102731474 -0.531387274 -0.47926820
#>  [18,]  0.978535209 -0.522161086 -0.0847779131 -0.268931724 -0.51961787
#>  [19,] -0.045684219 -0.177822398 -0.7099176641  0.730356535 -0.83841944
#>  [20,] -0.307701351 -0.610449381  0.1489500469 -0.506881853 -0.47520491
#>  [21,] -0.072712270  0.116262700 -0.1454382841 -0.869922499 -0.63830905
#>  [22,]  0.404922847  0.773168027 -0.0217280011  0.919977345  0.35450695
#>  [23,] -0.289726412 -0.995513851  0.5006480752  0.201352204 -0.38358623
#>  [24,] -0.942266500  0.681154082  0.7421134515 -0.697449488 -0.68047053
#>  [25,]  0.242739851  0.586110662  0.0002648872  0.181167826  0.68540786
#>  [26,]  0.457957103 -0.505388838 -0.2767110514 -0.715202319 -0.50575858
#>  [27,] -0.996762551  0.420863020 -0.2730571697 -0.431157043 -0.08496036
#>  [28,]  0.434566948 -0.297953018  0.0154462275  0.286591629  0.14973423
#>  [29,]  0.835469784 -0.671447816  0.3519387008  0.675582059 -0.49715210
#>  [30,]  0.423662357 -0.661516327 -0.4620087361 -0.517451298 -0.13995069
#>  [31,]  0.053547431  0.241826496  0.5991883366  0.738888100  0.18272890
#>  [32,]  0.841034076  0.069547267 -0.6376308706 -0.391556912 -0.99003348
#>  [33,]  0.830741182  0.012651775  0.0012985575  0.586613339  0.96852462
#>  [34,] -0.412170528 -0.242689956  0.2374251746 -0.026700801  0.16289584
#>  [35,] -0.687399506  0.948461022  0.8917105091  0.658537707  0.90656177
#>  [36,]  0.392808000  0.200507020  0.5088281892  0.319667588  0.03498964
#>  [37,]  0.948114914  0.823961017  0.9506253484 -0.037687303 -0.78515470
#>  [38,]  0.540602256 -0.689902600 -0.9113367517  0.876364983  0.03898303
#>  [39,] -0.909873552 -0.213688980  0.9847112959  0.497496008 -0.81896749
#>  [40,] -0.001355306 -0.627609426 -0.5861545508  0.339316515 -0.09795626
#>  [41,] -0.909219445 -0.887703370 -0.8397733849  0.282983555  0.78938824
#>  [42,] -0.369504268 -0.692323588  0.2317760126 -0.401211438 -0.17051431
#>  [43,]  0.022234960 -0.144076138  0.8153324258  0.449351988 -0.72651972
#>  [44,] -0.228392486 -0.626827429 -0.1651138715 -0.746282184  0.07820437
#>  [45,] -0.451176076 -0.754382748  0.2985194768  0.056072494  0.35565483
#>  [46,]  0.545758890 -0.036363684  0.1060178960  0.977405562  0.08211392
#>  [47,]  0.670227856  0.506082180 -0.2640988594  0.664266506  0.62562731
#>  [48,] -0.752628060 -0.850147025  0.4607485319  0.226103199 -0.39382281
#>  [49,]  0.501878045  0.316397030 -0.5499212597 -0.990506134 -0.19729438
#>  [50,] -0.060208515 -0.651061138 -0.8511976707 -0.391170784  0.79662209
#>  [51,] -0.026946679  0.203189359  0.5561638037  0.493274714  0.75338931
#>  [52,] -0.835982562 -0.697181403  0.9233151902 -0.691869813  0.02245052
#>  [53,] -0.434734423  0.235922256  0.5117976042  0.205196345  0.23405855
#>  [54,]  0.424065869  0.170790443 -0.1123522809 -0.865351277  0.71713214
#>  [55,]  0.721113597  0.706786333  0.0216599205 -0.569224494 -0.96591866
#>  [56,]  0.270832742  0.153325078  0.1030850471 -0.926010089  0.11716359
#>  [57,] -0.032494829  0.532869561  0.5242443886  0.195648483  0.23469527
#>  [58,]  0.459669210 -0.308433887 -0.7685295641  0.750330061  0.11621356
#>  [59,] -0.333119720 -0.870712732  0.8478342742  0.777965355 -0.26075814
#>  [60,]  0.526003285 -0.680006262  0.4064821303 -0.321640612 -0.28312381
#>  [61,]  0.968405825 -0.997022146  0.1612244565 -0.274715182  0.36211700
#>  [62,] -0.336904880 -0.151612386 -0.2117362227 -0.658333584  0.36306340
#>  [63,]  0.102749906 -0.407110808  0.9644131334 -0.298805470 -0.18615862
#>  [64,]  0.481526400  0.828415075 -0.7120492123 -0.946746509  0.26539231
#>  [65,] -0.148342904  0.181726634  0.9065529536 -0.059637844  0.13964456
#>  [66,] -0.622615059 -0.963691866  0.5167860175  0.586414101  0.20554305
#>  [67,]  0.068454988 -0.797006974  0.9012137959 -0.903701383  0.79345022
#>  [68,] -0.756583186 -0.517036839 -0.6334975688  0.514082284  0.65827323
#>  [69,] -0.517643237 -0.238973167  0.4921238190  0.137211358 -0.81915699
#>  [70,]  0.343193419  0.717675107 -0.4997548470  0.470144469  0.37300429
#>  [71,]  0.482009703 -0.258134944  0.7190201879 -0.412049270 -0.33212689
#>  [72,]  0.528665265  0.628689754 -0.7813475085  0.571272715 -0.94631781
#>  [73,] -0.350623517  0.637553052  0.2751664473  0.098208643 -0.74377360
#>  [74,] -0.932500047  0.696252210 -0.4051858638 -0.313419250  0.51897973
#>  [75,] -0.532214947  0.744644888 -0.1418317952 -0.005514279 -0.17139543
#>  [76,]  0.092347877  0.246214611 -0.2215953595  0.990247189 -0.31360705
#>  [77,] -0.741581760  0.861548984 -0.7736788420 -0.045020421  0.74408972
#>  [78,]  0.066552814 -0.659463539 -0.2426941609 -0.550848639  0.26985212
#>  [79,]  0.116802536 -0.132295886 -0.5902897008  0.960912513 -0.78824508
#>  [80,]  0.409134339  0.959927556  0.9982023635  0.753279740 -0.70951673
#>  [81,]  0.627523690 -0.008909382  0.9280836917 -0.955333895 -0.10800484
#>  [82,] -0.040052959  0.703511282  0.9014028125 -0.052470182 -0.14596747
#>  [83,] -0.761654230 -0.073792436  0.1019443301 -0.613581797 -0.86172075
#>  [84,]  0.172859640 -0.792757898  0.0070295720  0.018443746  0.77691862
#>  [85,] -0.081352080 -0.502672968 -0.0517036249  0.287954671 -0.51615488
#>  [86,] -0.575211667 -0.493794277 -0.2894647056  0.997753033 -0.57498686
#>  [87,]  0.566177801  0.792349019  0.6496504117  0.540547332 -0.86387627
#>  [88,]  0.448220468 -0.670571433 -0.4071059008  0.996032846 -0.16139523
#>  [89,] -0.595953035 -0.590764618  0.1357695838 -0.678222828 -0.06764311
#>  [90,]  0.465682780 -0.672473074  0.1285492820 -0.549506126  0.82391301
#>  [91,] -0.701680942  0.819910843  0.2599997343  0.901609030 -0.93626194
#>  [92,] -0.612155736 -0.228666216  0.0575025729  0.805973625 -0.03140135
#>  [93,]  0.951934969 -0.648735945 -0.4245211273  0.842025547 -0.70376396
#>  [94,] -0.698736229 -0.155392930  0.7804510086 -0.527444730  0.03267176
#>  [95,]  0.701102788  0.751261297  0.6798660606  0.360716371  0.79222230
#>  [96,] -0.490596444  0.697454020 -0.6267267233 -0.087367984 -0.36607590
#>  [97,] -0.221481024 -0.607221368  0.5827831514 -0.312738242 -0.21089883
#>  [98,] -0.802002205 -0.479781486 -0.5114918915 -0.071924973  0.34677772
#>  [99,] -0.472991794  0.138982618 -0.5675501092 -0.575532663  0.91115803
#> [100,] -0.902443602  0.183523925 -0.6628281255 -0.402944268 -0.16802661
#> 
#> $y
#>   [1] 0 1 0 0 0 0 0 1 0 1 1 1 1 0 0 1 0 0 1 1 1 0 0 0 1 1 0 1 0 1 1 0 0 1 0 1 0
#>  [38] 0 0 1 0 1 1 1 0 1 0 0 1 1 1 0 1 1 0 1 1 1 0 0 0 1 1 0 1 0 0 0 1 1 1 0 1 0
#>  [75] 0 1 0 1 1 0 1 1 1 0 1 1 0 0 0 0 0 1 0 1 0 0 1 0 1 0
#>