Simulate Data from Hyper-Sphere for Classification Problems

The classification problem Circle is described in the BARK paper (2008). Inputs are dim independent variables uniformly distributed on the interval \([-1,1]\), only the first 2 out of these dim are actually signals. Outputs are created according to the formula $$y = 1(x1^2+x2^2 \le 2/\pi)$$

Usage

sim_circle(n, dim = 5)

Arguments

n

number of data points to generate

dim

number of dimension of the problem, no less than 2

Value

Returns a list with components

x

input values (independent variables)

y

0/1 output values (dependent variable)

References

Ouyang, Zhi (2008) Bayesian Additive Regression Kernels. Duke University. PhD dissertation, Chapter 3.

See also

Examples

sim_circle(n=100, dim=5)
#> $x
#>                [,1]        [,2]         [,3]         [,4]          [,5]
#>   [1,] -0.386932792  0.53432087  0.388968448 -0.064451589  0.8055823483
#>   [2,] -0.751924175 -0.18935891  0.321951516  0.999485814  0.2548044468
#>   [3,] -0.286577923 -0.91867089  0.840319376  0.875023143 -0.7733738204
#>   [4,]  0.257046662  0.70683806  0.496701429 -0.133052980  0.4600299429
#>   [5,] -0.857645318 -0.92253976 -0.001380942  0.364419137  0.8719380759
#>   [6,] -0.677659263 -0.72006188  0.339368736 -0.594344947 -0.7446361170
#>   [7,] -0.486616151 -0.64137124  0.288405532  0.724600292 -0.6628916850
#>   [8,] -0.940552192 -0.20555076  0.589763162  0.646952742  0.3549043681
#>   [9,] -0.766705319  0.43285646 -0.020258514  0.374810794 -0.9613470146
#>  [10,] -0.921157052 -0.73838883  0.312537115 -0.375633228  0.3534791567
#>  [11,] -0.631149623  0.03708633 -0.186325469  0.370940889  0.3129079631
#>  [12,]  0.158638651 -0.88177516 -0.643619190 -0.059535943 -0.4987169965
#>  [13,] -0.155363825 -0.34653510  0.967428351 -0.011292305 -0.9917395590
#>  [14,] -0.236815264 -0.79803260  0.201460162  0.537097715  0.6989094075
#>  [15,]  0.872526129 -0.02487895  0.137240640  0.697654358  0.7682007677
#>  [16,] -0.413336233  0.16312896  0.983995556  0.869906761 -0.4101919490
#>  [17,]  0.814465357 -0.79199083 -0.318238711  0.305206730  0.0463861669
#>  [18,]  0.542721695  0.85305306  0.200419073 -0.260485320  0.4844592446
#>  [19,]  0.243809295  0.58927877 -0.101107806  0.131472406 -0.2921771030
#>  [20,]  0.444541095 -0.68051225 -0.566081539 -0.721566880  0.5214602067
#>  [21,] -0.573617728 -0.28897044 -0.752242912  0.694738319  0.3293661638
#>  [22,]  0.436697519 -0.99655483  0.613946704 -0.911825310 -0.3304925356
#>  [23,]  0.725585905 -0.59289836  0.024621009 -0.401948989  0.4949996267
#>  [24,] -0.497550083  0.90884989 -0.491607429  0.444095457 -0.4021770302
#>  [25,]  0.873590733  0.29786773  0.683245167  0.481950904  0.0852259160
#>  [26,] -0.572521088  0.13866188  0.163225019 -0.635767674  0.6047661258
#>  [27,]  0.417097027 -0.69362319 -0.123976045 -0.127926643  0.0614978117
#>  [28,] -0.316674679  0.86956713  0.758249367  0.768593170  0.9151572399
#>  [29,] -0.722616809 -0.41899353  0.772720272 -0.564377599  0.2851851592
#>  [30,] -0.679411423 -0.70983609 -0.837073569 -0.796795865 -0.1175919683
#>  [31,]  0.812732277  0.97251798 -0.308940165 -0.937977214 -0.4875512286
#>  [32,]  0.199482291  0.30060574 -0.876779931 -0.813240564 -0.4060061318
#>  [33,]  0.290080788  0.27376116 -0.926531999 -0.510744650 -0.9366859612
#>  [34,]  0.096153222 -0.30957708  0.495425627 -0.791413459  0.6186687169
#>  [35,] -0.101513299 -0.94906322  0.718449811  0.493907466 -0.9641722306
#>  [36,]  0.097340752  0.69100430 -0.972554897 -0.603875591 -0.7946286825
#>  [37,] -0.416532570  0.68987098 -0.459871160 -0.817279191  0.0195772238
#>  [38,]  0.022758116 -0.80839656  0.137453977  0.582628711  0.2472125511
#>  [39,] -0.635312588  0.91318046  0.887667215  0.227606361  0.5079922075
#>  [40,] -0.249768953 -0.33464154  0.421861134  0.587081965 -0.7893642788
#>  [41,]  0.486272177  0.13095309 -0.795046947  0.071787397  0.3552702218
#>  [42,]  0.118457824 -0.67451665 -0.359391568 -0.116440844  0.2993230182
#>  [43,]  0.864164203  0.81960291 -0.377041974 -0.845280624 -0.4143482950
#>  [44,] -0.847699909  0.82189709  0.810755105 -0.210289788  0.9112494374
#>  [45,]  0.115974055 -0.54598540  0.843321155  0.657196937  0.1972714625
#>  [46,] -0.091762993 -0.80064857 -0.785184785  0.122545455 -0.6640943880
#>  [47,]  0.820956153 -0.40762395 -0.214763972  0.373594853 -0.5081906840
#>  [48,]  0.969644789  0.85876702 -0.977024078  0.394567392  0.7376766955
#>  [49,]  0.216211447  0.64641950 -0.570583746  0.517105812 -0.0397657114
#>  [50,]  0.178619903 -0.93433125  0.163314418  0.321367751  0.5777147566
#>  [51,] -0.472999102  0.95424819  0.046372913  0.945981697 -0.4822479212
#>  [52,] -0.132839015 -0.71742164  0.922597400 -0.243380155 -0.3190213894
#>  [53,] -0.102784526 -0.69563213 -0.312341222 -0.542030210  0.4250549865
#>  [54,] -0.020910336  0.69416616 -0.934752813 -0.244727640  0.2984564248
#>  [55,]  0.087640309  0.84117450  0.250122065 -0.811403465  0.1494878358
#>  [56,] -0.824821203  0.11664245 -0.330422326 -0.154537545  0.4034007452
#>  [57,] -0.844417965  0.96573752  0.671494070 -0.212736643 -0.2199845742
#>  [58,] -0.431891041  0.79867435 -0.107413540 -0.052693453  0.2887311322
#>  [59,]  0.092798195  0.70030888 -0.982831589  0.147916616  0.1442131456
#>  [60,] -0.005202973 -0.03161370 -0.326448288  0.521552146  0.2199266171
#>  [61,] -0.880186781 -0.67900724  0.428619386 -0.657028402  0.6817153986
#>  [62,]  0.274940478 -0.16379219 -0.687128101 -0.209306731 -0.3539732094
#>  [63,]  0.073444578 -0.84246962 -0.730558281  0.183140540  0.5507886959
#>  [64,]  0.379475202  0.54260681  0.438520781 -0.031297598 -0.0945897708
#>  [65,]  0.582782359 -0.50984626 -0.803596628  0.008364052 -0.3661997817
#>  [66,] -0.790296025 -0.25313988  0.453068501 -0.481971049 -0.9128416055
#>  [67,]  0.398064425 -0.64596926  0.830403792  0.821400970 -0.6589777945
#>  [68,]  0.483852969 -0.74685288 -0.965336415 -0.547991517 -0.9068583446
#>  [69,]  0.872630597 -0.12964505 -0.110813649 -0.072982646 -0.8580943807
#>  [70,] -0.400006441 -0.70837389 -0.167852807  0.611105357 -0.4476502463
#>  [71,] -0.099504556  0.34101081 -0.240102678 -0.773683981 -0.5704620709
#>  [72,]  0.707910610 -0.58223229 -0.348802644  0.357201872  0.1888139728
#>  [73,]  0.899420677  0.97172557  0.205643102  0.265207066  0.1085693091
#>  [74,] -0.653040533  0.25508714 -0.623311666  0.187949366  0.9768558815
#>  [75,]  0.238879276  0.77039249 -0.500299192 -0.294984631 -0.7747200858
#>  [76,]  0.670102169  0.58006213  0.963891806  0.290586854  0.6414533784
#>  [77,]  0.591017828 -0.18134843 -0.589152415  0.772413834  0.6051176386
#>  [78,] -0.135048873  0.41966931  0.694564712  0.412624376  0.2823131429
#>  [79,]  0.762707782 -0.42975988  0.166363403  0.374645761  0.7241430040
#>  [80,] -0.076519708 -0.93716366  0.273942336 -0.185411660 -0.9070215961
#>  [81,] -0.475355107 -0.88363892  0.960788956  0.544458210  0.1585226976
#>  [82,] -0.638926488 -0.68185051  0.999397213  0.144274986  0.0369909015
#>  [83,]  0.796798848  0.59826467 -0.729345152  0.707702572 -0.0009459285
#>  [84,] -0.029413285 -0.58096507  0.138874001  0.328001860 -0.1960512316
#>  [85,] -0.718532462 -0.24729783  0.535171189 -0.796631015  0.4861785090
#>  [86,]  0.370777060  0.91275855 -0.975744836 -0.292144926  0.0052495822
#>  [87,]  0.800749464  0.54407526 -0.591142255  0.562916742  0.0977736097
#>  [88,]  0.426988938  0.37393609 -0.872215867 -0.820172298 -0.0758025395
#>  [89,]  0.358164650  0.13073162  0.617912168  0.766827783  0.4253958305
#>  [90,]  0.322863863  0.55856758 -0.014297876  0.541376987 -0.6241223705
#>  [91,] -0.473639661  0.51004382 -0.792250702  0.974423472 -0.3022505706
#>  [92,]  0.033126894  0.29669612  0.107152193 -0.024410836 -0.0761856893
#>  [93,] -0.835933838  0.39331987 -0.245450828  0.750419667  0.3579316763
#>  [94,]  0.688087012 -0.04170658  0.258967406 -0.730383323  0.7025636267
#>  [95,] -0.211266861  0.31530279 -0.508882122 -0.551943602 -0.1931218710
#>  [96,]  0.284182219  0.08119667  0.721922564 -0.141035294  0.2473391304
#>  [97,] -0.085463248  0.43119756  0.132852026  0.105364331 -0.2488214611
#>  [98,] -0.660149978 -0.71019251  0.283698387 -0.673060849 -0.2901238552
#>  [99,]  0.259603717  0.27852863 -0.408700037  0.005018179 -0.8712221947
#> [100,]  0.880856361  0.44107247 -0.581219010  0.825774965  0.2200641944
#> 
#> $y
#>   [1] 1 1 0 1 0 0 0 0 0 0 1 0 1 0 0 1 0 0 1 0 1 0 0 0 0 1 0 0 0 0 0 1 1 1 0 1 0
#>  [38] 0 0 1 1 1 0 0 1 0 0 0 1 0 0 1 1 1 0 0 0 0 1 1 0 1 0 1 1 0 1 0 0 0 1 0 0 1
#>  [75] 0 0 1 1 0 0 0 0 0 1 1 0 0 1 1 1 1 1 0 1 1 1 1 0 1 0
#>