Simulate Data from Hyper-Sphere for Classification Problems

The classification problem Circle is described in the BARK paper (2008). Inputs are dim independent variables uniformly distributed on the interval \([-1,1]\), only the first 2 out of these dim are actually signals. Outputs are created according to the formula $$y = 1(x1^2+x2^2 \le 2/\pi)$$

sim_circle(n, dim = 5)

Arguments

n

number of data points to generate

dim

number of dimension of the problem, no less than 2

Value

Returns a list with components

x

input values (independent variables)

y

0/1 output values (dependent variable)

References

Ouyang, Zhi (2008) Bayesian Additive Regression Kernels. Duke University. PhD dissertation, Chapter 3.

See also

Examples

sim_circle(n=100, dim=5)
#> $x
#>                [,1]        [,2]         [,3]        [,4]        [,5]
#>   [1,] -0.975428408  0.65244254 -0.237278809 -0.48444574 -0.19990392
#>   [2,] -0.803878927 -0.46924196  0.432989834 -0.40371698 -0.59869890
#>   [3,] -0.413334650 -0.14987450 -0.526682260  0.23623722  0.31329753
#>   [4,]  0.570217982  0.78823037 -0.079449179  0.74484742  0.45970951
#>   [5,] -0.002741765  0.07583624  0.674601219 -0.50281417  0.68267444
#>   [6,] -0.970562949 -0.17850905  0.887014559 -0.71831565  0.75601118
#>   [7,] -0.243975915 -0.45165870 -0.405455671 -0.34621900  0.41681195
#>   [8,] -0.862952315  0.54824869 -0.364725275 -0.76507882 -0.22256610
#>   [9,]  0.593913001 -0.41917005  0.311986668  0.32231221 -0.55510043
#>  [10,] -0.607559115 -0.99075723 -0.901132265  0.91221836  0.44453453
#>  [11,]  0.928587306  0.79791861  0.210278295 -0.71446164 -0.46742556
#>  [12,] -0.217317204 -0.55295769  0.825090309 -0.87834418  0.14572029
#>  [13,]  0.912517678  0.35288051 -0.141054176 -0.56768475 -0.62706978
#>  [14,]  0.301248433  0.68594923 -0.643918144  0.69454302 -0.64850436
#>  [15,]  0.959934343  0.98237769  0.463059989  0.49516109  0.10285986
#>  [16,] -0.408186715  0.13476861  0.680640382 -0.22197695 -0.71601265
#>  [17,]  0.640034088  0.40595884 -0.347920721 -0.33189677  0.84977198
#>  [18,] -0.436866438  0.75124612 -0.734253268 -0.02225004  0.80566838
#>  [19,]  0.906995629 -0.05283450  0.186797993 -0.62438547  0.98706575
#>  [20,]  0.403411150  0.78801011  0.140405280  0.75567772 -0.63359202
#>  [21,] -0.167170784  0.11602451 -0.906138843 -0.82812436 -0.90629471
#>  [22,]  0.562174477  0.15475158 -0.966624843 -0.85407648 -0.75615469
#>  [23,]  0.113701323  0.31732125 -0.968424704  0.97654139 -0.70842722
#>  [24,]  0.715420782 -0.32680254  0.180137096  0.27576210  0.53272108
#>  [25,]  0.814595661  0.62648288 -0.234509612  0.55810329 -0.76993763
#>  [26,]  0.715434684  0.30517828  0.002875701 -0.35693178 -0.73767117
#>  [27,]  0.322984166 -0.37123772  0.648008890 -0.77145343 -0.56268226
#>  [28,] -0.883689815 -0.70878501  0.140738558 -0.65826556 -0.14440020
#>  [29,]  0.841318299  0.74021567  0.477609480  0.84117046  0.94744220
#>  [30,]  0.871721553 -0.04957659  0.567158291 -0.15068440 -0.86160407
#>  [31,]  0.480602765  0.40876032 -0.819510886  0.93059883  0.51213609
#>  [32,]  0.285276127  0.77668233 -0.555750833  0.12272332  0.93530778
#>  [33,] -0.241217244 -0.05440112  0.515854450 -0.48683961 -0.80195284
#>  [34,]  0.430436695 -0.95146560 -0.078695299 -0.55370091  0.04462993
#>  [35,]  0.874541780 -0.39077496  0.610417759 -0.77526786  0.53896832
#>  [36,]  0.538061257  0.96429459  0.059005941  0.63488032 -0.22909138
#>  [37,]  0.555310683  0.76447766  0.371786917 -0.62724326  0.02420086
#>  [38,] -0.022947555 -0.75490985 -0.821124175  0.30023381  0.76436690
#>  [39,] -0.932506845 -0.64289546 -0.305589103 -0.75898992  0.69247239
#>  [40,]  0.097295891  0.45568295  0.595342907 -0.54111657 -0.20284791
#>  [41,]  0.823895580  0.81881819 -0.974640836  0.20140149 -0.01687192
#>  [42,] -0.835874869  0.98789123  0.742354868  0.54076985  0.85008196
#>  [43,]  0.807546852  0.36280765 -0.420503078 -0.12885564 -0.82770024
#>  [44,]  0.939838809 -0.08625251  0.718781227  0.25967784  0.48283656
#>  [45,] -0.835772085 -0.95210469  0.484384553  0.64003790 -0.20413609
#>  [46,]  0.687007510  0.93908640 -0.133215009  0.21058845  0.51703731
#>  [47,]  0.617300435 -0.26801105 -0.582530080  0.58066425  0.38497977
#>  [48,] -0.756106090 -0.12711019  0.267193080 -0.50981687  0.07200941
#>  [49,]  0.002050433 -0.28276523 -0.302232408  0.18814541  0.23110405
#>  [50,] -0.182183424  0.62174039 -0.708879992 -0.69837296 -0.40355579
#>  [51,] -0.687159008  0.42363083  0.109983310 -0.02306913 -0.22220913
#>  [52,] -0.009872360 -0.40807226 -0.978039252 -0.38969083  0.78739646
#>  [53,] -0.599355379 -0.51451550  0.295920947  0.54137156  0.41784590
#>  [54,]  0.331212321  0.49255596  0.417504140  0.40133546  0.45248506
#>  [55,] -0.857850099  0.24359480  0.084538423 -0.88536485 -0.02860728
#>  [56,]  0.200121499 -0.22966555  0.946808044  0.85579555  0.96334274
#>  [57,]  0.956927619  0.64940962  0.739329011 -0.38482926  0.78316143
#>  [58,]  0.393223607 -0.97312867  0.767397776  0.71267637 -0.29480394
#>  [59,] -0.865164358 -0.40890434  0.235042975  0.82785004  0.41367265
#>  [60,] -0.804029552 -0.76584965 -0.863042993 -0.28207288 -0.98177163
#>  [61,]  0.199012277  0.71512461 -0.495741217  0.29847823  0.62787895
#>  [62,]  0.147861580  0.79763428  0.918049685  0.15228677  0.10996582
#>  [63,] -0.568145473  0.70905499 -0.059405676  0.33705825 -0.17521887
#>  [64,] -0.367497563 -0.81335628  0.545055512 -0.14805384  0.25744968
#>  [65,]  0.309252134 -0.02228395 -0.738332410  0.42868757  0.25081571
#>  [66,]  0.878446430  0.04307647  0.229608684 -0.53408518 -0.68334074
#>  [67,] -0.206259479 -0.27020905 -0.853617292 -0.67943397  0.30572557
#>  [68,] -0.462968155 -0.12174406 -0.062067533 -0.87727356 -0.44109027
#>  [69,] -0.320666701 -0.12139109 -0.851385255 -0.15351286 -0.44558105
#>  [70,] -0.074596471 -0.96049795 -0.466156768 -0.52287566  0.32253749
#>  [71,]  0.679286839  0.85238409  0.995513891  0.36965261 -0.78356582
#>  [72,] -0.967119450 -0.33200856 -0.487100142 -0.23547970 -0.57116702
#>  [73,]  0.622397892 -0.08395576 -0.457972814  0.75159111 -0.71869894
#>  [74,] -0.452926689  0.13729431 -0.242629127 -0.86162305 -0.14830844
#>  [75,] -0.872455182  0.03097193  0.292210904 -0.34489698  0.41366608
#>  [76,]  0.755109103 -0.58470274 -0.128080368 -0.18241888  0.07250570
#>  [77,] -0.922418996  0.13692607  0.977875285 -0.08806290 -0.56443530
#>  [78,]  0.495478397 -0.22258966  0.349685735 -0.92419166 -0.70938815
#>  [79,] -0.031876044  0.68557723 -0.893996275  0.26270946 -0.83248594
#>  [80,] -0.769795240  0.37357220 -0.562223705 -0.91560526  0.70583454
#>  [81,]  0.256887013 -0.11881123 -0.482820117 -0.12868694 -0.65993511
#>  [82,]  0.537731700  0.22439418  0.287384606 -0.53825007 -0.46056197
#>  [83,]  0.018366255  0.42582186 -0.834303992  0.10900236 -0.07242080
#>  [84,]  0.168923318  0.45419145 -0.468588018 -0.20524436  0.76499829
#>  [85,]  0.203542070  0.62088692  0.465463309 -0.20921166  0.26687831
#>  [86,]  0.945584153  0.51219311 -0.211516907  0.42999097 -0.18548675
#>  [87,] -0.983115434 -0.74242300  0.471803720 -0.81198661  0.10863642
#>  [88,] -0.938548968  0.82620530 -0.883857276 -0.89967337 -0.39548250
#>  [89,]  0.021260323 -0.87445082 -0.056245452 -0.14307714 -0.70701594
#>  [90,] -0.311117006  0.88020330  0.404385379  0.35649252 -0.74852222
#>  [91,] -0.506128523 -0.22397258 -0.412466981  0.84999385  0.38315137
#>  [92,] -0.566475757  0.36157968  0.004522495 -0.39610517 -0.80509950
#>  [93,]  0.357749626  0.03769288 -0.678515138 -0.59205343 -0.99445303
#>  [94,] -0.731895090 -0.72736017 -0.988527166  0.69447546  0.90480526
#>  [95,]  0.873200540  0.16761876 -0.836091343 -0.75090929  0.94987667
#>  [96,]  0.694456519 -0.81686473 -0.289347669 -0.01426363 -0.11412844
#>  [97,] -0.216751014 -0.71574852  0.987583980  0.81130610  0.48125220
#>  [98,] -0.461122039 -0.35710085  0.964505991 -0.60590750 -0.65796932
#>  [99,] -0.803276152 -0.92163743 -0.565348006  0.42195464 -0.77819170
#> [100,] -0.338991499 -0.46080915  0.261580028 -0.55241325  0.88613337
#> 
#> $y
#>   [1] 0 0 1 0 1 0 1 0 1 0 0 1 0 1 0 1 1 0 0 0 1 1 1 1 0 1 1 0 0 0 1 0 1 0 0 0 0
#>  [38] 1 0 1 0 0 0 0 0 0 1 1 1 1 0 1 1 1 0 1 0 0 0 0 1 0 0 0 1 0 1 1 1 0 0 0 1 1
#>  [75] 0 0 0 1 1 0 1 1 1 1 1 0 0 0 0 0 1 1 1 0 0 0 1 1 0 1
#>