Simulate Data from Hyper-Sphere for Classification Problems

The classification problem Circle is described in the BARK paper (2008). Inputs are dim independent variables uniformly distributed on the interval \([-1,1]\), only the first 2 out of these dim are actually signals. Outputs are created according to the formula $$y = 1(x1^2+x2^2 \le 2/\pi)$$

sim_circle(n, dim = 5)

Arguments

n

number of data points to generate

dim

number of dimension of the problem, no less than 2

Value

Returns a list with components

x

input values (independent variables)

y

0/1 output values (dependent variable)

References

Ouyang, Zhi (2008) Bayesian Additive Regression Kernels. Duke University. PhD dissertation, Chapter 3.

See also

Examples

sim_circle(n=100, dim=5)
#> $x
#>                [,1]        [,2]         [,3]         [,4]        [,5]
#>   [1,] -0.809712808  0.96102983  0.918465209  0.894184201 -0.56817133
#>   [2,]  0.139093823  0.67635998  0.142552268 -0.515157051 -0.99192170
#>   [3,] -0.143131977 -0.04742160 -0.110887373 -0.371269562 -0.68249774
#>   [4,] -0.015364506  0.52535132  0.912392266  0.149775731 -0.93456638
#>   [5,]  0.195680846 -0.26439698 -0.193692747  0.900096769  0.15672384
#>   [6,]  0.189154381  0.08855193  0.818384330 -0.741730229  0.22984863
#>   [7,] -0.252877341 -0.06809957 -0.726642310 -0.180502085 -0.02538709
#>   [8,] -0.203153369  0.06862997 -0.946622496 -0.855757993  0.91095232
#>   [9,] -0.217931615  0.32626517 -0.232066425 -0.272847387  0.18179281
#>  [10,]  0.754664443  0.35464941 -0.403365406 -0.149015195  0.10693925
#>  [11,] -0.016455501 -0.38876773 -0.385469608 -0.114128690  0.63939609
#>  [12,] -0.590434983 -0.48362268  0.778881168  0.971040920  0.28968929
#>  [13,] -0.055802109 -0.87122396 -0.238456104  0.736830879  0.11128134
#>  [14,]  0.210964349  0.92243105 -0.149816419  0.507833621  0.65080414
#>  [15,]  0.038091619 -0.66615379  0.376501345 -0.818875627  0.47697400
#>  [16,]  0.309872727 -0.23479704  0.436077730 -0.296528981  0.22173819
#>  [17,]  0.484975211 -0.07204279 -0.401493207  0.494293949  0.93888197
#>  [18,]  0.824033434  0.58007644  0.413465630 -0.523557012 -0.79261801
#>  [19,]  0.154074718  0.42679198  0.373756572 -0.685893592  0.92138537
#>  [20,] -0.228769172  0.19042803  0.116829473 -0.148335384  0.31041664
#>  [21,]  0.611387563 -0.77107559 -0.417848705  0.654661885  0.52724913
#>  [22,]  0.608435597 -0.92424141  0.390366883 -0.327460016  0.96299314
#>  [23,]  0.559023435  0.65429547 -0.181054327  0.293175325 -0.11511875
#>  [24,]  0.550370046  0.66980876  0.825618497 -0.130946855  0.67348685
#>  [25,] -0.205948661 -0.32793272 -0.603439827 -0.390176272 -0.39152489
#>  [26,]  0.645006248 -0.18762517 -0.811151546 -0.649566070 -0.59097452
#>  [27,]  0.383132461 -0.09098725  0.464320680 -0.731019309 -0.14558474
#>  [28,]  0.105457095 -0.80150011 -0.218067541 -0.391212130  0.84411046
#>  [29,]  0.948408620  0.76377118  0.928967200 -0.372343370  0.58622354
#>  [30,] -0.682355029 -0.65415090 -0.661846513  0.125316923  0.23008726
#>  [31,]  0.354241957 -0.64620467 -0.653507882  0.796462982  0.83117308
#>  [32,]  0.006782439 -0.95614898 -0.163955771 -0.965216352  0.33305654
#>  [33,] -0.695238189  0.57056331  0.811966825  0.173057338  0.59090762
#>  [34,] -0.457711331 -0.03436814 -0.775548416  0.978158329  0.20959801
#>  [35,]  0.312099611 -0.42237680  0.271764772 -0.679343821 -0.01972278
#>  [36,] -0.273006861 -0.56758586 -0.292106458  0.942704535 -0.23347068
#>  [37,] -0.092434635  0.58932095 -0.706638973 -0.897734206  0.99825782
#>  [38,]  0.020996396  0.22097194  0.500567208  0.053038536  0.15463757
#>  [39,]  0.198479618 -0.36251156 -0.841888866 -0.861203577 -0.37202596
#>  [40,]  0.408012762 -0.85613742 -0.843183973  0.783270141  0.14608008
#>  [41,]  0.753511441 -0.11955556 -0.446209878 -0.434442615  0.27796862
#>  [42,]  0.402582663  0.93379848  0.826129395  0.677314067 -0.42221145
#>  [43,] -0.290305419 -0.45840998 -0.718597749  0.004554597  0.45605332
#>  [44,]  0.919799978 -0.18786479 -0.976036817  0.919998040  0.76623328
#>  [45,]  0.115119859  0.15090514 -0.349512055  0.558509992 -0.59146602
#>  [46,] -0.236838211  0.52192484  0.916214332  0.928983284 -0.40053043
#>  [47,]  0.851992940  0.87947256 -0.770026765  0.466854292 -0.32421439
#>  [48,] -0.280372303  0.33836054 -0.607452172  0.402107977 -0.17343619
#>  [49,]  0.395777829  0.90320860 -0.968162375  0.749072731 -0.51199223
#>  [50,] -0.137827527  0.41077756  0.052408144  0.316859528 -0.80185254
#>  [51,]  0.413091243  0.17139951  0.115409841 -0.594461706  0.14492771
#>  [52,] -0.996410556 -0.51858945  0.983863596 -0.467906620 -0.73230361
#>  [53,]  0.301300629  0.56378536 -0.176073196 -0.859033860  0.60967569
#>  [54,] -0.892360014  0.72432408  0.180301096  0.484870852 -0.83679562
#>  [55,] -0.480632443 -0.61384358  0.646675828  0.677778519 -0.68246748
#>  [56,] -0.914525840 -0.79685145  0.772338149 -0.960384623 -0.36369814
#>  [57,] -0.112933701 -0.97766701 -0.520235984 -0.375758713 -0.50170169
#>  [58,] -0.554919000  0.31030091  0.532615158  0.311504623 -0.65626152
#>  [59,] -0.425114122  0.71335187 -0.197026242 -0.903737333 -0.47454129
#>  [60,]  0.251638231  0.95040644 -0.473025842  0.915510040 -0.65658684
#>  [61,] -0.958250225  0.40097461 -0.595692552 -0.798205192  0.46968572
#>  [62,]  0.543013879  0.51800736 -0.199641542  0.310738507 -0.28844276
#>  [63,]  0.113019339  0.10290128 -0.516720041  0.364846304 -0.09345447
#>  [64,]  0.765625798  0.91405993 -0.769041261  0.437238640  0.10784712
#>  [65,]  0.026965859 -0.34718380  0.606274332  0.554996064 -0.09169745
#>  [66,]  0.476688805  0.49728370  0.328523142  0.914698770 -0.37205249
#>  [67,]  0.420199146 -0.97755143  0.015326133 -0.141333112  0.53944196
#>  [68,] -0.694059766 -0.85343193  0.835460423  0.956701854 -0.18109740
#>  [69,] -0.399374721 -0.21918194  0.694753095  0.623305630  0.78339995
#>  [70,] -0.637850372 -0.22315969 -0.975326202 -0.998014582  0.29814232
#>  [71,]  0.473494264  0.89689253 -0.489515900  0.760522804  0.39982770
#>  [72,] -0.324917668  0.42546488 -0.749031391 -0.753088805 -0.35535732
#>  [73,] -0.019464126  0.58710734  0.102513602 -0.077345474  0.17591497
#>  [74,] -0.843726803 -0.05637655  0.146445563  0.038621000 -0.29199451
#>  [75,]  0.848258916 -0.57416702  0.853911981 -0.780678248 -0.65444165
#>  [76,] -0.371898749 -0.29896777  0.042523068  0.392951628 -0.39495420
#>  [77,] -0.005534932  0.25288181  0.580826797  0.090923631  0.10300897
#>  [78,]  0.048752676  0.04407547 -0.594258361 -0.752834865  0.88029671
#>  [79,]  0.387303854 -0.78551238  0.009112360  0.473165419 -0.43266758
#>  [80,] -0.843105746  0.34059082  0.620731335 -0.321071478  0.31334091
#>  [81,] -0.911458303  0.45390589 -0.019598696 -0.126845754 -0.48175557
#>  [82,] -0.296170772 -0.65498170 -0.306170992 -0.014213279 -0.18403944
#>  [83,] -0.521583714 -0.96206085  0.247085679  0.485736208  0.77049661
#>  [84,]  0.593492937 -0.32054933 -0.309466982 -0.092312882  0.73409901
#>  [85,] -0.351492065 -0.55660202 -0.431446442 -0.439543457 -0.64252826
#>  [86,] -0.070597429 -0.17033775 -0.003838814 -0.129935928 -0.25545546
#>  [87,]  0.133941672 -0.98856280 -0.043522969 -0.443825210  0.98912979
#>  [88,]  0.953776794  0.34512049 -0.854962686  0.699715486  0.81560483
#>  [89,] -0.575977516  0.79711009  0.168178183  0.287449270 -0.09418446
#>  [90,]  0.190251970  0.15332954  0.924032463  0.882844055  0.74494577
#>  [91,]  0.657429880 -0.44581055 -0.677196202 -0.989297166 -0.02762596
#>  [92,] -0.953624083 -0.43226968 -0.688797112 -0.113095226 -0.97799430
#>  [93,] -0.521338935 -0.80938245  0.825936859  0.809975869 -0.44164494
#>  [94,]  0.706002304 -0.76967008 -0.085384372 -0.740106107 -0.05622265
#>  [95,] -0.075995276 -0.01103545 -0.647647210 -0.356273018 -0.87952003
#>  [96,]  0.807380313 -0.35831977  0.275897818 -0.027088553 -0.78208548
#>  [97,] -0.992901576  0.32818826 -0.184836163  0.227073783 -0.76656326
#>  [98,] -0.835534231  0.16866622  0.832686574 -0.530767915  0.27536025
#>  [99,]  0.525418647  0.46311944  0.325894458 -0.646905413 -0.66716993
#> [100,] -0.400902711  0.98004411 -0.903700145 -0.404410801 -0.28565189
#> 
#> $y
#>   [1] 0 1 1 1 1 1 1 1 1 0 1 1 0 0 1 1 1 0 1 1 0 0 0 0 1 1 1 0 0 0 1 0 0 1 1 1 1
#>  [38] 1 1 0 1 0 1 0 1 1 0 1 0 1 1 0 1 0 1 0 0 1 0 0 0 1 1 0 1 1 0 0 1 1 0 1 1 0
#>  [75] 0 1 1 1 0 0 0 1 0 1 1 1 0 0 0 1 1 0 0 0 1 0 0 0 1 0
#>