Simulate Data from Hyper-Sphere for Classification Problems

The classification problem Circle is described in the BARK paper (2008). Inputs are dim independent variables uniformly distributed on the interval \([-1,1]\), only the first 2 out of these dim are actually signals. Outputs are created according to the formula $$y = 1(x1^2+x2^2 \le 2/\pi)$$

Usage

sim_circle(n, dim = 5)

Arguments

n

number of data points to generate

dim

number of dimension of the problem, no less than 2

Value

Returns a list with components

x

input values (independent variables)

y

0/1 output values (dependent variable)

References

Ouyang, Zhi (2008) Bayesian Additive Regression Kernels. Duke University. PhD dissertation, Chapter 3.

See also

Examples

sim_circle(n=100, dim=5)
#> $x
#>               [,1]        [,2]        [,3]          [,4]         [,5]
#>   [1,] -0.45337321 -0.15440273 -0.36002544  0.2965862108  0.379158223
#>   [2,] -0.75432050 -0.23496595 -0.21087363 -0.2354900502  0.512914974
#>   [3,] -0.16320148 -0.72553825  0.96515244 -0.6194817340  0.337259047
#>   [4,] -0.20313187  0.39787434 -0.14479649  0.0286802617  0.342651539
#>   [5,]  0.79098854  0.90671301  0.89286910 -0.3711162526 -0.475230402
#>   [6,]  0.09526981  0.45209425  0.94957917  0.1369540496 -0.144677991
#>   [7,] -0.34787636  0.04741027 -0.89583559 -0.0802801382  0.635960056
#>   [8,] -0.38651389 -0.39099578  0.95520477 -0.4546341891  0.957648429
#>   [9,]  0.11159934  0.56801363 -0.94571212  0.0305866282 -0.054808905
#>  [10,] -0.84088103  0.44768299  0.78082639  0.4302023738 -0.027542224
#>  [11,]  0.15599902  0.92240135  0.91852301 -0.7116278498  0.173516517
#>  [12,] -0.19144462 -0.97108795 -0.18218678 -0.5959345079 -0.812701639
#>  [13,] -0.16629297 -0.43593358  0.54255149 -0.9993222812 -0.796454681
#>  [14,] -0.99546254  0.35908289  0.47038071  0.4381411737 -0.033432959
#>  [15,]  0.46262110 -0.33374895 -0.78632467  0.1946868836  0.494156646
#>  [16,]  0.88727595  0.95478335  0.76123985 -0.9007620169 -0.050229829
#>  [17,]  0.81665468 -0.88829849  0.12012865  0.6195017900  0.401596005
#>  [18,]  0.62332643  0.09530042  0.66001071  0.9333309764  0.630582414
#>  [19,] -0.73789413  0.05012169  0.99259732  0.1103853313  0.086742317
#>  [20,] -0.89522497 -0.05478181  0.97967900 -0.7016754304  0.325005409
#>  [21,]  0.60698285 -0.89401464  0.04448390 -0.4307164685  0.183996286
#>  [22,]  0.40330683  0.20376729 -0.20872624 -0.4854850164  0.635725028
#>  [23,] -0.72486247  0.91361771  0.26641132 -0.6783858240  0.640466545
#>  [24,] -0.88602051 -0.81831825 -0.06792286 -0.7608044120  0.207278097
#>  [25,]  0.61440364  0.55504856  0.15944965 -0.4107811791  0.133160319
#>  [26,]  0.40995382  0.15095569  0.52672645  0.2661542813  0.860638130
#>  [27,]  0.14858411  0.60627155  0.45710144  0.0487562302  0.349180221
#>  [28,] -0.54007368  0.32046097 -0.26118526 -0.3643341856 -0.695924314
#>  [29,] -0.13292037 -0.76455758 -0.59082399 -0.5488472925  0.697173988
#>  [30,]  0.26526455  0.09344026  0.70544924 -0.2597394804  0.167218067
#>  [31,] -0.27687297  0.56667009 -0.51107576  0.0454222211 -0.767915084
#>  [32,]  0.37899886 -0.70413237 -0.95143103 -0.5571308797 -0.592847197
#>  [33,] -0.63310667 -0.18167032  0.14501528 -0.2296891487  0.200808020
#>  [34,]  0.90468780  0.01387299  0.04888805  0.1666247211  0.332389263
#>  [35,]  0.79722436 -0.51696578 -0.18321450  0.1507908260  0.704807657
#>  [36,]  0.62170355  0.94014689 -0.48360551 -0.5368686756 -0.923800398
#>  [37,]  0.36893603 -0.49500872 -0.73411391 -0.7942445911 -0.909460187
#>  [38,]  0.97601533 -0.02801187 -0.27709866  0.0006011664 -0.260618537
#>  [39,] -0.70384831  0.49634307  0.82764166 -0.7970078834  0.441261544
#>  [40,] -0.90276143 -0.03688870  0.44351079  0.0335843959  0.549176718
#>  [41,] -0.35930583 -0.36475570  0.23426619 -0.5686651575 -0.514108508
#>  [42,]  0.57421994  0.54974498 -0.70268396  0.7702704426 -0.079432148
#>  [43,] -0.56789033 -0.06217495  0.77745230  0.5759576783  0.428539509
#>  [44,]  0.03664226 -0.23634400  0.51248718  0.0190110817  0.162831294
#>  [45,]  0.36323511 -0.58088707  0.13862890 -0.3676827261  0.363171799
#>  [46,]  0.35518307 -0.19535766  0.10639891 -0.4073229157  0.088241171
#>  [47,]  0.15689665  0.62398256  0.42565399  0.6844695029 -0.390098744
#>  [48,]  0.48752668 -0.95645346 -0.80854389 -0.9985002186  0.258477600
#>  [49,] -0.57521730 -0.72062148 -0.71189993  0.4726992529 -0.964012460
#>  [50,] -0.88539503 -0.90224405 -0.91799409  0.8085272452 -0.542757651
#>  [51,]  0.07589259  0.95485541 -0.37235941  0.8091940354 -0.076694184
#>  [52,]  0.77462396 -0.00939974  0.92250756 -0.9546828330 -0.030524314
#>  [53,]  0.76022727  0.26097110  0.61852350  0.1396334181  0.655589969
#>  [54,]  0.50148990 -0.31643511 -0.81758719  0.3404514259  0.468602492
#>  [55,]  0.72603690  0.78488582  0.43166432 -0.8801779724  0.514247772
#>  [56,] -0.16981335 -0.34815881 -0.61366663 -0.1532587148  0.434532279
#>  [57,]  0.02385805  0.13308378  0.46071157  0.1783426544  0.485319975
#>  [58,] -0.65278619 -0.19032078  0.32919008 -0.0632894337 -0.618860456
#>  [59,]  0.72412878  0.95598015  0.64131359 -0.8011124325  0.165687948
#>  [60,] -0.70278150 -0.04720645 -0.12711146  0.7085705777  0.608776283
#>  [61,]  0.55607910  0.60624120  0.05215305  0.0045237099 -0.266682438
#>  [62,]  0.18384078  0.07414109  0.96079414 -0.9790292257  0.483946027
#>  [63,] -0.96560780 -0.50500383  0.66894525  0.3242693394  0.057200287
#>  [64,]  0.94099722  0.72491512  0.35217257  0.2373679481  0.746377723
#>  [65,]  0.43340589 -0.10327687  0.44712064 -0.0357593722  0.499159813
#>  [66,]  0.10600398  0.64018711  0.90824622  0.1186515307  0.462256293
#>  [67,]  0.44434146  0.06752408 -0.01306416 -0.6337109506  0.274371781
#>  [68,] -0.40778587  0.53359927  0.67862677  0.4255832452  0.777928038
#>  [69,] -0.87427697  0.18196765 -0.27251047 -0.2269438994 -0.872651574
#>  [70,]  0.67484198 -0.13718799 -0.23732762 -0.8287443835 -0.931787701
#>  [71,]  0.83436015  0.06796591  0.84535042 -0.0546863596 -0.449642210
#>  [72,] -0.05385681  0.52158453 -0.73222420 -0.8872021516  0.120620433
#>  [73,] -0.37991629  0.40078726 -0.03872896 -0.0877896426 -0.089462562
#>  [74,]  0.62848692 -0.95639270 -0.13174584  0.9759464511  0.447836055
#>  [75,] -0.74520080 -0.32983699 -0.81059972  0.2139284974 -0.921622456
#>  [76,]  0.24795419  0.84019120 -0.74878629 -0.8277212610 -0.026088834
#>  [77,]  0.75681356  0.58813746 -0.10751715 -0.8055556682 -0.006010721
#>  [78,] -0.09718239  0.25824891 -0.44160139  0.4645843925 -0.994891685
#>  [79,] -0.53438758 -0.74189357 -0.84285105 -0.6936989906 -0.369179529
#>  [80,] -0.69007507  0.22232992  0.64146809 -0.6105852299 -0.859794447
#>  [81,] -0.18196341  0.67254846  0.74674686 -0.1406930094 -0.155743792
#>  [82,] -0.66178814  0.06327119  0.41280467 -0.5880092820  0.352546920
#>  [83,]  0.07484688  0.73377353 -0.32498555  0.9569837442  0.616602177
#>  [84,] -0.37478203  0.25180419  0.79850678  0.4001889653  0.073217004
#>  [85,] -0.29584172  0.41853441  0.98946593  0.6341814757 -0.733197206
#>  [86,]  0.75612358 -0.71256860 -0.32736340  0.5221771328  0.201462476
#>  [87,] -0.86396451  0.18056278 -0.73761174 -0.2315342580 -0.848354442
#>  [88,]  0.49725372 -0.74450471 -0.23928999  0.7400127612  0.654921120
#>  [89,] -0.33333806 -0.01161714 -0.70503038  0.3560835929  0.086897193
#>  [90,]  0.23588037  0.93192864 -0.63700961  0.2721007289  0.033374227
#>  [91,] -0.10265260 -0.59950622  0.18499510 -0.8863437213  0.741880945
#>  [92,]  0.11103633  0.99696677  0.76955436  0.6161472201  0.819635718
#>  [93,] -0.30424286 -0.60123302 -0.09213235 -0.7254038979 -0.215417898
#>  [94,]  0.78880296 -0.05664779 -0.26472335  0.8603283167  0.537574255
#>  [95,] -0.31853825 -0.24981631  0.68742021  0.0582864098 -0.847632356
#>  [96,]  0.33309272  0.48698617  0.17817454  0.2183894785  0.241813079
#>  [97,]  0.21159356 -0.61542830  0.30655516  0.6983710146  0.219508259
#>  [98,]  0.35633064  0.72558761 -0.65590137  0.7841508868 -0.156109639
#>  [99,] -0.48074649  0.18209403  0.77788993 -0.4599930276 -0.301042203
#> [100,] -0.69430901 -0.14227618 -0.08299870  0.9860050236 -0.341581785
#> 
#> $y
#>   [1] 1 1 1 1 0 1 1 1 1 0 0 0 1 0 1 0 0 1 1 0 0 1 0 0 0 1 1 1 1 1 1 0 1 0 0 0 1
#>  [38] 0 0 0 1 1 1 1 1 1 1 0 0 0 0 1 0 1 0 1 1 1 0 1 0 1 0 0 1 1 1 1 0 1 0 1 1 0
#>  [75] 0 0 0 1 0 1 1 1 1 1 1 0 0 0 1 0 1 0 1 1 1 1 1 0 1 1
#>