Simulate Data from Hyper-Sphere for Classification Problems

The classification problem Circle is described in the BARK paper (2008). Inputs are dim independent variables uniformly distributed on the interval \([-1,1]\), only the first 2 out of these dim are actually signals. Outputs are created according to the formula $$y = 1(x1^2+x2^2 \le 2/\pi)$$

sim_circle(n, dim = 5)

Arguments

n

number of data points to generate

dim

number of dimension of the problem, no less than 2

Value

Returns a list with components

x

input values (independent variables)

y

0/1 output values (dependent variable)

References

Ouyang, Zhi (2008) Bayesian Additive Regression Kernels. Duke University. PhD dissertation, Chapter 3.

See also

Examples

sim_circle(n=100, dim=5)
#> $x
#>                [,1]        [,2]         [,3]        [,4]        [,5]
#>   [1,] -0.001215152  0.73408336 -0.260836456 -0.75508631  0.50997690
#>   [2,] -0.480344629 -0.56403575 -0.376496369  0.73913528  0.53115242
#>   [3,]  0.243806154  0.69846102 -0.636369307  0.24839108  0.19767916
#>   [4,] -0.725914646 -0.80135166 -0.332852456 -0.17250828 -0.05748272
#>   [5,] -0.895816830  0.81589418 -0.030034706  0.37675161 -0.22422613
#>   [6,] -0.831019793 -0.63899748  0.492896495  0.62500670  0.98081151
#>   [7,] -0.872506893 -0.03985602  0.014845480  0.80895788 -0.17897743
#>   [8,]  0.375942257 -0.75566247 -0.126465174  0.26924980 -0.03911668
#>   [9,]  0.214065704 -0.68885185  0.806102249 -0.39683647 -0.69694443
#>  [10,]  0.517235354  0.71054838  0.247678780 -0.26325208 -0.43935720
#>  [11,]  0.313698396 -0.06932477  0.189762426  0.15972504  0.08979108
#>  [12,] -0.293020459  0.79268690 -0.452661353 -0.97877884 -0.05478371
#>  [13,] -0.844119248  0.88889921 -0.652681557 -0.69558095  0.03065685
#>  [14,]  0.104820895  0.24921307  0.302130131 -0.95550991  0.27084953
#>  [15,] -0.713477854 -0.49251846  0.351811817 -0.03639487  0.18512988
#>  [16,]  0.395056919 -0.33410901 -0.267484953  0.65012935  0.07571582
#>  [17,]  0.436493475 -0.72086313 -0.349386587  0.69946849  0.65908196
#>  [18,]  0.114676181  0.79349016 -0.907061885  0.42148050  0.71734969
#>  [19,] -0.624356721 -0.08955720 -0.952352970 -0.94033189  0.84084852
#>  [20,]  0.128939918  0.98148625  0.913371968  0.08448968  0.38848704
#>  [21,] -0.286559933  0.47712434 -0.562523311 -0.46608144 -0.37678464
#>  [22,]  0.009324590  0.09035238 -0.674821986 -0.91885704 -0.34039745
#>  [23,] -0.005874878  0.71922394  0.015158339  0.87978829  0.13643390
#>  [24,]  0.197140877  0.11168266 -0.655139879 -0.02648947  0.02522444
#>  [25,]  0.936250858  0.75903757  0.818588436 -0.39505805 -0.74673175
#>  [26,]  0.185174414 -0.64333331 -0.494555719  0.85678698  0.47238574
#>  [27,] -0.262761207 -0.47957724 -0.675217613 -0.68083135  0.28999064
#>  [28,]  0.682388327  0.37285796  0.756449374 -0.29428983 -0.21473396
#>  [29,]  0.022428131  0.85320090 -0.268075393 -0.55714402 -0.18129888
#>  [30,]  0.796372706 -0.44313962 -0.002921395 -0.96485004 -0.49047661
#>  [31,] -0.483005787  0.22539152 -0.075783272 -0.36657547  0.73486939
#>  [32,]  0.573537373  0.75350590  0.323604101  0.07768394  0.88253649
#>  [33,] -0.904936622 -0.20747946  0.836864589 -0.70911780 -0.45175327
#>  [34,]  0.067267823  0.64462253  0.856684565 -0.97833814  0.59320006
#>  [35,]  0.377161319 -0.11977377 -0.557073473  0.65619704 -0.85905514
#>  [36,] -0.154519288  0.14222670 -0.883234086 -0.21851275 -0.01172287
#>  [37,]  0.956407130  0.29969443  0.542153950 -0.77281673 -0.69063972
#>  [38,]  0.259402731  0.18206730 -0.808755321 -0.72812925 -0.29312968
#>  [39,] -0.527647981 -0.22985409  0.189658711  0.40001422  0.60769798
#>  [40,]  0.159611102  0.08176403  0.809622440  0.44778319  0.21798731
#>  [41,] -0.564128112 -0.42751663 -0.306747439 -0.17493316  0.19215473
#>  [42,] -0.356399388 -0.39542861  0.564461496  0.10723662 -0.59370367
#>  [43,]  0.934061371 -0.50605335 -0.752730233  0.49627516 -0.53604645
#>  [44,] -0.705833069 -0.48369430 -0.177028757  0.34216666 -0.83245985
#>  [45,]  0.103376440  0.86091508 -0.601813200 -0.73707072  0.53742671
#>  [46,]  0.646050845 -0.83460240  0.146530508 -0.86249618 -0.17660293
#>  [47,]  0.027862646 -0.15817775 -0.406281457 -0.36904872  0.01091563
#>  [48,] -0.448534701 -0.29436881 -0.615712848  0.41921888  0.25918363
#>  [49,]  0.240097537  0.94334184 -0.883823643  0.96954462 -0.77273614
#>  [50,]  0.129151926  0.30989729  0.248681174  0.46759769 -0.50058317
#>  [51,] -0.766744526 -0.27784983 -0.529190444 -0.48594239  0.65060861
#>  [52,]  0.698796783  0.94926931 -0.208306566  0.87079297  0.85873275
#>  [53,] -0.265042200  0.68988029 -0.310167233 -0.29578117 -0.29167833
#>  [54,] -0.508773001  0.97381099  0.273615711  0.09787281 -0.09945066
#>  [55,] -0.886171001 -0.35156158  0.208937999 -0.85169730  0.86913201
#>  [56,]  0.137797682 -0.69589943 -0.828191233  0.47025197  0.20556067
#>  [57,] -0.979141502  0.29432301  0.704136986  0.07544507  0.86483304
#>  [58,]  0.988488143 -0.06051191  0.333631342  0.04093041  0.06906792
#>  [59,]  0.140327265  0.16439471 -0.464180213 -0.08653321  0.14341993
#>  [60,]  0.830852555 -0.05995170 -0.815554859 -0.49767583  0.27383375
#>  [61,] -0.204909147 -0.79338514 -0.648727850  0.31643005 -0.50201937
#>  [62,]  0.876857380 -0.24853412 -0.497713450 -0.28163888 -0.61158184
#>  [63,]  0.101195170  0.55445548  0.873865087 -0.12546111 -0.65249977
#>  [64,] -0.593875798 -0.66078501 -0.611830922  0.62053840 -0.23260827
#>  [65,]  0.712441029  0.50664126 -0.038583992 -0.99104009  0.74501024
#>  [66,] -0.797614637 -0.33008987  0.458960587  0.13259817  0.26343508
#>  [67,] -0.545567634  0.64179596 -0.584981945  0.21537874  0.17617527
#>  [68,] -0.150789389  0.81337385  0.163398006 -0.81093611 -0.87749053
#>  [69,]  0.679818248  0.65230988 -0.534684231 -0.22589600 -0.26078535
#>  [70,]  0.954203975  0.50316456  0.311332047  0.04676675  0.88866552
#>  [71,]  0.041994283  0.53158876 -0.267502605 -0.08443877 -0.81386922
#>  [72,] -0.633053298  0.54159058 -0.795179320  0.45536863 -0.76658907
#>  [73,]  0.453376432 -0.32787437 -0.582013183 -0.49284694  0.95199349
#>  [74,] -0.843719837 -0.20948496  0.223764156  0.75157550 -0.47641522
#>  [75,]  0.199584746  0.88725397  0.335450390 -0.01528387  0.04865926
#>  [76,]  0.381488086  0.58879963  0.213719743 -0.46282606 -0.70882254
#>  [77,] -0.479194765  0.72505129  0.223801204  0.37794674  0.56007579
#>  [78,]  0.321511039  0.71102247  0.821112823 -0.70400249  0.70888986
#>  [79,] -0.400440515 -0.42973526 -0.844976743 -0.74627246 -0.71760121
#>  [80,]  0.703329386  0.19741025 -0.421526183  0.43225716 -0.27399435
#>  [81,] -0.906174336  0.74739926  0.282937620 -0.76454437  0.53340616
#>  [82,] -0.904994065 -0.90956882  0.709380164  0.26299156 -0.49241312
#>  [83,]  0.173935421  0.76519632  0.429774228  0.67156172 -0.09542485
#>  [84,]  0.819916599 -0.24090445 -0.416577686  0.16953229  0.82995840
#>  [85,]  0.193103533 -0.39081971  0.210395649  0.23004040 -0.21711677
#>  [86,] -0.795254443  0.04736169 -0.754837662  0.61560018  0.09109136
#>  [87,] -0.114912879  0.51449831 -0.406013417 -0.49962248 -0.02438276
#>  [88,] -0.200389622 -0.61390580  0.950602138  0.73823236 -0.73576532
#>  [89,]  0.279262420  0.00134912  0.347105844  0.13562331  0.64233425
#>  [90,] -0.966214093 -0.38671576  0.735728276 -0.40974289 -0.55651684
#>  [91,] -0.094786346 -0.58929071 -0.465086105  0.98957723  0.88387564
#>  [92,]  0.923744353  0.05358952 -0.536993736 -0.74366885 -0.54531656
#>  [93,] -0.943136320 -0.84259122 -0.072665819  0.14361959 -0.36868009
#>  [94,]  0.701735682 -0.22502293 -0.107786794  0.85563220 -0.82010344
#>  [95,]  0.988939390 -0.24293829 -0.626345345 -0.37749746  0.59757843
#>  [96,]  0.072236736 -0.96002888  0.418411460  0.39935174 -0.48122248
#>  [97,]  0.366072992  0.27786806  0.157403019  0.57672825 -0.92132835
#>  [98,]  0.274937118 -0.41313037  0.737423635  0.24213283 -0.86228969
#>  [99,]  0.085825276  0.76318671  0.599054171 -0.51541633  0.18575444
#> [100,]  0.188571826 -0.31534101 -0.279044837  0.30987331  0.51559729
#> 
#> $y
#>   [1] 1 1 1 0 0 0 0 0 1 0 1 0 0 1 0 1 0 0 1 0 1 1 1 1 0 1 1 1 0 0 1 0 0 1 1 1 0
#>  [38] 1 1 1 1 1 0 0 0 0 1 1 0 1 0 0 1 0 0 1 0 0 1 0 0 0 1 0 0 0 0 0 0 0 1 0 1 0
#>  [75] 0 1 0 1 1 1 0 0 1 0 1 1 1 1 1 0 1 0 0 1 0 0 1 1 1 1
#>