Simulate Data from Hyper-Sphere for Classification Problems

The classification problem Circle is described in the BARK paper (2008). Inputs are dim independent variables uniformly distributed on the interval \([-1,1]\), only the first 2 out of these dim are actually signals. Outputs are created according to the formula $$y = 1(x1^2+x2^2 \le 2/\pi)$$

sim_circle(n, dim = 5)

Arguments

n

number of data points to generate

dim

number of dimension of the problem, no less than 2

Value

Returns a list with components

x

input values (independent variables)

y

0/1 output values (dependent variable)

References

Ouyang, Zhi (2008) Bayesian Additive Regression Kernels. Duke University. PhD dissertation, Chapter 3.
http://stat.duke.edu/people/theses/OuyangZ.html

See also

Examples

sim_circle(n=100, dim=5)
#> $x
#>                [,1]        [,2]         [,3]        [,4]         [,5]
#>   [1,] -0.113667552 -0.53181286  0.640265322  0.13219207  0.380474854
#>   [2,] -0.215274225  0.44929422  0.221709032 -0.41176646  0.086903964
#>   [3,] -0.350317134 -0.63026584 -0.482139044 -0.32047920  0.657082712
#>   [4,] -0.058267030 -0.14016046 -0.449268313  0.81161383  0.573642068
#>   [5,] -0.712230732 -0.01364837 -0.044594378  0.34346368  0.255623545
#>   [6,] -0.992705520 -0.36073974 -0.195732383  0.27921153  0.555137289
#>   [7,]  0.671246497 -0.65169252 -0.189094327 -0.07781514 -0.654159016
#>   [8,]  0.949662668 -0.65353342 -0.840195959 -0.59247909 -0.263051353
#>   [9,] -0.660461289 -0.24493343  0.415231711 -0.99713333 -0.986487454
#>  [10,] -0.949231619  0.97073501 -0.008749646  0.98949950 -0.676722649
#>  [11,] -0.008393071 -0.15187539  0.969720819 -0.92686483 -0.277735056
#>  [12,]  0.393289914 -0.84137221  0.221287115  0.31776975 -0.190358563
#>  [13,]  0.204875504  0.45877303 -0.672137577 -0.46772661  0.004664798
#>  [14,]  0.616367102 -0.86113296  0.737951951 -0.22638342  0.313844040
#>  [15,]  0.912274456 -0.08302164 -0.630162160  0.60086817  0.594664427
#>  [16,]  0.371107244  0.14616101  0.581286129  0.55068899  0.825929047
#>  [17,] -0.993803271 -0.92022803 -0.365790766  0.67311326  0.036741795
#>  [18,]  0.800811521  0.07020302 -0.831055625  0.60376278 -0.148985322
#>  [19,] -0.931056913 -0.92744030  0.327064159  0.18925954  0.395910174
#>  [20,] -0.036160096  0.66975040  0.258543836  0.38822596 -0.038818521
#>  [21,] -0.712998214  0.53202891  0.755763472  0.78014003 -0.757724319
#>  [22,]  0.196964601 -0.83630433 -0.771672581  0.06539890 -0.364416456
#>  [23,] -0.987193059 -0.16859394  0.547028276  0.39805977  0.358758224
#>  [24,] -0.958543335 -0.72860345 -0.536981052  0.43586936 -0.989587583
#>  [25,] -0.304180143  0.05223340 -0.873426094 -0.45989219  0.553508809
#>  [26,]  0.489590838 -0.52196915  0.373309416  0.95432914 -0.135136322
#>  [27,] -0.421829743 -0.16610786 -0.969879464  0.31801099 -0.711861512
#>  [28,] -0.558002119  0.38576844  0.907600504 -0.68584985  0.899637277
#>  [29,]  0.622050932  0.05774742 -0.561245753 -0.01046062 -0.591283262
#>  [30,] -0.251560897  0.61464022  0.972436468  0.30946910 -0.966937094
#>  [31,]  0.400127660  0.57187412 -0.221524240  0.77661442  0.752619080
#>  [32,]  0.289606409 -0.83300517  0.701532424 -0.02632890  0.288546135
#>  [33,] -0.780790751  0.28177247  0.062594719  0.97602900 -0.224060675
#>  [34,] -0.202300222  0.80495651 -0.409761440  0.83122289 -0.017347605
#>  [35,] -0.036194184  0.54338149  0.515100114  0.87287450 -0.522012192
#>  [36,]  0.011294321  0.59522259 -0.175327440  0.79103178 -0.314530197
#>  [37,]  0.008898570 -0.79685058  0.866856011  0.39842094  0.512591828
#>  [38,]  0.258168914  0.75560912  0.969613762  0.84139496  0.999454417
#>  [39,] -0.916030950 -0.24107040 -0.294187745  0.60683145  0.006996394
#>  [40,] -0.926300449  0.74851386  0.653997979  0.56473673 -0.546440626
#>  [41,] -0.076272937 -0.38248222 -0.700817662  0.81249335  0.113604860
#>  [42,]  0.680036247 -0.70701285 -0.872645455 -0.56088292 -0.361256010
#>  [43,]  0.192297006  0.10561961  0.187587314 -0.90802511 -0.546270146
#>  [44,] -0.073972250  0.61035572  0.886145615 -0.27293604 -0.932527882
#>  [45,]  0.343035666 -0.71621225 -0.012086622 -0.37885868  0.780280875
#>  [46,]  0.274362228 -0.56999875 -0.962634407  0.85833169  0.129236604
#>  [47,]  0.290610872 -0.86216842 -0.223570747 -0.48010619 -0.840021391
#>  [48,] -0.607932943 -0.12820295 -0.494267648  0.96705186  0.527293571
#>  [49,]  0.363865348 -0.16815804  0.201907039  0.21801533 -0.960138757
#>  [50,]  0.776519602 -0.78664086  0.571537663  0.44849368 -0.299546542
#>  [51,] -0.171136237 -0.05381458  0.957663604 -0.75329634 -0.094938834
#>  [52,]  0.171166608 -0.95764201 -0.489278614  0.20457521  0.455004873
#>  [53,] -0.622744128  0.10218324  0.877781829 -0.75943185  0.441844022
#>  [54,] -0.837327761 -0.20053631 -0.271987131 -0.06032635 -0.339367005
#>  [55,]  0.560265763  0.42009009  0.675665520 -0.25729565  0.407361805
#>  [56,] -0.717339440  0.24292260  0.465176439 -0.88546087 -0.639685707
#>  [57,]  0.623833853  0.90608675  0.417630145 -0.94166125 -0.703709914
#>  [58,]  0.246534738  0.45049569  0.024868537 -0.50978856 -0.982739367
#>  [59,] -0.196731876  0.56255345 -0.339785819 -0.54587695 -0.668833214
#>  [60,]  0.389474596  0.16019743 -0.940430702  0.69516953  0.728091452
#>  [61,]  0.850944242  0.47575123  0.401503847  0.57076377 -0.665909389
#>  [62,] -0.944551819 -0.36856503 -0.984542543  0.88387325  0.649102162
#>  [63,] -0.014763159  0.62498006 -0.539191257 -0.14524374 -0.252106279
#>  [64,]  0.942831282 -0.22494675 -0.003561866 -0.91382113 -0.012937475
#>  [65,] -0.209329919  0.18042179 -0.646493263  0.10428415 -0.158034208
#>  [66,] -0.927904762 -0.23825455 -0.189013469 -0.99635583 -0.068950668
#>  [67,]  0.417548784 -0.75178241 -0.903993177  0.30451368 -0.034641409
#>  [68,]  0.461668056 -0.52443899  0.944991962  0.73342656  0.574693156
#>  [69,] -0.416929645 -0.54618655 -0.283155515 -0.08642325 -0.333937855
#>  [70,] -0.657765837 -0.60238098 -0.816336791 -0.41978456 -0.597926371
#>  [71,]  0.039903396  0.24321426 -0.746883558  0.36761337  0.736016776
#>  [72,] -0.301701047 -0.45685271  0.199509548 -0.47679562  0.375082528
#>  [73,]  0.696693826 -0.85216351 -0.024209694  0.31992306  0.588396349
#>  [74,]  0.452888055  0.84472931 -0.211094618  0.26309461  0.425948142
#>  [75,] -0.324038922 -0.02258854  0.528356529 -0.73563398 -0.815342498
#>  [76,] -0.249629572  0.91495979 -0.147153296 -0.78220191  0.375574594
#>  [77,] -0.701472237  0.35014197  0.149962763  0.76122818  0.377082149
#>  [78,]  0.972478047  0.78986778  0.768359248 -0.71058766 -0.053904519
#>  [79,]  0.489921512 -0.17362365 -0.070984899 -0.62889190  0.403417646
#>  [80,]  0.593727560 -0.58451712  0.317145441  0.68333205  0.657712665
#>  [81,] -0.146303119 -0.06781463  0.190632571 -0.96882614  0.060401077
#>  [82,]  0.537235709 -0.24363379 -0.443352312 -0.29686481  0.959477718
#>  [83,]  0.168016541  0.49649002  0.470808738  0.50467305 -0.691393541
#>  [84,] -0.156573534 -0.18322510 -0.977951272  0.69413495  0.931482047
#>  [85,]  0.798017999  0.52587149 -0.543808721 -0.58852806  0.084763352
#>  [86,] -0.925020670 -0.69706315 -0.920138297 -0.79942534  0.983011597
#>  [87,]  0.055761581  0.82256184 -0.528906657 -0.72206815 -0.468519005
#>  [88,]  0.805323745 -0.04849254  0.380360440 -0.32610206 -0.227007311
#>  [89,] -0.114533086  0.89618916  0.004096223 -0.03312316 -0.663988346
#>  [90,] -0.266407280 -0.78443963 -0.272004624 -0.45760766  0.206541871
#>  [91,] -0.602964651 -0.97897321 -0.623318510 -0.77888522  0.042104645
#>  [92,] -0.669262680 -0.31646686 -0.162221199 -0.39672334  0.287312281
#>  [93,]  0.536599267  0.26987391  0.907068547 -0.85475724 -0.425900656
#>  [94,] -0.744952501 -0.59957011  0.846076578 -0.15772891 -0.253780844
#>  [95,] -0.100447663  0.66868760  0.972428859 -0.14842816 -0.025177999
#>  [96,] -0.138569454  0.13412609 -0.448698825  0.77870557 -0.178615240
#>  [97,]  0.840925807  0.76958756  0.100672830 -0.81456669 -0.033116865
#>  [98,]  0.543882132 -0.49021856  0.247101011 -0.36915120 -0.055871479
#>  [99,]  0.409765646 -0.72483684 -0.882952926 -0.95642556  0.515869392
#> [100,]  0.680185847  0.45506374 -0.455635178  0.20939604  0.431975986
#> 
#> $y
#>   [1] 1 1 1 1 1 0 0 0 1 0 1 0 1 0 0 1 0 0 0 1 0 0 0 0 1 1 1 1 1 1 1 0 0 0 1 1 1
#>  [38] 0 0 0 1 0 1 1 1 1 0 1 1 0 1 0 1 0 1 1 0 1 1 1 0 0 1 0 1 0 0 1 1 0 1 1 0 0
#>  [75] 1 0 1 0 1 0 1 1 1 1 0 0 0 0 0 0 0 1 1 0 1 1 0 1 0 0
#>