Simulate Data from Hyper-Sphere for Classification Problems

The classification problem Circle is described in the BARK paper (2008). Inputs are dim independent variables uniformly distributed on the interval \([-1,1]\), only the first 2 out of these dim are actually signals. Outputs are created according to the formula $$y = 1(x1^2+x2^2 \le 2/\pi)$$

sim_circle(n, dim = 5)

Arguments

n

number of data points to generate

dim

number of dimension of the problem, no less than 2

Value

Returns a list with components

x

input values (independent variables)

y

0/1 output values (dependent variable)

References

Ouyang, Zhi (2008) Bayesian Additive Regression Kernels. Duke University. PhD dissertation, Chapter 3.
http://stat.duke.edu/people/theses/OuyangZ.html

See also

Examples

sim_circle(n=100, dim=5)
#> $x
#>               [,1]         [,2]        [,3]        [,4]          [,5]
#>   [1,] -0.04578811  0.483223097 -0.50947446 -0.43097781 -0.7261972702
#>   [2,] -0.96446384  0.691770026 -0.30418984 -0.14698319  0.2519525038
#>   [3,]  0.34588237  0.381753832  0.89954484 -0.19685145 -0.8004475837
#>   [4,]  0.18664035  0.878122823  0.32226439  0.76738737 -0.4268200486
#>   [5,]  0.76643743  0.209438925 -0.59897638 -0.68929962  0.8480300638
#>   [6,] -0.18511341  0.302477392 -0.07942831  0.58131055  0.7725216574
#>   [7,] -0.99670644 -0.526404845 -0.27719359 -0.62384343 -0.7676683422
#>   [8,]  0.57449919 -0.875172345 -0.26402239  0.10299543 -0.6045500147
#>   [9,]  0.55302061 -0.831505676  0.14110997  0.38655706  0.9701155038
#>  [10,] -0.64714232 -0.630396573 -0.06270043  0.43570487 -0.4060426983
#>  [11,] -0.35176660  0.836525708 -0.32547702  0.04869261 -0.8768354552
#>  [12,]  0.44947904 -0.196074213 -0.99703320 -0.33252280 -0.5471636648
#>  [13,] -0.98860457  0.672951689  0.51232620  0.92089999  0.3926591403
#>  [14,] -0.69197987  0.156653662  0.49104362  0.80127862 -0.9146999712
#>  [15,] -0.31529347 -0.768006967 -0.25340630 -0.63428934 -0.4656621967
#>  [16,] -0.27971943  0.002932833 -0.02880560  0.66846059  0.7982990053
#>  [17,] -0.15640185 -0.634861044  0.75277012 -0.12309264  0.0226434460
#>  [18,] -0.21909129  0.837411685  0.10982790 -0.12523446  0.0341656506
#>  [19,] -0.69319692 -0.978545373  0.92921511 -0.77233241 -0.4571297504
#>  [20,]  0.32146208 -0.841539944  0.94895985 -0.00918935 -0.0072296667
#>  [21,] -0.11057723 -0.694887863 -0.35341840 -0.21344284  0.3650433407
#>  [22,] -0.32629360  0.878055228  0.72101244 -0.27751350 -0.7668245556
#>  [23,] -0.37153082  0.015409946  0.50390111 -0.26087908 -0.4567414024
#>  [24,] -0.43738106 -0.056658589 -0.42871707 -0.81997630  0.3447432071
#>  [25,]  0.92886698 -0.691157440 -0.85864380 -0.35189090  0.5565725868
#>  [26,] -0.71838848  0.403715847  0.76561992 -0.44681415  0.8496118919
#>  [27,] -0.41588230  0.859162764 -0.43864270 -0.66509401  0.8555563767
#>  [28,] -0.97101911 -0.294583427  0.02076020  0.07738906  0.1227874863
#>  [29,]  0.06028844  0.128184951  0.94772159  0.31349560  0.0265728058
#>  [30,]  0.27308467  0.652917399 -0.36176862 -0.14610542 -0.0631045313
#>  [31,] -0.21248397  0.268022504 -0.22192715  0.64374797 -0.0003793319
#>  [32,] -0.73213151  0.727111407  0.95914261 -0.50353808  0.8966085692
#>  [33,]  0.83629620  0.033236289  0.79909363  0.84619364  0.2477420745
#>  [34,]  0.62439474 -0.329904364  0.03414377  0.27778389 -0.7326140213
#>  [35,] -0.12823788  0.017301821  0.01296183 -0.18206334 -0.6960310289
#>  [36,] -0.33207755  0.914305115 -0.61012532 -0.93052920  0.8749120184
#>  [37,]  0.07082272  0.352978114 -0.22840262 -0.55573770  0.0603006543
#>  [38,]  0.73169875 -0.654712683 -0.33178357  0.84556368  0.7729406869
#>  [39,] -0.61623505  0.585229559  0.62195260  0.13709300 -0.5354714640
#>  [40,]  0.92076188 -0.266516868 -0.16480664  0.97779292  0.6402078317
#>  [41,] -0.23147847 -0.338468505 -0.93603554  0.02897367 -0.3518101559
#>  [42,]  0.06274113 -0.286862188 -0.06313931 -0.48135315 -0.1642713402
#>  [43,]  0.05650434  0.095244014  0.86769949  0.14152540 -0.4142858605
#>  [44,]  0.36984459 -0.562402774  0.24791825  0.09813494  0.5717324228
#>  [45,] -0.81403957  0.162672645  0.26203791 -0.68474692 -0.5830433341
#>  [46,] -0.92010329  0.404879521 -0.90641867  0.77012273  0.0188231985
#>  [47,]  0.73650062  0.647322598  0.45038462 -0.25111785 -0.8010053826
#>  [48,] -0.73368897 -0.010483660  0.20374078  0.97803786  0.5338498214
#>  [49,]  0.08787172  0.662266301  0.27971156 -0.27191839  0.2951126150
#>  [50,]  0.60422534  0.520518110  0.24555658 -0.34443748  0.0046259062
#>  [51,] -0.62786574  0.333412443  0.45118739 -0.46777386  0.3475869391
#>  [52,] -0.08813954  0.839048575  0.14118683  0.64292755  0.7566810129
#>  [53,]  0.71992885  0.517440573  0.54122045 -0.68659808 -0.8754797825
#>  [54,] -0.77401217 -0.675459346  0.45010025 -0.27591269 -0.3348603626
#>  [55,]  0.72408206 -0.559446723  0.89716198 -0.06778241  0.2485273895
#>  [56,]  0.86783076  0.891246819 -0.41000948  0.51266838 -0.8790179575
#>  [57,] -0.73253921  0.982025023  0.12560017  0.64362816 -0.3391365018
#>  [58,] -0.48038374 -0.077041836  0.16534854  0.14383144 -0.6882034549
#>  [59,] -0.77126121  0.192437305  0.09439028  0.01111256  0.0805354454
#>  [60,]  0.23560898 -0.566905522 -0.18979825  0.11074137  0.5733498130
#>  [61,]  0.70705728  0.319356204 -0.86633658 -0.81360958 -0.9438234814
#>  [62,]  0.11862850  0.025499659  0.99456096 -0.45010358 -0.4200294456
#>  [63,]  0.25501797  0.811492267 -0.26939036  0.65486138  0.7107475456
#>  [64,]  0.83968209 -0.860907429 -0.97474105  0.60491093 -0.3885400076
#>  [65,]  0.45507289 -0.795910924  0.35221664  0.44259497 -0.4128178782
#>  [66,] -0.20415788  0.255343379  0.34294284  0.74146015 -0.6170757287
#>  [67,] -0.30025546 -0.339697664  0.51400729  0.20391003 -0.0079678721
#>  [68,]  0.61282719  0.971093122  0.34475616 -0.19572884 -0.9316730006
#>  [69,] -0.65394906  0.548748734  0.12331030  0.01939788  0.3013225324
#>  [70,]  0.53519666 -0.211374662  0.40574365 -0.24598117 -0.2918219049
#>  [71,]  0.68778202 -0.507895095 -0.58848968  0.91166191  0.7853231076
#>  [72,] -0.58398171  0.576873954 -0.95504033  0.73905492 -0.1059120754
#>  [73,] -0.17680306  0.299523985 -0.85820716  0.11045966 -0.5802269578
#>  [74,] -0.29927327 -0.024016516  0.20362269  0.66743795  0.3957718075
#>  [75,]  0.62600078  0.174816100 -0.09192944 -0.01924932  0.3549057650
#>  [76,]  0.97885255  0.075408690 -0.07939647  0.99139204 -0.4831643291
#>  [77,]  0.97403928 -0.586043901 -0.72418339 -0.15402052 -0.1015401511
#>  [78,] -0.46255400  0.382907332 -0.77682781 -0.51742301 -0.7321038069
#>  [79,] -0.53919927 -0.037775983 -0.14818808  0.47891080 -0.6270239786
#>  [80,]  0.37227386 -0.655264861 -0.29876327 -0.92246554 -0.2557588695
#>  [81,]  0.80980757 -0.778278416 -0.87017656  0.54753564 -0.8441804531
#>  [82,] -0.33013642 -0.732564371 -0.55965328  0.11272472 -0.2590920520
#>  [83,]  0.71970374 -0.619581450  0.38502384 -0.20221519  0.0454735965
#>  [84,]  0.95783678 -0.451964703 -0.70694406 -0.34244676  0.9453409570
#>  [85,]  0.65992284 -0.497585119  0.63911974  0.71285071 -0.3174894908
#>  [86,] -0.45643156  0.965448396 -0.62922217  0.30367956  0.5698656272
#>  [87,] -0.87322629 -0.852202795  0.04971789  0.21482999 -0.5777265434
#>  [88,] -0.91523185 -0.227298828 -0.17733183 -0.48759028 -0.0761155379
#>  [89,] -0.99115133 -0.996204128  0.20012244 -0.12795116  0.8560182247
#>  [90,] -0.64777152 -0.233679719  0.72246107  0.59434357  0.6875110473
#>  [91,] -0.57248364 -0.497643224 -0.20405606 -0.59323440  0.2167474665
#>  [92,] -0.89281831  0.771568047 -0.28193811  0.87304014 -0.5568175404
#>  [93,] -0.84377499 -0.102188933 -0.44993188 -0.60373230  0.0329591865
#>  [94,] -0.43864498  0.701245844  0.17942167  0.09915965  0.1814343627
#>  [95,] -0.74036405  0.709350963 -0.64036186  0.96739796  0.9893194716
#>  [96,] -0.31721415  0.377936119  0.46717534 -0.87887062 -0.3348350278
#>  [97,] -0.37395955  0.668517593  0.79250243  0.60496295  0.5890471591
#>  [98,] -0.92792216 -0.731249100 -0.19328046 -0.53719075 -0.2693445887
#>  [99,] -0.17686743 -0.052343646 -0.08181853  0.20841948  0.2545222943
#> [100,] -0.05152102  0.955521719 -0.36157335  0.41213240  0.9924058840
#> 
#> $y
#>   [1] 1 0 1 0 1 1 0 0 0 0 0 1 0 1 0 1 1 0 0 0 1 0 1 1 0 0 0 0 1 1 1 0 0 1 1 0 1
#>  [38] 0 0 0 1 1 1 1 0 0 0 1 1 1 1 0 0 0 0 0 0 1 1 1 1 1 0 0 0 1 1 0 0 1 0 0 1 1
#>  [75] 1 0 0 1 1 1 0 0 0 0 0 0 0 0 0 1 1 0 0 0 0 1 1 0 1 0
#>