Simulate Data from Hyper-Sphere for Classification Problems

The classification problem Circle is described in the BARK paper (2008). Inputs are dim independent variables uniformly distributed on the interval \([-1,1]\), only the first 2 out of these dim are actually signals. Outputs are created according to the formula $$y = 1(x1^2+x2^2 \le 2/\pi)$$

Usage

sim_circle(n, dim = 5)

Arguments

n

number of data points to generate

dim

number of dimension of the problem, no less than 2

Value

Returns a list with components

x

input values (independent variables)

y

0/1 output values (dependent variable)

References

Ouyang, Zhi (2008) Bayesian Additive Regression Kernels. Duke University. PhD dissertation, Chapter 3.

See also

Examples

sim_circle(n=100, dim=5)
#> $x
#>                [,1]         [,2]         [,3]         [,4]         [,5]
#>   [1,] -0.316516010 -0.313872406 -0.046118391  0.342168079  0.912812334
#>   [2,]  0.097642731 -0.055430877 -0.644420356 -0.624455030 -0.461950849
#>   [3,] -0.764194923  0.401478904  0.130416248  0.997432848 -0.323478231
#>   [4,]  0.610929998 -0.636729894 -0.840644281  0.631227239  0.823842058
#>   [5,]  0.385659677 -0.405506533 -0.027482272  0.780085204 -0.068138084
#>   [6,] -0.964603963  0.630009863  0.496320237  0.271944518  0.326051353
#>   [7,] -0.038789410 -0.294379758 -0.050819855  0.912043780 -0.852577007
#>   [8,] -0.999064467 -0.626268038  0.832621986  0.425621081 -0.804203650
#>   [9,] -0.308509839 -0.619404999  0.049560224  0.657255292  0.879143895
#>  [10,] -0.933535292  0.094450488 -0.633433359  0.624262368 -0.636755645
#>  [11,] -0.065287272  0.400023609 -0.951279534 -0.368834564 -0.633816550
#>  [12,]  0.429945722 -0.859521730 -0.660148414 -0.571188303  0.612609938
#>  [13,] -0.407445066 -0.750899777 -0.221908411 -0.500284588 -0.182718086
#>  [14,] -0.846778171  0.266522332 -0.437131529  0.082463943  0.551506477
#>  [15,]  0.125826573  0.712069195  0.759082668 -0.702373741 -0.336628904
#>  [16,]  0.386804246  0.807919017 -0.496729416 -0.643199832 -0.471589627
#>  [17,] -0.031817139  0.303977121  0.350963093  0.578756151  0.427159695
#>  [18,]  0.333261388 -0.583928981 -0.039343766 -0.765440958  0.322871210
#>  [19,] -0.002377823  0.015551843  0.333176800 -0.285986580  0.214378706
#>  [20,]  0.292052747  0.724078374 -0.532724343  0.912281959 -0.769878688
#>  [21,]  0.835448853  0.017862042 -0.158292810  0.142782329 -0.499237068
#>  [22,] -0.037117798 -0.547748038 -0.599718214  0.225079262  0.180089713
#>  [23,]  0.458260722  0.468179628 -0.638061466  0.402571347 -0.008493254
#>  [24,] -0.561693763  0.118476596 -0.750173298  0.522639398  0.954073910
#>  [25,]  0.852738872  0.078113865  0.183224386 -0.041279162  0.625962366
#>  [26,] -0.977629116  0.592047107 -0.650808973  0.561077314 -0.234270132
#>  [27,]  0.199377885 -0.396043516 -0.903894353 -0.258914144  0.962197233
#>  [28,]  0.433200088  0.409490694  0.231402598 -0.009685143  0.087550998
#>  [29,]  0.208533201 -0.263583099  0.244309383  0.470363068 -0.482829825
#>  [30,] -0.533563951 -0.739452578  0.250529964 -0.912093769  0.075680810
#>  [31,]  0.706721184 -0.219414852 -0.836050137  0.025261548  0.164607057
#>  [32,] -0.337316662  0.528080652  0.306247101  0.265523608  0.483027615
#>  [33,] -0.151894336  0.092402278 -0.760236708 -0.147552605 -0.794174255
#>  [34,] -0.900097999  0.144662424  0.232909055 -0.001104299 -0.124088373
#>  [35,] -0.377056411 -0.724643033 -0.117426409  0.583998716  0.513810707
#>  [36,]  0.932941622  0.934745498 -0.711984360 -0.550440076  0.633766062
#>  [37,] -0.354719015  0.443076195  0.132082422 -0.285245959 -0.321295949
#>  [38,]  0.053427189  0.586478412 -0.004941897  0.434704613 -0.582105082
#>  [39,] -0.413160948 -0.702761535  0.045590923 -0.384926134 -0.147583071
#>  [40,] -0.289940559 -0.446041801  0.355577600 -0.275822256  0.782506116
#>  [41,]  0.799629158  0.398103414  0.923428429 -0.098726468 -0.962266700
#>  [42,]  0.906715264 -0.705402491 -0.311722445  0.136584780  0.823281689
#>  [43,]  0.268514914 -0.100183372  0.391211294 -0.154602107  0.751918385
#>  [44,] -0.719931173 -0.448923951  0.527752378 -0.731625754 -0.584725457
#>  [45,] -0.762286288 -0.338568298  0.618871382  0.288153110  0.440260279
#>  [46,] -0.759165762 -0.771131782 -0.688834830  0.070155398  0.294487217
#>  [47,]  0.399573439 -0.025717137  0.005510363  0.201958694 -0.748278532
#>  [48,]  0.430569358 -0.924899045  0.750205254  0.282836037  0.557854300
#>  [49,] -0.485163096  0.570958537 -0.821601497  0.526915620  0.441412788
#>  [50,] -0.387646352  0.053342930 -0.158566421  0.070126996 -0.907699076
#>  [51,]  0.383611757 -0.258040454  0.387087831 -0.372341613  0.148131682
#>  [52,] -0.154708327  0.608091274  0.876902372 -0.417758160 -0.420678447
#>  [53,]  0.999150175 -0.537376775 -0.482546672 -0.927546313 -0.436393560
#>  [54,]  0.032114751  0.077277144  0.491438403 -0.936532470 -0.410003567
#>  [55,]  0.364423872  0.510072350  0.672689543  0.993713458 -0.292376481
#>  [56,] -0.783357395 -0.539928172  0.267396742 -0.843547400  0.382539227
#>  [57,]  0.769479622  0.049332017  0.456959025  0.737639529  0.069296571
#>  [58,]  0.709648444 -0.026116556 -0.474561379 -0.774210275  0.832341962
#>  [59,] -0.294753492  0.465480723  0.810254978 -0.852917031 -0.258990555
#>  [60,]  0.993436752  0.180259976 -0.692865309  0.513791865  0.759484684
#>  [61,] -0.199100919 -0.596517639  0.712716394  0.297874393  0.276514104
#>  [62,]  0.483980794 -0.138025975  0.440463708 -0.213601112  0.147506657
#>  [63,]  0.118676090  0.989110295 -0.390812581 -0.680438880 -0.427286095
#>  [64,]  0.431829629  0.090160803  0.590211364  0.646653134 -0.187337633
#>  [65,]  0.509301094 -0.916738814 -0.427120414 -0.084540211  0.771019087
#>  [66,] -0.593804721  0.825608273  0.449403487  0.717008708  0.747872569
#>  [67,] -0.405237270 -0.032013135  0.482519352  0.358336156  0.787146752
#>  [68,]  0.240071598  0.350543104 -0.475693423 -0.271271606  0.453367724
#>  [69,] -0.449623082 -0.038329741  0.095857855  0.600101922 -0.686849230
#>  [70,]  0.968696875 -0.466136529 -0.487992252  0.960151606 -0.912524543
#>  [71,]  0.559512048 -0.375842686 -0.442688214  0.356789089  0.149784051
#>  [72,] -0.694292629 -0.654550112  0.074797621  0.975776268  0.006504535
#>  [73,] -0.302213790  0.001181829 -0.903650224  0.530277735  0.090435477
#>  [74,]  0.788456499  0.264819008  0.430199918  0.444798755  0.101658341
#>  [75,]  0.490500185 -0.007473676 -0.334596517  0.081281517 -0.984838638
#>  [76,] -0.503696567  0.473016502 -0.003779869  0.705367355 -0.707695512
#>  [77,]  0.698563495 -0.564702231  0.479539698  0.014605942  0.114482866
#>  [78,] -0.622404487 -0.400750263  0.526940123  0.869280816  0.705506528
#>  [79,] -0.586736214  0.662561069 -0.733881017 -0.235192470 -0.525183739
#>  [80,]  0.138159016 -0.920269207 -0.010835303  0.910001337  0.734563868
#>  [81,] -0.269796766  0.371660549 -0.341619354 -0.147823915 -0.595531266
#>  [82,] -0.194645199  0.396830307 -0.506106199  0.183371592  0.980145716
#>  [83,]  0.101958337 -0.361272719  0.724892763  0.836558914  0.805257155
#>  [84,] -0.935154145  0.965221278 -0.327484888 -0.031504542  0.855739301
#>  [85,]  0.371963385  0.261810062 -0.490291001  0.164171454  0.070287446
#>  [86,]  0.173078193  0.276508885  0.224799681 -0.218039576  0.164154077
#>  [87,] -0.479097007 -0.928249944 -0.349797295  0.410219191 -0.308458930
#>  [88,]  0.058004168 -0.601555917  0.288925102  0.783078563 -0.281977192
#>  [89,]  0.834755742  0.483776765  0.914758848 -0.440929349  0.794379928
#>  [90,] -0.167792496  0.508368929  0.665760776  0.521866291 -0.946105462
#>  [91,] -0.408436841  0.084997236  0.215336121 -0.892456018  0.445693960
#>  [92,]  0.981480560 -0.586933174 -0.913783583 -0.424885517  0.170619208
#>  [93,]  0.301102093  0.956887649 -0.004745409 -0.976862396 -0.594305120
#>  [94,]  0.528391079  0.579275721 -0.107637200 -0.601999837  0.314918322
#>  [95,]  0.525262767 -0.305726563 -0.780376323 -0.944897068  0.502622397
#>  [96,] -0.305629270  0.283145391  0.803891741  0.065973404 -0.312394430
#>  [97,]  0.842919771  0.140647351  0.068735478  0.349747511  0.122824810
#>  [98,]  0.419715313 -0.144217241 -0.266857276 -0.298049267  0.003628739
#>  [99,]  0.328999164 -0.727612646  0.163963028  0.352292366  0.582202871
#> [100,] -0.047267366 -0.555702379 -0.502647877 -0.025954475 -0.845700740
#> 
#> $y
#>   [1] 1 1 0 0 1 0 1 0 1 0 1 0 0 0 1 0 1 1 1 1 0 1 1 1 0 0 1 1 1 0 1 1 1 0 0 0 1
#>  [38] 1 0 1 0 0 1 0 0 0 1 0 1 1 1 1 0 1 1 0 1 1 1 0 1 1 0 1 0 0 1 1 1 0 1 0 1 0
#>  [75] 1 1 0 1 0 0 1 1 1 0 1 1 0 1 0 1 1 0 0 1 1 1 0 1 0 1
#>