Simulate Data from Hyper-Sphere for Classification Problems

The classification problem Circle is described in the BARK paper (2008). Inputs are dim independent variables uniformly distributed on the interval \([-1,1]\), only the first 2 out of these dim are actually signals. Outputs are created according to the formula $$y = 1(x1^2+x2^2 \le 2/\pi)$$

Usage

sim_circle(n, dim = 5)

Arguments

n

number of data points to generate

dim

number of dimension of the problem, no less than 2

Value

Returns a list with components

x

input values (independent variables)

y

0/1 output values (dependent variable)

References

Ouyang, Zhi (2008) Bayesian Additive Regression Kernels. Duke University. PhD dissertation, Chapter 3.

See also

Examples

sim_circle(n=100, dim=5)
#> $x
#>                [,1]         [,2]        [,3]        [,4]         [,5]
#>   [1,] -0.009314896 -0.936458214 -0.72605748 -0.87280761  0.546015010
#>   [2,]  0.217768271 -0.704674458 -0.08533892 -0.31663328 -0.232626559
#>   [3,] -0.981114442  0.327642774  0.05544181  0.48831330  0.473853339
#>   [4,] -0.550100030  0.685334664 -0.56357442 -0.09710528 -0.528398104
#>   [5,]  0.846255206 -0.966243566 -0.58730739 -0.50390837 -0.924419826
#>   [6,]  0.873224417 -0.646599635 -0.45302305  0.83622496  0.379080525
#>   [7,]  0.034825728  0.003537993  0.56433691  0.16536366 -0.005657589
#>   [8,]  0.253026403 -0.052757000 -0.97803629 -0.56260505 -0.893588199
#>   [9,] -0.611670466 -0.022094898  0.83979278 -0.25294394  0.541349241
#>  [10,] -0.419246194 -0.850666844 -0.50828233  0.70482053 -0.528239690
#>  [11,]  0.658847910 -0.806303686  0.74483638 -0.85957238 -0.649000359
#>  [12,] -0.351355192 -0.071245487 -0.11160479  0.42177955  0.122114714
#>  [13,] -0.955194901  0.688207147  0.74702006 -0.33146965  0.559721141
#>  [14,] -0.664402531  0.920021769 -0.31438208 -0.65189261  0.075607995
#>  [15,]  0.504363137  0.890108620 -0.29438048  0.35260238 -0.689142838
#>  [16,]  0.091901076 -0.877643748  0.96859955 -0.96805836 -0.757146348
#>  [17,]  0.010678552  0.872937785  0.21580501  0.70146992 -0.068497914
#>  [18,] -0.672751472 -0.851343804 -0.38859698 -0.15633341 -0.598321556
#>  [19,] -0.178571651 -0.004652419 -0.10851844 -0.17504400  0.878980734
#>  [20,]  0.435542274  0.767881016  0.04625193  0.57241900 -0.280227200
#>  [21,]  0.105282869  0.169790709 -0.57925252 -0.76032802 -0.174400244
#>  [22,]  0.335269962 -0.246887191 -0.25560873 -0.24341910 -0.806845700
#>  [23,] -0.584252235 -0.867557572  0.96750762 -0.33935560 -0.862501512
#>  [24,]  0.516883244 -0.322608734  0.28546161 -0.61724682  0.409966339
#>  [25,]  0.122472360 -0.325197050 -0.41037878 -0.69000107 -0.429511876
#>  [26,] -0.988486317 -0.751771037  0.62748188 -0.01038682 -0.822640305
#>  [27,]  0.443866577  0.602500956  0.31223637 -0.53263222 -0.354444163
#>  [28,] -0.782828901  0.799917303 -0.53414499  0.58630082 -0.173794013
#>  [29,]  0.494727592 -0.139197256  0.44739789 -0.43072929  0.329890772
#>  [30,] -0.942552929  0.099577449  0.16344533  0.05801489 -0.590366639
#>  [31,] -0.022358634  0.748054421  0.57807692  0.87881372  0.904553738
#>  [32,] -0.006938151  0.727773537 -0.91363295  0.31933789  0.604556200
#>  [33,]  0.007693794 -0.843028903 -0.32442046 -0.51898687  0.541077449
#>  [34,] -0.306128252 -0.153121315 -0.05482152 -0.36027297  0.272236242
#>  [35,] -0.642333484  0.880075573  0.92093044  0.61682346 -0.819804145
#>  [36,]  0.365484235 -0.681785161 -0.75600011  0.68055495 -0.692604094
#>  [37,] -0.373756769  0.310510041  0.10849542 -0.25026687  0.989276360
#>  [38,] -0.615199784  0.619287097 -0.62597159  0.76292988 -0.018159370
#>  [39,] -0.391800703  0.713548936 -0.58513960 -0.10404044  0.728222640
#>  [40,] -0.535921062 -0.359970246 -0.84347084 -0.47964487  0.496478043
#>  [41,]  0.900196294 -0.542668434 -0.56695447 -0.02122420  0.789042762
#>  [42,]  0.105178278  0.111032852  0.42188543  0.73473985  0.902991473
#>  [43,] -0.448993514  0.529780813 -0.39920137  0.37437148 -0.132634425
#>  [44,]  0.506119221  0.785535885 -0.94042866  0.44732212 -0.604062750
#>  [45,]  0.376682212 -0.523070602 -0.33966434 -0.84507339 -0.232584530
#>  [46,] -0.026220590  0.672264708 -0.76920057  0.17241723  0.612525416
#>  [47,] -0.670929052  0.014618134 -0.51970073 -0.90117934  0.960401502
#>  [48,]  0.757618431 -0.594022621 -0.93743034  0.40635797  0.820694588
#>  [49,] -0.384708071 -0.473227429  0.23782697 -0.60889193 -0.926619707
#>  [50,] -0.015625448  0.777044745 -0.23515345  0.76484245 -0.838017552
#>  [51,] -0.125661494  0.510282549  0.97068442 -0.12288333 -0.466930473
#>  [52,]  0.319046638  0.006753485 -0.14810902 -0.75785419  0.382596790
#>  [53,]  0.477112786 -0.145973769 -0.53982101 -0.87207755  0.895914761
#>  [54,] -0.852344448 -0.941964139  0.84170877  0.29501007  0.220410824
#>  [55,]  0.200869630 -0.101411307 -0.32175126 -0.34362333 -0.378023186
#>  [56,] -0.149042251  0.774622511 -0.04888378  0.03413329 -0.760961849
#>  [57,] -0.427132969 -0.601187026  0.15770207  0.16137734 -0.075372883
#>  [58,] -0.605425566  0.336543174 -0.13856972  0.86662730  0.967631538
#>  [59,] -0.141657785 -0.319037152  0.20843843  0.17260007 -0.941085472
#>  [60,] -0.240615996  0.569642056  0.02768013  0.15382756  0.279651243
#>  [61,]  0.143538038  0.607122777 -0.25210597  0.02893768  0.579220597
#>  [62,] -0.717344281 -0.041143647 -0.15429569  0.18174790 -0.597589000
#>  [63,]  0.148010804  0.447779009 -0.24188480 -0.77605074  0.304969849
#>  [64,] -0.132926976  0.283866136 -0.88838848 -0.15848279 -0.924765389
#>  [65,]  0.778337066 -0.506581006 -0.37665387  0.37481927  0.247148101
#>  [66,] -0.709063193 -0.704399881 -0.57373252  0.21471695  0.557101619
#>  [67,] -0.395718277  0.800286921  0.72775347  0.63916624  0.683210168
#>  [68,]  0.128537804  0.127756641 -0.13411145  0.32144529 -0.230548081
#>  [69,] -0.379302404  0.751218189  0.49659766  0.34350953  0.397028311
#>  [70,]  0.844570576  0.518150769  0.19485153 -0.15751121 -0.991753665
#>  [71,] -0.920831681 -0.006198119 -0.91515526 -0.46862397 -0.524266432
#>  [72,]  0.234612913  0.661892195  0.82929629 -0.50257541  0.826511838
#>  [73,]  0.270512783  0.591596482  0.93579937 -0.53757485 -0.245548469
#>  [74,]  0.214253386  0.707227333 -0.74408015  0.33642092 -0.528681719
#>  [75,]  0.380926445 -0.708725430 -0.87119231  0.61200024  0.094714629
#>  [76,] -0.862096147  0.728035156 -0.28228241 -0.39753194 -0.759565629
#>  [77,]  0.845073927 -0.394476921  0.08636130  0.26118302  0.503801960
#>  [78,] -0.669081173 -0.861934748 -0.68551011 -0.66408208 -0.092619216
#>  [79,]  0.334287145 -0.568965792 -0.65860521 -0.08466521 -0.822102334
#>  [80,]  0.533402476  0.726841395 -0.96312185  0.67761455  0.992363891
#>  [81,]  0.947034341 -0.595343583 -0.80635518  0.17854512  0.703500279
#>  [82,]  0.175804201  0.043617457  0.64357963 -0.91063205  0.320503604
#>  [83,]  0.228505131 -0.799354987 -0.53761543  0.38657178  0.593371992
#>  [84,]  0.481099142 -0.375122853  0.55909165 -0.18936836  0.723181433
#>  [85,]  0.542799988  0.485022278  0.95486114  0.48270838  0.486654437
#>  [86,] -0.772912251  0.075011360  0.61151283 -0.09632495 -0.389677037
#>  [87,]  0.206157989  0.456240635  0.80289790 -0.10800794 -0.718553671
#>  [88,]  0.610448041 -0.921322216  0.96617611  0.82401919 -0.082219210
#>  [89,] -0.303602597 -0.492371197 -0.64608723 -0.41307999  0.931877806
#>  [90,] -0.184249680 -0.516340244  0.39919758 -0.52431503 -0.043699978
#>  [91,] -0.838187370 -0.593809386 -0.08893140 -0.37937453  0.571718332
#>  [92,]  0.308328617  0.454627117  0.94711574  0.71331471 -0.061480807
#>  [93,] -0.837670213  0.534777126 -0.92342696 -0.02036732 -0.999086555
#>  [94,] -0.038873910  0.097670702 -0.96142133  0.53593155  0.801184902
#>  [95,] -0.120712562 -0.214662395  0.10978916  0.83490470  0.038744634
#>  [96,]  0.845594524  0.987704055  0.42314850  0.79093161 -0.732202392
#>  [97,]  0.697560952 -0.745307203  0.96410850  0.28455023  0.176548688
#>  [98,]  0.842169723  0.026329961 -0.92837114  0.91645138  0.546562258
#>  [99,] -0.096774104 -0.588374128  0.51297051  0.57363949 -0.579568101
#> [100,] -0.370332560 -0.298619308 -0.76626291  0.16263654  0.067962024
#> 
#> $y
#>   [1] 0 1 0 0 0 0 1 1 1 0 0 1 0 0 0 0 0 0 1 0 1 1 0 1 1 0 1 0 1 0 1 1 0 1 0 1 1
#>  [38] 0 0 1 0 1 1 0 1 1 1 0 1 1 1 1 1 0 1 1 1 1 1 1 1 1 1 1 0 0 0 1 0 0 0 1 1 1
#>  [75] 0 0 0 0 1 0 0 1 0 1 1 1 1 0 1 1 0 1 0 1 1 0 0 0 1 1
#>