Simulate Data from Hyper-Sphere for Classification Problems

The classification problem Circle is described in the BARK paper (2008). Inputs are dim independent variables uniformly distributed on the interval \([-1,1]\), only the first 2 out of these dim are actually signals. Outputs are created according to the formula $$y = 1(x1^2+x2^2 \le 2/\pi)$$

Usage

sim_circle(n, dim = 5)

Arguments

n

number of data points to generate

dim

number of dimension of the problem, no less than 2

Value

Returns a list with components

x

input values (independent variables)

y

0/1 output values (dependent variable)

References

Ouyang, Zhi (2008) Bayesian Additive Regression Kernels. Duke University. PhD dissertation, Chapter 3.

See also

Examples

sim_circle(n=100, dim=5)
#> $x
#>                [,1]        [,2]        [,3]         [,4]        [,5]
#>   [1,]  0.142633062 -0.77036995 -0.79965413  0.771452300 -0.23595322
#>   [2,]  0.830519067  0.20579340  0.88245445  0.873596117  0.04780103
#>   [3,] -0.333462923  0.55284114  0.57298950  0.264954822  0.82235733
#>   [4,]  0.623769833 -0.66114648 -0.91362763  0.982301789  0.55216951
#>   [5,]  0.766102232 -0.10902854 -0.71143575 -0.059222542 -0.29318099
#>   [6,]  0.925897392 -0.49601301  0.61178672  0.393379514  0.02833709
#>   [7,] -0.813082522  0.10962529 -0.86765766 -0.165107894 -0.66074711
#>   [8,]  0.091599079  0.02371685  0.26839519 -0.291289350 -0.07465825
#>   [9,]  0.046958881  0.56073375  0.22249314 -0.109641530  0.98884614
#>  [10,] -0.654694219  0.30133619 -0.32024952 -0.226349058 -0.81059047
#>  [11,]  0.331845044 -0.14286842 -0.84421239 -0.779674542 -0.49192516
#>  [12,] -0.153899629  0.57532924  0.74088055 -0.006381945 -0.88809977
#>  [13,] -0.520395953  0.59116418 -0.29325661  0.731740236 -0.85415178
#>  [14,] -0.814639858 -0.66857605  0.75010649  0.768188514  0.62184088
#>  [15,]  0.166739652 -0.09515862 -0.67422872  0.551305606  0.03424341
#>  [16,]  0.473056738 -0.42811870  0.83206633 -0.049369047 -0.95209102
#>  [17,]  0.912868399 -0.98619969 -0.44534926 -0.971456590  0.31696144
#>  [18,]  0.448080544 -0.19639776 -0.26877974  0.288033073  0.63929015
#>  [19,] -0.260906925 -0.81176426  0.67801589 -0.938753437 -0.06343589
#>  [20,]  0.468061113 -0.89615219 -0.84136029 -0.096273922  0.21126873
#>  [21,]  0.899040127 -0.14406097  0.59212638  0.169240599  0.39928459
#>  [22,] -0.087003303 -0.11077722 -0.53627971  0.144168143 -0.69166831
#>  [23,] -0.253793887  0.02087163 -0.01913106 -0.307300669  0.77839240
#>  [24,]  0.930025224 -0.18393271  0.12586773 -0.197927651 -0.60479542
#>  [25,] -0.523953505  0.65441510  0.17963992 -0.487297562  0.46342988
#>  [26,]  0.241385678 -0.85976412  0.54922880 -0.555181224  0.74084387
#>  [27,] -0.662516236  0.69816490  0.83233640 -0.671266477  0.86749587
#>  [28,]  0.912854136  0.24994729 -0.63366971  0.852592736  0.97652424
#>  [29,] -0.401622468  0.29816753  0.22752978  0.618135571 -0.50271486
#>  [30,]  0.468073954  0.24783008  0.45992790 -0.281366602 -0.99522973
#>  [31,] -0.604383843  0.46843264  0.09759161  0.049480118  0.49280287
#>  [32,] -0.620866837  0.29622008  0.77067396 -0.656003464  0.41181058
#>  [33,] -0.142916765  0.07701893 -0.30440195  0.647984834  0.39916693
#>  [34,] -0.836728876 -0.83947997 -0.30637809  0.432373937 -0.85165218
#>  [35,]  0.755193468 -0.19239110 -0.70971862  0.721319743 -0.90260717
#>  [36,]  0.695467520 -0.23493014  0.52991468  0.412505731  0.36108508
#>  [37,] -0.530566255  0.26382060 -0.58167075 -0.192305882  0.61184739
#>  [38,]  0.916549878 -0.73603722  0.28287320  0.040454808  0.05901080
#>  [39,] -0.132046853  0.89437493  0.10837242  0.530934919  0.22396848
#>  [40,]  0.870218907 -0.90540768 -0.77385838 -0.751083270 -0.03039402
#>  [41,]  0.944405958  0.14993524  0.92590651 -0.812671488  0.46165049
#>  [42,] -0.638879494 -0.38592611 -0.40219124 -0.481095935 -0.73966068
#>  [43,]  0.529661827  0.18983601 -0.09275989 -0.543553011  0.12500832
#>  [44,] -0.861742162  0.20505134  0.08610550 -0.890167053  0.12428894
#>  [45,]  0.370367993 -0.26130162 -0.61890505 -0.778775957  0.19302414
#>  [46,] -0.434788791 -0.48483081 -0.60616350  0.876605899  0.85785441
#>  [47,] -0.648054005  0.79711118 -0.30229674 -0.553825056 -0.82123132
#>  [48,]  0.199225336  0.98420191  0.57817367  0.532484713 -0.35198905
#>  [49,]  0.391653501 -0.33282764 -0.56374181  0.074503344  0.25560439
#>  [50,]  0.015961934  0.53322282  0.50571594 -0.002299740  0.14083297
#>  [51,]  0.809723350 -0.08155938  0.14708178 -0.565191121 -0.39398500
#>  [52,] -0.861277375 -0.35849396 -0.13179062  0.898647263 -0.79418118
#>  [53,] -0.884184022 -0.91090230 -0.13289371  0.220336663  0.82031308
#>  [54,] -0.886759175 -0.16129603  0.09719197  0.166171316 -0.10249911
#>  [55,]  0.869406221  0.63926530  0.26900283 -0.432573723 -0.04379152
#>  [56,] -0.233975146 -0.88598159 -0.62685832 -0.810871189 -0.50528746
#>  [57,]  0.363468495 -0.23775303 -0.90273525  0.411881430  0.50350947
#>  [58,]  0.466674785 -0.16870150 -0.38237008 -0.966420168  0.50177026
#>  [59,] -0.390704125 -0.35476424  0.78809209 -0.866951658 -0.61014601
#>  [60,]  0.091110607 -0.96145407 -0.41537197 -0.783305010  0.09609003
#>  [61,]  0.999402990 -0.68044195 -0.84865413  0.625042743  0.83425979
#>  [62,]  0.013589891 -0.45325577  0.93990542  0.509293425 -0.54237157
#>  [63,] -0.781628148  0.62002113 -0.05320287  0.111800011 -0.22815653
#>  [64,] -0.316607465  0.13185368  0.23198367  0.004867096 -0.94926751
#>  [65,] -0.792947654 -0.52213964  0.39261722  0.574729621 -0.56658739
#>  [66,]  0.764044996  0.19911567 -0.02044117 -0.203633829 -0.42625762
#>  [67,]  0.129856044 -0.13366823 -0.98106073  0.588351133  0.59429214
#>  [68,]  0.637189941  0.88762618  0.40791839  0.815294667  0.01816516
#>  [69,] -0.924370285  0.28809515 -0.56151699  0.006740817  0.30884032
#>  [70,]  0.042567038  0.69648798  0.21259487 -0.924440044 -0.03655777
#>  [71,]  0.257354807 -0.47227444 -0.82945775 -0.859479273 -0.88909697
#>  [72,]  0.522720368 -0.94471247 -0.98578218  0.232879428 -0.46916259
#>  [73,]  0.908508929 -0.73036540  0.11973570  0.650894869  0.32355080
#>  [74,] -0.949620634  0.03596844 -0.08935675 -0.137152924  0.82873005
#>  [75,] -0.002536554  0.47499948 -0.10790405 -0.418774799 -0.42751403
#>  [76,] -0.553176770 -0.38265535  0.08611288  0.524761426  0.96243258
#>  [77,] -0.881040746  0.02642901 -0.08555274 -0.638272939 -0.27810247
#>  [78,]  0.817897565  0.41620415 -0.29803718 -0.192042837 -0.37032630
#>  [79,]  0.907469286 -0.01398535  0.04436836 -0.148460814  0.96251785
#>  [80,] -0.182049090 -0.91325671 -0.66467697  0.103042343  0.28825425
#>  [81,] -0.588087689 -0.04735887  0.53443331 -0.239419519  0.81406056
#>  [82,] -0.448711773 -0.42130154 -0.80620681  0.144958561 -0.96531189
#>  [83,]  0.405691970 -0.70687553 -0.35149671  0.386935054 -0.15512009
#>  [84,] -0.633187280  0.97806839  0.48293848  0.403366320  0.54760791
#>  [85,]  0.158015335 -0.79058916 -0.79265811 -0.849380189 -0.57111515
#>  [86,] -0.793139210  0.62394820 -0.24180736  0.161443631 -0.68472341
#>  [87,]  0.625206628  0.67562843 -0.65040529  0.711155528  0.40083444
#>  [88,] -0.010992314  0.06441818  0.98586066  0.059848744  0.09328704
#>  [89,]  0.861259794 -0.40177795  0.12695673 -0.997774912 -0.30450897
#>  [90,] -0.831662081 -0.88466955  0.68083625  0.976996366 -0.96276610
#>  [91,]  0.299219353 -0.09616220 -0.86729344 -0.335393128  0.51917819
#>  [92,]  0.665128532  0.35592743 -0.96352838 -0.310740368  0.54998690
#>  [93,]  0.202151067 -0.04183940  0.03130858 -0.525603449 -0.72489893
#>  [94,] -0.818149747  0.31014581  0.45160550 -0.530215672  0.82605405
#>  [95,] -0.618159100  0.83094544 -0.20543263 -0.158675457 -0.41833183
#>  [96,] -0.013934675 -0.96566870  0.63333841 -0.417629787  0.32840031
#>  [97,] -0.556050021 -0.48726110 -0.92147744  0.122889580 -0.31609178
#>  [98,] -0.599825300  0.53947200 -0.62063352  0.196445909 -0.99618393
#>  [99,]  0.677567201  0.30909057 -0.12558916 -0.531027821 -0.23497119
#> [100,]  0.863355857 -0.85605205 -0.31459479  0.516678538  0.92494853
#> 
#> $y
#>   [1] 1 0 1 0 1 0 0 1 1 1 1 1 1 0 1 1 0 1 0 0 0 1 1 0 0 0 0 0 1 1 1 1 1 0 1 1 1
#>  [38] 0 0 0 0 1 1 0 1 1 0 0 1 1 0 0 0 0 0 0 1 1 1 0 0 1 0 1 0 1 1 0 0 1 1 0 0 0
#>  [75] 1 1 0 0 0 0 1 1 0 0 0 0 0 1 0 0 1 1 1 0 0 0 1 0 1 0
#>