Simulate Data from Hyper-Sphere for Classification Problems

The classification problem Circle is described in the BARK paper (2008). Inputs are dim independent variables uniformly distributed on the interval \([-1,1]\), only the first 2 out of these dim are actually signals. Outputs are created according to the formula $$y = 1(x1^2+x2^2 \le 2/\pi)$$

sim_circle(n, dim = 5)

Arguments

n

number of data points to generate

dim

number of dimension of the problem, no less than 2

Value

Returns a list with components

x

input values (independent variables)

y

0/1 output values (dependent variable)

References

Ouyang, Zhi (2008) Bayesian Additive Regression Kernels. Duke University. PhD dissertation, Chapter 3.

See also

Examples

sim_circle(n=100, dim=5)
#> $x
#>               [,1]         [,2]        [,3]         [,4]        [,5]
#>   [1,] -0.08452664  0.101330275  0.72331683 -0.441592857 -0.63254772
#>   [2,]  0.15595596  0.895420386 -0.41793587  0.946491604  0.01295091
#>   [3,]  0.73276943  0.456990680 -0.44522915 -0.844699280 -0.55079706
#>   [4,]  0.43261663 -0.183990666  0.49818421 -0.223429658 -0.20988795
#>   [5,]  0.08184760 -0.769103600  0.58983817 -0.686279494 -0.24802775
#>   [6,]  0.44841298 -0.673371929  0.12122157  0.604762608  0.68996724
#>   [7,] -0.70007584  0.735792818  0.63817436 -0.474185892 -0.72528042
#>   [8,]  0.39778922 -0.774083593 -0.35851577  0.126804593  0.30177859
#>   [9,] -0.16475722  0.624400146  0.43253671  0.356198667 -0.96347542
#>  [10,] -0.34332060 -0.602164960  0.30299907 -0.567406399  0.97538222
#>  [11,]  0.42976578  0.387475038  0.20208126 -0.779453021  0.45962067
#>  [12,] -0.47096259  0.284490362 -0.45188461 -0.856824209 -0.48145752
#>  [13,] -0.80436966 -0.262076593  0.91167968 -0.835334292 -0.64907784
#>  [14,] -0.91393804 -0.268330650  0.69466965 -0.614090946  0.57564752
#>  [15,] -0.83684500 -0.210248753  0.71403112  0.273968129 -0.71815823
#>  [16,]  0.84020996 -0.294303737 -0.18196448 -0.015834797  0.64385445
#>  [17,] -0.79308640  0.997628835  0.08905327  0.340158387 -0.44165317
#>  [18,]  0.52854323  0.932122672  0.25628756  0.182537206  0.33268233
#>  [19,] -0.51701945 -0.290860876 -0.04209984 -0.210563725 -0.73574849
#>  [20,]  0.62730461 -0.001073826 -0.85551513 -0.693687586  0.79287392
#>  [21,]  0.98082810 -0.210243177  0.62636563 -0.246516807 -0.89633129
#>  [22,] -0.78618959  0.280161770 -0.21002910  0.715105188  0.01712816
#>  [23,] -0.81543761 -0.571704246 -0.57934071  0.338252876  0.84032866
#>  [24,] -0.36714401  0.202359763  0.54353177 -0.623819527 -0.60085892
#>  [25,] -0.61893666 -0.561476714 -0.15131230  0.469119145 -0.55061228
#>  [26,] -0.47087860 -0.954470000 -0.53668372 -0.176945971 -0.13440469
#>  [27,] -0.74509236  0.829711675 -0.92992438 -0.054717244 -0.32978074
#>  [28,]  0.52708509 -0.830918977  0.98012187 -0.952356084  0.68392246
#>  [29,]  0.09426028 -0.189178983  0.07848269 -0.895788156 -0.76925193
#>  [30,] -0.47160214 -0.448575716 -0.14559605 -0.978684469  0.83463023
#>  [31,] -0.38385390 -0.347508491 -0.72296185 -0.073631474  0.74966420
#>  [32,] -0.71682127  0.220399832  0.85680757 -0.455523405 -0.73890073
#>  [33,] -0.67942038  0.225858706  0.73878114 -0.605968850 -0.78055357
#>  [34,] -0.06395921 -0.649864627 -0.76896095  0.713090515  0.36832635
#>  [35,]  0.81779780 -0.947603768 -0.54777982  0.009302171  0.05911179
#>  [36,] -0.63328716  0.879974771  0.14934134  0.290471109  0.50848950
#>  [37,] -0.01906936  0.299628867  0.27533839  0.479107947  0.67176493
#>  [38,]  0.38352426  0.055539163  0.19554482  0.235570199  0.84058816
#>  [39,]  0.42139361 -0.135564244 -0.07419278 -0.149490992  0.60124368
#>  [40,] -0.94016235  0.983559057  0.03126201 -0.149744304  0.89346852
#>  [41,]  0.34276381 -0.728961418 -0.50855717  0.007562014 -0.59684666
#>  [42,]  0.99749825  0.438186191 -0.42302825 -0.437407248  0.12644007
#>  [43,] -0.35342559 -0.415652265  0.52524392  0.889833525 -0.20859526
#>  [44,] -0.04345677  0.765170050 -0.91445594  0.169316089 -0.43362202
#>  [45,]  0.53524141 -0.511079163  0.92313068 -0.157823738 -0.56915884
#>  [46,] -0.29020101 -0.442941658  0.85503095 -0.907775756  0.46699666
#>  [47,] -0.72069014  0.751513737  0.67762304 -0.829169889  0.68936450
#>  [48,]  0.30845330  0.656854424  0.69691063  0.611684017  0.16874025
#>  [49,]  0.05165923  0.754090013  0.17027128  0.953103513 -0.96770536
#>  [50,]  0.41988955  0.620482923  0.04219687  0.243748087  0.25661216
#>  [51,]  0.02945489 -0.459147224  0.18108575  0.698622149  0.13873774
#>  [52,] -0.20061506 -0.606237843 -0.01871559  0.579410545 -0.41929651
#>  [53,] -0.49721901  0.777766132 -0.51442575 -0.348908468 -0.09251269
#>  [54,]  0.40563091  0.295436271  0.78317597 -0.482817873 -0.22521746
#>  [55,] -0.69615108  0.350606656  0.01895616  0.721180127  0.99757215
#>  [56,] -0.61615156  0.119907879 -0.25677315  0.113397172 -0.30000755
#>  [57,] -0.81687360  0.107920416  0.66400921  0.372062648  0.20282718
#>  [58,]  0.39266702  0.043183082  0.54464624  0.022239228 -0.39506392
#>  [59,] -0.29542793 -0.293377575 -0.84140352 -0.003553580  0.91631294
#>  [60,]  0.75891196 -0.803516581 -0.15458770 -0.987833324  0.16783031
#>  [61,] -0.26863654  0.396748047  0.07784638 -0.986014022 -0.55170673
#>  [62,] -0.45628749  0.572788601  0.45428739 -0.250495860 -0.06664461
#>  [63,] -0.49060699 -0.283678907  0.65093045  0.541119278  0.09846941
#>  [64,]  0.27423417  0.338426222  0.91283682 -0.563331671  0.46279623
#>  [65,] -0.46033751  0.530280602 -0.31968306 -0.608260648  0.10671603
#>  [66,] -0.70533574 -0.327091372  0.13490932 -0.072657147 -0.50129536
#>  [67,]  0.05324964 -0.526464443 -0.92120511  0.711298154  0.38227072
#>  [68,] -0.47856949  0.872412733  0.29786543  0.247369515 -0.59324545
#>  [69,] -0.69212743 -0.275469359  0.68826075 -0.152498905 -0.14496092
#>  [70,] -0.16114711 -0.352707333  0.87924963  0.170609945  0.45765190
#>  [71,]  0.31672935 -0.221699340  0.50642927 -0.344493391 -0.60607173
#>  [72,]  0.19940201 -0.030890167 -0.79306709 -0.191715784 -0.46217010
#>  [73,]  0.38532305 -0.456427197  0.50769225 -0.083438330 -0.14621545
#>  [74,] -0.63761557  0.407431159  0.33817961  0.207666755  0.48738892
#>  [75,]  0.41077084  0.607955838  0.88167528 -0.357381352  0.99413746
#>  [76,] -0.60751430 -0.993970477  0.12184947 -0.538313330 -0.48633282
#>  [77,]  0.47287364  0.348646739 -0.03820868 -0.742978170  0.28336683
#>  [78,]  0.05071379  0.669419578 -0.43664986 -0.106161400  0.31295632
#>  [79,]  0.66434687 -0.251486531 -0.50471974 -0.986879635  0.23984100
#>  [80,]  0.92947299  0.445014133 -0.10579262 -0.710379221 -0.50534463
#>  [81,] -0.94142253 -0.570087194  0.77221680  0.985039417 -0.42070389
#>  [82,]  0.90311148 -0.105539032  0.30421265 -0.332931754 -0.11807742
#>  [83,] -0.80867318  0.849737684  0.45711161 -0.199890296  0.11966109
#>  [84,] -0.82129308 -0.067792511  0.93857922  0.770123595 -0.07516552
#>  [85,]  0.03930123 -0.370818816  0.14103538  0.581529843 -0.33629767
#>  [86,] -0.96528982  0.205326683 -0.36910108 -0.651995259  0.22855005
#>  [87,] -0.60553995 -0.890625484  0.67230480 -0.485751725  0.94501179
#>  [88,] -0.08641308 -0.316777518 -0.09043539 -0.583464809  0.24485894
#>  [89,] -0.42871812 -0.723145327 -0.52953307  0.965149643 -0.02570050
#>  [90,] -0.98939238 -0.863536765  0.13351262 -0.692239334 -0.52325959
#>  [91,] -0.14064522 -0.447449957 -0.64125892 -0.063854948  0.20389871
#>  [92,] -0.09594439  0.044509347  0.09112555 -0.410090148 -0.23932299
#>  [93,]  0.13116128 -0.603411872  0.99285592  0.373240705 -0.44624734
#>  [94,]  0.97225260 -0.114124035 -0.15133503  0.306285747 -0.59761365
#>  [95,]  0.33557820 -0.813099863 -0.69484393 -0.518849985 -0.20400691
#>  [96,] -0.79400739 -0.347439129  0.26677380 -0.059727194 -0.75285376
#>  [97,]  0.24304027 -0.976061678 -0.53744082  0.176702892  0.79987371
#>  [98,] -0.03106089 -0.816375311  0.91280126  0.947199382  0.26937416
#>  [99,]  0.70731285 -0.308611340 -0.90927261 -0.303513311  0.96282933
#> [100,]  0.56049119  0.715715812 -0.46534805 -0.130837847 -0.57691823
#> 
#> $y
#>   [1] 1 0 0 1 1 0 0 0 1 1 1 1 0 0 0 0 0 0 1 1 0 0 0 1 0 0 0 0 1 1 1 1 1 1 0 0 1
#>  [38] 1 1 0 0 0 1 1 1 1 0 1 1 1 1 1 0 1 1 1 0 1 1 0 1 1 1 1 1 1 1 0 1 1 1 1 1 1
#>  [75] 1 0 1 1 1 0 0 0 0 0 1 0 0 1 0 0 1 1 1 0 0 0 0 0 1 0
#>