Simulate Data from Hyper-Sphere for Classification Problems

The classification problem Circle is described in the BARK paper (2008). Inputs are dim independent variables uniformly distributed on the interval \([-1,1]\), only the first 2 out of these dim are actually signals. Outputs are created according to the formula $$y = 1(x1^2+x2^2 \le 2/\pi)$$

sim_circle(n, dim = 5)

Arguments

n
number of data points to generate

dim
number of dimension of the problem, no less than 2

Value

Returns a list with components

x

input values (independent variables)

y

0/1 output values (dependent variable)

References

Ouyang, Zhi (2008) Bayesian Additive Regression Kernels. Duke University. PhD dissertation, Chapter 3.
http://stat.duke.edu/people/theses/OuyangZ.html

See also

Examples

sim_circle(n=100, dim=5)

#> $x
#>               [,1]        [,2]         [,3]         [,4]         [,5]
#>   [1,] -0.79599389  0.52302095 -0.540724070  0.535288919  0.904768954
#>   [2,]  0.73316548  0.08237309 -0.739114863  0.028832033  0.572532528
#>   [3,]  0.79047651 -0.03349279  0.451766314 -0.514470804 -0.812082616
#>   [4,]  0.07140084  0.10156429  0.584054445  0.446144571 -0.253173853
#>   [5,] -0.20350930  0.10955254  0.080625156  0.651306992  0.674549972
#>   [6,] -0.83202940  0.96241837  0.678221125 -0.715251862 -0.034013377
#>   [7,] -0.99976584 -0.42121804 -0.554853144 -0.690064025  0.660875753
#>   [8,] -0.77154776  0.50906273  0.938156317  0.466509450 -0.886003392
#>   [9,]  0.34752901  0.22665718  0.266992510  0.141402486  0.437390365
#>  [10,]  0.38533177  0.02641961 -0.061437861 -0.112324316  0.182218903
#>  [11,] -0.82213746 -0.21150210  0.505861400  0.730023720 -0.014009876
#>  [12,] -0.41353104  0.54395652 -0.793106309 -0.608337866  0.568533760
#>  [13,] -0.17358651  0.69133968  0.039376027  0.229820617 -0.937751138
#>  [14,]  0.38660716  0.31441091  0.581296728 -0.816219128 -0.627124402
#>  [15,]  0.90030799  0.16718674 -0.229355379  0.945296223  0.062840974
#>  [16,] -0.94391867  0.95230898  0.146362919  0.590068278 -0.245815924
#>  [17,]  0.80668617 -0.22682934 -0.835069345  0.555434051 -0.269692111
#>  [18,] -0.23675579  0.06377376  0.977595280  0.200688989  0.706792763
#>  [19,] -0.06456553  0.28558144  0.278098208 -0.294208623 -0.115519361
#>  [20,] -0.08373448  0.46399253 -0.515796313  0.026922809 -0.474054892
#>  [21,] -0.43859406 -0.83742417  0.024437835 -0.410326131 -0.357145322
#>  [22,]  0.88165541 -0.63105576 -0.264557248  0.106700324  0.854831374
#>  [23,] -0.55333170  0.01817243 -0.196898384  0.835045838 -0.726006801
#>  [24,] -0.39203483 -0.03559085 -0.836011542 -0.461053860  0.645465994
#>  [25,]  0.45074022 -0.63745044  0.638172997 -0.329086827  0.019473762
#>  [26,]  0.46905486  0.28999956 -0.283016893 -0.448573759  0.372360803
#>  [27,] -0.51697006  0.30086607 -0.029266661 -0.285259130 -0.626770179
#>  [28,]  0.55774707  0.79167589 -0.108691894 -0.723902191 -0.577696600
#>  [29,] -0.57526843 -0.01967850 -0.103615826 -0.844216483  0.145259375
#>  [30,] -0.72140658  0.64885105  0.907152834 -0.725668510 -0.151816369
#>  [31,] -0.93630871 -0.57888956 -0.017239510  0.511562400 -0.394667353
#>  [32,]  0.11407688  0.06533773  0.183675310  0.155880141  0.993468265
#>  [33,]  0.52737944 -0.13226867 -0.889212684  0.962728161  0.310106163
#>  [34,]  0.76723122  0.49988176  0.321006180  0.317607594 -0.082730996
#>  [35,] -0.10459517  0.61303048 -0.985104760 -0.003078935 -0.682228411
#>  [36,] -0.82607438 -0.53211578 -0.072884298 -0.961218745 -0.590431037
#>  [37,] -0.21673681 -0.40820319 -0.494945958 -0.616459873 -0.250640166
#>  [38,] -0.53676122  0.18740912  0.849478340 -0.496995060 -0.078586721
#>  [39,] -0.77785525 -0.14719401 -0.107673313 -0.543280268 -0.477672768
#>  [40,]  0.43697718 -0.07629688  0.333781307 -0.935578096  0.346846997
#>  [41,] -0.22710883 -0.18392588 -0.988396603 -0.204484812 -0.839978388
#>  [42,]  0.70197230  0.80092815  0.253650262 -0.813288028 -0.672866588
#>  [43,]  0.78494649 -0.23190784 -0.618232606 -0.859390943 -0.337519940
#>  [44,]  0.59048471 -0.01812318  0.289423979 -0.414487966 -0.948846858
#>  [45,]  0.09823974  0.52064657  0.636486051 -0.823795689  0.238542670
#>  [46,] -0.82597122  0.82918764  0.390009434  0.261628569 -0.155205342
#>  [47,] -0.96374418 -0.21047070  0.897158876 -0.171749824 -0.742951598
#>  [48,]  0.55328807 -0.14425558  0.818005685 -0.229733804  0.250018244
#>  [49,] -0.73711528  0.43088965 -0.481549031 -0.362374019  0.170559105
#>  [50,]  0.59263891 -0.15695919  0.303798831 -0.782434708 -0.872693190
#>  [51,] -0.94133841 -0.70340985 -0.092056463  0.351446429  0.165804703
#>  [52,] -0.37791141 -0.53412835 -0.305980343  0.524756577 -0.256667087
#>  [53,]  0.26625854  0.83838713 -0.626961111 -0.668926401  0.205107555
#>  [54,] -0.99694907 -0.64184512 -0.440441916  0.546401564 -0.343171060
#>  [55,] -0.05717471 -0.72576396 -0.454312008  0.601825099  0.423819819
#>  [56,] -0.55501338 -0.97404513 -0.397154735  0.317145920 -0.820580288
#>  [57,] -0.64139941 -0.99696363  0.637510130 -0.363021735 -0.398199381
#>  [58,]  0.27461571 -0.59143754 -0.664868450 -0.999336319 -0.063082687
#>  [59,]  0.93195445  0.07482054 -0.023665487 -0.406523950  0.617701942
#>  [60,] -0.77008690 -0.09147235 -0.183432090  0.313433402 -0.595721704
#>  [61,] -0.27787770  0.94220834  0.314280409  0.645745598 -0.253945752
#>  [62,] -0.72145695 -0.35135414  0.307230261  0.233777322 -0.444423292
#>  [63,] -0.96928743  0.78854022 -0.050848250 -0.316560570  0.337641974
#>  [64,]  0.51490281 -0.03107788 -0.121031149  0.618705749 -0.543525795
#>  [65,] -0.60381061  0.36152059  0.858073372 -0.884973784 -0.680130776
#>  [66,] -0.75252195 -0.60232619 -0.250129257 -0.616832151 -0.021333609
#>  [67,]  0.93588771  0.56495169 -0.899016183  0.661988639  0.992908474
#>  [68,]  0.86252496  0.69032616  0.599549172  0.858992220 -0.542535445
#>  [69,] -0.15107148 -0.95583733  0.201717285 -0.020948558 -0.937260573
#>  [70,]  0.49037209  0.60978808  0.870836933 -0.289464201 -0.837713949
#>  [71,] -0.33770503 -0.07614774  0.019385949  0.219461026 -0.734467814
#>  [72,] -0.92024495  0.41184756  0.732434181 -0.569986583 -0.061209939
#>  [73,] -0.31596658  0.38989294 -0.137818443  0.652338374  0.539000568
#>  [74,] -0.14456081 -0.29826608  0.002082131 -0.991973092  0.037460430
#>  [75,] -0.81035426 -0.53163301  0.888121599  0.477638776  0.871195410
#>  [76,]  0.25628869  0.74047932 -0.178539956 -0.826522547 -0.866944777
#>  [77,] -0.72636470  0.11716831  0.657318381  0.422364273 -0.565410450
#>  [78,]  0.27073455  0.16622882 -0.970385673 -0.365342338  0.927674178
#>  [79,] -0.29114298 -0.32529270  0.403688138 -0.479136057  0.531721623
#>  [80,]  0.89614604 -0.58560708 -0.040712245 -0.374112774 -0.821286242
#>  [81,] -0.62580401  0.49069815 -0.814535300  0.463249215  0.483332386
#>  [82,]  0.21723429 -0.94381948 -0.849099585  0.251968155  0.518747735
#>  [83,] -0.45065672 -0.57580305 -0.891089183  0.546490790  0.812277310
#>  [84,]  0.53173575 -0.92057064  0.880004950 -0.266670106 -0.934170321
#>  [85,] -0.50820984 -0.49369993  0.914096502 -0.416980840 -0.087838813
#>  [86,]  0.92746429 -0.34348094  0.745237629  0.106787312 -0.620135018
#>  [87,]  0.58547678  0.39805264  0.199666512  0.139578837  0.851339859
#>  [88,] -0.23664496 -0.29555443  0.097213532  0.852905083  0.675776387
#>  [89,]  0.94462504 -0.37361370 -0.161335625  0.962189898  0.201233711
#>  [90,]  0.30914930  0.47959112  0.972053438 -0.393133589  0.633123261
#>  [91,] -0.07324089  0.07480017 -0.321647014  0.104286000 -0.009712472
#>  [92,]  0.42071461  0.78400963  0.897790799  0.158170812 -0.752829092
#>  [93,] -0.57720413 -0.54802061 -0.068966223 -0.496855915 -0.887923394
#>  [94,] -0.72300100 -0.04075246 -0.347559288  0.590405195  0.551628792
#>  [95,]  0.04541876  0.30290217 -0.204376775  0.808216333 -0.390205623
#>  [96,]  0.46848974 -0.94866630 -0.325201806  0.308435062 -0.407707535
#>  [97,] -0.26053688 -0.58354159 -0.536321489  0.005787554 -0.276976306
#>  [98,] -0.83909059  0.87123777 -0.126994552 -0.860047342 -0.281170816
#>  [99,]  0.63442380 -0.75226193 -0.525583990 -0.462731417  0.326210152
#> [100,] -0.93433723  0.84719883  0.698459758 -0.664912334  0.114647834
#> 
#> $y
#>   [1] 0 1 1 1 1 0 0 0 1 1 0 1 1 1 0 0 0 1 1 1 0 0 1 1 1 1 1 0 1 0 0 1 1 0 1 0 1
#>  [38] 1 1 1 1 0 0 1 1 0 0 1 0 1 0 1 0 0 1 0 0 1 0 1 0 0 0 1 1 0 0 0 0 1 1 0 1 1
#>  [75] 0 1 1 1 1 0 1 0 1 0 1 0 1 1 0 1 1 0 1 1 1 0 1 0 0 0
#>