Simulate Data from Hyper-Sphere for Classification Problems

The classification problem Circle is described in the BARK paper (2008). Inputs are dim independent variables uniformly distributed on the interval \([-1,1]\), only the first 2 out of these dim are actually signals. Outputs are created according to the formula $$y = 1(x1^2+x2^2 \le 2/\pi)$$

Usage

sim_circle(n, dim = 5)

Arguments

n

number of data points to generate

dim

number of dimension of the problem, no less than 2

Value

Returns a list with components

x

input values (independent variables)

y

0/1 output values (dependent variable)

References

Ouyang, Zhi (2008) Bayesian Additive Regression Kernels. Duke University. PhD dissertation, Chapter 3.

See also

Examples

sim_circle(n=100, dim=5)
#> $x
#>                [,1]        [,2]          [,3]         [,4]         [,5]
#>   [1,]  0.650197681  0.53532262  0.3316904875 -0.994320423  0.803080702
#>   [2,]  0.446298432  0.49126269 -0.9018608970  0.954881002  0.579974689
#>   [3,]  0.052820250  0.09670727 -0.5943873902  0.207454090 -0.227705631
#>   [4,]  0.256542898 -0.79670882  0.3305502362  0.509221357 -0.622816002
#>   [5,] -0.143597788 -0.07568780  0.8054930246  0.486863972  0.232268759
#>   [6,] -0.394028886 -0.63205460 -0.0932271066 -0.682653706  0.544351292
#>   [7,]  0.009063629  0.81361441  0.8576995409  0.860098628  0.626387807
#>   [8,]  0.921742515 -0.55550529 -0.8622727459 -0.870495277 -0.342103921
#>   [9,] -0.589452725  0.06642681  0.5991972536  0.107409098 -0.962223662
#>  [10,] -0.089160953 -0.35835669  0.5616340013  0.021952730  0.488443181
#>  [11,] -0.557682689 -0.17823060 -0.8309233692 -0.112695537  0.120383673
#>  [12,]  0.185703869 -0.58753883  0.0431763874 -0.908728813  0.122326742
#>  [13,] -0.157980524 -0.86352268  0.3897178862 -0.493233066 -0.351814002
#>  [14,]  0.039974548 -0.90514233 -0.7985444623  0.630688302  0.543042018
#>  [15,]  0.042828682 -0.48537775 -0.7913783723  0.044003569  0.232244812
#>  [16,]  0.847324266 -0.42982742  0.7713148776  0.895917530 -0.743899903
#>  [17,]  0.422351941 -0.70127726 -0.1455766265 -0.452414331 -0.598191860
#>  [18,] -0.593617069 -0.06566322  0.2139490587 -0.790794771 -0.881946801
#>  [19,] -0.985553520 -0.94815252  0.6835800996 -0.408709455  0.741117056
#>  [20,]  0.253659971  0.93329989 -0.7336702817 -0.288412653 -0.814218004
#>  [21,] -0.313296671  0.21557951 -0.5755559062 -0.779689575  0.710334142
#>  [22,]  0.641178046  0.14186285 -0.0347873759  0.152655785  0.306597209
#>  [23,]  0.503922857 -0.85493681  0.9644366256 -0.880111153 -0.839530224
#>  [24,]  0.413784668  0.46026428 -0.7900837143 -0.752909893  0.566323306
#>  [25,] -0.602096403  0.65584660  0.1283723176  0.143186367 -0.210202972
#>  [26,]  0.337742887  0.63799442 -0.6576864566  0.801054291  0.910707770
#>  [27,] -0.928091263 -0.97434778 -0.7460131734  0.428111113 -0.793234853
#>  [28,] -0.681183764 -0.29424570  0.5050693215  0.288671779  0.118180127
#>  [29,] -0.115176579  0.71645249 -0.1663641944  0.279783394  0.156325586
#>  [30,]  0.282614299 -0.81338526  0.5752487299  0.166971342  0.840722257
#>  [31,]  0.078516572  0.73162005  0.5022568018 -0.087077636 -0.532102093
#>  [32,]  0.737753501  0.75097332 -0.0550200767  0.328811169 -0.863685527
#>  [33,]  0.026588916  0.70429024  0.2472229842  0.478561764 -0.828534472
#>  [34,]  0.921551247 -0.97623336 -0.7415912538 -0.962192543  0.181949215
#>  [35,]  0.757840933 -0.63444633 -0.4002518887  0.130089264  0.533751161
#>  [36,]  0.855900755 -0.95120724 -0.2682217802 -0.756051903 -0.804447685
#>  [37,] -0.018656544 -0.64022815 -0.0005764146 -0.671306848  0.334156672
#>  [38,] -0.472052352  0.32711565  0.2845895444 -0.445791745  0.812289922
#>  [39,]  0.035574924  0.17681930 -0.4948922293 -0.480433498  0.783352986
#>  [40,]  0.538858874 -0.53868263  0.7678324580  0.736216504  0.484781040
#>  [41,]  0.722341719  0.74418892  0.5621087723  0.202809362  0.828461253
#>  [42,] -0.478820655 -0.27482160  0.8639999554 -0.306615112  0.054706520
#>  [43,]  0.189254971  0.71908432  0.2903315765  0.386611429 -0.459342070
#>  [44,] -0.616688946  0.80175352  0.5087602339 -0.769295226 -0.597489552
#>  [45,] -0.134045206 -0.38631186  0.9664437170 -0.312696971  0.405778379
#>  [46,] -0.117482134  0.54177776 -0.7366014593  0.572118578  0.718341673
#>  [47,]  0.553858417  0.52639710 -0.5610575997 -0.932296174 -0.224500587
#>  [48,]  0.202780287  0.52162645  0.7169307084  0.913127259 -0.461393044
#>  [49,] -0.917937334  0.34033062  0.4272754896 -0.985531483  0.005872388
#>  [50,]  0.727165694  0.71430660  0.1681610928 -0.527352449  0.969265754
#>  [51,] -0.954219996 -0.25797261  0.5051706419  0.566572719  0.520567303
#>  [52,] -0.811944278 -0.24381542 -0.6988238399 -0.036882049  0.485017265
#>  [53,]  0.474186624  0.38824798 -0.1042055669  0.021836862  0.848629714
#>  [54,]  0.471457149  0.14666420 -0.5358994529 -0.246213210 -0.792069888
#>  [55,] -0.211824899  0.55877062  0.3903445895  0.160745382  0.708560407
#>  [56,] -0.506493160  0.87739847  0.7081232951 -0.260544408  0.673591227
#>  [57,] -0.811294661 -0.35203192 -0.9356905622  0.765234418  0.689458491
#>  [58,]  0.370939522  0.12956514  0.7158577624  0.734163641  0.966418986
#>  [59,] -0.295346217 -0.44510025 -0.5489232698 -0.027509194  0.145217850
#>  [60,]  0.349495997  0.99452465 -0.5376990535  0.797323869  0.540192627
#>  [61,] -0.383391170  0.06606399 -0.0386093534 -0.292228674 -0.863729524
#>  [62,] -0.614285974  0.51013974  0.1591123408  0.047897939  0.257864938
#>  [63,]  0.870236219  0.48793987  0.4886861658  0.002782632  0.711288119
#>  [64,] -0.847395201 -0.29932441 -0.3804976796  0.049801271 -0.897203964
#>  [65,]  0.711645031  0.89856387  0.2249999903  0.596101913  0.576012792
#>  [66,]  0.714419194 -0.56869427 -0.4530908428  0.642518468 -0.297658500
#>  [67,] -0.152374936 -0.70291513  0.8377477820 -0.038859786 -0.392619197
#>  [68,] -0.455221022 -0.14605298 -0.8931302577  0.393743252 -0.841866529
#>  [69,]  0.311191367  0.15728060  0.0524810073  0.126852167  0.021895273
#>  [70,] -0.130448837 -0.78114515  0.0425432627 -0.016936057 -0.258164578
#>  [71,]  0.068034934 -0.14986805 -0.8859425974 -0.851902092  0.488589565
#>  [72,]  0.324932411  0.48025729  0.1978770699  0.647024796  0.479176424
#>  [73,]  0.215582126 -0.39617940  0.1682095281  0.703320131 -0.792556852
#>  [74,]  0.325687880 -0.75932744  0.8578832275  0.923702744  0.345665088
#>  [75,] -0.589891340 -0.23201723 -0.4112137067 -0.796262446 -0.985005027
#>  [76,] -0.394560944  0.10988378 -0.2042407407  0.164225255  0.259989579
#>  [77,] -0.587297324  0.88810423 -0.6810445623  0.796479017 -0.971496903
#>  [78,]  0.915457691  0.57017300  0.4422038873  0.599533502 -0.247727355
#>  [79,] -0.508325111 -0.73600815  0.1156101311  0.549186619  0.972044798
#>  [80,]  0.234444375 -0.37360046 -0.4274973557 -0.271122887 -0.760147349
#>  [81,]  0.487694048  0.93554837  0.0110969092  0.211223766  0.088013610
#>  [82,] -0.805770338 -0.58640922  0.9691611235 -0.683465274 -0.073875262
#>  [83,]  0.548692295  0.36662013  0.7752588410  0.755969626  0.105958310
#>  [84,] -0.737951796 -0.43753388 -0.0931523861 -0.519065110  0.416493917
#>  [85,] -0.078762360  0.89547628 -0.5719174887 -0.071282308 -0.076691134
#>  [86,] -0.984964395  0.27294845  0.5139587983 -0.706622405  0.466221122
#>  [87,]  0.603845900  0.32098094 -0.3668303569 -0.617296953 -0.848847335
#>  [88,]  0.228955221  0.51992620  0.1868416886  0.260681330  0.074514879
#>  [89,]  0.475486806  0.99332043  0.3056068737 -0.807969129 -0.462570075
#>  [90,]  0.145398589 -0.10797332 -0.8389755585  0.552986187 -0.800995099
#>  [91,] -0.169624841  0.84402241 -0.4013176593 -0.467764446  0.429607801
#>  [92,]  0.737292264 -0.57402570 -0.1702788984 -0.667360458  0.011572008
#>  [93,] -0.960718797 -0.04896125  0.9942990057  0.464593550  0.916459071
#>  [94,] -0.378644150 -0.46071662  0.0048731524 -0.531279893  0.201652104
#>  [95,] -0.214226720 -0.30791200  0.0772316819  0.065359149 -0.775442132
#>  [96,]  0.976102460 -0.55627189  0.3624162427 -0.689753617  0.441618594
#>  [97,]  0.341627233 -0.03476092  0.6471145079  0.335935014  0.386701020
#>  [98,]  0.133853036  0.52179818 -0.8133410304  0.634424517 -0.025132346
#>  [99,] -0.418290100  0.63882636 -0.0676669027 -0.978600293  0.335649102
#> [100,]  0.578694377  0.23329497  0.3936692751 -0.921451509  0.828374065
#> 
#> $y
#>   [1] 0 1 1 0 1 1 0 0 1 1 1 1 0 0 1 0 0 1 0 0 1 1 0 1 0 1 0 1 1 0 1 0 1 0 0 0 1
#>  [38] 1 1 1 0 1 1 0 1 1 1 1 0 0 0 0 1 1 1 0 0 1 1 0 1 0 0 0 0 0 1 1 1 1 1 1 1 0
#>  [75] 1 1 0 0 0 1 0 0 1 0 0 0 1 1 0 1 0 0 0 1 1 0 1 1 1 1
#>