Simulate Data from Hyper-Sphere for Classification Problems

The classification problem Circle is described in the BARK paper (2008). Inputs are dim independent variables uniformly distributed on the interval \([-1,1]\), only the first 2 out of these dim are actually signals. Outputs are created according to the formula $$y = 1(x1^2+x2^2 \le 2/\pi)$$

Usage

sim_circle(n, dim = 5)

Arguments

n

number of data points to generate

dim

number of dimension of the problem, no less than 2

Value

Returns a list with components

x

input values (independent variables)

y

0/1 output values (dependent variable)

References

Ouyang, Zhi (2008) Bayesian Additive Regression Kernels. Duke University. PhD dissertation, Chapter 3.

See also

Examples

sim_circle(n=100, dim=5)
#> $x
#>               [,1]         [,2]        [,3]         [,4]         [,5]
#>   [1,] -0.44798996  0.887781701  0.50032606 -0.036783415 -0.095311928
#>   [2,] -0.27235774 -0.591610830 -0.50063710 -0.279111995 -0.268660211
#>   [3,]  0.68719148  0.153921173  0.07957227  0.566339471 -0.415233553
#>   [4,]  0.84248151  0.769513681  0.90154792 -0.723003521 -0.711144258
#>   [5,]  0.83429346 -0.958684464 -0.07680428 -0.793397981  0.827509505
#>   [6,] -0.40750816  0.582310815  0.22473835 -0.378893956 -0.525050686
#>   [7,]  0.92306242  0.038203565  0.48068807  0.003682115  0.452551671
#>   [8,]  0.71172222 -0.318445952 -0.95530407  0.091826138 -0.314482357
#>   [9,]  0.47705279 -0.069879308 -0.84348424  0.191220722 -0.918251921
#>  [10,] -0.94834646  0.610031903  0.23752496 -0.237055255  0.574802334
#>  [11,]  0.07990591  0.337723746 -0.65621598 -0.210458796 -0.946810174
#>  [12,]  0.43992939 -0.716252020  0.84677237 -0.346186698 -0.653222429
#>  [13,] -0.48961450 -0.673163435 -0.87302492  0.606214647  0.352653137
#>  [14,]  0.05363279  0.160930662 -0.11479498  0.050630197 -0.202221741
#>  [15,] -0.07220222  0.490344910 -0.38389635 -0.076472146 -0.504288947
#>  [16,]  0.37014333 -0.925381143 -0.55851216 -0.248955594  0.368300793
#>  [17,]  0.63307717  0.248004905 -0.69840318  0.221526113  0.302672683
#>  [18,] -0.24125402 -0.757240416 -0.99859983  0.993098616 -0.886597446
#>  [19,] -0.72030752 -0.915223940  0.54659598  0.731760169 -0.565364418
#>  [20,]  0.76886439 -0.221243596  0.79586213 -0.505141176 -0.611433416
#>  [21,]  0.51457140 -0.817545205  0.22862316 -0.119520667 -0.433389386
#>  [22,]  0.02098366 -0.469587646  0.23713853  0.630361490 -0.897848406
#>  [23,]  0.54506486 -0.608589219  0.59219028 -0.731574601 -0.725823120
#>  [24,] -0.54160256 -0.183911304  0.22532491 -0.933842114  0.554972908
#>  [25,]  0.25382921 -0.601075227  0.16790896  0.707324158 -0.969631101
#>  [26,]  0.90699293  0.595480495  0.67996515 -0.917080383 -0.706569610
#>  [27,] -0.77628556 -0.013356651 -0.80864687 -0.619583694  0.190882293
#>  [28,] -0.04515709 -0.270310451  0.80118330 -0.745713489 -0.450597529
#>  [29,]  0.88224678 -0.212454860  0.63501963 -0.531985256  0.326165196
#>  [30,] -0.85393116 -0.692521604 -0.97651109 -0.309828041 -0.906251377
#>  [31,]  0.94224108 -0.172180320  0.12792513  0.191491011  0.378493038
#>  [32,]  0.96110270  0.698470925  0.64749800  0.880267780 -0.829474413
#>  [33,] -0.20351356  0.531205798 -0.98919923 -0.635228636 -0.620369774
#>  [34,]  0.70902307 -0.185370261  0.30822419 -0.494739036  0.595669979
#>  [35,] -0.38282555 -0.721350336  0.70604636 -0.019814931  0.795047282
#>  [36,] -0.52293518 -0.229876205  0.89743703  0.559099452  0.665718365
#>  [37,]  0.35093385  0.254715299 -0.45862887 -0.206457575  0.987466942
#>  [38,] -0.66525678  0.956259079 -0.84152434  0.747764723 -0.609799495
#>  [39,] -0.75597771 -0.100643741  0.14367140  0.749642025  0.955123559
#>  [40,]  0.22979998  0.285807000 -0.95624609  0.047616893  0.896261971
#>  [41,] -0.01887424 -0.982456230  0.34049636  0.023355893 -0.343989088
#>  [42,]  0.09046033  0.238620843  0.26350442  0.687961012  0.965184456
#>  [43,] -0.09388280 -0.395898765  0.76100660 -0.484245016 -0.187141252
#>  [44,] -0.96807600 -0.318386481  0.04036304  0.261479723  0.538742668
#>  [45,]  0.28305219  0.767082827 -0.85954390 -0.170208669 -0.686305186
#>  [46,]  0.94869875 -0.505279385 -0.85142670  0.823717277 -0.801235589
#>  [47,]  0.21269654 -0.403671815  0.89559054 -0.619787098  0.869519148
#>  [48,] -0.52107349  0.130998937 -0.53165724  0.152335969 -0.312852658
#>  [49,] -0.76310358  0.821325995  0.34307803  0.112220750 -0.659379022
#>  [50,] -0.30870734  0.656685076  0.57859501 -0.737187142 -0.955472175
#>  [51,]  0.85869156  0.286661315  0.13488128 -0.870724726 -0.338073511
#>  [52,]  0.96484026 -0.754657106  0.60080062  0.834352220  0.536331475
#>  [53,]  0.59056601 -0.197626086 -0.02402771  0.410236919  0.168786516
#>  [54,]  0.89793344  0.548184678 -0.03490133 -0.092434309 -0.881804559
#>  [55,] -0.08173655  0.088932613 -0.23951064 -0.358292147 -0.935069465
#>  [56,] -0.48811763  0.606478341  0.12134606 -0.526824014 -0.266602142
#>  [57,] -0.97243828  0.323272832  0.85443840  0.656781705  0.482178318
#>  [58,] -0.65371343 -0.818719879  0.26969835 -0.961987392  0.452436702
#>  [59,]  0.15693782 -0.803345890  0.79689663  0.725372063 -0.155189905
#>  [60,] -0.08417586  0.826186118  0.55107001 -0.015427922 -0.946964642
#>  [61,] -0.23319126 -0.299958311 -0.33628534 -0.602880985 -0.487229025
#>  [62,] -0.14329566 -0.849284658 -0.84787277  0.619055425  0.434009010
#>  [63,] -0.48294519 -0.716447484 -0.52731498 -0.686280306  0.014803486
#>  [64,]  0.15149708  0.462233167  0.73909456  0.533739107  0.723590482
#>  [65,] -0.39000238 -0.703607779 -0.50192594 -0.528651339 -0.095702532
#>  [66,] -0.62025026  0.481543687  0.63483309 -0.599206568  0.434763813
#>  [67,]  0.84984199  0.978466131 -0.10669360  0.371216664 -0.662678527
#>  [68,]  0.14650628  0.548247255  0.11515372 -0.875218988  0.766115344
#>  [69,]  0.16893507  0.002145600  0.74336369  0.939628824 -0.859675759
#>  [70,] -0.17046296  0.540599186  0.88512958 -0.422889035 -0.161383739
#>  [71,]  0.08400322  0.300761329  0.61815849  0.733512629 -0.588153501
#>  [72,] -0.01395907  0.473089214 -0.94843201  0.628646093  0.528956346
#>  [73,] -0.61708320  0.951872716 -0.04317708  0.581667305 -0.835597976
#>  [74,]  0.13400506 -0.780060512  0.17379212  0.873759882  0.860458349
#>  [75,] -0.09700078 -0.007235876  0.35859371 -0.247357960 -0.970791305
#>  [76,] -0.83756046  0.088949649 -0.39249488 -0.097080071 -0.563030198
#>  [77,] -0.39790748  0.020827930 -0.89808140  0.432307500 -0.809902047
#>  [78,] -0.96959852 -0.652522269  0.27729205 -0.599002922  0.660492742
#>  [79,]  0.01006254 -0.221268854 -0.57153475 -0.616976483 -0.587511557
#>  [80,]  0.96220996 -0.938183017 -0.71700495 -0.618722541  0.390006303
#>  [81,] -0.65272409 -0.693518549  0.84850958 -0.419677992  0.402943053
#>  [82,] -0.70152190  0.488793307  0.72779279 -0.592916223  0.024559204
#>  [83,] -0.54489922 -0.717911087  0.98079029  0.159234997  0.034794761
#>  [84,]  0.41686836  0.243674368  0.35382620 -0.004222989 -0.723780723
#>  [85,] -0.74947547 -0.647413715  0.72930723  0.900007425 -0.616273035
#>  [86,]  0.13424917 -0.941897228  0.94704757  0.070337578  0.405534233
#>  [87,] -0.69599645  0.421923236 -0.75900491 -0.746328937  0.390707167
#>  [88,]  0.10846737  0.647033427  0.45822846 -0.905587008  0.755442821
#>  [89,]  0.50293508  0.145177036 -0.50197961 -0.048670235  0.458025027
#>  [90,]  0.75879424 -0.266989279 -0.37455790 -0.380751216  0.039775973
#>  [91,] -0.33280689  0.537551888 -0.76589057  0.433638023 -0.168155979
#>  [92,] -0.88083320 -0.974994232  0.30640551 -0.051735559  0.669620309
#>  [93,]  0.10153897  0.041066101  0.75774814  0.479516878  0.001532208
#>  [94,]  0.59181053  0.021421195  0.61149796  0.425409401 -0.662136494
#>  [95,] -0.51685195 -0.966291080  0.26605832  0.885242105 -0.601991803
#>  [96,]  0.52585277  0.398513704  0.84814403 -0.487989563 -0.128974266
#>  [97,] -0.50809075 -0.361255295  0.96758102  0.570089599  0.846401559
#>  [98,] -0.06753227  0.109019820  0.26314642 -0.870320010  0.668090538
#>  [99,] -0.20429698 -0.317737533 -0.40361798  0.233200317  0.374154517
#> [100,] -0.62822035  0.562304129  0.75019886  0.187079878  0.337573907
#> 
#> $y
#>   [1] 0 1 1 0 0 1 0 1 1 0 1 0 0 1 1 0 1 1 0 0 0 1 0 1 1 0 1 1 0 0 0 0 1 1 0 1 1
#>  [38] 0 1 1 0 1 1 0 0 0 1 1 0 1 0 0 1 0 1 1 0 0 0 0 1 0 0 1 0 1 0 1 1 1 1 1 0 1
#>  [75] 1 0 1 0 1 0 0 0 0 1 0 0 0 1 1 0 1 0 1 1 0 1 1 1 1 0
#>