Simulate Data from Hyper-Sphere for Classification Problems

The classification problem Circle is described in the BARK paper (2008). Inputs are dim independent variables uniformly distributed on the interval \([-1,1]\), only the first 2 out of these dim are actually signals. Outputs are created according to the formula $$y = 1(x1^2+x2^2 \le 2/\pi)$$

sim_circle(n, dim = 5)

Arguments

n

number of data points to generate

dim

number of dimension of the problem, no less than 2

Value

Returns a list with components

x

input values (independent variables)

y

0/1 output values (dependent variable)

References

Ouyang, Zhi (2008) Bayesian Additive Regression Kernels. Duke University. PhD dissertation, Chapter 3.

See also

Examples

sim_circle(n=100, dim=5)
#> $x
#>                [,1]        [,2]        [,3]         [,4]         [,5]
#>   [1,]  0.520474344 -0.83731255 -0.13728013  0.234669687  0.211087651
#>   [2,] -0.107988938  0.42967879  0.19714968 -0.845352530  0.842944357
#>   [3,]  0.110507567 -0.44448340 -0.07289395  0.050552030 -0.495025038
#>   [4,] -0.371172499  0.59334956  0.67744469 -0.471961731  0.628955045
#>   [5,] -0.179795607 -0.68133591  0.57838645  0.908497988  0.223653275
#>   [6,]  0.417654119  0.52340544  0.67938957 -0.667737335  0.919417374
#>   [7,] -0.207638429 -0.75268057  0.67373663 -0.165878476 -0.666942458
#>   [8,] -0.256503146 -0.83839422 -0.88099816 -0.967397176 -0.231460593
#>   [9,]  0.675721768 -0.04647665  0.81999332 -0.210557226  0.098803518
#>  [10,]  0.521470004  0.71451786  0.52038559 -0.670331080  0.007446073
#>  [11,]  0.904728496 -0.17138171  0.94973221 -0.343914113  0.790512583
#>  [12,]  0.701828409 -0.86869970  0.69439540  0.865712882  0.657121383
#>  [13,] -0.389027392  0.67074596 -0.73051128 -0.034827504  0.983608456
#>  [14,]  0.557868232  0.38930550  0.08352697  0.153038220 -0.823888158
#>  [15,] -0.989352591 -0.58565315  0.15144316 -0.125695612  0.600839690
#>  [16,]  0.481340492  0.45509771  0.23367891  0.676105026 -0.586431023
#>  [17,]  0.714081015  0.38528315  0.25874720 -0.071054044 -0.968037014
#>  [18,] -0.619009384 -0.43674927 -0.37676448 -0.498062150 -0.788746716
#>  [19,]  0.291076679  0.57701632 -0.59543331  0.744438619  0.969700310
#>  [20,]  0.007406465  0.31626904  0.37834859  0.920877746 -0.871475534
#>  [21,] -0.749932975  0.65618922 -0.21527112 -0.084304955  0.872738116
#>  [22,]  0.787895860  0.54655688  0.73498920 -0.517446388 -0.717986648
#>  [23,] -0.001756745 -0.25621271 -0.12629625 -0.806890167 -0.971716559
#>  [24,]  0.540608949  0.62761129 -0.35154686 -0.875307492 -0.666748357
#>  [25,]  0.055308621  0.43431280 -0.67241892 -0.075295731 -0.711554360
#>  [26,] -0.992043889 -0.13723420  0.08166745  0.746841402 -0.545727760
#>  [27,]  0.490233105  0.69000592  0.22591296  0.318650584 -0.503215120
#>  [28,]  0.456101560  0.70731392 -0.83201051  0.497995367 -0.008717864
#>  [29,] -0.278206593  0.24419004 -0.88991080  0.337171404  0.402927151
#>  [30,]  0.805075432  0.19350741 -0.63611119  0.046664952 -0.932361206
#>  [31,] -0.886750321  0.92734545 -0.92447785 -0.340554992  0.296995895
#>  [32,]  0.202015623  0.64609027  0.64775987  0.573428513  0.578941401
#>  [33,]  0.971686076 -0.41775426 -0.16374681  0.801299487  0.328687050
#>  [34,]  0.706882812 -0.81723410 -0.48459198 -0.258699008 -0.339132447
#>  [35,] -0.621728785  0.90907857 -0.28624207  0.453291846  0.499762283
#>  [36,]  0.473612517 -0.56629240 -0.90831801  0.916004019  0.186020966
#>  [37,] -0.338435341  0.10839523 -0.45909133  0.640654855 -0.992085994
#>  [38,] -0.060884684 -0.29645904 -0.45262811 -0.059517265 -0.614430934
#>  [39,]  0.432098920 -0.26443086 -0.97588481 -0.036997285  0.904857348
#>  [40,]  0.085026646 -0.61897229  0.67075974 -0.566304340  0.265735691
#>  [41,]  0.172079384 -0.17291942 -0.37046013  0.777108712 -0.680955211
#>  [42,] -0.933384700 -0.09080046  0.67716789  0.293173858  0.605770391
#>  [43,] -0.916460813  0.54508400 -0.18610388  0.776727322 -0.996475351
#>  [44,] -0.567478193 -0.21190801  0.74268411 -0.901341053 -0.682889837
#>  [45,]  0.587319985  0.98740541  0.95738224 -0.474227521 -0.028606130
#>  [46,] -0.526414136 -0.02947723 -0.68870847  0.800303006  0.516527623
#>  [47,] -0.500694358  0.03283332  0.01429428 -0.871120160 -0.331143865
#>  [48,] -0.953564100 -0.25724052  0.84054195  0.003322601  0.216643678
#>  [49,]  0.117151512 -0.51633734 -0.72255766 -0.235465488  0.146085802
#>  [50,] -0.879900424  0.88795434  0.66035244 -0.764299150  0.496113806
#>  [51,]  0.522389191  0.18591124 -0.31722965  0.223930550  0.156403239
#>  [52,]  0.091632240 -0.15273041 -0.24614785  0.263142208  0.907169892
#>  [53,]  0.154287872 -0.01649539 -0.77441369  0.500951201  0.704512185
#>  [54,] -0.951507094  0.19343761  0.14702816 -0.476236396  0.202689035
#>  [55,] -0.034768599  0.87730562  0.14684078  0.295885394  0.501175156
#>  [56,] -0.530464156 -0.12203467 -0.58966219  0.647182118 -0.994687689
#>  [57,] -0.295503044 -0.26236484  0.00469014  0.026640694  0.174514872
#>  [58,]  0.262145196 -0.48489643 -0.02906036  0.334132474  0.598972738
#>  [59,] -0.134282541 -0.45811112 -0.93058306 -0.766410034  0.255523101
#>  [60,]  0.537098418 -0.02139605  0.89311065 -0.943793196  0.393677452
#>  [61,]  0.519321716  0.91625738 -0.42911775  0.865165106  0.804625337
#>  [62,] -0.978796622  0.75646675  0.13790905  0.887630999 -0.015880466
#>  [63,]  0.291666100 -0.36049526  0.65287177 -0.352189497  0.654123141
#>  [64,] -0.171123774  0.29911294 -0.64061707 -0.029803061 -0.194878797
#>  [65,] -0.479130479 -0.55699175 -0.89882448  0.800308696  0.499301168
#>  [66,]  0.769554487  0.18770446  0.38189244 -0.609205741  0.934873223
#>  [67,]  0.696240170 -0.86294942  0.88455521  0.516929521 -0.523936460
#>  [68,]  0.086408794  0.56010887  0.48999160  0.627784181 -0.126540491
#>  [69,]  0.867205064 -0.63722306  0.05960887 -0.060536909  0.922403495
#>  [70,]  0.301867474 -0.85353288  0.56888464 -0.800388754 -0.445214049
#>  [71,]  0.718706924  0.81755149  0.06927221  0.273155609  0.281623426
#>  [72,] -0.239120273  0.91563035 -0.12474722  0.847021273 -0.664718603
#>  [73,] -0.592495619  0.61291294 -0.85010342  0.660694678 -0.726649252
#>  [74,] -0.900985508 -0.89947741  0.58954524  0.070510778 -0.716431897
#>  [75,]  0.048283551 -0.14760298 -0.26290720  0.925494046 -0.600231794
#>  [76,]  0.694251601 -0.38327385  0.49804470 -0.113966536  0.496484384
#>  [77,] -0.469912315  0.40965577 -0.41300494  0.466248544 -0.919182553
#>  [78,]  0.382399655  0.63452380  0.97811686 -0.337024928  0.476583610
#>  [79,]  0.815938436  0.62446469  0.83461730 -0.602608117  0.677700929
#>  [80,]  0.876699240  0.67327231  0.11342371  0.500456848  0.620498043
#>  [81,]  0.076672181  0.01111919  0.87277628 -0.903603068  0.626560853
#>  [82,]  0.584156846  0.74806083 -0.03841150 -0.164858919 -0.927055695
#>  [83,]  0.515739589 -0.50874518  0.80342849  0.574219659 -0.119294951
#>  [84,]  0.727473218  0.52966464 -0.10338880  0.567745035 -0.237766340
#>  [85,] -0.904990921  0.37807185 -0.27372337 -0.275734426 -0.097390048
#>  [86,]  0.112804498 -0.12970224  0.67259292  0.338157583 -0.465235707
#>  [87,] -0.564600227 -0.02759953  0.84847124  0.691211635 -0.753747993
#>  [88,] -0.335164327 -0.46283812 -0.60978440 -0.380079885  0.163700269
#>  [89,] -0.132469858  0.36067566 -0.16796056 -0.810758380 -0.615528808
#>  [90,] -0.907637923 -0.17395157 -0.77568471 -0.671559540  0.974357685
#>  [91,] -0.191336662 -0.75410349 -0.85169759  0.363064759  0.411771934
#>  [92,] -0.050390258 -0.80164914  0.79822215  0.850479217  0.831902043
#>  [93,] -0.256164910 -0.76663339 -0.23347897  0.322901204 -0.219206915
#>  [94,] -0.734734375  0.74694932  0.18363615  0.993561779  0.385343338
#>  [95,] -0.770122869  0.20453742  0.04575705  0.224177040 -0.057827411
#>  [96,]  0.353745964 -0.08739316  0.63848693 -0.897092163 -0.309558927
#>  [97,]  0.964612151 -0.94345846 -0.78055173 -0.145043154 -0.861812916
#>  [98,] -0.247379538 -0.15896458 -0.97012519  0.514366796  0.262086481
#>  [99,] -0.177691914 -0.54501371  0.88708535  0.799829619 -0.084182553
#> [100,] -0.403434306 -0.01704826 -0.94875537 -0.435890809  0.388057383
#> 
#> $y
#>   [1] 0 1 1 1 1 1 1 0 1 0 0 0 1 1 0 1 0 1 1 1 0 0 1 0 1 0 0 0 1 0 0 1 0 0 0 1 1
#>  [38] 1 1 1 1 0 0 1 0 1 1 0 1 0 1 1 1 0 0 1 1 1 1 1 0 0 1 1 1 1 0 1 0 0 0 0 0 0
#>  [75] 1 1 1 1 0 0 1 0 1 0 0 1 1 1 1 0 1 0 0 0 1 1 0 1 1 1
#>