Simulate Data from Hyper-Sphere for Classification Problems

The classification problem Circle is described in the BARK paper (2008). Inputs are dim independent variables uniformly distributed on the interval \([-1,1]\), only the first 2 out of these dim are actually signals. Outputs are created according to the formula $$y = 1(x1^2+x2^2 \le 2/\pi)$$

Usage

sim_circle(n, dim = 5)

Arguments

n

number of data points to generate

dim

number of dimension of the problem, no less than 2

Value

Returns a list with components

x

input values (independent variables)

y

0/1 output values (dependent variable)

References

Ouyang, Zhi (2008) Bayesian Additive Regression Kernels. Duke University. PhD dissertation, Chapter 3.

See also

Examples

sim_circle(n=100, dim=5)
#> $x
#>               [,1]        [,2]         [,3]        [,4]          [,5]
#>   [1,]  0.27541662  0.73842348 -0.546405556 -0.08413837  0.2003096519
#>   [2,]  0.16683683  0.34693418  0.015996186 -0.94188520 -0.2250596588
#>   [3,]  0.01023620  0.23723603  0.307810749 -0.24142607 -0.2239676127
#>   [4,]  0.35282910 -0.76892379 -0.866911229  0.95695184 -0.0271255281
#>   [5,] -0.67208290  0.65762356 -0.784529430  0.57262793 -0.6749815154
#>   [6,] -0.03057291 -0.04136531 -0.697697517 -0.30412081 -0.9063444645
#>   [7,] -0.45159547 -0.05511296  0.098752826 -0.06260995 -0.2469213190
#>   [8,] -0.01597010  0.14546803  0.452006657  0.77389289 -0.8911703411
#>   [9,]  0.85253942  0.71060628  0.280920897  0.69707193  0.0005024183
#>  [10,] -0.48816325  0.25378587  0.146704090  0.11866621 -0.6994137797
#>  [11,]  0.59878191 -0.35501183  0.295156871 -0.46727775 -0.6526608695
#>  [12,]  0.39285420  0.70385878  0.588498115  0.43252647 -0.7802510774
#>  [13,] -0.54891330 -0.74416508 -0.596723865 -0.20367541 -0.7366492120
#>  [14,]  0.96283725 -0.83475879  0.200810403  0.77158340 -0.4802039284
#>  [15,] -0.59264571  0.75826483  0.965885012  0.50496020  0.6294245790
#>  [16,] -0.22254457 -0.95892746  0.890641325 -0.76621702  0.1497623501
#>  [17,]  0.17060484 -0.11249648  0.833562395 -0.28657406  0.1686657472
#>  [18,]  0.69659843 -0.70220528  0.833575253  0.93795061 -0.1751382081
#>  [19,] -0.13595064 -0.51645448  0.882180597 -0.42249614  0.6366903465
#>  [20,]  0.28283262  0.70672653 -0.084464875 -0.31009242  0.7895751782
#>  [21,]  0.48234016  0.39162606 -0.500226141 -0.95721505  0.4187824973
#>  [22,]  0.68038747  0.26766570  0.148954394  0.94824285  0.7470161384
#>  [23,] -0.72142470  0.74564186  0.539058947 -0.26211476 -0.1603043964
#>  [24,] -0.63831388  0.08858375 -0.888180770  0.76458387  0.2271075165
#>  [25,]  0.79529241  0.25917136 -0.939115958  0.95406421  0.3239523447
#>  [26,]  0.76037476 -0.51533960 -0.212074370 -0.99409617  0.6620545448
#>  [27,]  0.25002390  0.00466894  0.934140368  0.35750075 -0.5837626918
#>  [28,] -0.04080911  0.50125777  0.589562606 -0.10192650 -0.0003314172
#>  [29,] -0.88722040 -0.30037379  0.005894921 -0.51740794 -0.4726823298
#>  [30,] -0.47716083 -0.61191032 -0.309443533 -0.69050573 -0.6510186344
#>  [31,]  0.91453067 -0.41717829 -0.937109454 -0.18007465 -0.4441817375
#>  [32,]  0.95602019 -0.67705064 -0.168926840  0.42890082 -0.9418783183
#>  [33,]  0.78307877  0.91079349 -0.129682064  0.87795671 -0.0833904524
#>  [34,] -0.35332036 -0.55570654  0.257691880  0.26477165 -0.6307458058
#>  [35,]  0.69846639  0.18400360  0.130435386  0.18955708 -0.6575175971
#>  [36,] -0.96192682  0.75499921 -0.628311678 -0.33340009 -0.6034469232
#>  [37,]  0.90298560 -0.81592929  0.269436022 -0.21227945 -0.1431654566
#>  [38,]  0.77401248 -0.31841079  0.854955190  0.78496164  0.1442744317
#>  [39,] -0.82086818 -0.09089631 -0.645208552 -0.30436189 -0.2918875054
#>  [40,] -0.55678930  0.07083707 -0.908409787 -0.83453442 -0.5135752899
#>  [41,]  0.93467279  0.58808529  0.564376224  0.04477954  0.3720851759
#>  [42,] -0.91009574  0.67364362 -0.556174370 -0.18267051  0.4922916754
#>  [43,] -0.57870926 -0.38065819  0.551230425  0.85981650  0.5320804510
#>  [44,]  0.13360329  0.84905524  0.181928909 -0.43001970 -0.9636375187
#>  [45,]  0.52669564  0.64221891  0.481765935  0.76713098 -0.4651257289
#>  [46,] -0.64846718  0.88129687  0.597442476 -0.99852294 -0.8844283782
#>  [47,]  0.45949006  0.48011851 -0.333045681  0.89446867 -0.4582555373
#>  [48,]  0.12559889 -0.33663667 -0.115498350 -0.26487092  0.9061357705
#>  [49,]  0.17208427  0.94430964 -0.889564582 -0.06077192 -0.1593813272
#>  [50,] -0.50102671  0.62069307 -0.903242742 -0.81298798 -0.4911596985
#>  [51,]  0.45994881  0.72072641 -0.245723306  0.38766375  0.6689902255
#>  [52,]  0.70779428 -0.56257012 -0.149176762 -0.10560706  0.0569531140
#>  [53,] -0.28287124  0.47013121  0.909468332  0.92213246 -0.9739943603
#>  [54,] -0.58221743  0.87106310 -0.713813319  0.14360301 -0.2972560674
#>  [55,] -0.54227856  0.09557096 -0.772904322 -0.39016487  0.5047827554
#>  [56,] -0.33944620  0.80961911 -0.681535278  0.38556283 -0.0198598690
#>  [57,] -0.05830118 -0.44042793  0.274466838  0.66353817 -0.1849195575
#>  [58,]  0.19564274 -0.93938158  0.140423959  0.47613640 -0.8072478790
#>  [59,] -0.34430656  0.97367363 -0.679415409  0.85768711  0.1576568326
#>  [60,] -0.82473028 -0.27494053  0.423605665 -0.90267902  0.1830867166
#>  [61,] -0.97118330  0.76631782 -0.660376422  0.40860176 -0.8918468892
#>  [62,]  0.64329069 -0.84499320  0.806250605 -0.64441611  0.6584232152
#>  [63,] -0.08502647 -0.24213436 -0.701839640  0.55881988  0.5707650110
#>  [64,] -0.97363747  0.12394986  0.449023749  0.33761813  0.1434509554
#>  [65,] -0.59823867 -0.15891994  0.839959123  0.14828633 -0.0986248972
#>  [66,] -0.26205385  0.33896133  0.403890286  0.71862556  0.1885974207
#>  [67,] -0.13831131  0.71698399 -0.976080068 -0.75293260 -0.4618719341
#>  [68,] -0.84776481  0.01646776  0.806393036 -0.99099953 -0.5535192215
#>  [69,]  0.37945764  0.29416787  0.178300142 -0.81653403  0.8610591027
#>  [70,] -0.46585545 -0.75574726  0.857649603 -0.96611939 -0.5732851126
#>  [71,] -0.23460233 -0.99798408 -0.482880985  0.29301183 -0.5739200823
#>  [72,] -0.02620035 -0.82657804 -0.761103197  0.70630010  0.1594386245
#>  [73,]  0.44261450  0.08047205  0.723888047  0.18275868 -0.5069991294
#>  [74,]  0.47448635  0.42967162  0.901762541  0.68009935  0.9923244207
#>  [75,] -0.31130139  0.25365503 -0.388835569  0.62181237  0.6364997155
#>  [76,]  0.18339471  0.31450443  0.034053870  0.68153588 -0.7462277040
#>  [77,] -0.24256230 -0.42747470  0.948339412 -0.48898126  0.5221706112
#>  [78,] -0.50555149 -0.35012189 -0.535193027  0.87482147 -0.5679996111
#>  [79,] -0.34896398 -0.26516774  0.837369546 -0.27413914  0.5890050465
#>  [80,] -0.30781102 -0.77213600  0.771931740 -0.26273220  0.1585858762
#>  [81,] -0.42250348 -0.67688369 -0.525208137 -0.69801869 -0.6462934501
#>  [82,] -0.51475765  0.55505861 -0.104061283 -0.63289028 -0.4611705882
#>  [83,]  0.55109465  0.30530222  0.457964406 -0.15979354 -0.6988902777
#>  [84,]  0.52935947  0.86099476 -0.032113574 -0.76861687  0.9811893031
#>  [85,] -0.28856680 -0.76160364  0.238489840 -0.84092400 -0.5002017901
#>  [86,] -0.64476826  0.57146954  0.527525979  0.97155957 -0.4326326917
#>  [87,]  0.40422854 -0.57950373  0.182272754 -0.33249099  0.9407311589
#>  [88,] -0.40300586 -0.99982646  0.400773089  0.72776730 -0.5769531457
#>  [89,]  0.19210846 -0.87770296  0.089425715  0.67511364  0.3672286468
#>  [90,]  0.37115585  0.89358944  0.675032484  0.49820549  0.7208350897
#>  [91,] -0.66438525 -0.82772571 -0.526909029 -0.37651600 -0.7937996057
#>  [92,] -0.82674909 -0.68357181  0.934357282 -0.91792540 -0.6354457512
#>  [93,] -0.01648522 -0.21710179  0.559642100 -0.20314535 -0.0730503593
#>  [94,] -0.43550299 -0.40000203 -0.506886641 -0.03932372 -0.4086908130
#>  [95,]  0.61741105  0.30743414  0.951944877  0.78301811  0.5883932505
#>  [96,]  0.02314252 -0.40033554 -0.920604968  0.42877188 -0.3927406110
#>  [97,] -0.58572703  0.20094449 -0.959827492  0.41737688  0.3325095898
#>  [98,] -0.10437260 -0.58832544 -0.457817403  0.38856208  0.2412759410
#>  [99,] -0.50599232  0.56821677  0.167828629 -0.83504113 -0.9658793821
#> [100,]  0.85829260 -0.39747855  0.379116833 -0.60678259 -0.8428785531
#> 
#> $y
#>   [1] 1 1 1 0 0 1 1 1 0 1 1 0 0 0 0 0 1 0 1 1 1 1 0 1 0 0 1 1 0 1 0 0 0 1 1 0 0
#>  [38] 0 0 1 0 0 1 0 0 0 1 1 0 1 0 0 1 0 1 0 1 0 0 0 0 0 1 0 1 1 1 0 1 0 0 0 1 1
#>  [75] 1 1 1 1 1 0 0 1 1 0 0 0 1 0 0 0 0 0 1 1 1 1 1 1 1 0
#>