Simulate Data from Hyper-Sphere for Classification Problems

The classification problem Circle is described in the BARK paper (2008). Inputs are dim independent variables uniformly distributed on the interval \([-1,1]\), only the first 2 out of these dim are actually signals. Outputs are created according to the formula $$y = 1(x1^2+x2^2 \le 2/\pi)$$

sim_circle(n, dim = 5)

Arguments

n

number of data points to generate

dim

number of dimension of the problem, no less than 2

Value

Returns a list with components

x

input values (independent variables)

y

0/1 output values (dependent variable)

References

Ouyang, Zhi (2008) Bayesian Additive Regression Kernels. Duke University. PhD dissertation, Chapter 3.
http://stat.duke.edu/people/theses/OuyangZ.html

See also

Examples

sim_circle(n=100, dim=5)
#> $x
#>                 [,1]         [,2]        [,3]         [,4]         [,5]
#>   [1,]  1.871763e-01 -0.738491699 -0.18728731 -0.049401613 -0.314469191
#>   [2,] -2.491986e-01 -0.716745245  0.44203114 -0.021176855  0.487658152
#>   [3,] -4.694492e-01  0.817125425  0.19917789  0.541024325 -0.634352773
#>   [4,]  1.544428e-01 -0.178847570  0.60621194 -0.704064558  0.104734962
#>   [5,] -3.684966e-01 -0.188070510  0.95101064 -0.014899707  0.537847454
#>   [6,]  5.914430e-01 -0.673990234 -0.77516763  0.159065114  0.161214680
#>   [7,] -1.033033e-01 -0.325608386  0.77810283 -0.393385474 -0.660806565
#>   [8,] -4.997513e-01  0.506995993 -0.35975705  0.723882723 -0.795409505
#>   [9,] -9.466432e-01  0.980508330 -0.22117291 -0.941165815  0.611457064
#>  [10,]  5.036733e-01 -0.688256277 -0.88736854 -0.536625543 -0.012282068
#>  [11,]  2.533515e-01  0.401444993 -0.72557004 -0.060009826 -0.317855241
#>  [12,]  9.036689e-01  0.857920752  0.11817190 -0.423964284  0.341523265
#>  [13,]  5.910817e-01 -0.812597383  0.77934297  0.950965754 -0.804805823
#>  [14,] -6.932928e-01 -0.292591605 -0.43455255  0.120527061  0.039111029
#>  [15,]  9.330926e-01  0.652595908  0.40150070 -0.301652498 -0.680757541
#>  [16,] -3.294153e-01  0.859198475 -0.58931023  0.146855012  0.631641453
#>  [17,] -4.826904e-01  0.612369169 -0.94550125 -0.329742533 -0.902267595
#>  [18,] -8.473516e-01  0.895147598 -0.97365452  0.894011582 -0.085482932
#>  [19,] -6.054080e-01  0.659180502 -0.90830428 -0.196784295 -0.643330478
#>  [20,]  5.893176e-05 -0.448953266 -0.22308234  0.208484963  0.013252604
#>  [21,] -3.238788e-01  0.488387102  0.48580483 -0.392547018 -0.279231416
#>  [22,] -4.271732e-01  0.190433786 -0.53676948  0.619427054 -0.850171528
#>  [23,] -4.587008e-01 -0.810220657  0.47421065 -0.690256552  0.508675290
#>  [24,]  3.076414e-02 -0.843256520  0.05275254 -0.509899453  0.553388470
#>  [25,]  4.826016e-01  0.346683952 -0.23045077  0.001451641  0.566297506
#>  [26,]  9.757131e-01  0.113051523  0.67387123 -0.290358226 -0.410628229
#>  [27,] -6.264179e-01 -0.750629301 -0.15857410 -0.967232723  0.487934932
#>  [28,] -3.846681e-01 -0.307121497  0.02480394  0.353519700  0.678430195
#>  [29,]  8.164469e-01 -0.640311297 -0.56766458  0.962416237  0.021083383
#>  [30,] -3.475959e-01 -0.116914934  0.68777712  0.829320246  0.984692854
#>  [31,] -5.602201e-01 -0.375250139  0.23502268 -0.823474078 -0.548660501
#>  [32,]  7.568969e-01 -0.448674088  0.99187692  0.896339074 -0.492303941
#>  [33,] -3.669344e-01 -0.696352002 -0.99490472  0.565170535 -0.979416207
#>  [34,] -1.961313e-01  0.738614496 -0.40297715  0.614696562  0.942014337
#>  [35,]  3.584994e-01 -0.772777993  0.71241298  0.493451049  0.089688962
#>  [36,] -7.895646e-01 -0.661413015 -0.25959024 -0.514409209  0.788640399
#>  [37,] -2.787738e-01  0.846119496  0.53161291  0.612905860 -0.613822689
#>  [38,] -8.851427e-01  0.609442492 -0.59099877  0.671542356 -0.771239620
#>  [39,] -4.097528e-01  0.592490804 -0.59307708 -0.868419558  0.411198339
#>  [40,] -4.823107e-01  0.480976390  0.01626603 -0.084685247  0.834332249
#>  [41,] -7.174829e-01  0.928160166  0.76727889 -0.403307654 -0.558507730
#>  [42,] -3.246471e-01  0.871907501 -0.29719566  0.107294681  0.203397939
#>  [43,]  2.289261e-01  0.312264246 -0.55578095 -0.238625559 -0.395506666
#>  [44,] -9.296753e-01  0.368808480 -0.71804529  0.878219761 -0.647134370
#>  [45,]  2.881432e-01  0.334532191 -0.70742000 -0.509158341 -0.070166110
#>  [46,] -7.818248e-01  0.137145722 -0.45168271 -0.543997886 -0.918589870
#>  [47,] -6.300087e-01 -0.694807636 -0.16515283  0.686012591 -0.658199986
#>  [48,]  2.722855e-01 -0.360663002  0.69636801  0.427430409 -0.835477984
#>  [49,]  8.184612e-01  0.956800101  0.60812065  0.432237353 -0.128688749
#>  [50,] -2.708687e-01  0.037644053 -0.57784278  0.721135540  0.024730169
#>  [51,]  5.943050e-01 -0.275882640 -0.56800298  0.427278677  0.701902965
#>  [52,] -9.430874e-01 -0.928226499  0.13087953  0.479715131 -0.550802686
#>  [53,]  4.402492e-01 -0.436636275  0.44191529  0.239878529 -0.674138476
#>  [54,]  7.259095e-01  0.822460932 -0.48625291 -0.328875050 -0.171019926
#>  [55,]  8.128466e-01 -0.431472858  0.66998049 -0.645225286 -0.342776389
#>  [56,] -6.630099e-02 -0.730695344  0.43317063 -0.335045865 -0.315018152
#>  [57,] -3.619494e-01  0.811741938  0.17527403 -0.972545797 -0.558963022
#>  [58,]  1.693867e-01  0.923831746  0.29558158  0.682937643  0.116749377
#>  [59,]  8.591869e-01 -0.002045331  0.56819462  0.034978900  0.849530011
#>  [60,] -7.615605e-01 -0.764408813  0.56797845 -0.901562822  0.173474102
#>  [61,]  7.408187e-01  0.945065351  0.76900456 -0.785616364  0.985608277
#>  [62,] -7.068341e-01 -0.656825262  0.21402992 -0.374620367  0.710051999
#>  [63,]  1.948711e-01 -0.395564524  0.40771928 -0.416357393  0.318761814
#>  [64,] -2.159566e-01  0.056012675 -0.91417165 -0.131807635 -0.937870700
#>  [65,] -5.716465e-01 -0.655564852 -0.99153101 -0.658989378  0.590766601
#>  [66,]  6.332991e-02 -0.432582659  0.44636206  0.896395315 -0.020542574
#>  [67,]  2.897413e-01  0.959467463 -0.74241368  0.885722470  0.335414378
#>  [68,]  9.265806e-01 -0.591216687 -0.48201048  0.899862274 -0.780934124
#>  [69,]  2.122067e-02  0.408129001 -0.88549215  0.300825086 -0.561532416
#>  [70,] -9.457372e-01  0.844792413 -0.27143907 -0.058514057  0.803669722
#>  [71,] -2.962416e-01  0.639163221  0.18488223 -0.651793944 -0.945391078
#>  [72,] -2.778417e-01  0.154733868  0.25470219 -0.380586578  0.137190679
#>  [73,] -8.156792e-01  0.317023704 -0.41542843 -0.852047912 -0.741761419
#>  [74,]  9.110376e-01 -0.084439035 -0.43745238 -0.412092356 -0.658535372
#>  [75,] -5.366019e-02  0.311766841 -0.14925307 -0.238498243 -0.342690618
#>  [76,]  7.390590e-02 -0.211602347 -0.99625873  0.437622834  0.009406299
#>  [77,]  9.110773e-01 -0.531246896 -0.44484074  0.499964060  0.692533333
#>  [78,]  2.988849e-01 -0.183852139  0.15252614 -0.381588019 -0.098900646
#>  [79,]  8.084631e-01  0.724176013  0.90154327 -0.913538745 -0.736449129
#>  [80,] -1.097667e-01  0.027738762  0.24111479  0.641689055 -0.066691526
#>  [81,] -2.230567e-01  0.828449962 -0.68595787  0.765748644  0.414357461
#>  [82,]  4.740743e-01 -0.134631036 -0.74969312  0.438577093  0.033945378
#>  [83,] -3.812050e-01 -0.303136574  0.94302603 -0.576149116  0.279550482
#>  [84,] -7.226701e-01 -0.523677286  0.19396957  0.802317559 -0.847074366
#>  [85,] -4.862518e-01  0.377479784 -0.80057107  0.264778397  0.411192549
#>  [86,] -4.196230e-01 -0.260186781  0.89184300  0.965067043  0.450012394
#>  [87,]  9.959337e-01  0.247006686 -0.24288245  0.312016230  0.482593852
#>  [88,]  7.886617e-01  0.982689291  0.91217994 -0.196172458 -0.441409212
#>  [89,]  4.940806e-01  0.891884246  0.07842819  0.081894046  0.903410882
#>  [90,] -9.647330e-01  0.068450608 -0.30311417  0.988009321  0.396306979
#>  [91,] -5.280568e-01  0.045514598  0.64107967  0.834072208  0.443025634
#>  [92,]  2.819815e-01  0.516620049 -0.88868744 -0.251353662 -0.760346446
#>  [93,] -5.353556e-01 -0.893186650 -0.95997103  0.247823257  0.305075746
#>  [94,]  1.511494e-01  0.228262845  0.10707183 -0.324489131 -0.865943319
#>  [95,]  9.073619e-01 -0.396511829  0.26401867  0.728067994  0.534097685
#>  [96,]  4.365560e-01 -0.483391755 -0.51510677 -0.490364019 -0.384528407
#>  [97,]  5.934098e-01  0.592089546  0.01941695 -0.536768377  0.769205285
#>  [98,]  9.259206e-02  0.611647911  0.34701741  0.233253962  0.647940889
#>  [99,] -8.377304e-01  0.144113974 -0.26970204 -0.073116046  0.112699376
#> [100,]  7.443445e-01 -0.583566025  0.38488918 -0.682662499  0.864461002
#> 
#> $y
#>   [1] 1 1 0 1 1 0 1 1 0 0 1 0 0 1 0 0 1 0 0 1 1 1 0 0 1 0 0 1 0 1 1 0 1 1 0 0 0
#>  [38] 0 1 1 0 0 1 0 1 1 0 1 0 1 1 0 1 0 0 1 0 0 0 0 0 0 1 1 0 1 0 0 1 0 1 1 0 0
#>  [75] 1 1 0 1 0 1 0 1 1 0 1 1 0 0 0 0 1 1 0 1 0 1 0 1 0 0
#>