Simulate Data from Hyper-Sphere for Classification Problems

The classification problem Circle is described in the BARK paper (2008). Inputs are dim independent variables uniformly distributed on the interval \([-1,1]\), only the first 2 out of these dim are actually signals. Outputs are created according to the formula $$y = 1(x1^2+x2^2 \le 2/\pi)$$

sim_circle(n, dim = 5)

Arguments

n

number of data points to generate

dim

number of dimension of the problem, no less than 2

Value

Returns a list with components

x

input values (independent variables)

y

0/1 output values (dependent variable)

References

Ouyang, Zhi (2008) Bayesian Additive Regression Kernels. Duke University. PhD dissertation, Chapter 3.

See also

Examples

sim_circle(n=100, dim=5)
#> $x
#>                [,1]        [,2]         [,3]         [,4]         [,5]
#>   [1,]  0.325040199  0.32594270 -0.328010702  0.817617666 -0.923126321
#>   [2,] -0.415296027  0.75399597 -0.447000902 -0.143941563  0.705474217
#>   [3,]  0.721029098 -0.48398014  0.144597367  0.133591434  0.852595195
#>   [4,]  0.430826271 -0.71624400  0.786465541  0.646549499  0.853745170
#>   [5,] -0.756943604 -0.81335494 -0.205236849 -0.834943488  0.730894161
#>   [6,]  0.839556729  0.56981516 -0.828595658  0.671067378 -0.241335235
#>   [7,] -0.669934825  0.89743212  0.099599130  0.697450334 -0.848696220
#>   [8,] -0.314801582  0.70771202 -0.581160437 -0.418211094 -0.622019729
#>   [9,]  0.864912841 -0.20857485 -0.299963899  0.123767273 -0.551275863
#>  [10,] -0.753184668  0.42248826 -0.072700884 -0.350105314  0.215759957
#>  [11,]  0.913454564  0.40228067  0.663535715 -0.112881287  0.854214272
#>  [12,] -0.619266361  0.28777680  0.539825453  0.808923330  0.709726865
#>  [13,]  0.712623593 -0.56954175  0.845648487  0.740193722 -0.571327894
#>  [14,]  0.407844156  0.96297196 -0.322393435 -0.224828823  0.854569323
#>  [15,] -0.126592164  0.82079105 -0.426848793  0.222893114  0.247430644
#>  [16,] -0.714407205 -0.84737895  0.793938278 -0.280764046  0.635261865
#>  [17,] -0.747118250  0.60542262  0.554401263 -0.939650899 -0.309359720
#>  [18,]  0.099840999  0.79742234 -0.179938165 -0.870706650 -0.818717338
#>  [19,]  0.268370493  0.96568130  0.354057518 -0.013232108 -0.153586300
#>  [20,]  0.829330946  0.16463721  0.062072915 -0.269800076  0.931804724
#>  [21,] -0.537806964 -0.39355914  0.010514815  0.953104697  0.339494046
#>  [22,]  0.441812796 -0.33450884  0.179890050 -0.536294712 -0.523892551
#>  [23,] -0.307100325  0.89533288  0.388730046  0.574972395 -0.991740827
#>  [24,]  0.358339743  0.40580068 -0.360948174 -0.875962408 -0.163110047
#>  [25,] -0.421880733 -0.19611842 -0.381439076 -0.518931694  0.683404305
#>  [26,] -0.009431050 -0.93803818  0.960720509 -0.138104817  0.605036126
#>  [27,]  0.250439598 -0.82202780 -0.448755411 -0.014792455 -0.103738687
#>  [28,] -0.315170624 -0.58303840  0.817896641  0.890441507 -0.968688413
#>  [29,]  0.039521282  0.61927768  0.501503156 -0.319396820 -0.201285830
#>  [30,] -0.525648776  0.15415198  0.894825161  0.031786413 -0.652797045
#>  [31,]  0.405828094  0.74238341  0.933523110  0.796683216 -0.661354790
#>  [32,]  0.646793029 -0.99911794 -0.217450625  0.897956403  0.464545018
#>  [33,]  0.429500298  0.18436414 -0.048015444  0.601461390  0.865266955
#>  [34,] -0.784173680 -0.88022271 -0.788539278  0.551885094  0.323051895
#>  [35,] -0.297377051 -0.42507874  0.397922558 -0.309310534 -0.791173968
#>  [36,]  0.102860863 -0.87425204  0.660617840 -0.292000425 -0.083782867
#>  [37,] -0.074921106 -0.87786504  0.389935797 -0.240407658 -0.932785430
#>  [38,] -0.081156192 -0.07427110  0.501629675  0.864331724  0.365806283
#>  [39,] -0.712487309  0.31058397  0.950110510 -0.425380723 -0.963921829
#>  [40,] -0.742973737 -0.87473534  0.314960713  0.813216129 -0.380141581
#>  [41,] -0.689534688 -0.46068790  0.135887621 -0.982133043  0.110443089
#>  [42,]  0.075497744 -0.50679765 -0.648980380  0.515988651 -0.242141444
#>  [43,] -0.582076150 -0.83244008  0.149071285 -0.299935696  0.295952892
#>  [44,] -0.387688981  0.02713258 -0.839306764  0.048739982 -0.154857629
#>  [45,]  0.353399390  0.31781784  0.719484042 -0.531570959  0.737149171
#>  [46,]  0.422892805  0.21643526 -0.148412578  0.513384228 -0.521566193
#>  [47,]  0.448439721  0.54377700 -0.773641029 -0.032751150 -0.633453327
#>  [48,]  0.921171035  0.28328277 -0.367794365 -0.770399697 -0.590910539
#>  [49,]  0.652141321 -0.40609793  0.703980440 -0.377688694 -0.009147940
#>  [50,]  0.698989217 -0.06806710  0.306744919 -0.439119282 -0.013494755
#>  [51,] -0.079274662 -0.86178004  0.292090224  0.995137325 -0.565429319
#>  [52,]  0.687017091  0.26377861  0.149796194  0.837252917  0.035913222
#>  [53,] -0.887211452  0.24214240 -0.742903819  0.405447461 -0.561225051
#>  [54,] -0.801577084 -0.52520178  0.187207953 -0.571887114  0.189674391
#>  [55,]  0.878375507  0.61455326  0.930966808 -0.543433756 -0.262681012
#>  [56,]  0.704529749 -0.65047563 -0.277018260  0.638306341  0.913820232
#>  [57,]  0.559032932 -0.86337743 -0.143371172 -0.576633241 -0.796437166
#>  [58,]  0.037883618 -0.77166051  0.447031549 -0.619284433 -0.585700997
#>  [59,] -0.043633218 -0.10716118 -0.400506582  0.465316354  0.506053409
#>  [60,]  0.086407211  0.50563679 -0.254343097  0.237484475  0.570029265
#>  [61,] -0.365965855  0.46147277 -0.276483248  0.176874801 -0.285196160
#>  [62,] -0.884922961 -0.95380634 -0.337764289 -0.555623785  0.710175010
#>  [63,] -0.651525542  0.66632323 -0.228779279 -0.447395334 -0.321322744
#>  [64,]  0.237301683  0.19623023  0.077242150 -0.769965145  0.662327088
#>  [65,] -0.600930239  0.70308880 -0.836946502  0.135769105 -0.798432817
#>  [66,]  0.073220958 -0.34349863 -0.231833668  0.372114406  0.037790946
#>  [67,] -0.448546302  0.98123178 -0.817350792 -0.867418007  0.226681494
#>  [68,] -0.089054786  0.23237708  0.830795275 -0.352030581 -0.006909641
#>  [69,]  0.913049894 -0.63038157  0.398331837 -0.300549147  0.339169526
#>  [70,]  0.845319798 -0.91612507 -0.262475585  0.736552209 -0.848608491
#>  [71,] -0.117009664  0.85974705  0.212013303 -0.763268815  0.756459989
#>  [72,] -0.337672370 -0.02936554 -0.736562858 -0.369779987 -0.985045056
#>  [73,]  0.978766250  0.54797675 -0.992080315 -0.801812264  0.403946726
#>  [74,]  0.885005452  0.77558605 -0.243723765  0.268560664 -0.560701755
#>  [75,] -0.340043370  0.05005487  0.537928129  0.503769901  0.916538610
#>  [76,] -0.328717356  0.03850324 -0.018539562 -0.453220455 -0.700320565
#>  [77,]  0.002849699  0.10115455 -0.204344443  0.694826859 -0.413539357
#>  [78,]  0.697205036  0.19465724  0.560324763 -0.189733001  0.595630691
#>  [79,] -0.065055443 -0.16963676  0.985006765 -0.308182693 -0.988615015
#>  [80,] -0.201482621 -0.80381679 -0.614224868 -0.004584064 -0.233565301
#>  [81,] -0.013637914  0.41299120  0.805314421 -0.331578210 -0.002929828
#>  [82,]  0.011583295 -0.73222893 -0.960010644  0.263502871 -0.589774712
#>  [83,] -0.072145853  0.33223510 -0.318723440 -0.766793948 -0.721240599
#>  [84,] -0.471184000 -0.37474061  0.188723780 -0.258959861  0.370815011
#>  [85,]  0.653408352 -0.28312709  0.806079513 -0.859721497 -0.375245476
#>  [86,] -0.472360950 -0.02527741  0.427409086  0.498963187  0.059082258
#>  [87,]  0.046431860 -0.69909992 -0.380020315 -0.833855697  0.722017965
#>  [88,] -0.925109384 -0.72342173 -0.483970679  0.665486655  0.917008772
#>  [89,] -0.480758618  0.78995991 -0.824697822  0.069417059  0.842232635
#>  [90,] -0.277606051  0.97622623  0.951183341 -0.551580195  0.515430893
#>  [91,] -0.648741393 -0.06188290 -0.514559384 -0.858983939 -0.886415923
#>  [92,]  0.360981729  0.68388897 -0.557302044  0.195458729 -0.183458125
#>  [93,] -0.665734364  0.43722971 -0.795153632  0.543849769  0.945457901
#>  [94,]  0.967710743  0.97623032  0.089662903  0.699233803  0.652921261
#>  [95,] -0.752994601  0.46104054  0.379129677  0.082283149  0.678810947
#>  [96,] -0.102614847  0.05169768 -0.375801196 -0.378685164 -0.627826690
#>  [97,] -0.779192383 -0.17249947 -0.771442004 -0.835543864 -0.104342584
#>  [98,]  0.419997567  0.24204652  0.293874383  0.628888599  0.539131688
#>  [99,] -0.401275150  0.78691367  0.120681554 -0.424300779  0.934288196
#> [100,]  0.345166359  0.45323176  0.005322348  0.133587738  0.367030767
#> 
#> $y
#>   [1] 1 0 0 0 0 0 0 1 0 0 0 1 0 0 0 0 0 0 0 0 1 1 0 1 1 0 0 1 1 1 0 0 1 0 1 0 0
#>  [38] 1 1 0 0 1 0 1 1 1 1 0 1 1 0 1 0 0 0 0 0 1 1 1 1 0 0 1 0 1 0 1 0 0 0 1 0 0
#>  [75] 1 1 1 1 1 0 1 1 1 1 1 1 1 0 0 0 1 1 1 0 0 1 0 1 0 1
#>