Simulate Data from Hyper-Sphere for Classification Problems

The classification problem Circle is described in the BARK paper (2008). Inputs are dim independent variables uniformly distributed on the interval \([-1,1]\), only the first 2 out of these dim are actually signals. Outputs are created according to the formula $$y = 1(x1^2+x2^2 \le 2/\pi)$$

sim_circle(n, dim = 5)

Arguments

n

number of data points to generate

dim

number of dimension of the problem, no less than 2

Value

Returns a list with components

x

input values (independent variables)

y

0/1 output values (dependent variable)

References

Ouyang, Zhi (2008) Bayesian Additive Regression Kernels. Duke University. PhD dissertation, Chapter 3.

See also

Examples

sim_circle(n=100, dim=5)
#> $x
#>                 [,1]         [,2]        [,3]        [,4]        [,5]
#>   [1,]  0.0788157959  0.235568638 -0.43054051 -0.65675513 -0.68137309
#>   [2,]  0.8104555993  0.131676815  0.44906211  0.77251375 -0.33781647
#>   [3,] -0.0042161359 -0.426726838  0.20458572  0.68874254  0.57274037
#>   [4,] -0.7823198014 -0.493830219 -0.79384514  0.60894353  0.56144365
#>   [5,] -0.1017713048  0.576277959 -0.04437923 -0.03539522 -0.88251739
#>   [6,] -0.2165366365 -0.106551180  0.44965465  0.64772749  0.45526420
#>   [7,] -0.4173641577 -0.468726128  0.81594978  0.27617345  0.66031448
#>   [8,] -0.7645134758  0.108719731  0.58346807 -0.59649475 -0.44965557
#>   [9,] -0.2839505086 -0.549971094  0.01478695 -0.32391105 -0.13059540
#>  [10,]  0.5858955495  0.165553531 -0.45932401  0.89235843 -0.31602730
#>  [11,] -0.6085886429 -0.341385150  0.87023106  0.15001875  0.92620829
#>  [12,]  0.0358816124 -0.393897267 -0.43511032  0.61008265 -0.44186829
#>  [13,] -0.6325174747 -0.620381242 -0.72185811 -0.49024134 -0.53011732
#>  [14,]  0.4502306087  0.644549834 -0.80309123  0.51333189  0.09045961
#>  [15,]  0.4620157038  0.920016040 -0.19390276  0.96496918  0.52636239
#>  [16,]  0.7739798315  0.417444861  0.77275471 -0.37025931  0.63624618
#>  [17,]  0.8514895970  0.340289561 -0.38144334  0.55950621  0.69267517
#>  [18,] -0.9382727547  0.949537326 -0.12642116 -0.14870851  0.66439866
#>  [19,] -0.1230783653 -0.362008485  0.78555576  0.75692465  0.65993446
#>  [20,] -0.7280018101  0.524939822 -0.92333626 -0.89555527  0.99614339
#>  [21,] -0.2047133986  0.715905294  0.09689813 -0.12634293  0.93139269
#>  [22,] -0.3315623556 -0.957314056  0.44542612 -0.22871614 -0.80697563
#>  [23,] -0.4046669197 -0.123760161 -0.50798949  0.64733667  0.91021671
#>  [24,] -0.5106790923  0.493216621  0.89638680  0.35253472  0.25987298
#>  [25,] -0.5584829166 -0.422185418 -0.46613718  0.16485190 -0.85622494
#>  [26,] -0.5222799103  0.748416890  0.43092273 -0.71996124  0.38439309
#>  [27,]  0.3430654868 -0.754423497 -0.49458322 -0.37951550  0.05029225
#>  [28,]  0.0005380614  0.993277306 -0.42587702 -0.92599641 -0.03450191
#>  [29,]  0.4508523974 -0.010923983  0.93553275 -0.32242881  0.49745219
#>  [30,]  0.8113071802  0.841719561  0.04543558 -0.43405201  0.35157492
#>  [31,] -0.3000076646  0.458838556  0.81632543  0.94322637  0.91939461
#>  [32,]  0.3255188172  0.712008526  0.36036449 -0.58102219 -0.77737180
#>  [33,] -0.5752135850  0.784264043  0.36672689 -0.78792204  0.75312416
#>  [34,] -0.1715790839 -0.903595071  0.67351526  0.57237097  0.09822051
#>  [35,]  0.7494105962  0.505500313  0.16799528 -0.29397295  0.57380121
#>  [36,] -0.7206992959 -0.148820832 -0.62418864 -0.70287909  0.30495755
#>  [37,] -0.4093559231 -0.253150553  0.91727548 -0.32932333 -0.62098017
#>  [38,] -0.2194257560 -0.017959986 -0.89643853  0.35254346 -0.35330287
#>  [39,] -0.8629832296 -0.198275887  0.08849309  0.61328815 -0.52698233
#>  [40,]  0.5452662436  0.965637487 -0.76154415  0.02464950  0.40204989
#>  [41,]  0.5681676050  0.141980643 -0.60021090  0.89237382  0.70753588
#>  [42,] -0.0618545259  0.079816257 -0.79532922  0.45043547  0.01174007
#>  [43,] -0.5162933809 -0.315870967 -0.18498386  0.66366196  0.09368190
#>  [44,]  0.0798432995  0.121923591  0.18620896 -0.80017054 -0.25034675
#>  [45,] -0.1149805593 -0.176108246 -0.01912043  0.37794643  0.24378263
#>  [46,] -0.6464811377 -0.310185161  0.22276326 -0.27507008 -0.75122116
#>  [47,]  0.6334038349  0.738619399 -0.85215111  0.16031498  0.37241854
#>  [48,]  0.6541990256 -0.250368464 -0.49193273  0.31386215  0.59535613
#>  [49,] -0.3671474480 -0.182121385  0.90067163 -0.60570730 -0.93358343
#>  [50,] -0.1988331079  0.139785434  0.94272944  0.62836504  0.08193775
#>  [51,] -0.5481623919 -0.392169586  0.17343988 -0.21861452  0.25464822
#>  [52,]  0.6762874783  0.628989836 -0.03941426 -0.08896031 -0.23548198
#>  [53,]  0.6587070464  0.248837540  0.07000988  0.75308830 -0.15024901
#>  [54,] -0.8134497884  0.444532986  0.60050594 -0.26353106 -0.34773311
#>  [55,] -0.4245291371 -0.989485328  0.70733720 -0.91758839  0.53689270
#>  [56,]  0.2815793250  0.600770229  0.47802153  0.80159047  0.51694367
#>  [57,] -0.2747352873  0.669120808  0.54993717  0.70875073 -0.63984017
#>  [58,]  0.9986279914  0.248105599 -0.86719531  0.73118027 -0.30082026
#>  [59,]  0.6284217071 -0.954248329  0.24861497  0.56423584 -0.74877306
#>  [60,] -0.9377605990  0.140216769 -0.99582544 -0.18452676 -0.28847922
#>  [61,] -0.3259870517 -0.853186979 -0.50625332 -0.49694324  0.41208167
#>  [62,] -0.1378878835 -0.280739794 -0.63389543 -0.07600455 -0.36680696
#>  [63,]  0.5479337075  0.257187470 -0.34023525  0.21897919  0.96725743
#>  [64,] -0.7585166786  0.302332238 -0.34047018 -0.30685655 -0.20816397
#>  [65,] -0.2528486801 -0.929975058 -0.14794056 -0.40889082 -0.87683909
#>  [66,] -0.2877463503  0.124842304  0.54072045  0.84567580 -0.70668845
#>  [67,] -0.3767259852 -0.454963429  0.99165323 -0.49464682 -0.77783842
#>  [68,]  0.4054712974 -0.292556015 -0.21638384 -0.12072434  0.17269774
#>  [69,]  0.4924642839  0.219797110  0.68267824 -0.99797002  0.07977452
#>  [70,] -0.1116837296 -0.955030874 -0.20834934 -0.09131498  0.31834331
#>  [71,]  0.6518165013 -0.154020182 -0.95776368  0.82339290 -0.86658479
#>  [72,]  0.3784479182  0.555484067  0.53829961  0.37232803 -0.19689640
#>  [73,]  0.2813368691  0.182140275  0.13496871 -0.11047548  0.82304974
#>  [74,] -0.4386369106 -0.031430176  0.66046448  0.26776461 -0.87753160
#>  [75,]  0.0488948100  0.612696357 -0.22328560  0.37449303  0.64346634
#>  [76,] -0.6764174048  0.983127314 -0.56909688 -0.46123994  0.72061764
#>  [77,] -0.7389471917 -0.976224018 -0.73545725  0.43980649  0.68698746
#>  [78,] -0.6691939575 -0.025190824  0.58297658  0.60874655 -0.51685105
#>  [79,]  0.6048680213  0.703304056  0.25120100 -0.95443368  0.49791890
#>  [80,] -0.2856847504  0.388019740 -0.97857373  0.65251015 -0.69307657
#>  [81,]  0.5702205226 -0.899544618 -0.17715822  0.77289452 -0.94464257
#>  [82,] -0.4022338311 -0.586781760 -0.15810786  0.56530380  0.58597250
#>  [83,] -0.0109377443  0.167853335  0.40425456  0.54613055  0.31183079
#>  [84,] -0.4939554585  0.744484492 -0.96754891 -0.30678771  0.18129210
#>  [85,]  0.3866086644  0.610747463  0.18420308 -0.26979749  0.36236080
#>  [86,] -0.0636806609  0.768214728 -0.47884914 -0.96749842  0.88539232
#>  [87,] -0.2703961032  0.183598856 -0.72597854  0.44757588  0.60850120
#>  [88,]  0.2262744047  0.197390463  0.93908576 -0.44909785  0.24470251
#>  [89,]  0.4552209210 -0.217752466 -0.36227953  0.57663325  0.94197853
#>  [90,] -0.9613756132  0.271772119 -0.34703394  0.78766096  0.54438436
#>  [91,]  0.2189816423 -0.644553798  0.99143828  0.46136665  0.36785435
#>  [92,]  0.2270873068  0.480470179  0.79093922 -0.20564421  0.66942448
#>  [93,]  0.3401268413 -0.722896919 -0.66779935  0.04383393  0.35878107
#>  [94,] -0.4238561508  0.815893224  0.69919784  0.29053407 -0.85248748
#>  [95,] -0.4257608848  0.180997558  0.60836157 -0.98686058  0.39047581
#>  [96,]  0.5915699895 -0.001672821  0.88905911  0.83691511  0.85613783
#>  [97,] -0.3588946229 -0.409759722 -0.24898146  0.64193429 -0.41429103
#>  [98,] -0.8748507043 -0.049216712  0.67055229 -0.49452720  0.62617084
#>  [99,] -0.3996077632  0.854920411  0.05002504  0.42615370 -0.76344983
#> [100,]  0.9699967764 -0.854361940  0.31932018  0.85219193 -0.68696552
#> 
#> $y
#>   [1] 1 0 1 0 1 1 1 1 1 1 1 1 0 1 0 0 0 0 1 0 1 0 1 1 1 0 0 0 1 0 1 1 0 0 0 1 1
#>  [38] 1 0 0 1 1 1 1 1 1 0 1 1 1 1 0 1 0 0 1 1 0 0 0 0 1 1 0 0 1 1 1 1 0 1 1 1 1
#>  [75] 1 0 0 1 0 1 0 1 1 0 1 1 1 1 1 0 1 1 0 0 1 1 1 0 0 0
#>