Simulate Data from Hyper-Sphere for Classification Problems

The classification problem Circle is described in the BARK paper (2008). Inputs are dim independent variables uniformly distributed on the interval \([-1,1]\), only the first 2 out of these dim are actually signals. Outputs are created according to the formula $$y = 1(x1^2+x2^2 \le 2/\pi)$$

Usage

sim_circle(n, dim = 5)

Arguments

n

number of data points to generate

dim

number of dimension of the problem, no less than 2

Value

Returns a list with components

x

input values (independent variables)

y

0/1 output values (dependent variable)

References

Ouyang, Zhi (2008) Bayesian Additive Regression Kernels. Duke University. PhD dissertation, Chapter 3.

See also

Examples

sim_circle(n=100, dim=5)
#> $x
#>                [,1]         [,2]         [,3]        [,4]        [,5]
#>   [1,]  0.975821229 -0.065367376 -0.698348755 -0.99723307 -0.48890117
#>   [2,] -0.815052695  0.735533968 -0.095029953 -0.58409227  0.42798201
#>   [3,] -0.263012892  0.393151745  0.188804665 -0.47342255 -0.13591000
#>   [4,]  0.593501841 -0.912462131 -0.559658856 -0.03094744 -0.28581339
#>   [5,] -0.566189328  0.855882323  0.994283834  0.01369616  0.44241085
#>   [6,]  0.335130811  0.666525793 -0.921743468  0.35043599 -0.46498119
#>   [7,]  0.390097213 -0.611895834 -0.442335628 -0.08038949 -0.49055284
#>   [8,]  0.775166324 -0.973196564  0.004918071 -0.67056756 -0.83784962
#>   [9,]  0.845040507  0.036306241 -0.637934020 -0.92774838 -0.52190159
#>  [10,] -0.588194897 -0.672438187  0.973223836 -0.80801892 -0.29303144
#>  [11,] -0.509265625 -0.109510239 -0.782043473 -0.84144980 -0.24219264
#>  [12,] -0.163607672 -0.084936340 -0.818646852 -0.76605930  0.11098162
#>  [13,] -0.573537747 -0.695290095 -0.759193410 -0.17971130  0.39570591
#>  [14,] -0.630364189  0.238898743  0.697821736  0.87028271 -0.11381618
#>  [15,] -0.063845184 -0.150780298  0.418962287 -0.06003340  0.11751616
#>  [16,]  0.840706772 -0.675197136  0.926446602 -0.15913171 -0.04232543
#>  [17,]  0.193244335  0.511478401  0.222331763  0.55187687  0.83642365
#>  [18,]  0.291841084  0.908209098  0.590221951 -0.80870941 -0.50255247
#>  [19,]  0.828524870  0.661314778 -0.404194033 -0.77038140 -0.15609834
#>  [20,]  0.064816587  0.936050770 -0.525572568  0.34701269  0.81291865
#>  [21,] -0.993333779  0.827697689 -0.299121073  0.01039574  0.10070890
#>  [22,]  0.390423897 -0.685966813  0.568377191 -0.69648412 -0.42196756
#>  [23,] -0.534400884  0.219633265 -0.418356805  0.39547600  0.12678533
#>  [24,]  0.389583200  0.876172713  0.940034709  0.78337977  0.37713280
#>  [25,] -0.587283003  0.458973492  0.441363001 -0.60763344 -0.26380739
#>  [26,]  0.563267371  0.341706194  0.403475602 -0.87042125 -0.99525783
#>  [27,]  0.051921199  0.594847297  0.835958115  0.97042490  0.18843184
#>  [28,]  0.478587355 -0.176673979  0.109846320 -0.16939909  0.77147991
#>  [29,]  0.145051601  0.409302246 -0.070035895 -0.64911002 -0.67123425
#>  [30,] -0.919530912  0.149839206  0.693996111 -0.53930516 -0.91157181
#>  [31,]  0.370488130  0.696852513 -0.489506595  0.93929526 -0.01174439
#>  [32,] -0.332737794  0.794755812 -0.686019420  0.64967088  0.57655104
#>  [33,] -0.312313432  0.302346490  0.510987122 -0.65729316  0.43641980
#>  [34,]  0.434194757  0.291384620  0.687570885  0.04206299  0.79571139
#>  [35,] -0.371711892 -0.025778101  0.591172619  0.68272599  0.45476924
#>  [36,] -0.042330149  0.212297525 -0.279892285  0.23604613 -0.78634023
#>  [37,] -0.988397557  0.568528429  0.366947246  0.69593026  0.04134879
#>  [38,] -0.262641437 -0.522963679  0.658424382  0.95305412 -0.40574110
#>  [39,] -0.346340377 -0.620030962 -0.555712820  0.38185051  0.76368537
#>  [40,]  0.449994297  0.635341565  0.224615518  0.08873528  0.35018392
#>  [41,] -0.380312552  0.621431964 -0.545489703  0.53273834  0.53933062
#>  [42,] -0.696641868 -0.868211989 -0.991717428 -0.45327079  0.74313618
#>  [43,] -0.352516740  0.855951450  0.064392834 -0.76032783 -0.44142275
#>  [44,]  0.216768641 -0.656198630 -0.098391806 -0.77204431 -0.63409283
#>  [45,] -0.697359460  0.762149077  0.997162483 -0.94802138  0.39930339
#>  [46,] -0.908172235 -0.014678840  0.221568286  0.18849122 -0.61071021
#>  [47,] -0.616109109 -0.618729761  0.381824518 -0.34732270  0.24510249
#>  [48,]  0.529100116 -0.203632068 -0.832915975 -0.44581448 -0.25072662
#>  [49,] -0.265580708  0.754177281 -0.794811948 -0.35797404 -0.03609229
#>  [50,] -0.670169831  0.043330316  0.192044011 -0.11261825 -0.77591030
#>  [51,]  0.364536829  0.489465467 -0.607267810 -0.03865060 -0.25492098
#>  [52,]  0.397008236  0.911072383 -0.699721458 -0.14681978  0.04612984
#>  [53,]  0.111009216  0.292305168  0.399146435  0.90387390  0.39511879
#>  [54,] -0.107566810  0.822127958  0.903166850 -0.46761579  0.68891302
#>  [55,] -0.674955708 -0.072072838 -0.883673253 -0.48360353 -0.91957871
#>  [56,]  0.531570596 -0.132979871  0.934383042  0.50393567 -0.89729169
#>  [57,] -0.524264948 -0.612156021 -0.646059985 -0.57150397  0.38649992
#>  [58,] -0.370932793  0.528993536  0.842006715 -0.10169274 -0.80259820
#>  [59,] -0.528685960  0.459765212  0.641131049  0.51956314 -0.18923191
#>  [60,]  0.701635100  0.103467447  0.520580348  0.63809111 -0.56896992
#>  [61,]  0.352936728  0.261712031 -0.677813439 -0.55941460 -0.34514614
#>  [62,]  0.175238678 -0.855521625 -0.508453258  0.93992644  0.19414915
#>  [63,] -0.162050669  0.427968771 -0.609572145  0.39527977  0.59021475
#>  [64,]  0.190126361  0.802670914  0.576466386 -0.30634336 -0.70920388
#>  [65,] -0.819317132 -0.393665046  0.706119288 -0.05520323  0.87631015
#>  [66,]  0.120173351 -0.583094788 -0.331024010 -0.59588774 -0.50237791
#>  [67,] -0.388994395  0.171428238  0.416504533 -0.12379776 -0.04057315
#>  [68,]  0.613561470  0.132465205  0.096438240 -0.81792910 -0.66620721
#>  [69,]  0.363143626 -0.884682405  0.781623146  0.34551774 -0.81290066
#>  [70,]  0.289392569 -0.534602262  0.713496091  0.14275739 -0.44846397
#>  [71,] -0.299091981 -0.509686973  0.150556088  0.20051204 -0.45258852
#>  [72,] -0.767741716  0.863664456 -0.175509269  0.25746419  0.41278452
#>  [73,]  0.960667106  0.068894977 -0.258781706  0.80761249 -0.19831663
#>  [74,] -0.364517690 -0.434063018 -0.792336982  0.61876697  0.79228405
#>  [75,] -0.105796765 -0.006020301 -0.498024422 -0.30836178 -0.75109439
#>  [76,]  0.857822989  0.383285431  0.541935801  0.09927473 -0.12495130
#>  [77,] -0.790495770  0.789297533  0.607551902 -0.95820357 -0.21151420
#>  [78,] -0.951768174 -0.500869565  0.899198059 -0.70368817  0.72371713
#>  [79,] -0.018811516 -0.678665031  0.193551299  0.27815073 -0.19154966
#>  [80,]  0.514903677 -0.932731053  0.222350329  0.20027566  0.40014556
#>  [81,]  0.460864046 -0.241856859  0.561558809 -0.72223291 -0.73068500
#>  [82,] -0.447883463 -0.696356377  0.807053452 -0.85466298 -0.15351240
#>  [83,] -0.143587557 -0.077386705 -0.765955712  0.91099998  0.07738720
#>  [84,]  0.162434344 -0.774136157 -0.294119745  0.47265313  0.40476782
#>  [85,] -0.714669134 -0.633218145  0.760205682  0.15227896 -0.95286401
#>  [86,]  0.687203918  0.658018611 -0.686731548  0.24153293  0.34247138
#>  [87,]  0.570429718 -0.274873781  0.455093249  0.24917322  0.80538393
#>  [88,] -0.232973865  0.071494237 -0.904315461  0.18555306  0.72818883
#>  [89,] -0.828900163 -0.007902018  0.887183749 -0.71647963  0.42910186
#>  [90,] -0.535861402  0.078924255  0.824418414 -0.04099972  0.28863748
#>  [91,]  0.002707991  0.604709005  0.914816401 -0.01527257 -0.11048004
#>  [92,]  0.295062119 -0.301156681 -0.319818270  0.19532948  0.50169212
#>  [93,] -0.985003945 -0.818207680 -0.611722020  0.32608143  0.71379573
#>  [94,]  0.997573541 -0.427474576 -0.429997608 -0.41171462  0.28163536
#>  [95,] -0.783647766  0.168682031  0.497005584 -0.59552252 -0.12730819
#>  [96,] -0.221322501 -0.278914806 -0.425792261 -0.82962892 -0.38647306
#>  [97,] -0.166587968  0.818877855  0.955057937  0.77554383 -0.78697899
#>  [98,] -0.364830270 -0.106279697  0.672487618 -0.21990191 -0.12412769
#>  [99,] -0.576252072 -0.315302453 -0.075710230 -0.99328202  0.48901061
#> [100,]  0.915135166  0.309731346 -0.744181763 -0.52361277 -0.44153839
#> 
#> $y
#>   [1] 0 0 1 0 0 1 1 0 0 0 1 1 0 1 1 0 1 0 0 0 0 1 1 0 1 1 1 1 1 0 1 0 1 1 1 1 0
#>  [38] 1 1 1 1 0 0 1 0 0 0 1 0 1 1 0 1 0 1 1 0 1 1 1 1 0 1 0 0 1 1 1 0 1 1 0 0 1
#>  [75] 1 0 0 0 1 0 1 0 1 1 0 0 1 1 0 1 1 1 0 0 0 1 0 1 1 0
#>