Simulate Data from Hyper-Sphere for Classification Problems

The classification problem Circle is described in the BARK paper (2008). Inputs are dim independent variables uniformly distributed on the interval \([-1,1]\), only the first 2 out of these dim are actually signals. Outputs are created according to the formula $$y = 1(x1^2+x2^2 \le 2/\pi)$$

Usage

sim_circle(n, dim = 5)

Arguments

n

number of data points to generate

dim

number of dimension of the problem, no less than 2

Value

Returns a list with components

x

input values (independent variables)

y

0/1 output values (dependent variable)

References

Ouyang, Zhi (2008) Bayesian Additive Regression Kernels. Duke University. PhD dissertation, Chapter 3.

See also

Examples

sim_circle(n=100, dim=5)
#> $x
#>               [,1]         [,2]         [,3]        [,4]        [,5]
#>   [1,] -0.12359757  0.922689003 -0.326444965  0.74411807  0.88742355
#>   [2,] -0.21917430 -0.784869008 -0.107279066 -0.59409177 -0.04080846
#>   [3,] -0.44788973 -0.968567694  0.523446921  0.79130014 -0.29855906
#>   [4,]  0.70511695  0.257971094  0.460635819 -0.52602724  0.89241626
#>   [5,]  0.39016449  0.520275315  0.367298502 -0.08189511 -0.76986751
#>   [6,]  0.81891212  0.912116693 -0.285437633 -0.86340442  0.37303964
#>   [7,]  0.69745263 -0.310086520 -0.193504860 -0.02855286  0.38553525
#>   [8,]  0.50069405 -0.182804394 -0.592384688  0.45274873  0.16506186
#>   [9,]  0.68417447  0.192950007 -0.929444750 -0.10328378 -0.23505121
#>  [10,] -0.69585745  0.412322951  0.714900593 -0.69737435 -0.04993425
#>  [11,] -0.06932170 -0.821529852  0.449303890 -0.30266643  0.64606473
#>  [12,] -0.32250623 -0.844917905  0.524643654 -0.37009304  0.69313551
#>  [13,] -0.57181307  0.303407928 -0.765801719  0.46544439  0.33994889
#>  [14,]  0.61697300  0.305728688 -0.345273782 -0.88136522  0.22807670
#>  [15,] -0.42239474 -0.911529524 -0.320511890 -0.97490171 -0.76066716
#>  [16,]  0.11684931  0.695692490 -0.015175432  0.32139217  0.53870430
#>  [17,] -0.36463781  0.916055392  0.281583931 -0.19748941  0.04820844
#>  [18,]  0.08216765 -0.031187975 -0.288264855 -0.09306417 -0.64479680
#>  [19,] -0.63120873 -0.053087310 -0.342187898 -0.91925519  0.15008983
#>  [20,]  0.13419851  0.590434723  0.734032664  0.90035988 -0.17062524
#>  [21,]  0.84229306  0.460831632  0.008332543  0.72930953 -0.40485278
#>  [22,] -0.79624033 -0.529304579  0.435865663  0.58009481  0.98024878
#>  [23,] -0.71032452 -0.105660215  0.922267101 -0.45729879  0.75436495
#>  [24,] -0.15566552  0.541631325 -0.499134774 -0.07217538 -0.29364465
#>  [25,] -0.91616296 -0.706712647  0.573971734  0.73141854  0.15019656
#>  [26,]  0.87137130  0.446989172  0.538881365 -0.38717211  0.21430192
#>  [27,] -0.85574451 -0.977360194 -0.045044459  0.72129395 -0.17266975
#>  [28,]  0.85869644 -0.560754435  0.250974937 -0.27903683 -0.34612833
#>  [29,] -0.20677445  0.403794491  0.467669578 -0.36641249  0.56230354
#>  [30,]  0.85938943 -0.883982356 -0.789728421  0.84325779  0.45612931
#>  [31,] -0.43338383 -0.650742684  0.661835312 -0.10999104  0.31180071
#>  [32,] -0.03657798 -0.863402511 -0.676798178 -0.64209167 -0.16224460
#>  [33,]  0.90211927  0.181404112 -0.442326612  0.95499349  0.03975367
#>  [34,]  0.26966093  0.852520461 -0.120747759  0.56283968 -0.56756622
#>  [35,]  0.63155528  0.798387882 -0.268394520  0.83493194 -0.98925466
#>  [36,] -0.37475901  0.556541253  0.357309232  0.25680924 -0.91109816
#>  [37,] -0.50107653  0.829979799  0.096564085 -0.92142398  0.58216006
#>  [38,]  0.80780785 -0.474923547 -0.902941472  0.55826017 -0.54510006
#>  [39,]  0.59894333 -0.957645966 -0.054920278 -0.54822474 -0.37079296
#>  [40,]  0.16406870 -0.020175920 -0.870496906 -0.15646776 -0.76195830
#>  [41,] -0.77246280 -0.036457690 -0.496140248 -0.03539264  0.45517973
#>  [42,] -0.23727178 -0.791461727  0.636306655  0.86503763 -0.42535592
#>  [43,]  0.32461877  0.237472277 -0.482182024  0.58121245  0.62709173
#>  [44,]  0.80948446 -0.985048258 -0.340894286 -0.55507139 -0.03596961
#>  [45,] -0.78582890  0.714372830  0.599312741  0.47967291  0.15707826
#>  [46,] -0.13359836 -0.003758717 -0.108050612  0.76443889 -0.50894232
#>  [47,] -0.88130359 -0.693396080  0.099478780  0.89287266 -0.01877164
#>  [48,]  0.21278864 -0.539981028 -0.346816272 -0.23863541 -0.12582753
#>  [49,]  0.08833248  0.393496790 -0.563523312 -0.13777318 -0.94917047
#>  [50,]  0.69753165  0.321617883  0.735047666 -0.45690272 -0.77896464
#>  [51,] -0.28073147  0.146726094  0.868388333 -0.33624526  0.83192395
#>  [52,] -0.84136066 -0.632989134  0.323970678 -0.12848218 -0.60583204
#>  [53,] -0.14842597 -0.708499870 -0.166908252  0.55705929 -0.60835608
#>  [54,]  0.06072031 -0.686347189  0.835368610  0.63122412 -0.42225803
#>  [55,] -0.37953655 -0.378068733 -0.181563785  0.15403680 -0.22926582
#>  [56,] -0.72509380  0.183309960 -0.998553692 -0.21873319 -0.04667616
#>  [57,]  0.19119049 -0.318888661 -0.665810709  0.65926318  0.24582197
#>  [58,]  0.14560712 -0.147875552 -0.710073031 -0.10095580 -0.03661061
#>  [59,] -0.69603953  0.563460524 -0.267196281  0.42114393  0.84223442
#>  [60,] -0.66083404 -0.630596295 -0.004888769 -0.71947824 -0.94835053
#>  [61,]  0.39538113  0.825660586  0.565068464 -0.07019583 -0.45112209
#>  [62,]  0.75867031  0.483660102  0.982240422 -0.70224451 -0.55452912
#>  [63,] -0.94349296 -0.016436731 -0.009714753  0.14147293  0.41126345
#>  [64,] -0.43765499 -0.327828591 -0.761100843  0.33810024  0.37237294
#>  [65,]  0.78792474 -0.005509073  0.574949403  0.58959174 -0.16283517
#>  [66,] -0.76239706 -0.603333545 -0.378347182  0.76765873  0.29194100
#>  [67,] -0.07523249  0.617222787 -0.215247869  0.31589468  0.89655581
#>  [68,]  0.70154568 -0.415598104  0.585540064 -0.28298661 -0.79100330
#>  [69,]  0.90873179 -0.491947397 -0.144710365  0.68075662  0.15249788
#>  [70,]  0.93028818  0.109578053  0.834916749 -0.52962469  0.77231505
#>  [71,] -0.74033186 -0.363446318 -0.420177365  0.35847502  0.32712959
#>  [72,] -0.18753142 -0.536450001  0.633404538  0.16837989  0.37473631
#>  [73,]  0.16465496  0.034707603 -0.219299695  0.18500816  0.02509023
#>  [74,]  0.60124964 -0.277679607 -0.267868229  0.79124644 -0.56734495
#>  [75,]  0.41851332 -0.538339945 -0.768482260  0.82483588  0.81746639
#>  [76,]  0.93846763 -0.857405573 -0.273947529  0.77690720 -0.53883874
#>  [77,] -0.45005486 -0.616103075  0.564964783  0.34835305 -0.75118551
#>  [78,] -0.77940308  0.739204056 -0.224025260 -0.06580736 -0.79254363
#>  [79,]  0.57087423 -0.376464910  0.200438336  0.22414960  0.69398468
#>  [80,]  0.54647217  0.173747612 -0.667026144 -0.94860721 -0.07166722
#>  [81,]  0.67027501 -0.730590853 -0.826104189 -0.50314589 -0.34320719
#>  [82,] -0.90800378 -0.392956800  0.037031201  0.98328713 -0.91414233
#>  [83,]  0.04512694  0.130742949 -0.310909532 -0.94724854  0.04273469
#>  [84,]  0.39847350 -0.351403469 -0.195942956 -0.90963145 -0.61291170
#>  [85,]  0.58919011  0.193185041 -0.502136796 -0.03300798  0.48205885
#>  [86,]  0.91305484 -0.457411285 -0.218483171 -0.10404296  0.99583960
#>  [87,]  0.82956745  0.542904610 -0.972789431  0.20246018 -0.26085575
#>  [88,]  0.76867530  0.230375587  0.038248543 -0.63549221  0.20947404
#>  [89,]  0.43310779  0.068445115  0.519764286  0.99002956  0.75809966
#>  [90,]  0.40599545 -0.473236363  0.273937334 -0.81532536 -0.91543267
#>  [91,]  0.34354620  0.442513707 -0.862937579  0.59089289  0.87991589
#>  [92,] -0.71031434 -0.620317938 -0.391393537  0.46194206 -0.61472302
#>  [93,]  0.98308167 -0.768500855 -0.159118078 -0.88811994 -0.99344510
#>  [94,] -0.94254724 -0.058959303 -0.252432566  0.96548698 -0.19346924
#>  [95,]  0.35274435 -0.452549166 -0.867438928  0.04981405  0.28291735
#>  [96,] -0.14862125  0.864932194 -0.852787310 -0.11348482 -0.09637925
#>  [97,]  0.69380219  0.785533428  0.143073060  0.46870692  0.18489078
#>  [98,] -0.01577443  0.121818038  0.195659923 -0.77363747  0.99091977
#>  [99,]  0.34247039 -0.120309640  0.424062815  0.73547834  0.94163771
#> [100,]  0.14737432 -0.986926186 -0.517512314  0.56229943 -0.91812624
#> 
#> $y
#>   [1] 0 0 0 1 1 0 1 1 1 0 0 0 1 1 0 1 0 1 1 1 0 0 1 1 0 0 0 0 1 0 1 0 0 0 0 1 0
#>  [38] 0 0 1 1 0 1 0 0 1 0 1 1 1 1 0 1 1 1 1 1 1 0 0 0 0 0 1 1 0 1 0 0 0 0 1 1 1
#>  [75] 1 0 1 0 1 1 0 0 1 1 1 0 0 0 1 1 1 0 0 0 1 0 0 1 1 0
#>